% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aa_residues}
\alias{aa_residues}
\title{Masses of amino acid residues.}
\format{
A data frame with 26 rows and 6 variables:
\describe{
  \item{one_letter}{A one-letter representation of amino acid residues}
  \item{three_letter}{A three-letter representation of amino acid residues}
  \item{fullname}{The full name of amino acid residues}
  \item{monoisotopic_da}{Monoisotopic mass of a amino acid residue in Dalton}
  \item{average_da}{Average mass of a amino acid residue in Dalton}
  \item{composition}{Elemental composition of a amino acid residue}
}
}
\usage{
aa_residues
}
\description{
The dataset is based on Mascot.
}
\keyword{datasets}
