% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{add_cols_at}
\alias{add_cols_at}
\title{Pads columns to a placeholder data frame.}
\usage{
add_cols_at(df, df2, idx)
}
\arguments{
\item{df}{The original data frame.}

\item{df2}{The data frame to be inserted.}

\item{idx}{The index of \code{df} column for \code{df2} to be inserted
(after).}
}
\description{
Pads columns to a placeholder data frame.
}
