% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{add_complement_ions}
\alias{add_complement_ions}
\title{Adds the complimentary series of ions.}
\usage{
add_complement_ions(ms1, ms2)
}
\arguments{
\item{ms1}{The MS1 mass of precursor (neutral from \link{calc_pepmasses}).}

\item{ms2}{The series of MS2 product ions in one of a-, b- or c-ions.}
}
\description{
Adds the complimentary series of ions.
}
