% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_custom_entrez}
\alias{add_custom_entrez}
\title{Add custom Entrez IDs and optional overwriting species}
\usage{
add_custom_entrez(acc_lookup, entrez, acc_type)
}
\arguments{
\item{acc_lookup}{A data frame of protein accession lookups}

\item{entrez}{Character string(s) to the name(s) of entrez file(s) with
prepended directory path. At the \code{NULL} default, a convenience lookup
is available for species among \code{c("human", "mouse", "rat")}. For other
species, users need to provide the file path(s) and name(s) for the lookup
table(s). See also \code{\link{Uni2Entrez}} and \code{\link{Ref2Entrez}}
for preparing custom entrez files.}

\item{acc_type}{The type of protein accessions}
}
\description{
Add custom Entrez IDs and optional overwriting species
}
