% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{add_diann_pepseqmod}
\alias{add_diann_pepseqmod}
\title{Adds the \code{pep_seq_mod} field to DIANN \code{report.tsv}.}
\usage{
add_diann_pepseqmod(df = NULL, use_lowercase_aa = TRUE)
}
\arguments{
\item{df}{DIANN's \code{report.tsv} with complication of \code{pep_start}
etc.}

\item{use_lowercase_aa}{Logical; if TRUE, modifications in amino acid residues
will be abbreviated with lower-case and/or \code{^_~}. See the table below
for details. The default is TRUE.}
}
\description{
Adds the \code{pep_seq_mod} field to DIANN \code{report.tsv}.
}
