% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{add_empai}
\alias{add_empai}
\title{Add column "prot_empai"}
\usage{
add_empai(df, dat_dir)
}
\arguments{
\item{df}{A data frame of PSMs}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Add column "prot_empai"
}
