% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_entrez}
\alias{add_entrez}
\title{Adds Entrez IDs}
\usage{
add_entrez(acc_lookup, acc_type, warns = TRUE)
}
\arguments{
\item{acc_lookup}{A data frame of protein accession lookups}

\item{acc_type}{The type of protein accessions}

\item{warns}{Logical; if TRUE, show warning message(s).}
}
\description{
Adds Entrez IDs
}
