% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_entry_ids}
\alias{add_entry_ids}
\title{Add numeric indices for peptides or proteins}
\usage{
add_entry_ids(df, col_in = "pep_seq", col_out = "pep_index")
}
\arguments{
\item{df}{A data frame.}

\item{col_in}{The column key to be indexed.}

\item{col_out}{The name of the output column.}
}
\description{
Add numeric indices for peptides or proteins
}
