% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_fasta_attrs}
\alias{add_fasta_attrs}
\title{Helper of \link{parse_fasta}.}
\usage{
add_fasta_attrs(pat_acc, acc_type, acc_lookup, fasta_db)
}
\arguments{
\item{pat_acc}{A regular expression of protein accession.}

\item{acc_type}{The type of protein accession.}

\item{acc_lookup}{A look-up table of protein accessions.}

\item{fasta_db}{A database of fasta.}
}
\description{
Adds attributes to fasta entries.
}
