% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{add_fixvar_masses}
\alias{add_fixvar_masses}
\title{Helper to add modification masses to amino-acid residues.}
\usage{
add_fixvar_masses(mods, mod_type, aa_masses, add_varmasses = TRUE)
}
\arguments{
\item{mods}{A list of modifications.}

\item{mod_type}{The type of modification in one of \code{c("fmods", "vmods")}
where \code{fmods}: fixed modifications and \code{vmods}: variable
modifications.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{add_varmasses}{Logical; if TRUE, add mass delta to the corresponding
base mass for variable modifications.}
}
\value{
Lists of of amino-acid residues with modified mono-isotopic masses
  being incorporated.
}
\description{
It adds the masses of fixed, variable, and neutral-loss modifications to
amino-acid residues.
}
