% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{add_mascot_pepseqmod}
\alias{add_mascot_pepseqmod}
\title{Add the `pep_seq_mod` field to Mascot PSMs}
\usage{
add_mascot_pepseqmod(df, use_lowercase_aa)
}
\description{
Add the `pep_seq_mod` field to Mascot PSMs
}
