% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{add_maxquant_pepseqmod}
\alias{add_maxquant_pepseqmod}
\title{Add the `pep_seq_mod` field to MaxQuant PSMs}
\usage{
add_maxquant_pepseqmod(df, use_lowercase_aa)
}
\description{
Add the `pep_seq_mod` field to MaxQuant PSMs
}
