% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{add_mbr_psms}
\alias{add_mbr_psms}
\title{Adds MBR PSMs from psmC}
\usage{
add_mbr_psms(
  dfq,
  dfc,
  group_psm_by = "pep_seq_mod",
  cols = c("pep_ret_range"),
  mbr_ret_tol = 60
)
}
\arguments{
\item{dfq}{psmQ data.}

\item{dfc}{psmC data.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{cols}{Columns for use as unique identifiers of PSMs.}

\item{mbr_ret_tol}{Retention time tolerance (in seconds) for LFQ-MBR.}
}
\description{
The field of \code{pep_scan_title} from MBR does not contain new information
and purposely left as NA.
}
