% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{add_mod_conf}
\alias{add_mod_conf}
\title{Add column "pep_var_mod_conf"}
\usage{
add_mod_conf(df, dat_dir)
}
\arguments{
\item{df}{A data frame of PSMs}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Add column "pep_var_mod_conf"
}
