% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{add_modmasses}
\alias{add_modmasses}
\title{Add the masses of fixed and/or variable modifications to tryptic peptides.}
\usage{
add_modmasses(
  file = NULL,
  varmods = c("Acetyl (N-term)", "Oxidation (M)"),
  fixmods = c("Carbamidomethyl (C)")
)
}
\arguments{
\item{file}{A character string to the name of a peptide-mass file.}
}
\description{
Add the masses of fixed and/or variable modifications to tryptic peptides.
}
\examples{
\dontrun{
pepmasses <- add_modmasses("~/proteoQ/dbs/fasta/uniprot/pepmasses/pepmasses.rds")
}
}
