% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{add_n_neuloss}
\alias{add_n_neuloss}
\title{Calculates the number of neutral losses of peptides.}
\usage{
add_n_neuloss(df, group_psm_by, use_unique = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{use_unique}{Logical; if TRUE, filter data by uniqueness.}
}
\description{
Calculates the number of neutral losses of peptides.
}
