% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{add_prot_acc}
\alias{add_prot_acc}
\title{Adds prot_acc to a peptide table}
\usage{
add_prot_acc(df, out_path = "~/proteoQ/outs")
}
\arguments{
\item{df}{The results after scoring.}

\item{out_path}{An output path.}
}
\description{
Adds prot_acc to a peptide table
}
