% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_prot_desc}
\alias{add_prot_desc}
\title{Helper of \link{parse_fasta}.}
\usage{
add_prot_desc(acc_type, acc_lookup, fasta_db)
}
\arguments{
\item{acc_type}{The type of protein accession.}

\item{acc_lookup}{A look-up table of protein accessions.}

\item{fasta_db}{A database of fasta.}
}
\description{
Adds \code{prot_desc} by \code{acc_type}.
}
