% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_prot_icover}
\alias{add_prot_icover}
\title{Finds peptide abundance indexes.}
\usage{
add_prot_icover(df, id = "gene", pep_id = "pep_seq", max_len = 50)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{pep_id}{the column key of peptide sequences.}

\item{max_len}{The maximum length of peptide sequence for consideration.}
}
\description{
Finds peptide abundance indexes.
}
