% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{add_psm_apexes}
\alias{add_psm_apexes}
\title{Add PSM apexes by each \code{RAW_File}.}
\usage{
add_psm_apexes(
  df = NULL,
  trace_file = NULL,
  dat_dir = NULL,
  path_ms1 = NULL,
  cols = c("raw_file", "pep_tot_int", "pep_exp_mz", "pep_seq_mod", "pep_ret_range",
    "pep_apex_ret", "pep_scan_num", "pep_orig_scan", "pep_apex_scan", "pep_n_apexes")
)
}
\arguments{
\item{df}{A fraction of PSM table, e.g. \code{TMTSet1_LCMSinj1_PSM_N.txt},
under a single \code{RAW_File}.}

\item{trace_file}{A file name of MS1 traces (\code{ms1apexes_[...].rds}).}

\item{dat_dir}{A working directory.}

\item{path_ms1}{The path to the data of MS1 traces.}

\item{cols}{Column keys}
}
\description{
Add PSM apexes by each \code{RAW_File}.
}
