% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{add_quality_cols}
\alias{add_quality_cols}
\title{Add columns \code{pep_n_psm}, \code{prot_n_psm} and \code{prot_n_pep}.}
\usage{
add_quality_cols(df, group_psm_by, group_pep_by)
}
\arguments{
\item{df}{A data frame.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}
}
\description{
Add columns \code{pep_n_psm}, \code{prot_n_psm} and \code{prot_n_pep}.
}
