% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_refseq_gene}
\alias{add_refseq_gene}
\title{Adds gene IDs for RefSeq}
\usage{
add_refseq_gene(acc_lookup, acc_type)
}
\arguments{
\item{acc_lookup}{A data frame of protein accession lookups}

\item{acc_type}{The type of protein accessions}
}
\description{
Adds gene IDs for RefSeq
}
