% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{add_rptrs}
\alias{add_rptrs}
\title{Adds back reporter-ion intensities}
\usage{
add_rptrs(df = NULL, quant = "none", out_path = NULL)
}
\arguments{
\item{df}{Results from \link{calc_protfdr}.}

\item{quant}{A quantitation method. The default is "none". Additional choices
include \code{tmt6} etc. For other multiplicities of \code{tmt}, use the
compatible higher plexes, for example, \code{tmt16} for \code{tmt12} etc.
and \code{tmt10} for \code{tmt8} etc.}

\item{out_path}{The output path.}
}
\description{
Adds back reporter-ion intensities
}
