% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_scan_id}
\alias{add_scan_id}
\title{Add \code{pep_scan_id}}
\usage{
add_scan_id(df = NULL, fraction_scheme)
}
\arguments{
\item{df}{A data frame.}

\item{fraction_scheme}{The metadata of fraction scheme.}
}
\description{
For LFQ and MBR. For example, no MBR for fractions under the same
TMT_Set/Sample_ID and LCMS_Injection. Note the one-to-one correspondence
between TMT_Set and Sample_ID in LFQ.
}
