% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interims.R
\name{add_seions2}
\alias{add_seions2}
\title{Adds secondary ions of b0, y0 etc.}
\usage{
add_seions2(ms2s, type_ms2ions = "by", digits = 5L)
}
\arguments{
\item{ms2s}{A vector of theoretical MS2 m-over-z values.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Adds secondary ions of b0, y0 etc.
}
