% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{add_shared_genes}
\alias{add_shared_genes}
\title{Find peptides that are shared by genes.}
\usage{
add_shared_genes(df, key = "Proteins", sep = ";", fasta = NULL, entrez = NULL)
}
\arguments{
\item{df}{PSM data.}

\item{key}{the column key of mapped proteins.}

\item{sep}{Character string; the separator of the mapped proteins.}
}
\description{
For MaxQuant and MSFragger.
}
