% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{add_shared_prot_accs_mf}
\alias{add_shared_prot_accs_mf}
\title{Finds the shared protein accessions when using MSFragger.}
\usage{
add_shared_prot_accs_mf(df)
}
\arguments{
\item{df}{PSM data.}
}
\description{
Protein accessions will be used if the mapped accessions are present under
\code{prot_acc}; otherwise, \code{fasta_name} will be used. This allows quick
distinction whether mapped proteins are same-set/sub-set to the primary
entries or not.
}
