% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{add_sm_pepseqmod}
\alias{add_sm_pepseqmod}
\title{Adds the `pep_seq_mod` field to MaxQuant PSMs}
\usage{
add_sm_pepseqmod(df = NULL, use_lowercase_aa = TRUE)
}
\arguments{
\item{df}{A data frame containing the PSM table from database searches.}

\item{use_lowercase_aa}{Logical; if TRUE, modifications in amino acid
residues will be abbreviated with lower-case and/or \code{^_~}. See the
table below for details. The default is TRUE.}
}
\description{
Adds the `pep_seq_mod` field to MaxQuant PSMs
}
