% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{add_spec_counts}
\alias{add_spec_counts}
\title{Add spectrum counts}
\usage{
add_spec_counts(
  dat_dir,
  df_num,
  group_psm_by = "pep_seq_mod",
  group_pep_by = "gene",
  type = "protein"
)
}
\arguments{
\item{dat_dir}{A working directory.}

\item{df_num}{A data frame.}

\item{group_psm_by}{Group PSM data by.}

\item{group_pep_by}{Group peptide data by.}

\item{type}{The type of data.}
}
\description{
DO not aggregate from peptides to proteins. Missing values in peptide tables
can inflate protein spectrum counts: median(c(1, NA)) -> 1 not 0.5.
}
