% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{add_spn_pepseqmod}
\alias{add_spn_pepseqmod}
\title{Compile the \code{pep_seq_mod} field to Spectranaut results.}
\usage{
add_spn_pepseqmod(df = NULL, use_lowercase_aa = TRUE)
}
\arguments{
\item{df}{A Spectranaut peptide table with complication of \code{pep_start}
etc.}

\item{use_lowercase_aa}{Logical; if TRUE, modifications in amino acid residues
will be abbreviated with lower-case and/or \code{^_~}. See the table below
for details. The default is TRUE.}
}
\description{
Compile the \code{pep_seq_mod} field to Spectranaut results.
}
