% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspa.R
\name{add_subcell}
\alias{add_subcell}
\title{Add subcellular results to protein table.}
\usage{
add_subcell(
  df = NULL,
  dat_dir,
  group_pep_by = "gene",
  is_subcellular = FALSE,
  scale_log2r = TRUE,
  impute_na = FALSE
)
}
\arguments{
\item{df}{A data frame.}

\item{dat_dir}{The working directory.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{is_subcellular}{Logical; is subcellular data type or not.}
}
\description{
Add subcellular results to protein table.
}
