% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{add_term_mass}
\alias{add_term_mass}
\title{Helper in calculating peptide masses.}
\usage{
add_term_mass(aa_masses, peps)
}
\arguments{
\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{peps}{Lists of peptides, either "fwds" or "revs", from
\link{split_fastaseqs}.}
}
\description{
(2) "amods- tmod+ vnl- fnl-".
}
