% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlfq.R
\name{addApexes}
\alias{addApexes}
\title{Add retention time apex for a chunk of PSMs under a single RAW}
\usage{
addApexes(
  rts,
  mzs,
  trs,
  gap_bf = 250L,
  gap_af = 250L,
  step = 1e-05,
  yco = 100,
  min_y = 2e+06
)
}
\arguments{
\item{rts}{A vector of retention times at MS2 events (pep_ret_range).}

\item{mzs}{A vector of m-over-z values for PSMs.}

\item{gap_bf}{A leading gap of maximum allowance in retention times between
an MS2 scan and a peak apex.}

\item{gap_af}{A following gap of maximum allowance in retention times between
an MS2 scan and a peak apex.}

\item{step}{A step size in mass for mass error tolerance.}

\item{yco}{The cut-off in intensities.}

\item{min_y}{The minimum peak area.}
}
\description{
For parallel processing.
}
