% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlfq.R
\name{addApexRTs}
\alias{addApexRTs}
\title{Helper of adding apex retention times for one RAW}
\usage{
addApexRTs(
  df,
  file_ms1,
  path_ms1,
  from = 115,
  step = 1e-05,
  rt_size = 240,
  rt_margin = 240,
  yco = 100,
  min_y = 2e+06,
  n_para = 6L
)
}
\arguments{
\item{df}{A subset of PSM table under one RAW.}

\item{file_ms1}{The file name of an \code{ms1full_.rds}.}

\item{path_ms1}{The file path to \code{ms1full_.rds}.}

\item{from}{The starting point for calculating the bin indexes of masses.}

\item{step}{A step size in mass for binning.}

\item{rt_size}{The width of each LC retention times in seconds.}

\item{rt_margin}{The bracketing margin before and after an LC retention time
window.}

\item{yco}{The cut-off in intensities.}

\item{min_y}{The minimum peak area.}

\item{n_para}{The number of parallel processes.}
}
\description{
Inputs are at the levels of one RAW file.
}
