% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{addMBRpeps2}
\alias{addMBRpeps2}
\title{Add MBR peptide intensities}
\usage{
addMBRpeps2(file, dat_dir, group_psm_by = "pep_seq_mod")
}
\arguments{
\item{file}{A file name with prepending path to a peptide table.}

\item{dat_dir}{A working directory.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}
}
\description{
Add MBR peptide intensities
}
