% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{aggrLCMS_DIANN}
\alias{aggrLCMS_DIANN}
\title{Aggregates DIANN data from multiple LCMS injections.}
\usage{
aggrLCMS_DIANN(df)
}
\arguments{
\item{df}{A DIANN table.}
}
\description{
Aggregates DIANN data from multiple LCMS injections.
}
