% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggrLFQ}
\alias{aggrLFQ}
\title{Aggregation of LFQ inentisty.}
\usage{
aggrLFQ(f)
}
\arguments{
\item{f}{A function for data summarization.}
}
\description{
Summarizes \code{log2FC} and \code{intensity} by the descriptive statistics
of \code{c("mean", "median", "sum")}. Note that data are already subset by
top_n.
}
\examples{
\donttest{df_num <- aggrLFQ(sum)(df, prot_acc, na.rm = TRUE)}
}
