% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggrLFQs}
\alias{aggrLFQs}
\title{Aggregation of LFQ intensity.}
\usage{
aggrLFQs(f)
}
\arguments{
\item{f}{A function for data summary.}
}
\description{
Not currently used. Summarizes \code{log2FC} and \code{intensity} by the
descriptive statistics of \code{c("mean", "median", "sum")}. Note that data
are already subset by top_n.
}
\details{
\code{ids} are a list of column keys.
}
\examples{
\donttest{df_num <- aggrLFQ(sum)(df, prot_acc, na.rm = TRUE)}
}
