% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{aggrNumLCMS}
\alias{aggrNumLCMS}
\title{Aggregates data from the same TMT_Set at different LCMS_Injection.}
\usage{
aggrNumLCMS(df, group_psm_by = "pep_seq_mod", label_scheme_full)
}
\arguments{
\item{df}{A data frame.}

\item{label_scheme_full}{Metadata.}
}
\description{
Aggregates data from the same TMT_Set at different LCMS_Injection.
}
