% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggrNums}
\alias{aggrNums}
\title{Summarizes numeric values}
\usage{
aggrNums(f)
}
\arguments{
\item{f}{A function for data summarization.}
}
\description{
\code{aggrNums} summarizes \code{log2FC} and \code{intensity} by the
descriptive statistics of \code{c("mean", "median", "weighted.mean",
"top.3")}
}
\examples{
\donttest{df_num <- aggrNums(median)(df, prot_acc, na.rm = TRUE)}
}
