% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{alignPSMApexs}
\alias{alignPSMApexs}
\title{Align PSM apex retention times}
\usage{
alignPSMApexs(dfs = NULL, group_psm_by = "pep_seq_mod")
}
\arguments{
\item{dfs}{A list of PSM tables by TMTSet[i]LCMSinj[j].}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}
}
\description{
Also nullify retention-time outliers.
}
