% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{anal_pepOutlier}
\alias{anal_pepOutlier}
\title{Find if a protein contains peptide outliers under a contrast.}
\usage{
anal_pepOutlier(
  dat_dir = NULL,
  col_select = NULL,
  col_group = NULL,
  col_order = NULL,
  p_outlier = 0.05,
  min_n = 3L,
  p_contrast = 0.01,
  p_type = "adjP",
  cols_pep = c("pep_start", "pep_end", "pep_miss", "pep_phospho_locprob"),
  contr_pairs = NULL,
  group_renorm_by = NULL,
  scale_log2r = TRUE,
  df = NULL,
  filepath = NULL,
  filename = NULL,
  ...
)
}
\arguments{
\item{dat_dir}{A working directory.}

\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
At the \code{NULL} default, the column key of \code{Select} in
\code{expt_smry.xlsx} will be used. In the case of no samples being
specified under \code{Select}, the column key of \code{Sample_ID} will be
used. The non-empty entries under the ascribing column will be used in
indicated analysis.}

\item{col_group}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_group} will be
used for sample grouping in the indicated analysis. At the NULL default, the
column key \code{Group} will be used. No data annotation by groups will be
performed if the fields under the indicated group column is empty.}

\item{col_order}{Character string to a column key in \code{expt_smry.xlsx}.
Numeric values under which will be used for the left-to-right arrangement of
samples in graphic outputs or top-to-bottom arrangement in text outputs. At
the NULL default, the column key \code{Order} will be used. If values under
column \code{Order} are left blank, samples will be ordered by their names.}

\item{p_outlier}{Type-I error in outlier assessment. Set p = 1.0 to skip
outlier removals.}

\item{min_n}{The minimal number of peptides under a protein for
consideration.}

\item{p_contrast}{The cut-off in significant pVals from \link{proteoQ}. The
type of pVals is either original or BH-adjusted, indicated by
\code{p_type}.}

\item{p_type}{The type of p-values in one of \code{adjP, pVal}.}

\item{cols_pep}{The attribute columns of a peptide table to be included.}

\item{contr_pairs}{Contrast pairs. If \code{NULL}, all contrasts under a
formula will be included. To specify a subset of contrasts: 
contr_pairs = list(C1 = c("X-Y", "X-Z"), C2 = c("A-B", "C-B"))}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE. At
\code{scale_log2r = NA}, the raw \code{log2FC} without normalization will
be used.}

\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{filename}{A file name to output results. The default is
\code{Peptide_pVals.txt} for peptides and \code{Protein_pVals} for proteins.}

\item{...}{User-defined formulas for linear modeling. The syntax starts with a
tilde, followed by the name of an available column key in
\code{expt_smry.xlsx} and square brackets. The contrast groups are then
quoted with one to multiple contrast groups separated by commas. The
 default column key is \code{Term} in \code{expt_smry.xlsx}: \cr \code{~
 Term["A - C", "B - C"]}. \cr \cr Additive random effects are indicated by
\code{+ (1|col_key_1) + (1|col_key_2)}... Currently only a syntax of single
 contrast are supported for uses with random effects: \cr \code{~ Term["A -
 C"] + (1|col_key_1) + (1|col_key_2)} \cr \cr \code{filter_}: Logical
expression(s) for the row filtration against data in a primary file of
\code{Peptide[_impNA].txt} or \code{Protein[_impNA].txt}. See also
\code{\link{normPSM}} for the format of \code{filter_} statements.}

\item{p}{A significance level for assessing outliers.}
}
\description{
Find if a protein contains peptide outliers under a contrast.
}
