% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-nmf.R
\name{anal_prnNMF}
\alias{anal_prnNMF}
\title{Protein NMF Classification}
\usage{
anal_prnNMF(...)
}
\description{
NMF Classification of protein \code{log2-ratios}
}
\examples{
prnNMF(
  scale_log2r = TRUE,
  xmin = -1,
  xmax = 1,
  x_margin = 0.1,
  annot_cols = c("Peptide_Yield", "TMT_Set", "Group"),
  r = 6,
  nrun = 200
)


}
\seealso{
\code{\link{proteoNMF}} for parameters
}
