% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-string.R
\name{anal_prnString}
\alias{anal_prnString}
\title{STRING outputs of protein-protein interactions}
\usage{
anal_prnString(scale_log2r = TRUE, complete_cases = FALSE,
  impute_na = FALSE, db_path = "~\\\\proteoQ\\\\dbs\\\\string",
  score_cutoff = 0.7, df = NULL, filepath = NULL, filename = NULL,
  ...)
}
\arguments{
\item{scale_log2r}{Not currently used. Values before and after scaling will be
both reported.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used. The default is FALSE.}

\item{db_path}{Character string; the local path for database(s). The default
is \code{"~\\proteoQ\\dbs\\string"}.}

\item{score_cutoff}{Numeric; the threshold in the \code{combined_score} of
protein-protein interaction. The default is 0.7.}

\item{df}{The name of input data file. By default, it will be determined
automatically after matching the data type with an \code{id} among
\code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}.}

\item{filepath}{Use system default.}

\item{filename}{Use system default. Otherwise, the basename will be prepended
to \code{_[species]_ppi.tsv} for network data and \code{_[species]_expr.tsv}
for expression data.}

\item{...}{\code{filter_}: Logical expression(s) for the row filtration of
data against the column keys in \code{Protein[_impNA]_pVals.txt}; also see
\code{\link{normPSM}}. \cr \code{arrange_}: Logical expression(s) for the
row ordering of data; also see \code{\link{prnHM}}.}
}
\description{
\code{anal_prnString} prepares the data of both
\code{\href{https://string-db.org/}{STRING}} protein-protein interactions
(ppi) and companion protein expressions.
}
\details{
Convenience features, such as data row filtration via \code{filter_} varargs,
are available. The ppi file, \code{..._ppi.tsv}, and the expression file,
\code{..._expr.tsv}, are also compatible with a third-party
\code{\href{https://cytoscape.org/}{Cytoscape}}.
}
\examples{
# ===================================
# String DB
# ===================================

## !!!require the brief working example in `?load_expts`

string_dir <- "~\\\\proteoQ\\\\dbs\\\\string"
dir.create(string_dir, recursive = TRUE, showWarnings = FALSE)

## download
dl_stringdbs(
  species = human,
  db_path = string_dir,
)

## ppi and expression
# human and mouse and probably 'unknown' species
anal_prnString(
  db_path = string_dir,
  score_cutoff = .9,
  filter_prots_by = exprs(prot_n_pep >= 2),
)

# cleaner outputs by specifying species
# ('unknown' will be removed)
anal_prnString(
  db_path = string_dir,
  score_cutoff = .9,
  filter_by_sp = exprs(species == "human"),
  filter_prots_by = exprs(prot_n_pep >= 2),
)

anal_prnString(
  db_path = string_dir,
  score_cutoff = .9,
  filter_by_sp = exprs(species == "mouse"),
  filter_prots_by = exprs(prot_n_pep >= 2),
)

# custom filename with outputs in: 
# prot_npep2 W16_vs_W2.pVal (W16-W2)_human_expr.tsv
# prot_npep2 W16_vs_W2.pVal (W16-W2)_human_ppi.tsv
# prot_npep2 W16_vs_W2.pVal (W16-W2)_mouse_expr.tsv
# prot_npep2 W16_vs_W2.pVal (W16-W2)_mouse_ppi.tsv
anal_prnString(
  db_path = string_dir,
  score_cutoff = .9,
  complete_cases = TRUE, 
  filter_by = exprs(species \%in\% c("human", "mouse"), 
                    prot_n_pep >= 2, 
                    `W16_vs_W2.pVal (W16-W2)` <= 1e-6),
  filename = "prot_npep2 W16_vs_W2.pVal (W16-W2).txt",
)
}
\seealso{
\code{\link{dl_stringdbs}} for database downloads. \cr
 \code{\link{prnSig}} for significance tests \cr
}
