% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{analOutlier}
\alias{analOutlier}
\title{Outlier analysis of peptide data.}
\usage{
analOutlier(
  df,
  dat_dir = NULL,
  id = "pep_seq_mod",
  group_pep_by = "gene",
  col_select = NULL,
  col_group = NULL,
  col_order = NULL,
  label_scheme_sub = NULL,
  p_outlier = 0.05,
  min_n = 3L,
  min_n_per_group = 3L,
  p_contrast = 0.01,
  p_type = "adjP",
  cols_pep = c("pep_start", "pep_end", "pep_miss", "pep_phospho_locprob"),
  contr_pairs = NULL,
  impute_group_na = FALSE,
  group_renorm_by = NULL,
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  filepath,
  filename = "Peptide_Outlier_Z.tsv",
  anal_type = "Outlier",
  ...
)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{dat_dir}{The current working directory.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
At the \code{NULL} default, the column key of \code{Select} in
\code{expt_smry.xlsx} will be used. In the case of no samples being
specified under \code{Select}, the column key of \code{Sample_ID} will be
used. The non-empty entries under the ascribing column will be used in
indicated analysis.}

\item{col_group}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_group} will be
used for sample grouping in the indicated analysis. At the NULL default, the
column key \code{Group} will be used. No data annotation by groups will be
performed if the fields under the indicated group column is empty.}

\item{col_order}{Character string to a column key in \code{expt_smry.xlsx}.
Numeric values under which will be used for the left-to-right arrangement of
samples in graphic outputs or top-to-bottom arrangement in text outputs. At
the NULL default, the column key \code{Order} will be used. If values under
column \code{Order} are left blank, samples will be ordered by their names.}

\item{p_outlier}{Type-I error in outlier assessment. Set p = 1.0 to skip
outlier removals.}

\item{min_n_per_group}{The minimum number of data points per group for
consideration of outlier assessment.}

\item{impute_group_na}{Logical; if TRUE, impute (intensity) values that are
exclusively NA under a sample group. The sample grouping is defined under
\code{label_scheme}.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE. At
\code{scale_log2r = NA}, the raw \code{log2FC} without normalization will
be used.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used. The default is FALSE.}

\item{filepath}{Use system default.}

\item{filename}{A representative file name to outputs. By default, it will be
determined automatically by the name of the current call.}

\item{...}{\code{filter_}: Variable argument statements for the row
filtration against data in a primary file linked to \code{df}. See also
\code{\link{normPSM}} for the format of \code{filter_} statements. \cr \cr
\code{arrange_}: Variable argument statements for the row ordering against
data in a primary file linked to \code{df}. See also \code{\link{prnHM}}
for the format of \code{arrange_} statements. \cr \cr Additional arguments
for \link[e1071]{cmeans}, \link[stats]{kmeans}, \link[cluster]{clara},
\link[cluster]{pam}. Note that \code{centers} in \link[e1071]{cmeans} or
\link[stats]{kmeans} is replaced with \code{n_clust}. The same applies to
\code{k} in \link[cluster]{clara} or \link[cluster]{pam}. \cr With
\link[e1071]{cmeans}, \code{m} is according to Schwaemmle and Jensen if not
provided; \cr \code{x} is disabled with input data being determined
automatically.}
}
\description{
Outlier analysis of peptide data.
}
