% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-string.R
\name{annot_stringdb_old}
\alias{annot_stringdb_old}
\title{Annotates protein STRING ids by species}
\usage{
annot_stringdb_old(species, df, db_path, id, score_cutoff)
}
\arguments{
\item{db_path}{Character string; the local path for database(s). The default
is \code{"~\\proteoQ\\dbs\\string"}.}

\item{score_cutoff}{Numeric; the threshold in the \code{combined_score} of
protein-protein interaction. The default is 0.7.}
}
\description{
Annotates protein STRING ids by species
}
