% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annotKin}
\alias{annotKin}
\title{Adds kinase annotation}
\usage{
annotKin(df, acc_type)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{acc_type}{Character string; the type of protein accessions in one of
c("refseq_acc", "uniprot_acc", "uniprot_id")}
}
\description{
Adds kinase annotation
}
