% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{annotPeppos}
\alias{annotPeppos}
\title{Add peptide start and end positions}
\usage{
annotPeppos(df, fasta)
}
\description{
\code{annotPeppos} annotates the start and the end positions of peptides in
ascribed proteins description based on the \code{fasta}.
}
