% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{annotPeppos_mq}
\alias{annotPeppos_mq}
\title{Add start and end positions and preceding and succeeding residues of peptides}
\usage{
annotPeppos_mq(df, fasta)
}
\description{
\code{annotPeppos_mq} annotates start and end positions and the preceding and
succeeding residues of peptides.
}
