% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{annotPSM}
\alias{annotPSM}
\title{Annotates PSM results}
\usage{
annotPSM(expt_smry = "expt_smry.xlsx", rm_krts = FALSE,
  plot_violins = TRUE)
}
\arguments{
\item{expt_smry}{A character string to the \code{.xlsx} file containing the
metadata of TMT experiments. The default is "expt_smry.xlsx".}

\item{rm_krts}{Logical; if TRUE, removes keratin entries from the output.}

\item{plot_violins}{Logical; if TRUE, prepares the violin plots of
reporter-ion intensities.}
}
\description{
\code{annotPSM} adds fields of annotation to PSM tables.
}
