% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{annotPSM_sm}
\alias{annotPSM_sm}
\title{Annotates PSM results}
\usage{
annotPSM_sm(fasta = NULL, expt_smry = "expt_smry.xlsx",
  rm_krts = FALSE, plot_rptr_int = TRUE)
}
\arguments{
\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search.}

\item{expt_smry}{A character string to the \code{.xlsx} file containing the
metadata of TMT experiments. The default is "expt_smry.xlsx".}

\item{plot_rptr_int}{Logical; if TRUE, prepares the violin plots of
reporter-ion intensities.}
}
\description{
\code{annotPSM_sm} adds fields of annotation to PSM tables.
}
