% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{arrangers_in_call}
\alias{arrangers_in_call}
\title{Arrange rows}
\usage{
arrangers_in_call(.df, ..., .na.last = TRUE)
}
\arguments{
\item{.df}{a data frame.}

\item{...}{Arguments for \link[dplyr]{arrange}}

\item{.na.last}{The same as \link[base]{order}}
}
\description{
Arrange rows
}
