% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{as_lgldist}
\alias{as_lgldist}
\title{Simplified \link[stats]{as.dist} for memory efficiency.}
\usage{
as_lgldist(m, diag = FALSE, upper = FALSE)
}
\arguments{
\item{m}{An object with distance information to be converted to a
    \code{"dist"} object.  For the default method, a \code{"dist"}
    object, or a matrix (of distances) or an object which can be coerced
    to such a matrix using \code{\link{as.matrix}()}.  (Only the lower
    triangle of the matrix is used, the rest is ignored).}

\item{diag}{logical value indicating whether the diagonal of the
    distance matrix should be printed by \code{print.dist}.}

\item{upper}{logical value indicating whether the upper triangle of the
    distance matrix should be printed by \code{print.dist}.}
}
\description{
Assumed the input is already a symmetric matrix.
}
