% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{assign_duppeps}
\alias{assign_duppeps}
\title{Assign duplicated peptides to a leading protein}
\usage{
assign_duppeps(
  df,
  group_psm_by = "pep_seq",
  group_pep_by = "pep_seq_mod",
  use_duppeps = TRUE,
  duppeps_repair = "denovo"
)
}
\arguments{
\item{df}{A PSM data frame}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{use_duppeps}{Logical; if TRUE, re-assigns double/multiple dipping
peptide sequences to the most likely proteins by majority votes.}

\item{duppeps_repair}{Not currently used (or only with \code{majority}).
  Character string; the method of reparing double-dipping peptide sequences
  upon data pooling.

  For instance, the same sequence of PEPTIDE may be assigned to protein
  accession PROT_ACC1 in data set 1 and PROT_ACC2 in data set 2. At the
  \code{denovo} default, the peptide to protein association will be
  re-established freshly. At the \code{majority} alternative, a majority rule
  will be applied for the re-assignments.}
}
\description{
Assign duplicated peptides to a leading protein
}
