% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtests.R
\name{base_sigtest_y}
\alias{base_sigtest_y}
\title{Calculate baseline intensities.}
\usage{
base_sigtest_y(
  dfR,
  dfI,
  elements = NULL,
  key_col = NULL,
  label_scheme_sub_sub,
  perc_baseline_intensity = 1e-04,
  abs_baselne_intensity = 10^5.5,
  seed = NULL
)
}
\arguments{
\item{dfR}{The complete data frame of log2Ratios.}

\item{dfI}{The corresponding intensity data frame.}

\item{elements}{The elements of groups in contrast fits.}

\item{key_col}{The key column in contrast fits, e.g., column \code{Term}.}

\item{label_scheme_sub_sub}{The metadata corresponding \code{dfR}.}

\item{perc_baseline_intensity}{An estimated percentage of baseline intensity
in relative to the measured intensities. The setting will be modified to an
absolute value in a later version.}

\item{abs_baselne_intensity}{The absolute level of intensity for uses at
\code{impute_group_na = TRUE}.}

\item{seed}{A seed for reproducible random number generations.}

\item{ys_base}{The baseline intensities. The length is equal to the number of
columns in \code{dfR}.}
}
\description{
Calculate baseline intensities.
}
