% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{batchPSMheader}
\alias{batchPSMheader}
\title{Processes Mascot PSMs}
\usage{
batchPSMheader(
  filename = NULL,
  dat_dir = NULL,
  TMT_plex = 10L,
  proc_extdata = TRUE,
  warns = TRUE
)
}
\arguments{
\item{filename}{The name of a PSM file.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{TMT_plex}{Numeric; the multiplexity of TMT, i.e., 10, 11 etc.}

\item{proc_extdata}{Logical; if TRUE, processes extended PSM outputs;
otherwise, skips them. Setting the value to FALSE allows bypassing
time-consuming steps.}

\item{warns}{Logical; if TRUE, show warning message(s).}
}
\description{
Processes Mascot PSMs
}
