% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{bin_ms1masses}
\alias{bin_ms1masses}
\title{Helper in binning precursor masses}
\usage{
bin_ms1masses(res, min_mass, max_mass, ppm_ms1, out_path)
}
\arguments{
\item{res}{The results from \link{calc_pepmasses2}.}

\item{min_mass}{A minimum mass of precursors.}

\item{max_mass}{A maximum mass of precursors.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}

\item{out_path}{The output path.}
}
\description{
Helper in binning precursor masses
}
