% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{bin_theopeps}
\alias{bin_theopeps}
\title{Helper: separates theoretical peptides into mass groups.}
\usage{
bin_theopeps(peps, min_mass = 350L, max_mass = 1700L, ppm = 20L)
}
\arguments{
\item{peps}{A list of theoretical peptides with masses.}

\item{min_mass}{Numeric; the minimum MS1 mass.}

\item{max_mass}{Numeric; the maximum MS1 mass.}

\item{ppm}{Numeric; the eror tolerance of MS1 mass in ppm.}
}
\description{
Helper: separates theoretical peptides into mass groups.
}
