% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{byfile_gspahm}
\alias{byfile_gspahm}
\title{A helper function for gspaHM by input file names}
\usage{
byfile_gspahm(
  ess_in,
  meta_in,
  fml_nm,
  filepath,
  filename,
  scale_log2r,
  impute_na,
  ...
)
}
\description{
A helper function for gspaHM by input file names
}
