% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{byfml_gspahm}
\alias{byfml_gspahm}
\title{A helper function for gspaHM by formula names}
\usage{
byfml_gspahm(fml_nm, filepath, filename, scale_log2r, impute_na, ...)
}
\description{
A helper function for gspaHM by formula names
}
