% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanos.R
\name{byfml_volcano}
\alias{byfml_volcano}
\title{Formula specific volcano plots}
\usage{
byfml_volcano(
  fml_nm = NULL,
  gspval_cutoff,
  gslogFC_cutoff,
  topn_gsets,
  df,
  df2,
  col_ind,
  id,
  filepath,
  filename,
  adjP,
  topn_labels,
  anal_type,
  show_sig,
  gset_nms,
  scale_log2r,
  complete_cases,
  impute_na,
  highlights = NULL,
  theme = NULL,
  ...
)
}
\arguments{
\item{fml_nm}{A character string; the name of \code{fml}.}

\item{gspval_cutoff}{Numeric value or vector for uses with
\code{\link{gspaMap}}. \code{Gene sets} with enrichment \code{pVals} less
significant than the threshold will be excluded from volcano plot
visualization. The default significance is 0.05 for all formulas matched to
or specified in argument \code{fml_nms}. Formula-specific threshold is
allowed by supplying a vector of cut-off values.}

\item{gslogFC_cutoff}{Numeric value or vector for uses with
\code{\link{gspaMap}}. \code{Gene sets} with absolute enrichment
\code{log2FC} less than the threshold will be excluded from volcano plot
visualization. The default magnitude is \code{log2(1.2) } for all formulas
matched to or specified in argument \code{fml_nms}. Formula-specific
threshold is allowed by supplying a vector of absolute values in
\code{log2FC}.}

\item{topn_gsets}{Numeric value or vector; top entries in gene sets ordered by
 increasing \code{pVal} for visualization. The default is to use all
 available entries.

 Note that it is users' responsibility to ensure that the custom gene sets
 contain terms that can be found from the one or multiple preceding analyses
 of \code{\link{prnGSPA}}. For simplicity, it is generally applicable to
 include \emph{all} of the data bases that have been applied to
 \code{\link{prnGSPA}} and in that way no terms will be missed for
 visualization. See also \code{\link{prnGSPA}} for examples of custom data
 bases.}

\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{df2}{Character vector or string; the name(s) of secondary data file(s).
An informatic task, i.e. \code{anal_prnTrend(...)} against a primary
\code{df} generates secondary files such as
\code{Protein_Trend_Z_nclust6.txt} etc. See also \code{\link{prnHist}} for
the description of a primary \code{df}; \code{\link{normPSM}} for the lists
of \code{df} and \code{df2}.}

\item{col_ind}{Numeric vector; the indexes of columns for the ascribed \code{fml_nm}.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{filename}{A representative file name to outputs. By default, the
name(s) will be determined automatically. For text files, a typical file
extension is \code{.txt}. For image files, they are typically saved via
\code{\link[ggplot2]{ggsave}} or \code{\link[pheatmap]{pheatmap}} where the
image type will be determined by the extension of the file name.}

\item{adjP}{Logical; if TRUE, use Benjamini-Hochberg pVals in volcano plots.
The default is FALSE.}

\item{topn_labels}{A non-negative integer; the top-n species for labeling in a
plot. At \code{topn_labels = 0}, no labels of proteins/peptides will be
shown. The default is to label the top-20 species with the lowest p-values.}

\item{anal_type}{Character string; the type of analysis that are preset for
method dispatch in function factories. The value will be determined
automatically. Exemplary values include \code{anal_type = c("PCA",
"Corrplot", "EucDist", "GSPA", "Heatmap", "Histogram", "MDS", "Model",
"NMF", "Purge", "Trend", "LDA", ...)}.}

\item{show_sig}{Character string indicating the type of significance values to
be shown with \code{\link{gspaMap}}. The default is \code{"none"}.
Additional choices are from \code{c("pVal", "qVal")} where \code{pVal} or
\code{qVal} will be shown, respectively, in the facet grid of the plots.}

\item{gset_nms}{Character string or vector containing the shorthanded name(s),
full file path(s), or both, to gene sets for enrichment analysis. For
species among \code{"human", "mouse", "rat"}, the default of
\code{c("go_sets", "c2_msig", "kinsub")} will utilize terms from gene
ontology (\code{GO}), molecular signatures (\code{MSig}) and
kinase-substrate network (\code{PSP Kinase-Substrate}). Custom \code{GO},
\code{MSig} and other data bases at given species are also supported. See
also: \code{\link{prepGO}} for the preparation of custom \code{GO};
\code{\link{prepMSig}} for the preparation of custom \code{MSig}. For other
custom data bases, follow the same format of list as \code{GO} or
\code{MSig}.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE. At
\code{scale_log2r = NA}, the raw \code{log2FC} without normalization will
be used.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used. The default is FALSE.}

\item{highlights}{A list of entries for highlighting. See also \code{filter_}
varargs for the format.}

\item{theme}{A
\href{https://ggplot2.tidyverse.org/reference/ggtheme.html}{ggplot2} theme,
i.e., theme_bw(), or a custom theme. At the NULL default, a system theme
will be applied.}

\item{...}{\code{filter_}: Variable argument statements for the row filtration
against data in a primary file linked to \code{df}. See also
\code{\link{normPSM}} for the format of \code{filter_} statements. \cr \cr
Additional parameters for plotting: \cr \code{xco}, the cut-off lines of
fold changes at position \code{x}; the default is at \eqn{-1.2} and
\eqn{+1.2}. \cr \code{yco}, the cut-off line of \code{pVal} at position
\code{y}; the default is \eqn{0.05}. \cr \code{width}, the width of plot;
\cr \code{height}, the height of plot. \cr \code{nrow}, the number of rows
in a plot. \cr \code{xmin}, the minimum \code{x}. \cr \code{xmax}, the
maximum \code{x}. \cr \code{ymin}, the minimum \code{y}. \cr \code{ymax},
the maximum \code{y}. \cr \code{x_label}, the label on \code{x}. \cr
\code{y_label}, the label on \code{y}.}
}
\description{
Formula specific volcano plots
}
