% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{c2_msig_hs}
\alias{c2_msig_hs}
\title{Entrez IDs of human molecular signatures.}
\format{A list at a length of 5501 
\describe{
  \item{hs_...}{human entries}
}}
\source{
\url{https://www.gsea-msigdb.org/gsea/index.jsp}
}
\usage{
c2_msig_hs
}
\description{
A dataset containing human entrez IDs by the gene sets of molecular signatures.
}
\keyword{datasets}
