% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{c2_msig_mm}
\alias{c2_msig_mm}
\title{Entrez IDs of mouse molecular signatures.}
\format{A list at a length of 5501 
\describe{
  \item{mm_...}{mouse entries}
}}
\source{
\url{https://www.gsea-msigdb.org/gsea/index.jsp}
}
\usage{
c2_msig_mm
}
\description{
A dataset containing mouse entrez IDs by the gene sets of molecular signatures.
}
\keyword{datasets}
