% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-datasets.R
\docType{data}
\name{c2_msig_rn}
\alias{c2_msig_rn}
\title{Entrez IDs of rat molecular signatures.}
\format{A list at a length of 5501 
\describe{
  \item{rn_...}{rat entries}
}}
\source{
\url{https://www.gsea-msigdb.org/gsea/index.jsp}
}
\usage{
c2_msig_rn
}
\description{
A dataset containing rat entrez IDs by the gene sets of molecular signatures.
}
\keyword{datasets}
