% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{calc_allaxions}
\alias{calc_allaxions}
\title{Masses of all a- and x-ions.}
\usage{
calc_allaxions(ntmod, ctmod, aas, aa_masses, digits = 5)
}
\arguments{
\item{ntmod}{The \code{N-term} modification.}

\item{ctmod}{The \code{C-term} modification.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
a, a2, a*, a*2, a0, a*2; x, x2
}
