% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{calc_allbyions}
\alias{calc_allbyions}
\title{Masses of all b- and y-ions.}
\usage{
calc_allbyions(ntmod, ctmod, aas, aa_masses, digits = 5)
}
\arguments{
\item{ntmod}{The parsed \emph{variable} \code{N-term} modifications from
\code{aa_masses}.}

\item{ctmod}{The parsed \emph{variable} \code{C-term} modifications from
\code{aa_masses}.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Calculates the masses of b, b2, b*, b*2, b0, b*2; y, y2, y*, y*2, y0, y*2
ions. In general, the utility should not be called but use \link{calc_byions}
and derive the remaining (with MS1 precursor mass being known).
}
\seealso{
\link{add_complement_ions}
}
