% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{calc_bystar2ions}
\alias{calc_bystar2ions}
\title{Masses of doubly-charged b*- and y*-ions.}
\usage{
calc_bystar2ions(ntmod, ctmod, aas, aa_masses, digits = 5)
}
\arguments{
\item{ntmod}{The parsed \emph{variable} \code{N-term} modifications from
\code{aa_masses}.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Masses of doubly-charged b*- and y*-ions.
}
\seealso{
\link{add_complement_ions}
}
