% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{calc_cions}
\alias{calc_cions}
\title{Masses of c-ions}
\usage{
calc_cions(ntmod, ctmod, aas, aa_masses, z, digits)
}
\arguments{
\item{ntmod}{The \code{N-term} modification.}

\item{ctmod}{The \code{C-term} modification.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{z}{Integer; the charge state of an MS1 ion. Currently only for postive
charges.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Singly charged.
}
