% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{calc_czions}
\alias{calc_czions}
\title{Masses of singly-charged c- and z-ions.}
\usage{
calc_czions(ntmod, ctmod, aas, aa_masses, digits = 5)
}
\arguments{
\item{ntmod}{The \code{N-term} modification.}

\item{ctmod}{The \code{C-term} modification.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Masses of singly-charged c- and z-ions.
}
