% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{calc_lfq_log2r}
\alias{calc_lfq_log2r}
\title{Helper: calculates LFQ log2FC.}
\usage{
calc_lfq_log2r(df, type, refChannels)
}
\arguments{
\item{df}{A data frame.}

\item{type}{The type of intensity data.}

\item{refChannels}{The reference channels.}
}
\description{
Helper: calculates LFQ log2FC.
}
