% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{calc_lfq_pepnums}
\alias{calc_lfq_pepnums}
\title{Calculates log2FC of peptides based on LFQ intensity.}
\usage{
calc_lfq_pepnums(df, omit_single_lfq)
}
\arguments{
\item{df}{A data frame.}

\item{omit_single_lfq}{Logical for MaxQuant LFQ; if TRUE, omits LFQ entries
with single measured values across all samples. The default is TRUE.}
}
\description{
With LFQ, values of \code{log2_R000 (...)} and \code{N_log2_R000 (...)} in
\code{df_num} are not yet filled after \link{calc_tmt_nums}. This utility
calculates the ratio using the values of LFQ intensity.
}
