% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlfq.R
\name{calc_mat_log2s_to_refs}
\alias{calc_mat_log2s_to_refs}
\title{LFQ helper: calculate log2FC values from a matrix of samples and references}
\usage{
calc_mat_log2s_to_refs(mat, are_smpls, are_refs)
}
\arguments{
\item{mat}{A numeric matrix.}

\item{are_smpls}{Logical; are samples or not.}

\item{are_refs}{Logical; are references or not.}
}
\description{
LFQ helper: calculate log2FC values from a matrix of samples and references
}
