% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlfq.R
\name{calc_mat_sds}
\alias{calc_mat_sds}
\title{LFQ helper: calculate SD for matrix data}
\usage{
calc_mat_sds(mat)
}
\arguments{
\item{mat}{A data matrix.}
}
\description{
LFQ helper: calculate SD for matrix data
}
