% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calc_mono_pep}
\alias{calc_mono_pep}
\title{Peptide molecular weight}
\usage{
calc_mono_pep(aa_seq, digits = 4)
}
\arguments{
\item{aa_seq}{Character string; the sequence of a peptide with one-letter
representation of amino acids.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Calculates the mono-isotopic molecular weight of a polypeptide (neutral
species).
}
\examples{
\dontrun{
calc_mono_pep("AAIDWFDGKEFSGNPIK")
}

}
