% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calc_monopep}
\alias{calc_monopep}
\title{Calculates molecular weight of a polypeptide ([MH]+).}
\usage{
calc_monopep(aa_seq, digits = 5)
}
\arguments{
\item{aa_seq}{Character string; the sequence of a peptide with one-letter
representation of amino acids.}

\item{digits}{Integer; the number of decimal places to be used. Not currently
used.}
}
\description{
Calculates molecular weight of a polypeptide ([MH]+).
}
\examples{
\dontrun{
x <- calc_monopep("AAIDWFDGKEFSGNPIK")
}
}
