% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{calc_monopep2}
\alias{calc_monopep2}
\title{With position combinations}
\usage{
calc_monopep2(
  aa_seq,
  aa_masses,
  vmods_ps,
  amods,
  tmod,
  ntmod,
  ctmod,
  mod_indexes = NULL,
  maxn_vmods_per_resid = 3,
  maxn_vmods_sitescombi_per_pep = 64,
  digits = 5
)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{vmods_ps}{A named list of positions and sites with positions being names
and sites being values.}

\item{amods}{Anywhere modifications subset from \code{vmods_ps}.}

\item{tmod}{A terminal modification subset from \code{vmod_ps}.}

\item{ntmod}{A N-term modification subset from \code{tmod}.}

\item{ctmod}{A C-term modification subset from \code{tmod}.}

\item{maxn_vmods_sitescombi_per_pep}{Integer; the maximum number of
combinatorial variable modifications per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
For future usage.
}
