% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils_to_be_removed.R
\name{calc_more_psm_sd}
\alias{calc_more_psm_sd}
\title{Add Z_log2_R, sd_N_log2_R and sd_Z_log2_R}
\usage{
calc_more_psm_sd(df, group_psm_by, range_log2r, range_int, set_idx,
  injn_idx)
}
\description{
Add Z_log2_R, sd_N_log2_R and sd_Z_log2_R
}
