% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{calc_mq_log2r}
\alias{calc_mq_log2r}
\title{Helper of calculating MaxQuang log2FC}
\usage{
calc_mq_log2r(df, type, refChannels)
}
\arguments{
\item{df}{A data frame.}

\item{type}{of of intensity data.}

\item{refChannels}{Reference channels.}
}
\description{
Helper of calculating MaxQuang log2FC
}
