% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{calc_ms2ions}
\alias{calc_ms2ions}
\title{Helper With position combinations}
\usage{
calc_ms2ions(
  aa_seq,
  mass,
  aa_masses,
  mod_indexes,
  type_ms2ions = "byall",
  vmods_ps,
  amods,
  tmod,
  ntmod,
  ctmod,
  maxn_vmods_per_pep = 3,
  maxn_sites_per_vmod = 5,
  maxn_vmods_sitescombi_per_pep = 32,
  digits = 5
)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}.}

\item{vmods_ps}{A named list of positions and sites with positions being names
and sites being values.}

\item{amods}{Anywhere modifications subset from \code{vmods_ps}.}

\item{tmod}{A terminal modification subset from \code{vmod_ps}.}

\item{ntmod}{A N-term modification subset from \code{tmod}.}

\item{ctmod}{A C-term modification subset from \code{tmod}.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{maxn_vmods_sitescombi_per_pep}{Integer; the maximum number of
combinatorial variable modifications per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Helper With position combinations
}
