% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_a1_vnl1_fnl0.R
\name{calc_ms2ions_a1_vnl1_fnl0}
\alias{calc_ms2ions_a1_vnl1_fnl0}
\title{Calculates}
\usage{
calc_ms2ions_a1_vnl1_fnl0(
  vmods_combi,
  vnl_combi,
  aas2,
  aa_masses,
  ntmass,
  ctmass,
  type_ms2ions,
  digits
)
}
\arguments{
\item{ms1_mass}{The mass of a theoretical MS1 (for subsetting).}
}
\description{
Calculates
}
