% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{calc_pep_retsd}
\alias{calc_pep_retsd}
\title{Calculates the standard deviation in the retention times of peptides.}
\usage{
calc_pep_retsd(df, group_psm_by = "pep_seq_mod", use_unique = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{use_unique}{Logical; if TRUE, filter data by uniqueness.}
}
\description{
Different charge states can have small differences in \code{pep_apex_ret},
e.g., due to the spikiness of peaks.
}
\details{
Sets \code{use_unique = TRUE} when used with \link{normPSM} where the same
\code{group_psm_by} can be redundant by \code{prot_acc}. Sets
\code{use_unique = FALSE} when used with \link{mergePep} where the same
\code{group_psm_by} from different samples and LCMS series are all taken into
account.
}
