% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_pepfdr}
\alias{calc_pepfdr}
\title{Calculates the cut-off score at a peptide FDR.}
\usage{
calc_pepfdr(out, nms, target_fdr = 0.01, fdr_type = "psm")
}
\arguments{
\item{out}{The output from \link{calc_pepscores}.}

\item{nms}{The name(s) of \code{out} that correspond(s) to decoy results.}

\item{target_fdr}{Numeric; the levels of false-discovery rate.}

\item{fdr_type}{Character string; the type of FDR controlling. The value is
in one of c("psm", "peptide", "protein").}
}
\description{
Calculates the cut-off score at a peptide FDR.
}
