% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{calc_peploc}
\alias{calc_peploc}
\title{Calculates the delta scores of `pep_seq`.}
\usage{
calc_peploc(x)
}
\arguments{
\item{x}{The result from \link{calc_pepscores}.}
}
\description{
A score delta between the best and the second best.
}
\details{
There should not be any duplicated rows by the combination of
c("pep_isdecoy", "scan_num", "raw_file", "pep_seq", "pep_ivmod")
}
