% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_probi}
\alias{calc_probi}
\title{Helper for score calculations}
\usage{
calc_probi(
  mts,
  expts,
  N,
  type_ms2ions = "by",
  topn_ms2ions = 100,
  penalize_sions = FALSE,
  ppm_ms2 = 25,
  digits = 5
)
}
\arguments{
\item{mts}{Nested data frame of \code{theo} and matched \code{expt} m-over-z.}

\item{expts}{Nested data frames of \code{expt}, match and unmatched
experimental m-over-z, and \code{int}, the corresponding experimental
intensity.}

\item{N}{Numeric; the number of MS2 features in an MGF query.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz".}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches.}

\item{penalize_sions}{Logical; if TRUE, penalizes secondary ions of b0, y0
etc. with lower weights in peptide scoring.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Helper for score calculations
}
