% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_probi_bypep}
\alias{calc_probi_bypep}
\title{Helper for score calculations}
\usage{
calc_probi_bypep(
  mts,
  nms,
  expt_moverzs,
  expt_ints,
  N,
  type_ms2ions,
  topn_ms2ions,
  penalize_sions,
  ppm_ms2,
  digits
)
}
\arguments{
\item{mts}{Nested data frame of \code{theo} and matched \code{expt} m-over-z.}

\item{nms}{The names (of peptides).}

\item{expt_moverzs}{Nested list of match and unmatched experimental m-over-z.}

\item{expt_ints}{Nested list of match and unmatched experimental intensity.}

\item{N}{Numeric; the number of MS2 features in an MGF query.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{penalize_sions}{Logical; if TRUE, penalizes secondary ions of b0, y0
etc. with lower weights in peptide scoring.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
By peptides.
}
