% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_protfdr_i_orig}
\alias{calc_protfdr_i_orig}
\title{Helper of \link{calc_protfdr}.}
\usage{
calc_protfdr_i_orig(td, target_fdr = 0.01)
}
\arguments{
\item{td}{A data frame with paired target-decoys.}

\item{target_fdr}{Numeric; the levels of false-discovery rate (FDR).}
}
\description{
Helper of \link{calc_protfdr}.
}
