% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_protscore_i}
\alias{calc_protscore_i}
\title{Helper of \link{calc_protfdr}.}
\usage{
calc_protscore_i(df)
}
\arguments{
\item{df}{A data subset at a given \code{prot_n_pep}.}
}
\description{
Helper of \link{calc_protfdr}.
}
