% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_sd_fcts_psm}
\alias{calc_sd_fcts_psm}
\title{Calculates PSM SDs}
\usage{
calc_sd_fcts_psm(
  df,
  range_log2r = c(5, 95),
  range_int = c(5, 95),
  set_idx,
  injn_idx
)
}
\arguments{
\item{df}{A data frame.
range of the \code{log2FC} for use in the scaling normalization of standard
deviation across samples. The default is between the 10th and the 90th
quantiles.}

\item{range_int}{Numeric vector at length two. The argument specifies the
range of the \code{intensity} of reporter ions (including \code{I000}) for
use in the scaling normalization of standard deviation across samples. The
default is between the 5th and the 95th quantiles.}

\item{set_idx}{A set index.}

\item{injn_inx}{A injection index.}
}
\description{
The standard deviations are based on samples under each TMT set and LCMS.
}
