% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{calc_tmtint}
\alias{calc_tmtint}
\title{Reporter-ion quantitation.}
\usage{
calc_tmtint(
  scores,
  quant = c("none", "tmt6", "tmt10", "tmt11", "tmt16"),
  ppm_ms2 = 25,
  out_path = "~/proteoQ/outs"
)
}
\arguments{
\item{scores}{The results from \link{calc_pepscores}.}

\item{quant}{A quantitation method. The default is "none".}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{out_path}{The output path.}
}
\description{
Reporter-ion quantitation.
}
