% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlfq.R
\name{calc_vec_log2s_to_refs}
\alias{calc_vec_log2s_to_refs}
\title{LFQ helper: calculate log2FC values from a vector of samples and references}
\usage{
calc_vec_log2s_to_refs(vals, are_smpls, are_refs)
}
\arguments{
\item{vals}{A numeric vector.}

\item{are_smpls}{Logical; are samples or not.}

\item{are_refs}{Logical; are references or not.}
}
\description{
LFQ helper: calculate log2FC values from a vector of samples and references
}
