% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{calc_zions}
\alias{calc_zions}
\title{Masses of singly-charged z-ions}
\usage{
calc_zions(ctmod, aas, aa_masses, digits = 5)
}
\arguments{
\item{ctmod}{The parsed \emph{variable} \code{C-term} modifications from
\code{aa_masses}.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Singly charged.
}
\seealso{
\link{add_complement_ions}
}
