% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{calcLFQPepNums}
\alias{calcLFQPepNums}
\title{Calculates log2FC of peptides based on LFQ intensity.}
\usage{
calcLFQPepNums(
  df,
  label_scheme,
  tmt_plex = 0L,
  min_int = 1000,
  imp_refs = FALSE,
  imp_vals = FALSE
)
}
\arguments{
\item{df}{A data frame.}

\item{label_scheme}{Meta data.}

\item{tmt_plex}{The multiplexicty of TMT.}

\item{min_int}{Minimal Y intensity to replace 0 values.}

\item{imp_refs}{Logical; impute missing references or not.}
}
\description{
With LFQ, values of \code{log2_R000 (...)} and \code{N_log2_R000 (...)} in
\code{df_num} are not yet filled after \link{spreadPepNums}. This utility
calculates the ratio using the values of LFQ intensity.
}
\details{
For SILAC, intensity of heave, light etc. have been spread into multiple
columns and thus the log2Ratios can be calculated like regular LFQ.
}
