% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{calcLFQPeptide}
\alias{calcLFQPeptide}
\title{Calculates Mzion peptide data for individual samples}
\usage{
calcLFQPeptide(
  df = NULL,
  group_psm_by = "pep_seq_mod",
  group_pep_by = "prot_acc",
  dat_dir = NULL,
  set_idx = 1L,
  injn_idx = 1L,
  tmt_plex = 1L,
  ok_lfq = TRUE,
  rm_allna = FALSE,
  type_sd = "log2_R"
)
}
\arguments{
\item{df}{A data frame.}

\item{dat_dir}{A working directory.}

\item{set_idx}{A set index.}

\item{injn_idx}{Not used. An injection index.}

\item{tmt_plex}{The multiplexity of TMT.}

\item{ok_lfq}{Logical; performs LFQ or not.}

\item{rm_allna}{Logical; if TRUE, removes data rows that are exclusively NA
across ratio columns of \code{log2_R126} etc. The setting also applies to
\code{log2_R000} in LFQ.}

\item{type_sd}{Character string; the type of log2Ratios for SD calculations.
The value is one \code{log2_R}, \code{N_log2_R} or \code{Z_log2_R}.}
}
\description{
Argument \code{injn_idx} does not currently used.
}
