% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{calcms1mass_noterm_bypep}
\alias{calcms1mass_noterm_bypep}
\title{Helper of \link{calcms1mass_noterm_byprot}.}
\usage{
calcms1mass_noterm_bypep(
  aa_seq,
  aa_masses,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  digits = 4L
)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
For single protein.
}
