% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{calcms1mass_noterm_byprot}
\alias{calcms1mass_noterm_byprot}
\title{Helper of \link{calcms1mass_noterm}.}
\usage{
calcms1mass_noterm_byprot(
  prot_peps,
  aa_masses,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  digits = 4L
)
}
\arguments{
\item{prot_peps}{Lists of peptides under a proteins.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
For single protein.
}
