% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{calcms2_a1_t1_nl1}
\alias{calcms2_a1_t1_nl1}
\title{Helper in calculating MS2 ion masses.}
\usage{
calcms2_a1_t1_nl1(
  vmods_combi,
  nl_combi,
  ntmod,
  ctmod,
  aas,
  aa_masses,
  type_ms2ions,
  mod_indexes = NULL,
  digits
)
}
\arguments{
\item{vmods_combi}{Lists of variable modifications.}

\item{nl_combi}{Lists of combinations of neutral losses for corresponding
\code{vmods_combi}. Each list contains a table where each column
corresponds to a set of neutral loss. The first column corresponds to the
combination without NLs.}

\item{ntmod}{The parsed \emph{variable} \code{N-term} modifications from
\code{aa_masses}.}

\item{ctmod}{The parsed \emph{variable} \code{C-term} modifications from
\code{aa_masses}.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{mod_indexes}{Integer; the indexes of fixed and/or variable
modifications.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
(10) "amods+ tmod+ vnl+ fnl-".
}
\details{
Updates \code{aas} (since `amods+`) and runs column-wisely through the table
of \code{nl_combi} (variable) for corresponding masses of residues.
}
