% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calcpep_0}
\alias{calcpep_0}
\title{Helper in calculating the mass of an unmodified peptide.}
\usage{
calcpep_0(aa_seq, aas, aa_masses, digits)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
(1) "amods- tmod- vnl- fnl-".
}
