% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calcpep_a0_t1_nl0}
\alias{calcpep_a0_t1_nl0}
\title{Helper in calculating peptide masses.}
\usage{
calcpep_a0_t1_nl0(aa_seq, aas, aa_masses, ntmod, ctmod, digits)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{ntmod}{The parsed \emph{variable} \code{N-term} modifications from
\code{aa_masses}.}

\item{ctmod}{The parsed \emph{variable} \code{C-term} modifications from
\code{aa_masses}.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
(2) "amods- tmod+ vnl- fnl-".
}
