% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{calcPepide}
\alias{calcPepide}
\title{Calculate peptide data for individual TMT experiments
Argument injn_idx not used}
\usage{
calcPepide(df, label_scheme, id, method_psm_pep, group_pep_by, set_idx,
  injn_idx)
}
\description{
Calculate peptide data for individual TMT experiments
Argument injn_idx not used
}
