% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calcpepmass}
\alias{calcpepmass}
\title{Helper of \link{calc_mvmods_masses}.}
\usage{
calcpepmass(aas, aa_masses, ntmod, ctmod)
}
\arguments{
\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{ntmod}{The \code{N-term} modification.}

\item{ctmod}{The \code{C-term} modification.}
}
\description{
Helper of \link{calc_mvmods_masses}.
}
