% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calcpepmass_nl0}
\alias{calcpepmass_nl0}
\title{Helper of peptide-mass calculation..}
\usage{
calcpepmass_nl0(aas, aa_masses, ntmod, ctmod)
}
\arguments{
\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{ntmod}{The parsed \emph{variable} \code{N-term} modifications from
\code{aa_masses}.}

\item{ctmod}{The parsed \emph{variable} \code{C-term} modifications from
\code{aa_masses}.}
}
\value{
A numeric scalar
}
\description{
Without NLs: "fnl-" or "vnl-".
}
