% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calcpepmass_vnl1}
\alias{calcpepmass_vnl1}
\title{Helper of peptide-mass calculation..}
\usage{
calcpepmass_vnl1(aas, nl_combi, aa_masses, ntmod, ctmod)
}
\arguments{
\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{nl_combi}{Lists of combinations of neutral losses for corresponding
\code{vmods_combi}. Each list contains a table where each column
corresponds to a set of neutral loss. The first column corresponds to the
combination without NLs.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{ntmod}{The parsed \emph{variable} \code{N-term} modifications from
\code{aa_masses}.}

\item{ctmod}{The parsed \emph{variable} \code{C-term} modifications from
\code{aa_masses}.}
}
\value{
A numeric vector
}
\description{
With NLs: "vnl+".
}
\details{
The calculation goes through the columns in \code{nl_combi}.
}
