% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{calcPeptide}
\alias{calcPeptide}
\title{Calculates peptide data for individual TMT experiments}
\usage{
calcPeptide(
  df = NULL,
  group_psm_by = "pep_seq",
  method_psm_pep = "median",
  group_pep_by = "prot_acc",
  dat_dir = NULL,
  set_idx = 1L,
  injn_idx = 1L,
  TMT_plex = 10L,
  lfq_ret_tol = 60L,
  rm_allna = FALSE,
  type_sd = "log2_R",
  engine = "mz"
)
}
\arguments{
\item{df}{A data frame containing the PSM table from database searches.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{method_psm_pep}{Character string; the method to summarize the
\code{log2FC} and the \code{intensity} of \code{PSMs} by peptide entries.
The descriptive statistics includes \code{c("mean", "median",
"weighted_mean", "top_3_mean", "lfq_max", "lfq_top_2_sum", "lfq_top_3_sum",
"lfq_all")} with \code{median} being the default for TMT and
\code{lfq_top_2_sum} for LFQ. The \code{log10-intensity} of reporter (or
LFQ) ions at the \code{PSMs} levels will be the weight when summarizing
\code{log2FC} with various \code{"top_n"} statistics or
\code{"weighted_mean"}.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{set_idx}{Numeric.  The index of a multiplex TMT experiment in metadata
files such as \code{label_scheme.xlsx} and \code{frac_scheme.xlsx}.}

\item{injn_idx}{Numeric. The index of \code{LCMS_Inj} in metadata files such
as \code{label_scheme.xlsx} and \code{frac_scheme.xlsx}.}

\item{TMT_plex}{Numeric; the multiplexity of TMT, i.e., 10, 11 etc.}

\item{lfq_ret_tol}{The tolerance of retention time (in seconds) for the
aggregation of LFQ data.}

\item{rm_allna}{Logical; if TRUE, removes data rows that are exclusively NA
across ratio columns of \code{log2_R126} etc. The setting also applies to
\code{log2_R000} in LFQ.}

\item{type_sd}{Character string; the type of log2Ratios for SD calculations.
The value is one \code{log2_R}, \code{N_log2_R} or \code{Z_log2_R}.}

\item{engine}{The name of search engine.}
}
\description{
Argument \code{injn_idx} does not currently used.
}
