% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialty.R
\name{calcTMTLabs}
\alias{calcTMTLabs}
\title{Calculates labeling efficiency}
\usage{
calcTMTLabs(df, file)
}
\arguments{
\item{df}{A PSM table}

\item{file}{A file name.}
}
\description{
Calculates labeling efficiency
}
