% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{calcTMTLFQPeptide}
\alias{calcTMTLFQPeptide}
\title{Calculates peptide data for individual TMT experiments}
\usage{
calcTMTLFQPeptide(
  df = NULL,
  group_psm_by = "pep_seq",
  method_psm_pep = "median",
  group_pep_by = "prot_acc",
  dat_dir = NULL,
  set_idx = 1L,
  injn_idx = 1L,
  tmt_plex = 10L,
  rm_allna = FALSE,
  type_sd = "log2_R",
  engine = "mz"
)
}
\arguments{
\item{df}{A data frame.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{method_psm_pep}{Depreciated. Character string; the method to summarize
the \code{log2FC} and the \code{intensity} of \code{PSMs} by peptide
entries. The \code{log10-intensity} of reporter ions at the \code{PSMs}
levels will be the weight when summarizing \code{log2FC} with various
\code{"top_n"} statistics or \code{"weighted_mean"}. The choices of summary
statistics for LFQ are depreciated at the users' levels.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{dat_dir}{A working directory.}

\item{set_idx}{A set index.}

\item{injn_idx}{Not used. An injection index.}

\item{tmt_plex}{The multiplexity of TMT.}

\item{rm_allna}{Logical; if TRUE, removes data rows that are exclusively NA
across ratio columns of \code{log2_R126} etc. The setting also applies to
\code{log2_R000} in LFQ.}

\item{type_sd}{Character string; the type of log2Ratios for SD calculations.
The value is one \code{log2_R}, \code{N_log2_R} or \code{Z_log2_R}.}

\item{engine}{A search engine name.}
}
\description{
For TMT and also for LFQ other than Mzion to flow through. Entries under the
same \code{pep_seq_mod} at different \code{z} are collapsed.
}
