% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{calib_ms1rt}
\alias{calib_ms1rt}
\title{Calibrate MS1 retention times}
\usage{
calib_ms1rt(tmat)
}
\arguments{
\item{tmat}{A matrix of retention times. Column correspondance: pep_seq_modz;
row correspondance: individual peptide tables.}
}
\description{
Calibrate MS1 retention times
}
