% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{calibMS1RT}
\alias{calibMS1RT}
\title{Calibrate MS1 retention times}
\usage{
calibMS1RT(tmat)
}
\arguments{
\item{tmat}{A matrix of retention times. Column correspondence: pep_seq_modz;
row correspondence: individual peptide tables.}
}
\description{
Calibrate MS1 retention times
}
