% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{calibPSMApexs2}
\alias{calibPSMApexs2}
\title{Calibrate PSM apex retention times under \code{apex_ts}}
\usage{
calibPSMApexs2(df = NULL, fit = NULL, key = "apex_ts")
}
\arguments{
\item{df}{A subset of PSM table, e.g. \code{TMTSet1_LCMSinj1_PSM_N.txt},
under a single \code{RAW_File}.}

\item{fit}{A model for RT calibration.}

\item{key}{A column key for RT calibration.}
}
\description{
Calibrate PSM apex retention times under \code{apex_ts}
}
