% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{cbind_theopepes}
\alias{cbind_theopepes}
\title{Combines theoretical peptides after binning.}
\usage{
cbind_theopepes(out, out_nm)
}
\arguments{
\item{out}{A list of binned theoretical peptides.}

\item{out_nm}{The output file path and name.}
}
\description{
Combines theoretical peptides after binning.
}
