% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{center_df}
\alias{center_df}
\title{Updates df after normalization.}
\usage{
center_df(df, label_scheme_fit, cf_x_fit, sd_coefs_fit)
}
\arguments{
\item{df}{A data frame of peptide or protein table.}

\item{label_scheme_fit}{Experiment summary for samples being selected for
fitting.}

\item{cf_x_fit}{A data frame containing the \code{x} positions for each
samples indicated in label_scheme_fit.}

\item{sd_coefs_fit}{The standard deviations for each samples indicated in
label_scheme_fit.}
}
\description{
At a given centering method, the following three take place:
}
\details{
(1) \code{Z_log2_R} centered and SD adjusted;
(2) \code{N_log2_R} centered;
(3) \code{N_I} scaled.
}
