% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{channelInfo}
\alias{channelInfo}
\title{Extracts the channel information in TMT experiments}
\usage{
channelInfo(dat_dir = NULL, set_idx = NULL, injn_idx = 1)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{set_idx}{Numeric.  The index of a multiplex TMT experiment in metadata
files such as \code{label_scheme.xlsx} and \code{frac_scheme.xlsx}.}

\item{injn_idx}{Numeric. The index of \code{LCMS_Inj} in metadata files such
as \code{label_scheme.xlsx} and \code{frac_scheme.xlsx}.}
}
\value{
Three lists of indexes: \code{refChannels}, reference channels(s);
  \code{emptyChannels}, empty channel(s) that were not used for sample
  labeling; \code{labeledChannels}, non-empty channels including both
  reference(s) and sample(s).
}
\description{
A function returns the indexes of TMT channels that are associated to
reference(s), sample(s) and probable unused void(s).
}
