% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_aes_length}
\alias{check_aes_length}
\title{Checks lengths of aesthetics.}
\usage{
check_aes_length(
  label_scheme = NULL,
  x = "Size",
  aes = "size_manual",
  vals = c(3, 5)
)
}
\arguments{
\item{label_scheme}{A metadata.}

\item{x}{A column name in metadata.}

\item{aes}{The name of an aesthetics.}

\item{vals}{Values of manual aesthetics.}
}
\description{
The length between metadata and manual aesthetics.
}
