% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_channel_prefix}
\alias{check_channel_prefix}
\title{Adds the \code{TMT-} prefix.}
\usage{
check_channel_prefix(x, is_tmt = TRUE)
}
\arguments{
\item{x}{Values under \code{TMT_Channel} in metadata.}

\item{is_tmt}{Logical; is TMT experiments or not.}
}
\description{
Adds the \code{TMT-} prefix.
}
