% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_depreciated_args}
\alias{check_depreciated_args}
\title{check depreciated arguments}
\usage{
check_depreciated_args(blacklist = NULL, ...)
}
\arguments{
\item{blacklist}{A character vector of variable names.}

\item{...}{A list of arguments for checking.}
}
\description{
check depreciated arguments
}
