% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{check_dixon_outliers}
\alias{check_dixon_outliers}
\title{Check the existence of an outlier using Dixon's method.}
\usage{
check_dixon_outliers(x, y = NULL, p = 0.05, p_co = 0.01)
}
\arguments{
\item{p}{A type-I error tolerance.}
}
\description{
Check the existence of an outlier using Dixon's method.
}
