% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dots}
\alias{check_dots}
\title{Checks duplicate argments in 'dots'}
\usage{
check_dots(blacklist = NULL, ...)
}
\arguments{
\item{blacklist}{A character vector of variable names.}

\item{...}{A list of arguments for checking.}
}
\description{
Checks duplicate argments in 'dots'
}
