% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{check_dup_unimods}
\alias{check_dup_unimods}
\title{Checks duplicated \code{Anywhere} variable modifications.}
\usage{
check_dup_unimods(varmods)
}
\arguments{
\item{varmods}{A vector of variable modifications.}
}
\description{
A helper for the option of \code{use_lowercase_aa}.
}
