% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_dups_at_lcms_and_sid}
\alias{check_dups_at_lcms_and_sid}
\title{Checks duplicated Sample_IDs in different LCMS_Injection's for LFQ.}
\usage{
check_dups_at_lcms_and_sid(df)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}
}
\description{
Supposed under a "LCMS_Injection + Sample_ID", there should be only one row.
If there is more than one row, likely the RAW_File is duplicated.
}
