% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_empty_rows}
\alias{check_empty_rows}
\title{Removes rows of all NA values}
\usage{
check_empty_rows(df)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}
}
\description{
Removes rows of all NA values
}
