% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_exptfrac_raws}
\alias{check_exptfrac_raws}
\title{Checks the consistency in RAW_File between expt_smry.xlsx and frac_smry.xlsx.}
\usage{
check_exptfrac_raws(
  fraction_scheme = NULL,
  label_scheme_full = NULL,
  filename = "frac_smry.xlsx",
  dat_dir = NULL
)
}
\arguments{
\item{fraction_scheme}{The metadata of frac_smry.}

\item{label_scheme_full}{The metadata of expt_smry.}

\item{filename}{A file name of fraction_scheme.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Checks the consistency in RAW_File between expt_smry.xlsx and frac_smry.xlsx.
}
