% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_formalArgs2}
\alias{check_formalArgs2}
\title{Check conflicts in function arguments.}
\usage{
check_formalArgs2(w = anal_prnTrend, f = NMF::nmf, excludes = NULL)
}
\arguments{
\item{w}{Expression; the wrapper function.}

\item{f}{Expression; the function to be wrapped.}

\item{excludes}{Character string; arguments from \code{f} to be excluded for
checking.}
}
\description{
Conflicts between wrapper and function to be wrapped. Note that \code{f}
contains both function name and package name.
}
