% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_frac_multipsms}
\alias{check_frac_multipsms}
\title{Checks the PSM_File in frac_smry.xlsx}
\usage{
check_frac_multipsms(df = NULL, filename = "frac_smry.xlsx", dat_dir = NULL)
}
\arguments{
\item{df}{A data frame of metadata.}

\item{filename}{A file name of fraction_scheme.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Also made changes to the PSM_File.
}
