% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspa.R
\name{check_gspa_subcellular}
\alias{check_gspa_subcellular}
\title{Check the suitability of subcellular GSPA.}
\usage{
check_gspa_subcellular(
  fml_nm = NULL,
  dat_dir = NULL,
  label_scheme_sub,
  col_subcellular = NULL,
  col_group = NULL
)
}
\arguments{
\item{fml_nm}{A formula name}

\item{dat_dir}{The working directory.}

\item{label_scheme_sub}{Metadata.}

\item{col_group}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_group} will be
used for sample grouping in the indicated analysis. At the NULL default, the
column key \code{Group} will be used. No data annotation by groups will be
performed if the fields under the indicated group column is empty.}
}
\description{
Check the suitability of subcellular GSPA.
}
