% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_lfq_exptraws}
\alias{check_lfq_exptraws}
\title{Checks duplicated RAW_Files in expt_smry.xlsx for LFQ.}
\usage{
check_lfq_exptraws(df)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}
}
\description{
Also makes changes to the input \code{df}.
}
