% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_metadata_integers}
\alias{check_metadata_integers}
\title{Checks the integerish of columns.}
\usage{
check_metadata_integers(df, is_tmt, filename = "expt_smry.xlsx")
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}

\item{is_tmt}{Logical; is TMT experiments or not.}

\item{filename}{A name of metadata file (e.g., expt_smry.xlsx).}
}
\description{
Used integerish (to handle NA values).
}
