% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{check_mq_df}
\alias{check_mq_df}
\title{Compare sample IDS between MaxQuant results and expt_smry.xlsx}
\usage{
check_mq_df(df, label_scheme)
}
\arguments{
\item{df}{A data frame of MaxQuant peptides.txt or proteinGroups.txt.}
}
\description{
Compare sample IDS between MaxQuant results and expt_smry.xlsx
}
