% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{check_ms1_mass_vmods}
\alias{check_ms1_mass_vmods}
\title{Check the MS1 mass for proceeding with MS2 matches.}
\usage{
check_ms1_mass_vmods(
  vmods_combi,
  aas,
  aa_masses,
  ntmod,
  ctmod,
  ms1_mass,
  tol = 2e-05
)
}
\arguments{
\item{vmods_combi}{Lists of variable modifications.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{ntmod}{The parsed \emph{variable} \code{N-term} modifications from
\code{aa_masses}.}

\item{ctmod}{The parsed \emph{variable} \code{C-term} modifications from
\code{aa_masses}.}

\item{ms1_mass}{The mass of a theoretical MS1 (for subsetting).}

\item{tol}{The tolerance in mass.}
}
\description{
Check the MS1 mass for proceeding with MS2 matches.
}
