% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_a1_vnl0_fnl0.R
\name{check_ms1_mass_vmods2}
\alias{check_ms1_mass_vmods2}
\title{Check the MS1 mass for proceeding with MS2 matches.}
\usage{
check_ms1_mass_vmods2(
  vmods_combi,
  aas2,
  aa_masses,
  ntmod,
  ctmod,
  ms1_mass,
  tol = 0.001
)
}
\arguments{
\item{vmods_combi}{Lists of variable modifications.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{ntmod}{The parsed \emph{variable} \code{N-term} modifications from
\code{aa_masses}.}

\item{ctmod}{The parsed \emph{variable} \code{C-term} modifications from
\code{aa_masses}.}

\item{ms1_mass}{The mass of a theoretical MS1 (for subsetting).}

\item{tol}{The tolerance in mass.}
}
\description{
Maybe missed if the "matched" occurred after `maxn_vmods_sitescombi_per_pep`.
}
