% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{check_outliers}
\alias{check_outliers}
\title{Check the existence of an outlier.}
\usage{
check_outliers(x, y = NULL, p = 0.05, min_n = 3L, p_co = 0.01)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A second vector, e.g., p-values.}

\item{p}{A type-I error tolerance.}

\item{min_n}{The minimal number of values in \code{x} for consideration.}

\item{p_co}{A y (e.g., pVals from \link{pepSig}) cutoff for assessing the
significance of \code{x}. Only effective when \code{!is.null(y)}.}
}
\description{
Check the existence of an outlier.
}
