% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{check_raws}
\alias{check_raws}
\title{Find mismatches in RAW file names}
\usage{
check_raws(df)
}
\arguments{
\item{df}{A data frame containing the PSM table from database searches.}
}
\description{
\code{check_raws} finds mismatched RAW files between expt_smry.xlsx and
PSM outputs.
}
