% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_required_cols}
\alias{check_required_cols}
\title{Checks the presence of required columns in metadata.}
\usage{
check_required_cols(df, filename = NULL)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}

\item{filename}{A name of metadata file (e.g., expt_smry.xlsx).}
}
\description{
Checks the presence of required columns in metadata.
}
