% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{check_rosner_outliers}
\alias{check_rosner_outliers}
\title{Check the existence of an outlier using Rosner's method.}
\usage{
check_rosner_outliers(x, y = NULL, k = 1L, p = 0.05, p_co = 0.01)
}
\arguments{
\item{k}{An allowance of the number of outliers.}

\item{p}{A p-value cut-off.}
}
\description{
Check the existence of an outlier using Rosner's method.
}
