% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{check_tmod_p}
\alias{check_tmod_p}
\title{Checks the compatibility between terminal and anywhere modifications.}
\usage{
check_tmod_p(aas, residue, p, ntmod, ctmod)
}
\arguments{
\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{residue}{The residue (e.g. "M").}

\item{p}{The position indexes of \code{residue}.}

\item{ntmod}{The parsed fixed or variable \code{N-term} modifications from
\code{aa_masses}.}

\item{ctmod}{The parsed fixed or variable \code{C-term} modifications from
\code{aa_masses}.}
}
\value{
Updated position \code{p}.
}
\description{
For both fixed and variable terminal modifications
}
