% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_tmt126_row}
\alias{check_tmt126_row}
\title{Checks the rows TMT-126 in metadata.}
\usage{
check_tmt126_row(df)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}
}
\description{
Checks the rows TMT-126 in metadata.
}
