% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_tmt_chans_vs_levs}
\alias{check_tmt_chans_vs_levs}
\title{Checks the consistency between TMT_Channel's and TMT_level's.}
\usage{
check_tmt_chans_vs_levs(df, TMT_levels, filename = "expt_smry.xlsx")
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}

\item{TMT_levels}{The levels of TMT channels.}

\item{filename}{A name of metadata file (e.g., expt_smry.xlsx).}
}
\description{
Checks the consistency between TMT_Channel's and TMT_level's.
}
