% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_tmt_emptyids}
\alias{check_tmt_emptyids}
\title{Checks the uniqueness of Sample_ID for Empty's.}
\usage{
check_tmt_emptyids(df)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}
}
\description{
Under the same TMT_Set but different LCMS_Injection, it is allowed that one
is \code{sample_1} and another is \code{Empty.xxx}.
}
