% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_tmt_mixplexes}
\alias{check_tmt_mixplexes}
\title{Checks TMT_Channel after \link{check_channel_prefix} where "126" becomes
"TMT-126" etc.}
\usage{
check_tmt_mixplexes(df, TMT_levels, TMT_plex)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}

\item{TMT_levels}{The levels of TMT channels.}

\item{TMT_plex}{The multiplexicity of TMT.}
}
\description{
For example, 16-plex TMT may become 10-plex after Sample_ID removals. whereas
TMT_Channel may include 134 etc. The utility checks the compatibilty of
TMT_Channel's at mix-plexes.
}
