% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_tmt_nc}
\alias{check_tmt_nc}
\title{Standardizes TMT channel name.}
\usage{
check_tmt_nc(df)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}
}
\description{
For trailing N or C.
}
