% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_tmt_plex}
\alias{check_tmt_plex}
\title{Standardizes TMT-plexes.}
\usage{
check_tmt_plex(df, TMT_plex)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}

\item{TMT_plex}{The multiplexity of TMT.}
}
\description{
Standardizes TMT-plexes.
}
