% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_tmt_rawfiles}
\alias{check_tmt_rawfiles}
\title{Checks the uniqueness of RAW_File per TMT_Set and LCMS_Injection.}
\usage{
check_tmt_rawfiles(df)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}
}
\description{
Checks the uniqueness of RAW_File per TMT_Set and LCMS_Injection.
}
