% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_tmt_ref}
\alias{check_tmt_ref}
\title{Checks the \code{Reference} channels in metadata}
\usage{
check_tmt_ref(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Checks the \code{Reference} channels in metadata
}
