% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_tmt_sampleids}
\alias{check_tmt_sampleids}
\title{Checks the counts of Sample_ID under each TMT_Set and LCMS_Injection.
 
Should equal to the TMT_plex.}
\usage{
check_tmt_sampleids(df, TMT_plex)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}

\item{TMT_plex}{Numeric; the multiplexity of TMT.}
}
\description{
Checks the counts of Sample_ID under each TMT_Set and LCMS_Injection.
 
Should equal to the TMT_plex.
}
