% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{checkdots_analTrend}
\alias{checkdots_analTrend}
\title{Check the dot-dot-dot of \link{anal_prnTrend}.}
\usage{
checkdots_analTrend(choice, ...)
}
\arguments{
\item{choice}{Character string; the clustering method in \code{c("cmeans",
"clara", "kmeans", "pam", "fanny")}. The default is "cmeans".}

\item{...}{A character vector of argument(s) in \code{dots}.}
}
\description{
Stop if \code{...} contains disabled argument(s) for a given \code{choice}.
}
