% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{chunk_groupsplit}
\alias{chunk_groupsplit}
\title{Splits data by groups then into chunks.}
\usage{
chunk_groupsplit(data, f, n_chunks)
}
\arguments{
\item{data}{Input data.}

\item{f}{A factor; see also base \code{split}.}

\item{n_chunks}{The number of chunks.}
}
\description{
Splits data by groups then into chunks.
}
