% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{chunksplit}
\alias{chunksplit}
\title{Splits data into chunks by length.}
\usage{
chunksplit(data, n_chunks = 5L, type = "list")
}
\arguments{
\item{data}{Input data.}

\item{n_chunks}{The number of chunks.}

\item{type}{The type of data for splitting.}
}
\description{
Splits data into chunks by length.
}
