% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{chunksplitLB}
\alias{chunksplitLB}
\title{Splits data into chunks with approximately equal sizes.}
\usage{
chunksplitLB(data, n_chunks = 5L, nx = 100L, type = "list")
}
\arguments{
\item{data}{Input data.}

\item{n_chunks}{The number of chunks.}

\item{nx}{Positive integer; an arbitrarily large number for data to be split
into for estimating the cumulative sizes.}

\item{type}{The type of data for splitting.}
}
\description{
Splits data into chunks with approximately equal sizes.
}
