% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{cleanPSMYTS}
\alias{cleanPSMYTS}
\title{Clean up PSM intensity (Y), retention time (T) and scan values (S)}
\usage{
cleanPSMYTS(
  dfs,
  ymat,
  tmat,
  smat,
  fmat,
  nmat,
  xsmat = NULL,
  ysmat = NULL,
  tsmat = NULL,
  ssmat = NULL,
  fsmat = NULL,
  nsmat = NULL,
  key = "pep_seq_modz",
  update = TRUE,
  update2 = FALSE,
  cols = c("pep_tot_int", "N_I000", "I000", "pep_apex_ret", "pep_apex_scan",
    "pep_apex_fwhm", "pep_apex_n")
)
}
\arguments{
\item{dfs}{A list of PSM tables.}

\item{ymat}{A matrix of apex intensities.}

\item{tmat}{A matrix of apex retention times.}

\item{smat}{A matrix of apex scan numbers.}

\item{fmat}{A matrix of apex FWHM values.}

\item{nmat}{A matrix of apex peak widths measured by the number of MS1 scans.}

\item{xsmat}{A matrix of sets of moverz values.}

\item{ysmat}{A matrix of sets of apexes intensities.}

\item{tsmat}{A matrix of sets of retention times.}

\item{ssmat}{A matrix of sets of scan numbers.}

\item{fsmat}{A matrix of sets of FWHM values.}

\item{nsmat}{A matrix of sets of  peak widths measured by the number of MS1
scans.}

\item{key}{The column key for matching.}

\item{update}{Logical; update retention times or not. FALSE if dfs are
already RT calibrated and the purpose is to nullify bad entries according
to tmat.}

\item{cols}{The column keys for value updates.}
}
\description{
Clean up PSM intensity (Y), retention time (T) and scan values (S)
}
