% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{cleanRTChim}
\alias{cleanRTChim}
\title{Clean PSM data with multiple precursors}
\usage{
cleanRTChim(df = NULL)
}
\arguments{
\item{df}{A PSM table with \code{pep_n_apexes > 1L}.}
}
\description{
Matched by mass (among chimeric entries).
}
