% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{cleanupPSM}
\alias{cleanupPSM}
\title{Cleans Up PSM results}
\usage{
cleanupPSM(rm_outliers = FALSE)
}
\arguments{
\item{rm_outliers}{Logical; if TRUE, PSM outlier removals will be performed
for peptides with more than two identifying PSMs. Dixon's method will be
used when \eqn{2 < n \le 25} and Rosner's method will be used when \eqn{n >
25}. The default is FALSE.}
}
\description{
\code{cleanupPSM} removes PSM outliers after \code{splitPSM},
\code{splitPSM_mq} or \code{splitPSM_sm}. The outlier removals will be
assessed at the basis of per peptide per TMT channel.
}
\details{
Dixon's method will be used when \eqn{2 < n \le 25}; Rosner's method will be
used when \eqn{n > 25}.
}
