% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluego_api.R
\name{cluego}
\alias{cluego}
\title{ClueGO via API}
\usage{
cluego(
  df2 = "Protein_Trend_Z_nclust5.txt",
  species = c(human = "Homo Sapiens"),
  n_clust = 3L
)
}
\arguments{
\item{df2}{Character string; the name of a secondary data file from
\code{\link{anal_prnTrend}}, such as Protein_Trend_Z_nclust6.txt etc.}

\item{species}{Named character string . The default is \code{c(human = "Homo
Sapiens")}. In this example, \code{human} is the value that can be found
under the \code{species} column in a Trend result table and \code{"Homo
Sapiens"} is the corresponding species name in ClueGO.}

\item{n_clust}{Numeric vector; the cluster ID(s) corresponding to
\code{\link{anal_prnTrend}} for visualization.}
}
\description{
\code{cluego} sends \code{\link{anal_prnTrend}} findings in, e.g.
\code{Protein_Trend_[...].txt} etc. to
\code{\href{http://apps.cytoscape.org/apps/cluego}{ClueGO}} via API.
}
\examples{
\donttest{
# Make sure CytoSpace is opened, and 
# yFiles layouts and ClueGO plug-in are installed.

# Human
cluego(
  df2 = Protein_Trend_Z_nclust5.txt, 
  species = c(human = "Homo Sapiens"), 
  n_clust = c(3, 5)
)

# Mouse
cluego(
  df2 = Protein_Trend_Z_nclust5.txt, 
  species = c(mouse = "Mus Musculus"), 
  n_clust = c(3:4)
)
}
}
