% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular.R
\name{cluster_svmprobs}
\alias{cluster_svmprobs}
\title{Cluster SVM probabilities.}
\usage{
cluster_svmprobs(
  df_prob = NULL,
  sub_type = NULL,
  dat_dir = NULL,
  filepath = NULL,
  id = "gene",
  n_clust = NULL,
  choice = c("cmeans", "clara", "kmeans", "pam", "fanny"),
  ...
)
}
\arguments{
\item{df_prob}{A data frame of SVM probabilities (under column
\code{loc_score}).}

\item{dat_dir}{The working directory.}

\item{filepath}{A filepath for inputs and outputs.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{n_clust}{The number of clusters.}

\item{choice}{A clustering method.}
}
\description{
Cluster SVM probabilities.
}
