% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{cmbn_meta}
\alias{cmbn_meta}
\title{Combine data with metadata}
\usage{
cmbn_meta(data, metadata)
}
\arguments{
\item{data}{A data frame}

\item{metadata}{Another data frame}
}
\description{
Combine data with metadata
}
