% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{collapse_psm_charges}
\alias{collapse_psm_charges}
\title{Collapse PSMs at good proximity of retention times}
\usage{
collapse_psm_charges(
  df,
  cols_out = c("pep_seq_mod", "pep_seq", "pep_tot_int", "pep_apex_ret", "pep_apex_scan")
)
}
\arguments{
\item{df}{A PSM table with a unique \code{pep_seq_mod}. Also intensities are
in a descending order.}

\item{cols_out}{Output columns}
}
\description{
Each \code{df} is unique by \code{pep_seq_mod}. The PSMs include duplicated
entries, e.g., at different prot_acc.
}
