% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{combi_vmods}
\alias{combi_vmods}
\title{The combinations of variable modifications (single site)}
\usage{
combi_vmods(aa_seq, residue_mods, aa_masses, max_nvmods_per_site, digits)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{residue_mods}{A residue with \code{Anywhere} modification(s).}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{max_nvmods_per_site}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
The combinations of variable modifications (single site)
}
