% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_a1_vnl0_fnl0.R
\name{combi_vmods2}
\alias{combi_vmods2}
\title{The combinations of variable modifications (single site)}
\usage{
combi_vmods2(
  aas,
  residue_mods,
  aa_masses,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 32L,
  digits = 4L
)
}
\arguments{
\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{residue_mods}{A residue with \code{Anywhere} modification(s).}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{maxn_vmods_sitescombi_per_pep}{Integer; the maximum number of
combinatorial variable modifications per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
The combinations of variable modifications (single site)
}
