% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlfq.R
\name{combine_lfq123}
\alias{combine_lfq123}
\title{Combine values before or after when collapsing adjacent bins of retention
times}
\usage{
combine_lfq123(
  xs_cr,
  ys_cr,
  zs_cr,
  xs_bf,
  ys_bf,
  zs_bf,
  xs_af,
  ys_af,
  zs_af,
  iboth,
  are_refs,
  are_smpls,
  spc
)
}
\arguments{
\item{xs_cr}{A vector of retention-time values of the current bin.}

\item{ys_cr}{A vector of intensity values of the current bin.}

\item{zs_cr}{A vector of scan numbers of the current bin.}

\item{xs_bf}{A vector of retention-time values of the preceding bin.}

\item{ys_bf}{A vector of intensity values of the preceding bin.}

\item{zs_bf}{A vector of scan numbers of the preceding bin.}

\item{xs_af}{A vector of retention-time values of the following bin.}

\item{ys_af}{A vector of intensity values of the following bin.}

\item{zs_af}{A vector of scan numbers of the following bin.}

\item{iboth}{A vector of confounding retention times present in both the
preceding and the following columns.}

\item{are_refs}{A logical vector; are references or not.}

\item{are_smpls}{A logical vector; are samples or not.}

\item{spc}{Species center(s).}
}
\description{
Combine values before or after when collapsing adjacent bins of retention
times
}
