% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{combine_olr_by_contr}
\alias{combine_olr_by_contr}
\title{Helper: combined contrast and group outliers by contrast.}
\usage{
combine_olr_by_contr(
  df_contri,
  fml_nm,
  df_olr,
  group_pep_by = "gene",
  filepath = NULL,
  filename = NULL
)
}
\arguments{
\item{df_contri}{A single-column data frame of outliers by contrast, e.g.
\code{A-C}.}

\item{fml_nm}{A formula name.}

\item{df_olr}{A reference data frame of outliers by groups, e.g., \code{A, B,
C}.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{filepath}{An output file path.}

\item{filename}{An output file name.}
}
\description{
Helper: combined contrast and group outliers by contrast.
}
