% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{combine_pepvecs}
\alias{combine_pepvecs}
\title{Combines peptide scores to table.}
\usage{
combine_pepvecs(data, outcol = "pep_score")
}
\arguments{
\item{data}{Nested list of table from ion matches. Each table contains
columns \code{theo} and \code{expt}.}

\item{outcol}{The output column name.}
}
\description{
Combines peptide scores to table.
}
