% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{compile_hdr_mods}
\alias{compile_hdr_mods}
\title{Compiles Mascot modification indexes.}
\usage{
compile_hdr_mods(dupvmods, univmods)
}
\arguments{
\item{dupvmods}{A table of duplicated variable modifcations.}

\item{univmods}{A table of unique variable modifications.}
}
\description{
Compiles Mascot modification indexes.
}
