% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{concat_fml_dots}
\alias{concat_fml_dots}
\title{Concatenate formula(e) to varargs of dots}
\usage{
concat_fml_dots(fmls = NULL, fml_nms = NULL, dots = NULL, anal_type = "zzz")
}
\description{
Concatenate formula(e) to varargs of dots
}
