% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{concat_peps}
\alias{concat_peps}
\title{Concatenates two adjacent peptides in a list.}
\usage{
concat_peps(peps, n = 1)
}
\arguments{
\item{peps}{A list of peptide sequences with a one-letter representation of
amino acid residues.}

\item{n}{The number of miss cleavages.}
}
\description{
Concatenates two adjacent peptides in a list.
}
\examples{
\dontrun{
peps <- c("MESNHK", "SGDGLSGTQK", "EAALR", "ALVQR", "TGYSLVQENGQR")
concat_peps(peps)
}
}
