% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ions.R
\name{contain_pos_site}
\alias{contain_pos_site}
\title{Function factories for checking the existence of an amino-acid residue at a
position.}
\usage{
contain_pos_site(pos)
}
\arguments{
\item{pos}{The Unimod position of a variable modification.}
}
\value{
A function for checking the existence of a residue at the position
  specified by the argument \code{pos}. For each function, it takes a list of
  variable modifications specified by argument \code{vmods} as inputs.
}
\description{
If the the value in a \code{vmods} is one of the LETTERS -> a site is
specified in the corresponding variable modification. Otherwise, a site is in
\code{c("C-term", "N-term")}.
}
