% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ions.R
\name{contain_protntany}
\alias{contain_protntany}
\alias{contain_anyntany}
\alias{contain_protctany}
\alias{contain_anyctany}
\title{Checks if variable modifications are on terminal positions.}
\usage{
contain_protntany(vmods)

contain_anyntany(vmods)

contain_protctany(vmods)

contain_anyctany(vmods)
}
\arguments{
\item{vmods}{A named list of variable modifications.}
}
\description{
Flowchart (2-nt): `Acetyl (Protein N-term)`

Flowchart (4-nt): `Acetyl (N-term)`

Flowchart (2-ct): `Amidated (Protein C-term)`

Flowchart (4-ct): `Amidated (C-term)`
}
\examples{
# `Acetyl (Protein N-term)`
sites <- list(`Acetyl (Protein N-term)` = "N-term", `Oxidation (M)` = "M", `Deamidated (N)` = "N")
positions <- c("Protein N-term", "Anywhere", "Anywhere")
vmods <- purrr::map2(sites, positions, ~ setNames(.x, .y))
contain_protntany(vmods)

# `Acetyl (N-term)`
sites <- list(`Acetyl (N-term)` = "N-term", 
              `Oxidation (M)` = "M", 
              `Deamidated (N)` = "N")
positions <- c("Any N-term", "Anywhere", "Anywhere")
vmods <- purrr::map2(sites, positions, ~ setNames(.x, .y))
contain_anyntany(vmods)

# `Amidated (Protein C-term)`
sites <- list(`Amidated (Protein C-term)` = "C-term", 
              `Oxidation (M)` = "M", 
              `Deamidated (N)` = "N")
positions <- c("Protein C-term", "Anywhere", "Anywhere")
vmods <- purrr::map2(sites, positions, ~ setNames(.x, .y))
contain_protctany(vmods)

# `Amidated (C-term)`
sites <- list(`Amidated (C-term)` = "C-term", 
              `Oxidation (M)` = "M", 
              `Deamidated (N)` = "N")
positions <- c("Any C-term", "Anywhere", "Anywhere")
vmods <- purrr::map2(sites, positions, ~ setNames(.x, .y))
contain_anyctany(vmods)

}
