% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ions.R
\name{contain_termpos_any}
\alias{contain_termpos_any}
\title{Function factories for checking the existence of an amino-acid residue at a
\code{terminal} position.}
\usage{
contain_termpos_any(pos)
}
\arguments{
\item{pos}{The position. It must be terminal in \code{c("Protein N-term",
"Any N-term", "Protein C-term", "Any C-term")}.}
}
\value{
A function for checking the existence of a residue at the
  \code{terminal} position specified by the argument \code{pos}. For each
  function, it takes a list of variable modifications specified by argument
  \code{vmods} as inputs.
}
\description{
A \code{pos} must be terminal in one of \code{c("Protein N-term", "Any
N-term", "Protein C-term", "Any C-term")}.
}
