% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs2.R
\name{conv_to_pepids}
\alias{conv_to_pepids}
\title{Converts protein accessions and peptide sequences to integers.}
\usage{
conv_to_pepids(data, out_path, type = "fwd")
}
\arguments{
\item{out_path}{An output path}

\item{type}{An indicator of either forward or reversed sequences.}

\item{pep_seqs}{Lists of peptide sequences by protein accessions. Peptide
sequences are names and monoisotopic masses in values.}
}
\description{
Converts protein accessions and peptide sequences to integers.
}
\examples{
\donttest{
pep_seqs <- list(ALG2_HUMAN = c("-MAEEQGR" = 819.35450, 
                                "-MAEEQGRER" = 1122.50877 ))

out <- conv_to_pepids(pep_seqs, out_path, "rev")
}
}
