% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{convert_genes_human_to_mouse}
\alias{convert_genes_human_to_mouse}
\title{Convert human genes to mouse.}
\usage{
convert_genes_human_to_mouse(genes, species = "Mus musculus", human = TRUE)
}
\arguments{
\item{genes}{A list of human genes.}

\item{species}{A target species.}

\item{human}{Logical; are the input genes belong to human or not.}
}
\description{
Convert human genes to mouse.
}
