% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{cos_sim}
\alias{cos_sim}
\title{Cosine similarity.}
\usage{
cos_sim(M)
}
\arguments{
\item{M}{An input Matrix.

Vectors are in rows: mtcars[1, ].}
}
\description{
Cosine similarity.
}
\examples{
sim <- cos_sim(as.matrix(mtcars[1:3, ]))

# distances (against normalized vectors)
as.dist(1 - sim)
}
