% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_phosphopeps}
\alias{count_phosphopeps}
\title{Counts phospho.}
\usage{
count_phosphopeps(collapses = c("mn"))
}
\arguments{
\item{collapses}{The modified residues to be collapsed with unmodified
counterparts.}
}
\description{
Not currently used.
}
