% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{countSpecs}
\alias{countSpecs}
\title{Spectrum counts at peptide levels}
\usage{
countSpecs(
  dat_dir = NULL,
  label_scheme_full = NULL,
  type = "PSM",
  filelist = NULL,
  basenames = NULL,
  set_idxes = 1L,
  injn_idxes = 1L,
  group_psm_by = "pep_seq_mod",
  group_pep_by = "prot_acc",
  pep_col = "pep_n_specs",
  prot_col = "prot_n_specs"
)
}
\arguments{
\item{dat_dir}{The working directory.}

\item{label_scheme_full}{The full label scheme.}

\item{type}{The data type.}

\item{filelist}{A list of file names of TMTset1_LCMSinj1_Peptide_N.txt etc
with prepending path.}

\item{basenames}{The base names of file names in \code{filelist}.}

\item{set_idxes}{The indexes of \code{TMT_Set}'s corresponding to
\code{basenames}.}

\item{injn_idxes}{The indexes of \code{LCMS_Inj}'s corresponding to
\code{basenames}.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{pep_col}{The output column name for peptide summary statistics.}

\item{prot_col}{The output column name for protein summary statistics.}
}
\description{
Spectrum counts at peptide levels
}
