% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{create_db_path}
\alias{create_db_path}
\title{Helper to create `db_path`}
\usage{
create_db_path(db_path)
}
\arguments{
\item{db_path}{Character string; the local path for database(s). The default
is \code{"~\\proteoQ\\dbs\\go"}.}
}
\description{
Helper to create `db_path`
}
