% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{create_os_lookup}
\alias{create_os_lookup}
\title{create a lookup table and remove duplicated entries}
\usage{
create_os_lookup(species, os_name, overwrite = FALSE)
}
\arguments{
\item{species}{Character string; the name of a species.}

\item{os_name}{An organism name by UniProt.}

\item{overwrite}{Logical; if TRUE, overwrite the downloaded database(s). The
default is FALSE.}
}
\description{
create a lookup table and remove duplicated entries
}
