% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{cut_protgrps}
\alias{cut_protgrps}
\title{Cuts proteins into groups.}
\usage{
cut_protgrps(mat, out_path)
}
\arguments{
\item{mat}{A logical matrix; peptides in rows and proteins in columns.}

\item{out_path}{A file pth to outputs.}
}
\description{
By the number of shared peptides.
}
