% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{cut_protgrps2}
\alias{cut_protgrps2}
\title{Cuts proteins into groups.}
\usage{
cut_protgrps2(mat = NULL, out_path = NULL)
}
\arguments{
\item{mat}{A logical matrix; peptides in rows and proteins in columns.}

\item{out_path}{A file pth to outputs.}
}
\description{
By the number of shared peptides.
}
