% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-normalization.R
\name{dblTrim}
\alias{dblTrim}
\title{Data normalization}
\usage{
dblTrim(df, range_log2r, range_int)
}
\arguments{
\item{range_log2r}{The range of the \code{log2FC} of peptide/protein entries
for use in the scaling normalization of standard deviation across samples.
The default is between the 20th and the 90th quantiles.}

\item{range_int}{The range of the \code{intensity} of reporter ions for use in
the scaling normalization of standard deviation across samples. The default
is between the 5th and the 95th quantiles.}
}
\value{
A data frame.
}
\description{
\code{dblTrim} doubly trims the \code{log2FC} and reporter-ion intensity by
the given ranges.
}
