% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mzion.R
\name{deduce_mod_indexes}
\alias{deduce_mod_indexes}
\title{Deduces mod_indexes.}
\usage{
deduce_mod_indexes(file = "psmQ.txt", dat_dir)
}
\arguments{
\item{file}{A psmQ file.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Not applicable to fixed modifications or terminal modifications.
}
