% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{delete_files}
\alias{delete_files}
\title{Deletes files under a directory.}
\usage{
delete_files(path, ignores = NULL, ...)
}
\arguments{
\item{path}{A file path.}

\item{ignores}{The file extensions to be ignored.}
}
\description{
The directory will be kept.
}
