% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{delta_ms1_a0_fnl1}
\alias{delta_ms1_a0_fnl1}
\title{Helper of peptide-mass calculation..}
\usage{
delta_ms1_a0_fnl1(fnl_combi, aas, aa_masses)
}
\arguments{
\item{fnl_combi}{A data.frame of combinations of neutral losses for fixed
modifications. Each row corresponds to a set of neutral loss. The first row
corresponds to the combination without NLs (all zeros).}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}
}
\value{
A numeric vector
}
\description{
(5) "amods- tmod+ vnl- fnl+"; (6) "amods- tmod- vnl- fnl+".
}
\details{
The calculation goes through the rows in \code{fnl_combi}.
}
