% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{detect_cores}
\alias{detect_cores}
\title{Detects and suggests the number of CPU cores.}
\usage{
detect_cores(max_n_cores = NULL)
}
\arguments{
\item{max_n_cores}{The maximum number of cores for uses.}
}
\description{
Detects and suggests the number of CPU cores.
}
