% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{distri_fpeps}
\alias{distri_fpeps}
\title{Distributes peptides by fixed modifications.}
\usage{
distri_fpeps(data, max_miss, is_fixed_protnt, is_fixed_protct)
}
\arguments{
\item{data}{Lists of peptides by prot_accs.}

\item{max_miss}{The maximum number of mis-cleavages per peptide sequence.}

\item{is_fixed_protnt}{Logical; is protein N-terminal modification fixed.}

\item{is_fixed_protct}{Logical; is protein C-terminal modification fixed.}
}
\description{
Distributes peptides by fixed modifications.
}
