% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{distri_peps}
\alias{distri_peps}
\title{Distributes peptides by variable modifications.}
\usage{
distri_peps(aa_masses, peps, max_miss)
}
\arguments{
\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{peps}{Lists of peptides, either "fwds" or "revs", from
\link{split_fastaseqs}.}

\item{max_miss}{The maximum number of mis-cleavages per peptide sequence.}
}
\description{
Distributes peptides by variable modifications.
}
