% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{dixon_test}
\alias{dixon_test}
\title{Outlier removals with Dixon's method}
\usage{
dixon_test(x, type = 0, opposite = FALSE, two.sided = TRUE)
}
\arguments{
\item{x}{A data frame.}

\item{type}{The type of Dixon method. See also outliers::dixon.test.}

\item{opposite}{The high or low. See also outliers::dixon.test.}

\item{two.sided}{Logical; is two-sided or not. See also
outliers::dixon.test.}
}
\description{
Outlier removals with Dixon's method
}
