% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{dl_gene_homo}
\alias{dl_gene_homo}
\title{Helper to download gene homology data.}
\usage{
dl_gene_homo(
  url = NULL,
  db_path = "~/proteoQ/dbs/gene_homology",
  overwrite = FALSE
)
}
\arguments{
\item{db_path}{Character string; the local path for database(s). The default
is \code{"~/proteoQ/dbs/msig"}.}

\item{overwrite}{Logical; if TRUE, overwrite the downloaded database(s). The
default is FALSE.}
}
\description{
Helper to download gene homology data.
}
