% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-go.R
\name{dl_msig}
\alias{dl_msig}
\title{Helper to download `gmt`}
\usage{
dl_msig(
 
    msig_url = "https://data.broadinstitute.org/gsea-msigdb/msigdb/release/7.0/c2.all.v7.0.entrez.gmt",
  db_path = "~\\\\proteoQ\\\\dbs\\\\msig",
  overwrite = FALSE
)
}
\description{
Helper to download `gmt`
}
