% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-string.R
\name{dl_stringdbs}
\alias{dl_stringdbs}
\title{Downloads STRING databases}
\usage{
dl_stringdbs(
  species = "human",
  db_path = "~\\\\proteoQ\\\\dbs\\\\string",
  overwrite = FALSE
)
}
\arguments{
\item{species}{Character string; the species. The currently supported species
include \code{human, mouse, rat, fly, cow, dog}. The default is
\code{human}.}

\item{overwrite}{Logical; if TRUE, overwrite the downloaded databse(s). The
default is FALSE.}
}
\description{
Downloads STRING databases
}
\seealso{
\code{\link{anal_prnString}} for protein-protein interaction
 networks.
}
