% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{env_where}
\alias{env_where}
\title{Find the environment of a function}
\usage{
env_where(name, env = rlang::caller_env())
}
\arguments{
\item{name}{The name of a function.}

\item{env}{The environment.}
}
\description{
Find the environment of a function
}
