% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{essential_gsets_interim}
\alias{essential_gsets_interim}
\title{Greedy set cover of GSPA}
\usage{
essential_gsets_interim(out, df, filepath, filename, formula, all_sets,
  greedy_sets)
}
\description{
Greedy set cover of GSPA
}
