% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{exclude_n_misses}
\alias{exclude_n_misses}
\title{Exclude mis-cleavages in a vector.}
\usage{
exclude_n_misses(x, n)
}
\arguments{
\item{x}{A vector of data.}

\item{n}{Integer. The number of mis-cleavages.}
}
\description{
Exclude mis-cleavages in a vector.
}
\seealso{
keep_n_misses
}
