% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{exclude_one_to_n}
\alias{exclude_one_to_n}
\title{Exclude the first \eqn{n + 1} elements in a vector.}
\usage{
exclude_one_to_n(x, n)
}
\arguments{
\item{x}{A vector of data.}

\item{n}{Integer. The number of mis-cleavages.}
}
\description{
Exclude the first \eqn{n + 1} elements in a vector.
}
