% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{ext_aa_masses}
\alias{ext_aa_masses}
\title{Calculates molecular weight of a polypeptide ([MH]+).}
\usage{
ext_aa_masses(
  fixedmods = c("TMT6plex (N-term)", "TMT6plex (K)", "Carbamidomethyl (. = C)"),
  varmods = c("Acetyl (Protein N-term)", "Oxidation (M)", "Deamidated (N)",
    "Gln->pyro-Glu (N-term = Q)"),
  max_ncomb_mods = 64
)
}
\arguments{
\item{max_ncomb_mods}{Integer; the maximum number of combinatorial
variable modifications allowed per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Calculates molecular weight of a polypeptide ([MH]+).
}
\examples{
\donttest{
library(purrr)

x <- ext_aa_masses()
x_att <- map(aa_masses_all, attributes)
names(x_att[[1]])
x_vmods <- map(x_att, `[`, c("vmods"))

x <- ext_aa_masses(c("TMT6plex (N-term)", "TMT6plex (K)", "Carbamidomethyl (C)"), c("Acetyl (N-term)", "Gln->pyro-Glu (N-term = Q)", "Oxidation (M)"))

x <- ext_aa_masses(fixedmods = NULL)
x <- ext_aa_masses(fixedmods = NULL, varmods = NULL)

# Fixed mod, no NL
x <- ext_aa_masses(fixedmods = c("TMT6plex (N-term)", "Carbamidomethyl (. = C)"), varmods = NULL)

# Fixed mod + NL
x <- ext_aa_masses(fixedmods = c("TMT6plex (N-term)", "Carbamidomethyl (. = M)"), varmods = NULL)

# Fixed mod, no NL; var mod, no NL
x <- ext_aa_masses(fixedmods = c("TMT6plex (N-term)", "Carbamidomethyl (. = C)"), varmods = c("Acetyl (N-term)", "Gln->pyro-Glu (N-term = Q)"))

# Fixed mod + NL; var mod + NL
x <- ext_aa_masses(c("TMT6plex (N-term)", "Carbamidomethyl (. = M)",
"Deamidated (. = R)"), c("Acetyl (N-term)", "Gln->pyro-Glu (N-term = Q)",
"Hex(5)HexNAc(2) (N)"))
}
\dontrun{
# conflicts
x <- ext_aa_masses(c("Carbamidomethyl (N-term)", "TMT2plex (N-term)"), NULL)
}
}
