% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{ext_aamasses}
\alias{ext_aamasses}
\title{Calculates molecular weight of a polypeptide ([MH]+).}
\usage{
ext_aamasses(
  fixedmods = c("TMT6plex (N-term)", "TMT6plex (K)", "Carbamidomethyl (. = C)"),
  varmods = c("Acetyl (Protein N-term)", "Oxidation (M)", "Deamidated (N)",
    "Gln->pyro-Glu (N-term = Q)"),
  max_ncomb_mods = 64,
  compatible_mods = TRUE
)
}
\arguments{
\item{fixedmods}{A character vector of fixed modifications. See also
\link{parse_unimod} for grammars.}

\item{varmods}{A character vector of variable modifications.}

\item{max_ncomb_mods}{Integer; the maximum number of combinatorial variable
modifications allowed per peptide sequence.}

\item{compatible_mods}{Logical. At the TRUE default, the combinations with
  more than one variable modification to the same site of a peptide sequence
  will be excluded. In addition, fixed modifications will be coerced to the
  corresponding variable modifications, should there be other variable
  modifications to the same sites.

  In general, the argument will be at TRUE. In the case that a joint
  modification to the same site is desirable, the construction of a new entry
  in Unimod may be suitable.}
}
\description{
Calculates molecular weight of a polypeptide ([MH]+).
}
\examples{
\donttest{
library(purrr)

x <- ext_aamasses()
x_att <- map(x, attributes)
names(x_att[[1]])
x_vmods <- map(x_att, `[`, c("vmods"))

x <- ext_aamasses(c("TMT6plex (N-term)", "TMT6plex (K)", "Carbamidomethyl (C)"), c("Acetyl (N-term)", "Gln->pyro-Glu (N-term = Q)", "Oxidation (M)"))

x <- ext_aamasses(fixedmods = NULL)
x <- ext_aamasses(fixedmods = NULL, varmods = NULL)

# Fixed mod, no NL
x <- ext_aamasses(fixedmods = c("TMT6plex (N-term)", "TMT6plex (K)", "Carbamidomethyl (. = C)"), varmods = NULL)

# Fixed mod + NL
x <- ext_aamasses(fixedmods = c("TMT6plex (N-term)", "TMT6plex (K)", "Carbamidomethyl (. = M)"), varmods = NULL)

# Fixed mod, no NL; var mod, no NL
x <- ext_aamasses(fixedmods = c("TMT6plex (N-term)", "TMT6plex (K)", "Carbamidomethyl (. = C)"), varmods = c("Acetyl (N-term)", "Gln->pyro-Glu (N-term = Q)"))

# Fixed mod + NL; var mod + NL
x <- ext_aamasses(c("TMT6plex (N-term)", "TMT6plex (K)", "Carbamidomethyl (. = M)",
"Deamidated (. = R)"), c("Acetyl (N-term)", "Gln->pyro-Glu (N-term = Q)",
"Hex(5)HexNAc(2) (N)"))
}
\dontrun{
# conflicts
x <- ext_aamasses(c("Carbamidomethyl (N-term)", "TMT2plex (N-term)"), NULL)
}
}
