% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{extract_matches_col}
\alias{extract_matches_col}
\title{Extracts the lists of theoretical or experimental m-over-z's.}
\usage{
extract_matches_col(data, col = "theo")
}
\arguments{
\item{data}{Nested list of table from ion matches. Each table contains
columns \code{theo} and \code{expt}.}

\item{col}{The name of a column where data will be extracted.}
}
\description{
Extracts the lists of theoretical or experimental m-over-z's.
}
