% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{extract_mq_ints}
\alias{extract_mq_ints}
\title{Extracts MaxQuant intensity values and calculates log2FC.}
\usage{
extract_mq_ints(df)
}
\arguments{
\item{df}{A data frame of MaxQuant results.}
}
\description{
Extracts MaxQuant intensity values and calculates log2FC.
}
