% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{extract_psm_raws}
\alias{extract_psm_raws}
\title{Extracts the names of RAW MS files from PSM data}
\usage{
extract_psm_raws(dat_dir = NULL)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
\code{extract_psm_raws} extracts the MS file names from the PSM data under
the working directory.
}
\examples{
\donttest{
extract_psm_raws()
}

}
