% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{extract_theos}
\alias{extract_theos}
\title{Extracts the lists of theoretical m-over-z's.}
\usage{
extract_theos(data)
}
\arguments{
\item{data}{Nested list of table from ion matches. Each table contains
columns \code{theo} and \code{expt}.}
}
\description{
Extracts the lists of theoretical m-over-z's.
}
