% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{extractRAW}
\alias{extractRAW}
\title{Extract RAW file names}
\usage{
extractRAW(rptr_intco = 1000, rm_craps = FALSE, plot_violins = TRUE)
}
\description{
\code{extractRAW} extracts the RAW file names from the PSM result files under
the current working directory.
}
\examples{
extractRAW()

}
