% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlfq.R
\name{fillMBRnaRTs}
\alias{fillMBRnaRTs}
\title{Fill NA retention times with the MS2 triggering retention times}
\usage{
fillMBRnaRTs(mbr_rets, mbr_peps, dat_dir)
}
\arguments{
\item{mbr_rets}{A vector of retention times.}

\item{mbr_peps}{A vector of \code{pep_seq_modz}.}

\item{dat_dir}{The working directory.}
}
\description{
To obtain a reference point for reassess apexes and MBR. NA \code{mbr_rets}
occur at \code{mbr_peps} with all missing apex retention times (cannot be
determined in the first pass and \code{pep_tot_int = 0}). No need to call
this utility if data were filtered by \code{pep_tot_int > 0}.
}
