% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-sigtests.R
\name{filterData}
\alias{filterData}
\title{Filter rows by variance quantiles}
\usage{
filterData(df, cols = NULL, var_cutoff = 0.001)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{cols}{Column indexes}

\item{var_cutoff}{Numeric; the cut-off in the variances of \code{log2FC}.
Entries with variances smaller than the threshold will be excluded from
linear modeling. The default is 1E-3.}
}
\description{
Filter rows by variance quantiles
}
