% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filters_in_call}
\alias{filters_in_call}
\title{Filters rows}
\usage{
filters_in_call(df, ...)
}
\arguments{
\item{df}{a data frame.}

\item{...}{Arguments for \link[dplyr]{filter}}
}
\description{
Filters rows
}
