% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{find_aa_masses}
\alias{find_aa_masses}
\title{Finds the existence of \code{aa_masses_all.rds}.}
\usage{
find_aa_masses(
  out_path = NULL,
  fixedmods = NULL,
  varmods = NULL,
  maxn_vmods_setscombi = 64L,
  index_mods = FALSE
)
}
\arguments{
\item{out_path}{The output path.}

\item{fixedmods}{A character vector of fixed modifications. See also
\link{parse_unimod} for grammars.}

\item{varmods}{A character vector of variable modifications.}

\item{maxn_vmods_setscombi}{Integer; the maximum number of combinatorial variable
modifications and neutral losses.}

\item{index_mods}{Logical; if TRUE, converts the names of modifications to
indexes. Not currently used.}
}
\description{
Finds the existence of \code{aa_masses_all.rds}.
}
