% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{find_aa_site}
\alias{find_aa_site}
\title{Finds the site of an AA residue.}
\usage{
find_aa_site(pos_site)
}
\arguments{
\item{pos_site}{A named value. Position in name and site in value.}
}
\description{
Finds the site of an AA residue.
}
