% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{find_acc_type}
\alias{find_acc_type}
\title{Helper for \link{load_fasta2}.}
\usage{
find_acc_type(acc_pattern)
}
\arguments{
\item{acc_pattern}{Regular expression(s) describing the patterns to separate
the header lines of fasta entries. At the \code{NULL} default, the pattern
will be automated when \code{acc_type} are among c("uniprot_acc",
"uniprot_id", "refseq_acc", "other").}
}
\description{
Not used for custom acc_pattern, i.e. acc_pattern = "...".
}
