% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{find_acctype}
\alias{find_acctype}
\title{Finds the Accession_Type}
\usage{
find_acctype(label_scheme_full)
}
\description{
\code{find_acctype} finds the \code{Accession_Type} that will be used in the
annotation of protein IDs.
}
