% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtests.R
\name{find_baseline_slope}
\alias{find_baseline_slope}
\title{Helper of \link{base_sigtest_y}.}
\usage{
find_baseline_slope(x, y, abs_baselne_intensity = 1e+05)
}
\arguments{
\item{x}{A vector of row variances of intensities from each entry, e.g.,
protein.}

\item{y}{A data frame of intensities. Each column correspondes to a sample.}

\item{abs_baselne_intensity}{The absolute level of intensity for uses at
\code{impute_group_na = TRUE}.}
}
\description{
Finds the slope for the imputation of baseline intensities.
}
