% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlfq.R
\name{find_best_lfqvec}
\alias{find_best_lfqvec}
\title{Find the best LFQ vector}
\usage{
find_best_lfqvec(d1, d2, dc)
}
\arguments{
\item{d1}{A vector of distance.}

\item{d2}{Another vector of distance.}

\item{dc}{A vector of distance by spectrum counts.}
}
\description{
Find the best LFQ vector
}
