% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{find_best_mbry}
\alias{find_best_mbry}
\title{Find the best intensity}
\usage{
find_best_mbry(xs, ys, mbr_mz, step = 1e-05)
}
\arguments{
\item{xs}{Vectors of X values over a range of retention times.}

\item{ys}{Vectors of Y values over a range of retention times.}

\item{mbr_mz}{The Value of an m-over-z for MBR.}

\item{step}{The mass error in \code{ppm / 1e6}.}
}
\description{
To handle the case of multiple matched Y values in a scan.
}
