% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{find_callarg_val}
\alias{find_callarg_val}
\title{Finds the setting of a arguments.}
\usage{
find_callarg_val(
  time = ".2021-05-21_211227",
  path = "~/proteoQ/.MSearch/Cache/Calls",
  fun = "calc_pepmasses",
  arg = "fasta"
)
}
\arguments{
\item{time}{The time stamp.}

\item{path}{A (system) file path.}

\item{fun}{The name of function being saved.}

\item{arg}{Argument to be matched.}
}
\description{
Not currently used.
}
