% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{find_callarg_vals}
\alias{find_callarg_vals}
\title{Finds the values of a list of arguments.}
\usage{
find_callarg_vals(
  time = ".2021-05-21_211227.rda",
  path = "~/proteoQ/.MSearch/Cache/Calls",
  fun = "calc_pepmasses",
  args = c("fasta", "max_miss")
)
}
\arguments{
\item{time}{The time stamp.}

\item{path}{A (system) file path.}

\item{fun}{The name of function being saved.}

\item{args}{Arguments to be matched.}
}
\value{
The time stamp of a matched cache results.
}
\description{
Finds the values of a list of arguments.
}
