% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ions.R
\name{find_cmodtree}
\alias{find_cmodtree}
\title{Find and subset peptides.}
\usage{
find_cmodtree(peps, vmods)
}
\arguments{
\item{peps}{A list of peptide sequences with a one-letter representation of
amino acid residues.}

\item{vmods}{A named list of variable modifications.}
}
\description{
(1-ct) Protein C-term + site -> (2-ct) Any protein C-term -> (3-ct) Any
C-term + site -> (4-ct) Any C-term.
}
