% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtests.R
\name{find_contrast_elements}
\alias{find_contrast_elements}
\title{Helper to find contrast elements}
\usage{
find_contrast_elements(contrs)
}
\arguments{
\item{contrs}{A vectors of contrast groups.}
}
\description{
Helper to find contrast elements
}
