% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{find_df_species}
\alias{find_df_species}
\title{Finds the species from result tables}
\usage{
find_df_species(df, acc_type)
}
\description{
\code{find_df_species} find the species from PSM or peptide data.
}
