% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{find_dir}
\alias{find_dir}
\title{Finds a file directory.}
\usage{
find_dir(path, create = FALSE)
}
\arguments{
\item{path}{A file path.}

\item{create}{Logical; if TRUE create the path if not yet existed.}
}
\description{
With the option of creating a direcotry.
}
