% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{find_ess_prots}
\alias{find_ess_prots}
\title{Helper of \link{greedysetcover}.}
\usage{
find_ess_prots(mat)
}
\arguments{
\item{mat}{A logical matrix; protein accession in column names and peptide
sequences in row names.}
}
\description{
Helper of \link{greedysetcover}.
}
