% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_fn_bases}
\alias{find_fn_bases}
\title{Finds the base names of files}
\usage{
find_fn_bases(filenames)
}
\arguments{
\item{filenames}{A character vector of filenames}
}
\description{
(not currently used)
}
