% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{find_intercombi}
\alias{find_intercombi}
\title{Finds the combinations across residues.}
\usage{
find_intercombi(intra_combis)
}
\arguments{
\item{intra_combis}{The results from \link{unique_mvmods}.}
}
\description{
For multiple residues (each residue one to multiple modifications).
}
