% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_a1_vnl0_fnl0.R
\name{find_intercombi_p2}
\alias{find_intercombi_p2}
\title{Finds the combinations of positions and sites across residues.}
\usage{
find_intercombi_p2(intra_combis, maxn_vmods_sitescombi_per_pep)
}
\arguments{
\item{intra_combis}{The results from \link{unique_mvmods}.}

\item{maxn_vmods_sitescombi_per_pep}{Integer; the maximum number of
combinatorial variable modifications per peptide sequence.}
}
\description{
For multiple residues (each residue one to multiple modifications).
}
\seealso{
\link{find_intercombi}
}
