% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{find_mascot_plex}
\alias{find_mascot_plex}
\title{Finds the plexes according to the Mascot outputs.}
\usage{
find_mascot_plex(df = NULL)
}
\arguments{
\item{df}{A data frame of Mascot outputs with headers being removed.}
}
\description{
Finds the plexes according to the Mascot outputs.
}
