% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{find_mascot_tmtplex}
\alias{find_mascot_tmtplex}
\title{Finds the TMT-plex from Mascot PSM exports}
\usage{
find_mascot_tmtplex(df, pep_seq_rows = NULL)
}
\arguments{
\item{df}{A PSM export from Mascot}

\item{pep_seq_rows}{The row(s) contain character string "pep_seq".}
}
\description{
Finds the TMT-plex from Mascot PSM exports
}
