% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{find_mascot_vmods}
\alias{find_mascot_vmods}
\title{Finds the variable modifications.}
\usage{
find_mascot_vmods(df_header)
}
\arguments{
\item{df_header}{Mascot output header.}
}
\description{
For Mascot.
}
