% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{find_masct_tmtplex}
\alias{find_masct_tmtplex}
\title{Find the TMT-plex from Mascot PSM exports}
\usage{
find_masct_tmtplex(df, pep_seq_rows = NULL)
}
\arguments{
\item{df}{A PSM export from Mascot}

\item{pep_seq_rows}{The row(s) contain character string "pep_seq".}
}
\description{
Find the TMT-plex from Mascot PSM exports
}
