% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{find_max_mass}
\alias{find_max_mass}
\title{Finds the existence of \code{max_mass.rds}.}
\usage{
find_max_mass(out_path = NULL, .path_fasta = NULL, .time_stamp = NULL)
}
\arguments{
\item{out_path}{The output path.}

\item{.time_stamp}{The time stamp to cache.}

\item{.path_cache}{The file path to cache.}
}
\description{
Finds the existence of \code{max_mass.rds}.
}
