% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_proteom.R
\name{find_mod_indexesQ}
\alias{find_mod_indexesQ}
\title{Finds mod_indexes.}
\usage{
find_mod_indexesQ(dat_dir)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\value{
A named vector with modifications in names and indexes in values.
}
\description{
Finds mod_indexes.
}
