% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{find_modmass}
\alias{find_modmass}
\title{Find the mono-isotopic mass of a modification}
\usage{
find_modmass(mod_nm = "Carbamidomethyl")
}
\arguments{
\item{mod_nm}{The name of a modification.}
}
\description{
Find the mono-isotopic mass of a modification
}
