% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{find_ms1_interval}
\alias{find_ms1_interval}
\title{Calculates the frame number for an experimental mass by intervals.}
\usage{
find_ms1_interval(from = 350, mass = 1714.81876, ppm = 20)
}
\arguments{
\item{from}{Numeric; the starting MS1 mass.}

\item{mass}{Numeric; an MS1 mass.}

\item{ppm}{Numeric; the ppm for data binning.}
}
\description{
Needs correct \code{from}.
}
