% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{find_ms1filepath}
\alias{find_ms1filepath}
\title{Finds the filepath of MS1 traces}
\usage{
find_ms1filepath(dat_dir, pat = "ms1full", type = 1L)
}
\arguments{
\item{dat_dir}{The current working directory.}

\item{pat}{A pattern for recognizing full MS1 files.}

\item{type}{The exit type: 0, warning; 1, stop.}
}
\description{
Finds the filepath of MS1 traces
}
