% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_base.R
\name{find_ms2_bycombi}
\alias{find_ms2_bycombi}
\title{Helper of \link{find_ms2_bypep}}
\usage{
find_ms2_bycombi(theos, expts, ex, ppm_ms2, min_ms2mass)
}
\arguments{
\item{theos}{Numeric vector; one to multiple series of theoretical MS2s.}

\item{expts}{Numeric vector; one series of experimental MS2s.}

\item{ex}{Indexed \code{expts}.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}
}
\description{
Helper of \link{find_ms2_bypep}
}
