% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_base.R
\name{find_ms2_bypep}
\alias{find_ms2_bypep}
\title{Helper: finds the the outer products for vectors of MS2 ions.}
\usage{
find_ms2_bypep(theos, expts, ppm_ms2 = 25L, min_ms2mass = 110L)
}
\arguments{
\item{theos}{Numeric vector; one to multiple series of theoretical MS2s.}

\item{expts}{Numeric vector; one series of experimental MS2s.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}
}
\description{
The same theoretical peptides at different position permutations.
}
\details{
Run time about 41 us.
}
\examples{
\donttest{
# experimental `322.18704` fit to both b- and y- ions
expts <- c(101.07140,102.05540,107.04956,110.07165,111.07500,
           112.08729,113.07130,115.08693,116.07092,120.08105,
           121.08428,126.12794,127.12494,127.13121,128.12825,
           128.13452,129.13164,129.13789,130.06493,130.09772,
           130.13495,130.14124,131.13831,132.14160,134.07635,
           136.07573,157.10826,158.09238,159.09117,170.12067,
           173.12846,173.14980,175.11896,176.12245,176.15956,
           184.11806,186.15303,188.15977,190.09743,193.63914,
           207.11292,210.12289,229.16670,230.17030,231.17410,
           235.10779,240.14383,248.18073,262.15305,265.67874,
           273.21210,285.13416,301.20700,305.16055,312.17740,
           314.69138,322.18704,365.23856,369.24496,371.70316,
           374.18283,376.27573,392.19308,393.23337,394.23743,
           399.20920,400.27567,400.72501,401.22650,401.27139,
           401.58405,401.72778,409.21918,410.22241,423.24564,
           433.29709,452.27066,462.25473,480.26578,481.26828,
           498.27670,530.35126,572.28278,573.28625,599.33899,
           600.34039,609.32202,626.29218,627.29303,627.33948,
           628.33417,629.30463,630.30219,643.31946,644.31763,
           644.35913,645.32520,646.32880,647.32825,648.32892)

theos <- c(114.0913,251.1503,322.1874,423.2350,480.2565,
           627.3249,783.4260,175.1190,322.1874,379.2088,
           480.2565,551.2936,688.3525,801.4366)

nms <- unlist(stringr::str_split(pep, ""))

names(theos) <- c(nms, rev(nms))
theos <- list(`0000000` = theos)

x <- find_ms2_bypep(theos, expts)
}
}
