% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{find_n_comp}
\alias{find_n_comp}
\title{Checks \code{n_comp}}
\usage{
find_n_comp(df, n_comp, method_align)
}
\arguments{
\item{df}{An input data frame}

\item{n_comp}{Integer; the number of Gaussian components to be used with
\code{method_align = MGKernel}. A typical value is 2 or 3. The variable
\code{n_comp} overwrites the argument \code{k} in
\code{\link[mixtools]{normalmixEM}}.}

\item{method_align}{Character string indicating the method in aligning
\code{log2FC} across samples. \code{MC}: median-centering; \code{MGKernel}:
the kernel density defined by multiple Gaussian functions
(\code{\link[mixtools]{normalmixEM}}). At the \code{MC} default, the ratio
profiles of each sample will be aligned in that the medians of the
\code{log2FC} are zero. At \code{MGKernel}, the ratio profiles of each
sample will be aligned in that the \code{log2FC} at the maximums of kernel
density are zero.}
}
\description{
Checks \code{n_comp}
}
