% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ions.R
\name{find_nmodtree}
\alias{find_nmodtree}
\title{Find and subset peptides.}
\usage{
find_nmodtree(peps, vmods)
}
\arguments{
\item{peps}{A list of peptide sequences with a one-letter representation of
amino acid residues.}

\item{vmods}{A named list of variable modifications.}
}
\description{
(1-nt) Protein N-term + site -> (2-nt) Any protein N-term -> (3-nt) Any
N-term + site -> (4-nt) Any N-term -> (5) Anywhere + site.
}
