% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_NorZ}
\alias{find_NorZ}
\title{Finds the patterns of log2FC columns.}
\usage{
find_NorZ(scale_log2r = TRUE)
}
\arguments{
\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE. At
\code{scale_log2r = NA}, the raw \code{log2FC} without normalization will
be used.}
}
\description{
Finds the patterns of log2FC columns.
}
