% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{find_nterm_mass}
\alias{find_nterm_mass}
\title{Finds N-terminal mass.}
\usage{
find_nterm_mass(aa_masses)
}
\arguments{
\item{aa_masses}{Amino-acid lookup at a given combination of fixed and
variable.}
}
\description{
Not yet used.
}
