% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{find_padding_pos}
\alias{find_padding_pos}
\title{Finds the padding positions of TMT channels.}
\usage{
find_padding_pos(this_plex, TMT_plex)
}
\arguments{
\item{this_plex}{Numeric; the multiplexity of TMT, i.e., 10, 11 etc. before
padding.}

\item{TMT_plex}{Numeric; the multiplexity of TMT, i.e., 10, 11 etc.}
}
\description{
Finds the padding positions of TMT channels.
}
