% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{find_pep_pos}
\alias{find_pep_pos}
\title{Find peptide start and end positions}
\usage{
find_pep_pos(prot_acc, pep_seq, fasta)
}
\description{
\code{find_pep_pos} finds the start and the end positions of peptides in
ascribed proteins description based on the \code{fasta}.
}
