% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{find_pepoutliers}
\alias{find_pepoutliers}
\title{Helper to find if a protein contains peptide outliers under a contrast.}
\usage{
find_pepoutliers(
  dat_dir = NULL,
  col_select = "Select",
  col_group = "Group",
  group_psm_by = "pep_seq_mod",
  group_pep_by = "gene",
  fml_nms = NULL,
  contr_pairs = NULL,
  p_outlier = 0.05,
  min_n = 3L,
  p_contrast = 0.01,
  p_type = "adjP",
  cols_pep = c("pep_start", "pep_end", "pep_miss", "pep_phospho_locprob"),
  out_name = "pepoutliers_btw_contrasts.tsv"
)
}
\arguments{
\item{dat_dir}{A working directory.}

\item{fml_nms}{Formula names. If \code{NULL}, all formulas from \link{pepSig}
will be included.}

\item{contr_pairs}{Contrast pairs. If \code{NULL}, all contrasts under a
formula will be included.}

\item{p_outlier}{A significance level for assessing outliers.}

\item{min_n}{The minimal number of peptides under a protein for
consideration.}

\item{p_contrast}{The cut-off in significant pVals from \link{pepSig}. The
type of pVals is either original or BH-adjusted, indicated by
\code{p_type}.}

\item{p_type}{The type of p-values from \link{pepSig} in one of \code{adjP,
pVal}.}

\item{cols_pep}{The attribute columns of a peptide table to be included.}

\item{out_name}{An output file name for being saved as an R object.}
}
\description{
Helper to find if a protein contains peptide outliers under a contrast.
}
