% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{find_pepoutliers_by_contr}
\alias{find_pepoutliers_by_contr}
\title{Helper to find if a protein contains peptide outliers under a formula and
then a contrast.}
\usage{
find_pepoutliers_by_contr(
  contr = NULL,
  fml_nm = NULL,
  df = NULL,
  group_psm_by = "pep_seq_mod",
  group_pep_by = "gene",
  p_outlier = 0.05,
  min_n = 3L,
  p_contrast = 0.01,
  p_type = "adjP"
)
}
\arguments{
\item{contr}{For example, a contrast defined in \link{pepSig}.}

\item{fml_nm}{A formula name.}

\item{df}{A data frame.}

\item{p_outlier}{A significance level for assessing outliers.}

\item{min_n}{The minimal number of peptides under a protein for
consideration.}

\item{p_contrast}{The cut-off in significant pVals from \link{pepSig}. The
type of pVals is either original or BH-adjusted, indicated by
\code{p_type}.}

\item{p_type}{The type of p-values in once of \code{adjP, pVal}.}
}
\description{
Helper to find if a protein contains peptide outliers under a formula and
then a contrast.
}
