% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{find_ppm_error}
\alias{find_ppm_error}
\title{Finds the numeric difference in ppm.}
\usage{
find_ppm_error(x = 1000, y = 1000.01)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The difference between \eqn{x} and \eqn{y} in ppm.
}
\description{
Finds the numeric difference in ppm.
}
