% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{find_ppm_outer_bycombi}
\alias{find_ppm_outer_bycombi}
\title{Finds the the outer products for a vector of MS2 ions at a given ion series.}
\usage{
find_ppm_outer_bycombi(theos, expts, ppm_ms2 = 25L)
}
\arguments{
\item{theos}{Numeric vector; one series of theoretical MS2s.}

\item{expts}{Numeric vector; one series experimental MS2s.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}
}
\description{
A theoretical peptide at a given position permutation.
}
\examples{
\donttest{
expts <- c(101.0714, 102.0554, 110.0717, 115.0505, 126.1279, 127.0504, 
           127.1249, 127.1312, 128.1283, 128.1346, 129.0660, 129.1316, 
           129.1379, 130.1350, 130.1412, 131.1383, 133.0431, 133.0609, 
           134.0448, 136.0757, 145.0608, 158.0924, 173.1496, 175.1190, 
           176.1594, 191.0663, 193.0971, 198.0873, 201.0869, 201.1233, 
           202.0821, 230.1702, 248.0876, 248.1806, 257.1527, 298.6711, 
           312.6687, 335.1190, 361.1768, 361.6685, 367.2292, 369.6903, 
           370.1821, 370.6830, 376.2756, 377.2790, 384.6952, 412.7264, 
           420.2143, 423.2597, 475.2341, 475.3442, 476.3469, 484.2382, 
           484.7389, 487.7817, 488.2842, 496.2717, 523.3182, 537.3159, 
           537.8178, 572.8343, 573.3373, 623.3586, 624.3298, 636.8528, 
           637.3558, 637.8575, 638.4089, 645.3453, 645.8500, 646.3501, 
           687.8795, 688.3799, 692.8696, 695.4283, 701.3759, 701.8768, 
           702.3782, 702.8797, 737.3956, 737.8962, 738.3671, 739.3572, 
           740.3575, 745.9085, 746.4121, 752.4502, 801.4272, 809.4722, 
           810.4739, 852.4418, 967.4690, 968.4714, 969.5046, 1084.5293, 
           1132.5676, 1133.5686, 1134.5735, 1135.5786)
           
theos <- c(116.0342, 203.0662, 290.0983, 418.1569, 475.1783, 572.2311, 
           732.2617, 861.3043, 958.3571, 1071.4412, 1168.4939, 1225.5154, 
           1322.5681, 1435.6522, 1536.6999, 1664.7585, 1761.8112, 1917.9123, 
           175.1190, 272.1717, 400.2303, 501.2780, 614.3620, 711.4148, 
           768.4363, 865.4890, 978.5731, 1075.6259, 1204.6684, 1364.6991, 
           1461.7519, 1518.7733, 1646.8319, 1733.8639, 1820.8960, 1935.9229)
           
names(theos) <- c("D", "S", "S", "Q", "G", "P", "C", "E", "P", 
                  "L", "P", "G", "P", "L", "T", "Q", "P", "R", 
                  "R", "P", "Q", "T", "L", "P", "G", "P", "L", 
                  "P", "E", "C", "P", "G", "Q", "S", "S", "D")

find_ppm_outer_bycombi(theos, expts)
}
}
