% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{find_ppm_outer_bypep}
\alias{find_ppm_outer_bypep}
\title{Helper: finds the the outer products for vectors of MS2 ions.}
\usage{
find_ppm_outer_bypep(theos, expts, ppm_ms2)
}
\arguments{
\item{theos}{Numeric vector; one to multiple series of theoretical MS2s.}

\item{expts}{Numeric vector; one series of experimental MS2s.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}
}
\description{
The same theoretical peptides at different position permutations.
}
