% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{find_preceding_colnm}
\alias{find_preceding_colnm}
\title{Helper: finds the column name before \code{to_move}.}
\usage{
find_preceding_colnm(df, to_move)
}
\arguments{
\item{df}{The original data frame.}

\item{to_move}{The column to be moved.}
}
\description{
To keep columns at the same order after descriptive summary.
}
