% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_psmQ_files}
\alias{find_psmQ_files}
\title{Finds the list of psmQ files.}
\usage{
find_psmQ_files(dat_dir)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Finds the list of psmQ files.
}
