% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_ratio_cols}
\alias{find_ratio_cols}
\title{Finds the columns of reporter-ion ratios.}
\usage{
find_ratio_cols(TMT_plex)
}
\arguments{
\item{TMT_plex}{Numeric; the multiplexity of TMT, i.e., 10, 11 etc.}
}
\description{
Finds the columns of reporter-ion ratios.
}
