% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{find_reporter_ints}
\alias{find_reporter_ints}
\title{Finds the intensities of reporter-ions.}
\usage{
find_reporter_ints(ms2_moverzs, ms2_ints, theos, ul, ppm_ms2, len, nms)
}
\arguments{
\item{ms2_moverzs}{Numeric vector; a series of experimental MS2 m-over-z's
(in the region of reporter ions).}

\item{ms2_ints}{Numeric vector; a series of experimental MS2 intensities (in
the region of reporter ions).}

\item{theos}{The theoretical m-over-z of reporter ions.}

\item{ul}{The upper and lower bound for reporter-ion m-over-z's.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{len}{The length of reporter-ion plexes.}

\item{nms}{The names of reporter-ion channels.}
}
\description{
Finds the intensities of reporter-ions.
}
