% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{find_reporters_ppm}
\alias{find_reporters_ppm}
\title{Finds the indexes of reporter ions.}
\usage{
find_reporters_ppm(theos, expts, ppm_reporters = 10, len, nms)
}
\arguments{
\item{theos}{The theoretical m-over-z of reporter ions.}

\item{expts}{Numeric vector; a series of experimental MS2s (in the region of
reporter ions).}

\item{ppm_reporters}{The mass tolerance of MS2 reporter ions.}

\item{len}{The length of reporter-ion plexes.}

\item{nms}{The names of reporter-ion channels.}
}
\value{
A vector of indexes
}
\description{
Finds the indexes of reporter ions.
}
