% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_search_engine}
\alias{find_search_engine}
\title{Finds the type of search engine}
\usage{
find_search_engine(dat_dir = NULL)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Finds the type of search engine
}
