% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-informatics.R
\name{find_sec_df}
\alias{find_sec_df}
\title{helper for finding input `df`
not currently used}
\usage{
find_sec_df(df = NULL, anal_type = NULL, id = NULL, ...)
}
\description{
helper for finding input `df`
not currently used
}
