% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{find_shared_prots}
\alias{find_shared_prots}
\title{Finds peptides that are shared by proteins.}
\usage{
find_shared_prots(df = NULL, pep_id = "pep_seq", prot_id = "prot_acc")
}
\arguments{
\item{df}{PSM data.}

\item{pep_id}{the column key of peptide sequences.}

\item{prot_id}{The column key of protein accessions.}
}
\description{
Only for PSM tables where the same \code{pep_seq}  duplicated in rows with
different \code{prot_acc}'s . This for now means Mascot with same-set and
sub-set proteins enabled. Otherwise, e.g. when applying to other search
engines without such redundancy, every peptide would turn out to be unique.
}
\details{
Output rows are unique at the levels of \code{c("prot_acc", "pep_seq")}.
}
