% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{find_sm_psmraws}
\alias{find_sm_psmraws}
\title{Finds the names of Spectrum Mill PSM files.}
\usage{
find_sm_psmraws(filelist = NULL, dat_dir = NULL)
}
\arguments{
\item{filelist}{A list of PSM files.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Finds the names of Spectrum Mill PSM files.
}
