% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{find_species}
\alias{find_species}
\title{Find the Species}
\usage{
find_species(label_scheme_full)
}
\description{
\code{find_species} find the species that will be used in analysis such as GSEA.
}
