% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{find_species_centers2}
\alias{find_species_centers2}
\title{Finds the data center of log2FC by species and samples}
\usage{
find_species_centers2(ymat, are_refs = NULL, are_smpls = NULL, sps)
}
\arguments{
\item{ymat}{An intensity matrix.}

\item{are_refs}{A logical vector indicating reference status.}

\item{are_smpls}{A logical vector indicating reference status.}

\item{sps}{The species values corresponding the column peptides.}
}
\description{
Input matrix: columns, peptides; rows, samples.
}
