% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{find_trend_df2}
\alias{find_trend_df2}
\title{Finds the input file for Trend plots.}
\usage{
find_trend_df2(
  df2 = NULL,
  filepath = NULL,
  n_clust = NULL,
  pat = "Trend_[ONZ]_.*nclust\\\\d+\\\\.txt$",
  ext = "txt",
  scale_log2r = TRUE,
  impute_na = FALSE
)
}
\arguments{
\item{df2}{Filename(s) of trend inputs.}

\item{filepath}{Input and output filepath.}

\item{n_clust}{The number of clusters.}

\item{pat}{The pattern of secondary files (df2).}

\item{ext}{The filename extension.}
}
\description{
Finds the input file for Trend plots.
}
