% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{find_trend_m}
\alias{find_trend_m}
\title{Helper to compile arguments for trend analysis}
\usage{
find_trend_m(dots, choice = "cmeans", n_clust = 5L, df_mean_log2r)
}
\arguments{
\item{dots}{Arguments of trend clustering.}

\item{choice}{Character string; the clustering method in \code{c("cmeans",
"clara", "kmeans", "pam", "fanny")}. The default is "cmeans".}

\item{n_clust}{Numeric vector; the number(s) of clusters that data will be
divided into. At the NULL default, it will be determined by the gap method
in \code{\link[cluster]{clusGap}}. The \code{n_clust} overwrites the
argument \code{centers} in \code{\link[e1071]{cmeans}}.}

\item{df_mean_log2r}{Data of mean log2Ratio.}
}
\description{
Helper to compile arguments for trend analysis
}
