% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{find_unimod}
\alias{find_unimod}
\title{Find a Unimod.}
\usage{
find_unimod(unimod = "Carbamidomethyl (C)")
}
\arguments{
\item{unimod}{The name of a \href{https://www.unimod.org/}{Unimod} modification.}
}
\description{
Find the mono-isotopic mass, position, site and neutral losses of a
modification.
}
\details{
In the field of \code{position_site}, \code{position} is the name and
\code{site} is the value.
}
\examples{
\donttest{
x <- find_unimod("Carbamidomethyl (C)")
x <- find_unimod("Carbamidomethyl (M)")
x <- find_unimod("Acetyl (Protein N-term)")
x <- find_unimod("Gln->pyro-Glu (N-term = Q)")
x <- find_unimod("Hex(5)HexNAc(2) (N)")
}

\dontrun{
# Prefer an "=" sign between 'N-term' and 'Q'
x <- find_unimod("Gln->pyro-Glu (N-term Q)")
}
}
