% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{find_unique_sets}
\alias{find_unique_sets}
\title{Finds the sets of \code{n} elements in \code{positions}.}
\usage{
find_unique_sets(ps = c(1:5), ns = c("A", "B", "C"))
}
\arguments{
\item{ps}{A vector of positions.}

\item{ns}{The names to be filled into \code{p}.}
}
\description{
At least one occupancy for each elements in \code{ns}.
}
