% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fixBrukerMGF}
\alias{fixBrukerMGF}
\title{Fix the \code{File} field in \code{Title} lines.}
\usage{
fixBrukerMGF(file)
}
\arguments{
\item{file}{A file name with prepending path.}
}
\description{
The first tilde in the \code{Title} lines needs to before \code{File}.
}
