% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{fml2mass}
\alias{fml2mass}
\title{Standarizes mixtured MSGF formulas and masses to monoisotopic masses.}
\usage{
fml2mass(mods)
}
\arguments{
\item{mods}{Fixed or variable modifications.}
}
\description{
Standarizes mixtured MSGF formulas and masses to monoisotopic masses.
}
