% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_base.R
\name{fuzzy_match_one}
\alias{fuzzy_match_one}
\title{Fuzzy matches with +/-1 offsets.}
\usage{
fuzzy_match_one(x, y)
}
\arguments{
\item{x}{A vector to be matched.}

\item{y}{A vector to be matched against.}
}
\description{
Fuzzy matches with +/-1 offsets.
}
