% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-volcanos.R
\name{gageMap}
\alias{gageMap}
\title{Volcano Plots of Protein \code{log2-ratios} under Given Gene Sets}
\usage{
gageMap(...)
}
\description{
Volcano Plots of Protein \code{log2-ratios} under Given Gene Sets
}
\seealso{
\code{\link{proteoVolcano}} for parameters
}
