% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gage.R
\name{gageTest}
\alias{gageTest}
\title{Perform significance tests}
\usage{
gageTest(df, id, label_scheme_sub, filepath, filename, complete_cases,
  method = "limma", gset_nm = "go_sets", var_cutoff = 0.001,
  pval_cutoff = 0.05, logFC_cutoff = log2(1.1), ...)
}
\description{
Perform significance tests
}
