% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtests.R
\name{gen_randoms}
\alias{gen_randoms}
\title{Generate random intensity values.}
\usage{
gen_randoms(n = 3L, mu = 0, sigma = 1.8, seed = NULL)
}
\arguments{
\item{n}{The number of random values.}

\item{mu}{The mean.}

\item{sigma}{The standard deviation.}

\item{seed}{A seed for reproducible random number generations.}
}
\description{
Generate random intensity values.
}
