% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gene_homo_slim}
\alias{gene_homo_slim}
\title{Gene homology across human, mouse and rat.}
\format{
A list at a length of 21929
\describe{
  \item{hs_...}{human entries}
}
}
\source{
\url{https://www.informatics.jax.org/downloads/reports/index.html#homology}
}
\usage{
gene_homo_slim
}
\description{
A dataset containing human, moue and rat gene symbols.
}
\keyword{datasets}
