% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{geom_lower_text}
\alias{geom_lower_text}
\title{Geom plot for ggpairs}
\usage{
geom_lower_text(data, mapping, params, show_ids, ...)
}
\arguments{
\item{data}{Input data.}

\item{mapping}{The mapping for ggplot2.}

\item{params}{Additional parameters for geom_point.}

\item{show_ids}{Logical; if TRUE, shows the sample IDs in \code{MDS/PCA}
plots. The default is TRUE.}

\item{...}{\code{filter_}: Variable argument statements for the row filtration
against data in a primary file linked to \code{df}. See also
\code{\link{normPSM}} for the format of \code{filter_} statements. \cr \cr
\code{arrange_}: Variable argument statements for the row ordering against
data in a primary file linked to \code{df}. See also \code{\link{prnHM}} for
the format of \code{arrange_} statements. \cr \cr Additional parameters for
\code{prcomp}: \cr \code{center}, data centering; \cr \code{tol}, the
tolerance \cr \code{...} \cr \cr Additional parameters for
\code{ggsave}: \cr \code{width}, the width of plot; \cr \code{height}, the
height of plot \cr \code{...}}
}
\description{
Geom plot for ggpairs
}
