% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanos.R
\name{geom_table}
\alias{geom_table}
\title{Print table in ggplot2 images}
\usage{
geom_table(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{The same as ggplot2.}

\item{data}{same as ggplot2.}

\item{stat}{The same as ggplot2.}

\item{position}{same as ggplot2.}

\item{na.rm}{The same as ggplot2.}

\item{show.legend}{same as ggplot2.}

\item{inherit.aes}{The same as ggplot2.}

\item{...}{same as ggplot2.}
}
\description{
Print table in ggplot2 images
}
