% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{gg_imgname}
\alias{gg_imgname}
\title{Check file names for ggsave()}
\usage{
gg_imgname(filename)
}
\arguments{
\item{filename}{Character string; An output file name.}
}
\description{
Check file names for ggsave()
}
