% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{go_sets_hs}
\alias{go_sets_hs}
\title{Entrez IDs of human gene-ontology annotations.}
\format{
A list at a length of 18320 
\describe{
  \item{hs_...}{human entries}
}
}
\source{
\url{http://current.geneontology.org/products/pages/downloads.html}
}
\usage{
go_sets_hs
}
\description{
A dataset containing human entrez IDs by the gene sets of GO.
}
\keyword{datasets}
