% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-datasets.R
\docType{data}
\name{go_sets_mm}
\alias{go_sets_mm}
\title{Entrez IDs of mouse gene-ontology annotations.}
\format{A list at a length of 18364 
\describe{
  \item{mm_...}{mouse entries}
}}
\source{
\url{http://current.geneontology.org/products/pages/downloads.html}
}
\usage{
go_sets_mm
}
\description{
A dataset containing mouse entrez IDs by the gene sets of GO.
}
\keyword{datasets}
