% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{go_sets_rn}
\alias{go_sets_rn}
\title{Entrez IDs of rat gene-ontology annotations.}
\format{A list at a length of 18726 
\describe{
  \item{rn_...}{rat entries}
}}
\source{
\url{http://current.geneontology.org/products/pages/downloads.html}
}
\usage{
go_sets_rn
}
\description{
A dataset containing rat entrez IDs by the gene sets of GO.
}
\keyword{datasets}
