% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{\%+\%}
\alias{\%+\%}
\title{Sums elements across lists.}
\usage{
x \%+\% y
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\description{
Each list has the same length. NA values are removed.
}
