% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{greedysetcover}
\alias{greedysetcover}
\title{Greedy set cover.}
\usage{
greedysetcover(df, both = FALSE)
}
\arguments{
\item{df}{A two-column data frame.}

\item{both}{Logical; if TRUE, outputs both the \code{sets} and the
comprising elements.}
}
\description{
Greedy set cover.
}
