% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{greedysetcover2}
\alias{greedysetcover2}
\title{Greedy set cover.}
\usage{
greedysetcover2(mat)
}
\arguments{
\item{mat}{A matrix of protein (cols)-peptide (rows) map.}
}
\value{
A list of proteins.
}
\description{
A matrix input. Output essential sets only (no elements).
}
