% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{greedysetcover3}
\alias{greedysetcover3}
\title{Greedy set cover.}
\usage{
greedysetcover3(mat)
}
\arguments{
\item{mat}{A matrix of protein (cols)-peptide (rows) map.}
}
\value{
A two-column data frame of prot_acc and pep_seq.
}
\description{
A matrix input. Output both essential sets and elements.
}
