% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{greedysetcoverM}
\alias{greedysetcoverM}
\title{Greedy set cover.}
\usage{
greedysetcoverM(df)
}
\arguments{
\item{mat}{A matrix of protein (cols)-peptide (rows) map.}
}
\value{
A two-column data frame of prot_acc and pep_seq.
}
\description{
Expands a two-column input to a matrix input. The output table contains both
essential sets and elements.
}
