% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{groupMZPepZ}
\alias{groupMZPepZ}
\title{Group Mzion peptides by charge states}
\usage{
groupMZPepZ(df, group_psm_by = "pep_seq_mod", sdco = sqrt(9))
}
\arguments{
\item{df}{A Mzion peptide table.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{sdco}{A standard deviaiton cut-off.}
}
\description{
Group Mzion peptides by charge states
}
