% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{groupMZPepZ2}
\alias{groupMZPepZ2}
\title{Group Mzion peptides by charge states}
\usage{
groupMZPepZ2(df_num)
}
\arguments{
\item{df_num}{A Mzion peptide table.}
}
\description{
Group Mzion peptides by charge states
}
