% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{groupMZPSM1}
\alias{groupMZPSM1}
\title{Group PSMs}
\usage{
groupMZPSM1(df, group_psm_by = "pep_seq_mod", yco = 1e+07, sdco = sqrt(9))
}
\arguments{
\item{df}{A Mzion PSM table.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{yco}{An intensity cut-off.}

\item{sdco}{A standard deviaiton cut-off.}
}
\description{
For \code{lfq_mbr = TRUE}. Delayed \code{pep_seq_modz} to \code{pep_seq_mod}
conversion.
}
