% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{groupProts}
\alias{groupProts}
\title{Groups proteins by shared peptides.}
\usage{
groupProts(df, out_path = NULL)
}
\arguments{
\item{df}{Interim results from \link{matchMS}.}

\item{out_path}{The output path.}
}
\description{
Adds columns \code{prot_hit_num} and \code{prot_family_member} to
\code{psm.txt}.
}
