% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{grp_prots}
\alias{grp_prots}
\title{Helper of \link{groupProts}.}
\usage{
grp_prots(out, out_path = NULL)
}
\arguments{
\item{out}{The data frame from upstream steps.}

\item{out_path}{The output path.}
}
\description{
Helper of \link{groupProts}.
}
