% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{grubbs_test}
\alias{grubbs_test}
\title{Outlier removals with Grubbs' method}
\usage{
grubbs_test(x, type = 10, opposite = FALSE, two.sided = FALSE)
}
\arguments{
\item{x}{A data frame.}

\item{type}{The type of Dixon method. See also outliers::dixon.test.}

\item{opposite}{The high or low. See also outliers::dixon.test.}

\item{two.sided}{Logical; is two-sided or not. See also
outliers::dixon.test.}
}
\description{
See also outliers::grubbs.test.
}
