% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{gspa_summary_limma}
\alias{gspa_summary_limma}
\title{gspa pVal calculations using limma}
\usage{
gspa_summary_limma(gset, df, min_size = 10, min_delta = 4)
}
\description{
gspa pVal calculations using limma
}
