% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-volcanos.R
\name{gsvaVol}
\alias{gsvaVol}
\title{Volcano Plots of GSVA Terms}
\usage{
gsvaVol(...)
}
\description{
Volcano Plots of GSVA Terms
}
\seealso{
\code{\link{proteoVolcano}} for parameters
}
