% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-volcanos.R
\name{gsVolcano}
\alias{gsVolcano}
\title{Volcano plots of protein \code{log2FC} under given gene sets}
\usage{
gsVolcano(
  df2 = NULL,
  df = NULL,
  contrast_groups = NULL,
  gsea_key = "term",
  gsets = NULL,
  theme = NULL,
  fml_nm = NULL,
  filepath = NULL,
  filename = NULL,
  adjP = FALSE,
  show_labels = TRUE,
  show_sig = "none",
  gspval_cutoff = 1e-06,
  gslogFC_cutoff = log2(1.2),
  topn = Inf,
  ...
)
}
\description{
Volcano plots of protein \code{log2FC} under given gene sets
}
