% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{hadd_psm_apexes}
\alias{hadd_psm_apexes}
\title{Helper of adding PSM apexes}
\usage{
hadd_psm_apexes(
  file = NULL,
  dfs = NULL,
  trace_files = NULL,
  fits = NULL,
  dat_dir = NULL,
  path_ms1 = NULL,
  lfq_ret_tol = 60L,
  ok_mbr = TRUE
)
}
\arguments{
\item{file}{A PSM file name, e.g., \code{TMTSet1_LCMSinj1_PSM_N.txt}.}

\item{dfs}{A list split of PSM data under the \code{file} at different
\code{RAW_File}.}

\item{trace_files}{The file names of \code{ms1apexes_} corresponding to
\code{dfs}.}

\item{fits}{The models of retention-time fittings corresponding to
\code{dfs}.}

\item{dat_dir}{A working directory.}

\item{path_ms1}{The file path to MS1 data from Mzion.}

\item{lfq_ret_tol}{Retention time tolerance for LFQ.}

\item{ok_mbr}{Logical; OK to perform MBR or not.}
}
\description{
By each PSM file of \code{TMTSet1_LCMSinj1.txt}, \code{TMTSet2_LCMSinj1.txt}
etc.
}
