% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlfq.R
\name{haddApexRTs_allsets}
\alias{haddApexRTs_allsets}
\title{Helper of adding apex retention times for all PSM tables}
\usage{
haddApexRTs_allsets(
  filelist = NULL,
  ms1full_files = NULL,
  dat_dir = NULL,
  path_ms1 = NULL,
  max_n_apexes = 2L,
  data_type = "raw"
)
}
\arguments{
\item{filelist}{The file names of PSM tables, e.g.,
TMTSet1_LCMSinj1_PSM_N.txt etc.}

\item{ms1full_files}{The file names of \code{ms1full_.rds}.}

\item{dat_dir}{The working directory.}

\item{path_ms1}{The file path to \code{ms1full_.rds}.}

\item{max_n_apexes}{The maximum number of apexes for consideration.}

\item{data_type}{The type of MS data.}
}
\description{
Inputs are at the levels of all TMTSet_LCMSInj.txt files.
}
