% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular.R
\name{hclassify_trends}
\alias{hclassify_trends}
\title{Helper of \link{classify_trends}.}
\usage{
hclassify_trends(
  df_med = NULL,
  col_subtype = NULL,
  col_fraction = NULL,
  col_group = NULL,
  col_order = NULL,
  col_cluster = "cluster",
  col_x = NULL,
  col_y = NULL,
  tie_method = "none"
)
}
\arguments{
\item{df_med}{A median description of localization scores by modules.}

\item{col_subtype}{Only applicable to subcellular data analysis for assessing
sub-group fraction purity etc. Character string to a column key in
\code{expt_smry.xlsx} indicating sample subtypes linked to subcellular
fractions. This is needed, e.g., when each subcellular fraction contain
multiple subtypes of control, treated etc. and purity will be asssessed
independently.}

\item{col_fraction}{Only applicable to subcellular data analysis for
  assessing fraction purity etc. Character string to a column key in
  \code{expt_smry.xlsx} indicating subcellular fractions. No calculation of
  subcellular location score at the \code{NULL} default.

  Suggest, e.g., the corresponding column key \code{Fraction} in metadata.}

\item{col_group}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_group} will be
used for sample grouping in the indicated analysis. At the NULL default, the
column key \code{Group} will be used. No data annotation by groups will be
performed if the fields under the indicated group column is empty.}

\item{col_order}{Character string to a column key in \code{expt_smry.xlsx}.
Numeric values under which will be used for the left-to-right arrangement of
samples in graphic outputs or top-to-bottom arrangement in text outputs. At
the NULL default, the column key \code{Order} will be used. If values under
column \code{Order} are left blank, samples will be ordered by their names.}

\item{col_cluster}{The column key of cluster nodes.}

\item{col_x}{The column name of X values.}

\item{col_y}{The column name of Y values for summarization.}

\item{tie_method}{A method handling ties in localization scores.}
}
\description{
Helper of \link{classify_trends}.
}
