% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_proteom.R
\name{hdeduce_mod_indexes}
\alias{hdeduce_mod_indexes}
\title{Helper of \link{deduce_mod_indexes}.}
\usage{
hdeduce_mod_indexes(df)
}
\arguments{
\item{df}{A subset of psmQ table with single modification.}
}
\description{
Helper of \link{deduce_mod_indexes}.
}
