% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{hfind_fit_nms}
\alias{hfind_fit_nms}
\title{Helper of \link{find_fit_nms}.}
\usage{
hfind_fit_nms(
  field = "^N_log2_R",
  nms,
  sub_sids,
  pat0 = "[0-9]{3}[NC]{0,1}\\\\s+\\\\("
)
}
\arguments{
\item{field}{A field of name prefix.}

\item{nms}{A vector of column names with sample IDs in parentheses.}

\item{sub_sids}{A subset of sample IDs for match with \code{nms}.}

\item{pat0}{A pattern of regular expression.}
}
\description{
Helper of \link{find_fit_nms}.
}
