% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{hfind_shared_prots}
\alias{hfind_shared_prots}
\title{Helper of find_shared_prots.}
\usage{
hfind_shared_prots(df, pep_id = "pep_seq", prot_id = "prot_acc")
}
\arguments{
\item{df}{PSM data.}

\item{pep_id}{the column key of peptide sequences.}

\item{prot_id}{The column key of protein accessions.}
}
\description{
Helper of find_shared_prots.
}
