% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspa.R
\name{hgspa_summary_mean}
\alias{hgspa_summary_mean}
\title{Helper of \link{ok_min_size}.}
\usage{
hgspa_summary_mean(
  dfs,
  min_size = 10L,
  min_delta = 5L,
  gsscore_cutoff = 5,
  gspval_cutoff = 0.01,
  gslogFC_cutoff = log2(1.2),
  id = "gene",
  id_gspa = "entrez"
)
}
\arguments{
\item{dfs}{A set of data frames.}

\item{min_size}{Numeric value or vector; minimum number of protein entries for
consideration in gene set tests. The number is after data filtration by
\code{pval_cutoff}, \code{logFC_cutoff} or varargs expressions under
\code{filter_}. The default is 10 for all formulas matched to or specified
in argument \code{fml_nms}. Formula-specific threshold is allowed by
supplying a vector of sizes.}

\item{min_delta}{Numeric value or vector; the minimum count difference between
the up- and the down-expressed group of proteins for consideration in gene
set tests. For example at \code{min_delta = 4}, a gene set will 6
upregulated proteins and 2 down-expressed proteins, or vice versa, will be
assessed. The number is after data filtration by \code{pval_cutoff},
\code{logFC_cutoff} or varargs expressions under \code{filter_}. The default
is 4 for all formulas matched to or specified in argument \code{fml_nms}.
Formula-specific threshold is allowed by supplying a vector of sizes.}

\item{gspval_cutoff}{Depreciated. Numeric value or vector; the cut-off in
gene-set significance \code{pVal}. Only enrichment terms with \code{pVals}
more significant than the threshold will be reported. The default is 0.05
for all formulas matched to or specified in argument \code{fml_nms}.
Formula-specific threshold is allowed by supplying a vector of cut-off
values.}

\item{gslogFC_cutoff}{Numeric value or vector; the cut-off in gene-set
enrichment fold change. Only enrichment terms with absolute fold change
greater than the threshold will be reported. The default magnitude is
\code{log2(1.2)} for all formulas matched to or specified in argument
\code{fml_nms}. Formula-specific threshold is allowed by supplying a vector
of absolute values in \code{log2FC}.}

\item{id}{Currently only "entrez".}

\item{id_gspa}{Character string; the ID type used in GSPA. The default is
\code{entrez}.}
}
\description{
Helper of \link{ok_min_size}.
}
