% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtests.R
\name{himpute_baseline}
\alias{himpute_baseline}
\title{Helper of \link{impute_baseline_int}.}
\usage{
himpute_baseline(dfRg, dfIg, g, b, dfsR, dfsI, grps)
}
\arguments{
\item{dfRg}{The data frame of log2Ratios belongs to the group \code{g}.}

\item{dfIg}{The data frame of intensities belongs to the group \code{g}.}

\item{g}{The current group.}

\item{b}{The current baseline intensities. Each value corresponds to one
sample.}

\item{dfsR}{All data frames of log2Ratios.}

\item{dfsI}{All data frames of intensities.}

\item{grps}{All groups}
}
\description{
Helper of \link{impute_baseline_int}.
}
