% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_a0_vnl0_fnl1.R, R/ms2_a1_vnl0_fnl0.R,
%   R/ms2_a1_vnl0_fnl1.R, R/ms2_a1_vnl1_fnl0.R, R/ms2_base.R
\name{hms2_a0_vnl0_fnl1}
\alias{hms2_a0_vnl0_fnl1}
\alias{hms2_a1_vnl0_fnl0}
\alias{hms2_a1_vnl0_fnl1}
\alias{hms2_a1_vnl1_fnl0}
\alias{hms2_base}
\title{Searches MGF frames.}
\usage{
hms2_a0_vnl0_fnl1(
  mgf_frames,
  theopeps,
  aa_masses,
  ntmass,
  ctmass,
  fmods_nl,
  mod_indexes,
  type_ms2ions = "by",
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 32L,
  minn_ms2 = 7L,
  ppm_ms1 = 20L,
  ppm_ms2 = 25L,
  min_ms2mass = 110L,
  digits = 4L
)

hms2_a1_vnl0_fnl0(
  mgf_frames,
  theopeps,
  aa_masses,
  ntmod = NULL,
  ctmod = NULL,
  ntmass,
  ctmass,
  amods,
  mod_indexes,
  type_ms2ions = "by",
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 32L,
  minn_ms2 = 7L,
  ppm_ms1 = 20L,
  ppm_ms2 = 25L,
  min_ms2mass = 110L,
  digits = 4L
)

hms2_a1_vnl0_fnl1(
  mgf_frames,
  theopeps,
  aa_masses,
  ntmod = NULL,
  ctmod = NULL,
  ntmass,
  ctmass,
  amods,
  fmods_nl,
  mod_indexes,
  type_ms2ions = "by",
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 32L,
  minn_ms2 = 7L,
  ppm_ms1 = 20L,
  ppm_ms2 = 25L,
  min_ms2mass = 110L,
  digits = 4L
)

hms2_a1_vnl1_fnl0(
  mgf_frames,
  theopeps,
  aa_masses,
  ntmod = NULL,
  ctmod = NULL,
  ntmass,
  ctmass,
  amods,
  vmods_nl,
  mod_indexes,
  type_ms2ions = "by",
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 32L,
  minn_ms2 = 7L,
  ppm_ms1 = 20L,
  ppm_ms2 = 25L,
  min_ms2mass = 110L,
  digits = 4L
)

hms2_base(
  mgf_frames,
  theopeps,
  aa_masses,
  ntmass,
  ctmass,
  mod_indexes,
  type_ms2ions = "by",
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 32L,
  minn_ms2 = 7L,
  ppm_ms1 = 20L,
  ppm_ms2 = 25L,
  min_ms2mass = 110L,
  digits = 4L
)
}
\arguments{
\item{mgf_frames}{A group of mgf frames. Each frame contains one to multiple
mgf queries with the same frame number.}

\item{theopeps}{Binned theoretical peptides corresponding to an i-th
\code{aa_masses}.}

\item{aa_masses}{Amino-acid lookup at a given combination of fixed and
variable.}

\item{ntmass}{The mass of N-terminal.}

\item{ctmass}{The mass of C-terminal.}

\item{fmods_nl}{The attribute of \code{fmods_nl} from an \code{aa_masses}.}

\item{mod_indexes}{Integer; the indexes of fixed and/or variable
modifications.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{amods}{The attribute of \code{amods} from an \code{aa_masses}.}

\item{vmods_nl}{The attribute of \code{vmods_nl} from an \code{aa_masses}.}
}
\description{
(5) "amods- tmod- vnl- fnl+", (6) "amods- tmod+ vnl- fnl+"

(7) "amods+ tmod- vnl- fnl-", (8) "amods+ tmod+ vnl- fnl-"

(11) "amods+ tmod- vnl- fnl+", (12) "amods+ tmod+ vnl- fnl+"

(9) "amods+ tmod- vnl+ fnl-", (10) "amods+ tmod+ vnl+ fnl-"

(1) "amods- tmod- vnl- fnl-", (2) "amods- tmod+ vnl- fnl-"
}
