% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{hpepMBR}
\alias{hpepMBR}
\title{Spreads Mzion peptide numbers.}
\usage{
hpepMBR(
  filelist,
  basenames,
  set_idxes,
  injn_idxes,
  are_refs,
  are_smpls,
  dfs = NULL,
  df_sps = NULL,
  prot_spec_counts = NULL,
  dat_dir = NULL,
  path_ms1 = NULL,
  ms1files = NULL,
  temp_dir = NULL,
  mbr_ret_tol = 25,
  max_mbr_fold = 20,
  sp_centers_only = FALSE,
  imp_refs = FALSE,
  group_psm_by = "pep_seq_modz",
  group_pep_by = "gene"
)
}
\arguments{
\item{filelist}{The name of peptide files (TMTset1_LCMSinj1_Peptide_N.txt)
with prepending path.}

\item{basenames}{The basenames corresponding to \code{filelist}.}

\item{set_idxes}{The TMT_Set indexes corresponding to \code{basenames}.}

\item{injn_idxes}{The LCMS_Injection indexes corresponding to
\code{basenames}.}

\item{are_refs}{A logical vector corresponding to \code{basenames},
indicating the reference status.}

\item{are_smpls}{A logical vector corresponding to \code{basenames},
indicating the sample status.}

\item{dfs}{Lists of peptide tables in correspondence to \code{filelist}.}

\item{df_sps}{A look-up table between peptide sequences and species.}

\item{prot_spec_counts}{A data frame of protein spectrum counts.}

\item{dat_dir}{The working directory.}

\item{path_ms1}{The path to MS1 data files.}

\item{ms1files}{The names of MS1 data files.}

\item{temp_dir}{The path to temporary files.}

\item{mbr_ret_tol}{The tolerance in MBR retention time in seconds.}

\item{max_mbr_fold}{The maximum absolute fold change in MBR.}

\item{sp_centers_only}{Logical; for a side-effect to return only the values
of species centers.}

\item{imp_refs}{Logical; impute missing references or not.}

\item{group_psm_by}{Group PSMs by.}

\item{group_pep_by}{Group peptides by.}

\item{df}{A data frame of peptide table, ordered by ascending \code{TMT_Set}.
Helper of \link{peptideMBR}}
}
\description{
No aggregation of data at the same TMT_Set but different LCMS_Injs.
}
\details{
Spreads fields of numeric values: sd_log2_R, log2_R, log2_R, I, N_I by TMT
sets.

Also works for LFQ as each sample corresponds to a TMT set.

For single SILAC sample, the values of log2Ratios spreads into
\emph{MULTIPLE} columns of heavy, light etc. Despite, log2Ratios remains NA,
just like regular single-sample LFQ. The log2Ratios will be later calculated
with \link{calcLFQPepNums} that are based on intensity values.
}
