% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{hrm_lowIntPSMs}
\alias{hrm_lowIntPSMs}
\title{Helper for removals of low-intensity PSMs}
\usage{
hrm_lowIntPSMs(dfs, key = "pep_seq_modz", yfrac = 0.02, max_n_apexes = 1L)
}
\arguments{
\item{dfs}{A list of PSM tables.}

\item{key}{The name of a column key.}

\item{yfrac}{The minimum percentage of intensity in relative to the base peak
for considerations.}

\item{max_n_apexes}{The maximum number of apexes for consideration.}
}
\description{
Single primary apex under the same \code{pep_seq_modz} since
\code{max_n_apexes = 1L}.
}
