% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{hsub_by_bestz}
\alias{hsub_by_bestz}
\title{Helper for subsetting PSMs by the best charge state under a
\code{pep_seq_mod}}
\usage{
hsub_by_bestz(
  dfs,
  cols = c("pep_seq_mod", "pep_exp_z", "pep_apex_ret", "I000")
)
}
\arguments{
\item{dfs}{A set of PSM tables.}

\item{cols}{The column keys needed for determining the best charge states.}
}
\description{
Helper for subsetting PSMs by the best charge state under a
\code{pep_seq_mod}
}
