% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{hsubColnames}
\alias{hsubColnames}
\title{Helper of \link{newColnames}.}
\usage{
hsubColnames(
  id = NULL,
  df,
  pattern = "[RI][0-9]{3}[NC]{0,1}",
  TMT_Set = 1L,
  label_scheme_sub
)
}
\arguments{
\item{id}{A character string; for example, "heavy" or "light" that can be
found from column \code{pep_group} in \code{df}. The value of \code{id} is
NULL if column \code{pep_group} is not in \code{df}.}

\item{df}{Data frame; log2FC data.}

\item{pattern}{Regex pattern for capturing intensity and ratio columns.}

\item{TMT_Set}{Integer; the index of TMT experiment.}

\item{label_scheme_sub}{Experiment summary at \code{TMT_Set}.}
}
\description{
Helper of \link{newColnames}.
}
