% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular.R
\name{hupdate_subcellular}
\alias{hupdate_subcellular}
\title{Update subcellular protein and peptide tables.}
\usage{
hupdate_subcellular(
  df_trend,
  dat_dir,
  id = "gene",
  label_scheme_sub,
  col_fraction = NULL,
  col_subtype = NULL
)
}
\arguments{
\item{df_trend}{Trend results}

\item{dat_dir}{The working directory.}

\item{id}{Identifier.}

\item{label_scheme_sub}{Metadata.}

\item{col_fraction}{The column key in metadata linking to subcellular
values.}

\item{col_subtye}{The column key metadata linking to sample sub-types.}
}
\description{
Not used.
}
