% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identical_dots}
\alias{identical_dots}
\title{Compare dot-dot-dot between prior and current}
\usage{
identical_dots(call_nm, curr_dots, pattern)
}
\arguments{
\item{call_nm}{The name of a function call.}

\item{curr_dots}{The values of the current dots.}

\item{pattern}{The pattern for comparison.}
}
\description{
Compare dot-dot-dot between prior and current
}
