% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{impPepNA}
\alias{impPepNA}
\title{Imputes NA values in a peptide table}
\usage{
impPepNA(df, cols_num)
}
\arguments{
\item{df}{A intensity data frame (not matrix) of peptides.}

\item{cols_num}{The indexes of numeric columns in \code{df}.}
}
\description{
Imputes NA values in a peptide table
}
