% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{impPepNA_byrow}
\alias{impPepNA_byrow}
\title{Imputes NA values in a peptide table}
\usage{
impPepNA_byrow(df, fold = 20)
}
\arguments{
\item{df}{A intensity data frame (not matrix) of peptides.}

\item{fold}{The fold difference to mean.}
}
\description{
Not used.
}
