% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{impRefNA}
\alias{impRefNA}
\title{Impute NA for references}
\usage{
impRefNA(df, cols_ref, cols_smpl, refChannels, group_psm_by = "pep_seq_mod")
}
\arguments{
\item{df}{A data frame.}

\item{cols_ref}{The colume keys of reference IDs.}

\item{cols_smpl}{The column keys of sample IDs.}

\item{refChannels}{The reference channels.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}
}
\description{
Impute NA for references
}
