% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtests.R
\name{impute_baseline_ints}
\alias{impute_baseline_ints}
\title{Imputation baseline intensities.}
\usage{
impute_baseline_ints(
  dfsR,
  dfsI,
  dfR,
  dfI,
  ys_base,
  sample_ids,
  max_fold = 100,
  impute_low_qualities = FALSE,
  replace_lows_by = 0
)
}
\arguments{
\item{dfsR}{List of log2Ratio data frames. Data are grouped according to the
formulas in \link{prnSig}.}

\item{dfsI}{The corresponding intensity data frames.}

\item{dfR}{The grand log2Ratio data.}

\item{dfI}{The grand intensity data.}

\item{ys_base}{The baseline intensities. The length is equal to the number of
columns in \code{dfR}.}

\item{sample_ids}{The sample IDs for joining data at the original order.}

\item{max_fold}{The maximum fold change allowed.}

\item{impute_low_qualities}{Logical; impute low-quality data entries or not.}

\item{replace_lows_by}{A replacement value of log2Ratios at
\code{impute_low_qualities = TRUE}.}
}
\description{
The values of log2Ratios are derived accordingly
}
