% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{imputeNA}
\alias{imputeNA}
\title{Imputes NA}
\usage{
imputeNA(id, ...)
}
\arguments{
\item{id}{Character string to indicate the type of data. Peptide data will be
used at \code{id = pep_seq} or \code{pep_seq_mod}, and protein data at
\code{id = prot_acc} or \code{gene}.}

\item{...}{Parameters from \code{\link[mice]{mice}}}
}
\value{
\code{Peptide_impNA.txt} for peptide data and \code{Protein_impNA.txt}
 for protein data.
}
\description{
Imputes NA
}
\examples{
imputeNA(id = gene, m = 5, maxit = 5)

\dontrun{
}
}
