% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ins_cols_after}
\alias{ins_cols_after}
\title{Re-order columns in a data frame}
\usage{
ins_cols_after(df = NULL, idx_bf = ncol(df), idx_ins = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{idx_bf}{A column index for the insertion of columns after.}

\item{idx_ins}{Column index(es) for the columns to be inserted after
\code{idx_bf}.}
}
\description{
\code{ins_cols_after} re-orders columns in a data frame.
}
