% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_trivial_col}
\alias{is_trivial_col}
\title{Finds a trivial column}
\usage{
is_trivial_col(df, col)
}
\arguments{
\item{df}{A data frame of \code{log2FC} and \code{intensity}.}

\item{col}{The index of column.}
}
\description{
Finds a trivial column
}
\examples{
\donttest{is_trivial_col(mtcars, 1)}
}
