% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{join_mgfs}
\alias{join_mgfs}
\title{Joins MGF data to MSGF outputs.}
\usage{
join_mgfs(dat_dir = NULL, mgf_path = NULL, target_fdr = 0.01)
}
\arguments{
\item{dat_dir}{The working directory.}

\item{mgf_path}{The path to MGF files. The default is
\code{file.path(dat_dir, "mgf")}.}

\item{target_fder}{A targeted false-discovery rate (FDR).}
}
\description{
Joins MGF data to MSGF outputs.
}
