% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{keep_n_misses}
\alias{keep_n_misses}
\title{Find mis-cleavages in a vector.}
\usage{
keep_n_misses(x, n)
}
\arguments{
\item{x}{A vector of data.}

\item{n}{Integer. The number of mis-cleavages.}
}
\description{
A convenience utility may be used to extract the first \eqn{n+1} peptides
from 0 to n mis-cleavages. It also assumes that the data were already sorted
in a desirable way.
}
