% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{keep_one_to_n}
\alias{keep_one_to_n}
\title{Find the first n elements in a vector}
\usage{
keep_one_to_n(x, n)
}
\arguments{
\item{x}{A vectot of data.}

\item{n}{Integer. The first n elements by index in \code{x} to be kept.}
}
\description{
Find the first n elements in a vector
}
