% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{keep_psm_bestint}
\alias{keep_psm_bestint}
\title{Keep the highest intensity PSM}
\usage{
keep_psm_bestint(dfs)
}
\arguments{
\item{dfs}{PSM tables unique by \code{pep_seq_modz}.}
}
\description{
Keep the highest intensity PSM
}
