% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-datasets.R
\docType{data}
\name{kegg_sets_hs}
\alias{kegg_sets_hs}
\title{Entrez IDs of human KEGG annotations.}
\format{A list at a length of 229 
\describe{
  \item{hsa...}{human entries}
}}
\source{
\url{https://www.genome.jp/kegg/pathway.html}
}
\usage{
kegg_sets_hs
}
\description{
A dataset containing human entrez IDs by the gene sets of KEGG pathways.
}
\keyword{datasets}
