% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{kegg_sets_mm}
\alias{kegg_sets_mm}
\title{Entrez IDs of mouse KEGG annotations.}
\format{A list at a length of 225 
\describe{
  \item{mmu...}{mouse entries}
}}
\source{
\url{https://www.genome.jp/kegg/pathway.html}
}
\usage{
kegg_sets_mm
}
\description{
A dataset containing mouse entrez IDs by the gene sets of KEGG pathways.
}
\keyword{datasets}
