% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-datasets.R
\docType{data}
\name{kegg_sets_rn}
\alias{kegg_sets_rn}
\title{Entrez IDs of rat KEGG annotations.}
\format{A list at a length of 225 
\describe{
  \item{mmu...}{rat entries}
}}
\source{
\url{https://www.genome.jp/kegg/pathway.html}
}
\usage{
kegg_sets_rn
}
\description{
A dataset containing rat entrez IDs by the gene sets of KEGG pathways.
}
\keyword{datasets}
