% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-datasets.R
\docType{data}
\name{kinase_lookup}
\alias{kinase_lookup}
\title{Lookups of human or mouse kinases.}
\format{A data frame with 2050 rows and 9 variables:
\describe{
  \item{refseq_acc}{RefSeq accession number}
  \item{gene}{gene names}
  \item{kin_attr}{the attribute being a kinase or not}
  \item{kin_class}{the class of a kinase; for example: TK, tyrosine kinase}
  \item{kin_order}{the number order of a kinase class by the kinase tree}
  \item{uniprot_acc}{UniProt accession number}
  \item{uniprot_id}{UniProt ID}
  \item{prot_desc}{protein description}
  \item{entrez}{Entrez ID}
}}
\source{
\url{http://kinase.com/human/kinome/phylogeny.html}
}
\usage{
kinase_lookup
}
\description{
A dataset containing the \code{refseq}, \code{uniprot}, \code{gene names} of
human or mouse kinases.
}
\keyword{datasets}
