% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{kinsub_hs}
\alias{kinsub_hs}
\title{Kinase (in gene name) and the the entrez IDs of substrates}
\format{
A list at a length of 385 
\describe{
  \item{hu...}{human entries}
}
}
\source{
\url{https://www.phosphosite.org/}
}
\usage{
kinsub_hs
}
\description{
A dataset containing human entrez IDs by the gene sets of kinase-substrate
networks.
}
\keyword{datasets}
