% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{kinsub_mm}
\alias{kinsub_mm}
\title{Kinase (in gene name) and the the entrez IDs of substrates}
\format{
A list at a length of 217 
\describe{
  \item{mm...}{mouse entries}
}
}
\source{
\url{https://www.phosphosite.org/}
}
\usage{
kinsub_mm
}
\description{
A dataset containing mouse entrez IDs by the gene sets of kinase-substrate
networks.
}
\keyword{datasets}
