% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{kinsub_rn}
\alias{kinsub_rn}
\title{Kinase (in gene name) and the the entrez IDs of substrates}
\format{
A list at a length of 100 
\describe{
  \item{rn...}{rat entries}
}
}
\source{
\url{https://www.phosphosite.org/}
}
\usage{
kinsub_rn
}
\description{
A dataset containing rat entrez IDs by the gene sets of kinase-substrate
networks.
}
\keyword{datasets}
