% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\name{lgl_cleanup}
\alias{lgl_cleanup}
\title{Updates data groups of `log2_R_` and `I_` etc. by logical matrix
`lgl_log2_...`.}
\usage{
lgl_cleanup(df, type = "sd", rm_allna = FALSE)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{type}{Type of logical matrix.}
}
\description{
Updates data groups of `log2_R_` and `I_` etc. by logical matrix
`lgl_log2_...`.
}
