% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{list_leftmatch}
\alias{list_leftmatch}
\title{Matches two lists.}
\usage{
list_leftmatch(a, b)
}
\arguments{
\item{a}{The left vector.}

\item{b}{The right vector.}
}
\description{
Without making a data frame.
}
\examples{
\donttest{
a <- c(3, 4, 1, 2, 5)
b <- 2

list_leftmatch(a, b)
}
}
