% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{lm_gspa}
\alias{lm_gspa}
\title{significance tests of pVals between the up and the down groups}
\usage{
lm_gspa(df, min_delta)
}
\description{
significance tests of pVals between the up and the down groups
}
