% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtests.R
\name{lm_summary}
\alias{lm_summary}
\title{Model summary}
\usage{
lm_summary(pvals, log2rs, pval_cutoff, logFC_cutoff, padj_method = "BH")
}
\arguments{
\item{pvals}{A data frame of pVal}

\item{log2rs}{A data frame of log2FC}

\item{pval_cutoff}{Numeric; the cut-off in significance \code{pVal}. Entries
with \code{pVals} smaller than the threshold will be excluded from multiple
test corrections. The default is at \code{1} to include all entries.}

\item{logFC_cutoff}{Numeric; the cut-off in \code{log2FC}. Entries with
absolute \code{log2FC} smaller than the threshold will be removed from
multiple test corrections. The default is at \code{log2(1)} to include all
entries.}

\item{padj_method}{Character string; the method of multiple-test corrections
for uses with \link[stats]{p.adjust}. The default is "BH". See
?p.adjust.methods for additional choices.}
}
\description{
Model summary
}
