% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_craps}
\alias{load_craps}
\title{Load cRAPs}
\usage{
load_craps(acc_types)
}
\arguments{
\item{acc_types}{The types of protein accessions}
}
\description{
Load the cRAP proteins by the types of accessions
}
