% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{load_dbs}
\alias{load_dbs}
\title{Loads species-specific Databases}
\usage{
load_dbs(gset_nms = NULL, species = NULL)
}
\description{
A function loads a set of precompiled tables in protein accessions, \code{GO}
data and
\code{\href{http://software.broadinstitute.org/gsea/msigdb}{molecular
signatures}}.
}
\examples{
load_dbs("human")

}
\seealso{
\code{\link{load_expts}} for supported species.
}
