% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{load_fasta}
\alias{load_fasta}
\title{Loads fasta}
\usage{
load_fasta(fasta = NULL)
}
\arguments{
\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search. There is no default and users need to provide the
correct file path(s) and name(s).}
}
\description{
Loads fasta
}
\examples{
\donttest{
fasta_db <- load_fasta("~/proteoQ/dbs/fasta/uniprot/uniprot_hs_2020_05.fasta")
}
}
