% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{load_fasta2}
\alias{load_fasta2}
\title{Loads fasta (with parsing rule).}
\usage{
load_fasta2(fasta = NULL, acc_type = NULL, acc_pattern = NULL)
}
\arguments{
\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search. There is no default and users need to provide the
correct file path(s) and name(s).}

\item{acc_type}{Character string(s); the types of protein accessions in one
of c("uniprot_acc", "uniprot_id", "refseq_acc", "other"). For custom names,
the corresponding regular expressions need to be supplied via argument
\code{acc_pattern}.}

\item{acc_pattern}{Regular expression(s) describing the patterns to separate
the header lines of fasta entries. At the \code{NULL} default, the pattern
will be automated when \code{acc_type} are among c("uniprot_acc",
"uniprot_id", "refseq_acc", "other").}
}
\description{
The length of \code{acc_type} needs to match the length of \code{fasta};
otherwise, the first value will be used for all \code{fasta} files.
}
\examples{
\donttest{
fasta_db <- load_fasta2(
              c("~/proteoQ/dbs/fasta/uniprot/uniprot_hs_2020_05.fasta",
                "~/proteoQ/dbs/fasta/crap/crap.fasta"),
              c("uniprot_acc", "other")
)

# Need `acc_pattern` as "crap" is not one of the default acc_type
load_fasta2(
   c("~/proteoQ/dbs/fasta/uniprot/uniprot_hs_2020_05.fasta",
     "~/proteoQ/dbs/fasta/crap/crap.fasta"),
   c("uniprot_acc", "crap")
)

# ok
fasta_db2 <- load_fasta2(
               c("~/proteoQ/dbs/fasta/uniprot/uniprot_hs_2020_05.fasta",
                        "~/proteoQ/dbs/fasta/crap/crap.fasta"),
               c("uniprot_acc", "crap"),
               c("^>..\\\\|([^\\\\|]+)\\\\|[^\\\\|]+", "(.*)")
)

fasta_db3 <- load_fasta2(
               c("~/proteoQ/dbs/fasta/uniprot/uniprot_hs_2020_05.fasta",
                        "~/proteoQ/dbs/fasta/crap/crap.fasta"),
               c("my_acc", "crap"),
               c("^>..\\\\|([^\\\\|]+)\\\\|[^\\\\|]+", "(.*)")
)

stopifnot(identical(fasta_db, fasta_db2),
          identical(fasta_db, fasta_db3))
}
}
