% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_ls_group}
\alias{load_ls_group}
\title{Loads label_scheme or label_scheme_group.}
\usage{
load_ls_group(dat_dir, filename = "label_scheme", prefer_group = TRUE)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{filename}{A file name of "label_scheme" or "label_scheme_full".}

\item{prefer_group}{Logical; if TRUE, prefer label_scheme_group over
label_scheme etc.}
}
\description{
Prefer label_scheme_group over label_scheme.
}
