% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{load_mgfs}
\alias{load_mgfs}
\title{Helper in loading MGFs.}
\usage{
load_mgfs(mgf_path, min_mass, topn_ms2ions, ppm_ms1)
}
\arguments{
\item{mgf_path}{The file path to a list of MGF files. There is no default and
the experimenters need to supply the files.}

\item{min_mass}{A minimum mass of precursors.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}
}
\description{
Helper in loading MGFs.
}
