% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{load_stringdbs}
\alias{load_stringdbs}
\title{Loads species-specific databases of STRING}
\usage{
load_stringdbs(db_nms = NULL)
}
\arguments{
\item{db_nms}{Character string(s) to the name(s) of STRING database(s) with
prepended directory path. The \code{STRING} database(s) need to match those
generated from \code{\link{prepString}}. There is no default and users need
to provide the correct file path(s) and name(s).}
}
\description{
A function loads a set of precompiled gene sets of 
\href{http://current.geneontology.org/products/pages/downloads.html}{GO}
and
\href{http://string-db.org/}{String}.
}
\examples{
\donttest{
prepString(human)
prepString(mouse)

load_stringdbs(
  db_nms = c("~\\\\proteoQ\\\\dbs\\\\string\\\\string_hs.rds",
             "~\\\\proteoQ\\\\dbs\\\\string\\\\string_mm.rds")
)
}

}
\seealso{
\code{\link{load_expts}} for supported species.

\code{\link{load_dbs}} for loading databases of \code{GO} and
 \code{MSig}.
}
