% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{locate_outliers}
\alias{locate_outliers}
\title{Locates the positions of outliersf}
\usage{
locate_outliers(df, range_colRatios)
}
\arguments{
\item{df}{A data frame containing the PSM table from Mascot.}

\item{range_colRatios}{The range of columns.}
}
\value{
A data frame.
}
\description{
Locates the positions of outliersf
}
\examples{
locate_outliers(df, 2:3)
}
