% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{locate_vec_outlier}
\alias{locate_vec_outlier}
\title{Locate outliers from a vector.}
\usage{
locate_vec_outlier(x, p = 0.05, k = 1L)
}
\arguments{
\item{x}{A numeric vector.}

\item{p}{The level of type-I error.}

\item{k}{The number of suspected outliers.}
}
\description{
Locate outliers from a vector.
}
