% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm.R
\name{make_dotplot_hm}
\alias{make_dotplot_hm}
\title{Makes heat maps as dot plots.}
\usage{
make_dotplot_hm(
  df,
  df_lgr,
  df_int,
  id = "gene",
  plot_sids,
  sample_ids,
  xmin = -2,
  xmax = -xmin,
  filepath,
  fn_prefix,
  fn_suffix,
  dots
)
}
\arguments{
\item{df}{A data frame of both intensities and log2Ratios.}

\item{df_lgr}{A data frame of log2Ratios.}

\item{df_int}{A data frame of intensities.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{plot_sids}{Ordered sample IDs for plotting.}

\item{sample_ids}{Sample_ids from metadata.}

\item{xmin}{The minimum X value.}

\item{xmax}{The maximum X value.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{fn_prefix}{The prefix of output file name.}

\item{fn_suffix}{The suffix of output file name.}

\item{dots}{Additional arguments.}
}
\description{
Makes heat maps as dot plots.
}
