% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{make_fastapeps0}
\alias{make_fastapeps0}
\title{Make peptide sequences from FASTA databases.}
\usage{
make_fastapeps0(fasta_db, max_miss = 2L)
}
\arguments{
\item{fasta_db}{Fasta database(s).}

\item{max_miss}{The maximum number of mis-cleavages per peptide sequence.}
}
\description{
Not yet concatenating peptides by the number of mis-cleavages.
}
