% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{make_frac_smry}
\alias{make_frac_smry}
\title{Makes new frac_smry.xlsx.}
\usage{
make_frac_smry(
  frac_smry = "frac_smry.xlsx",
  expt_smry = "expt_smry",
  dat_dir = NULL,
  label_scheme_full = NULL,
  tbl_mascot = NULL,
  is_tmt = TRUE
)
}
\arguments{
\item{frac_smry}{A character string to a \code{.xlsx} file containing peptide
fractionation summary. The default is \code{frac_smry.xlsx}.}

\item{expt_smry}{A character string to a \code{.xlsx} file containing the
metadata of TMT or LFQ experiments. The default is \code{expt_smry.xlsx}.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{label_scheme_full}{The metadata of expt_smry.}

\item{tbl_mascot}{A table of message for Mascot workflow.}

\item{is_tmt}{Logical; is TMT experiments or not.}
}
\description{
warning: data in a auto-generated \code{frac_smry.xlsx} currently may not fix
itself if they were based on wrong information from \code{expt_smry.xlsx}.
May need more thorough checking of \code{frac_smry.xlsx} upon any update or
correction of \code{expt_smry.xlsx}
}
