% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{make_mascot_ratios}
\alias{make_mascot_ratios}
\title{Defines the ratio columns of Mascot outputs.}
\usage{
make_mascot_ratios(TMT_plex = 10L, nrow = 1L)
}
\arguments{
\item{TMT_plex}{Numeric; the multiplexity of TMT, i.e., 10, 11 etc.}

\item{nrow}{Positive integer; the number of rows in the data.}
}
\description{
Defines the ratio columns of Mascot outputs.
}
