% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_mq_meta2}
\alias{make_mq_meta2}
\title{Process MaxQuant mqpar.xml}
\usage{
make_mq_meta2(
  dat_dir = proteoQ:::get_gl_dat_dir(),
  mqpar = "mqpar.xml",
  filename = "mq_meta.txt",
  out = "new_mqpar.xml"
)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{mqpar}{The name of .xml file. The default is "mqpar.xml".}

\item{filename}{The name of metadata file.}

\item{out}{The name of output .xml file.}
}
\description{
Process MaxQuant mqpar.xml
}
\examples{
\dontrun{
dat_dir <- "Y:/qzhang/MaxQuant"
make_mq_meta2()
}
}
