% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{make_renorm_data}
\alias{make_renorm_data}
\title{Renormalize data by group.}
\usage{
make_renorm_data(
  df_lgr = NULL,
  df_log2r = NULL,
  df_int = NULL,
  group_renorm_by = NULL,
  group_fct_int = NULL,
  label_scheme_sub = NULL
)
}
\arguments{
\item{df_lgr}{A data frame of log2Ratios.}

\item{df_log2r}{A data frame of log2Ratios with clustering difference made
for heat map visualization.}

\item{df_int}{A data frame of intensities.}

\item{group_renorm_by}{A metadata column key for re-normalization data within
each groups.}

\item{group_fct_int}{Used in conjunction with \code{group_renorm_by}; a
vector for intensity scaling by group.}

\item{label_scheme_sub}{Metadata.}
}
\description{
For defined scope of data.
}
