% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlfq.R
\name{makeLFQSpecies}
\alias{makeLFQSpecies}
\title{Makes the look-ups of shared peptides across species}
\usage{
makeLFQSpecies(
  mbr_peps = NULL,
  mbr_sps = NULL,
  nms_sp,
  new_na_species = ".other",
  dat_dir = NULL,
  to_list = TRUE
)
}
\arguments{
\item{mbr_peps}{A vector of \code{pep_seq_modz}.}

\item{mbr_sps}{A vector of peptide species in the universe.}

\item{nms_sp}{The names of species (for conversion from character strings to
integers).}

\item{new_na_species}{A replace value for NA species.}

\item{dat_dir}{A working directory.}

\item{to_list}{Logical; convert output to list or not.}
}
\description{
Makes the look-ups of shared peptides across species
}
