% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{makeLFQXYTS}
\alias{makeLFQXYTS}
\title{Prepare the XYTS matrices for LFQ}
\usage{
makeLFQXYTS(
  pep_seq_modzs = NULL,
  pep_exp_ints = NULL,
  pep_exp_mzs = NULL,
  pep_apex_rets = NULL,
  pep_apex_scans = NULL
)
}
\arguments{
\item{pep_seq_modzs}{List of peptide sequences with modifications and charge
states.}

\item{pep_exp_ints}{List of intensity (Y) values.}

\item{pep_exp_mzs}{List of m-over-z (X) values.}

\item{pep_apex_rets}{List of retention time (T) values.}

\item{pep_apex_scans}{List of apex scan numbers (S).}
}
\description{
Prepare the XYTS matrices for LFQ
}
