% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{makePepDIANN}
\alias{makePepDIANN}
\title{Make DIANN peptide table}
\usage{
makePepDIANN(
  dat_dir = NULL,
  group_pep_by = "gene",
  fasta = NULL,
  entrez = NULL,
  use_lowercase_aa = FALSE
)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search. There is no default and users need to provide the
correct file path(s) and name(s).}

\item{entrez}{Character string(s) to the name(s) of entrez file(s) with
prepended directory path. At the \code{NULL} default, a convenience lookup
is available for species among \code{c("human", "mouse", "rat")}. For other
species, users need to provide the file path(s) and name(s) for the lookup
table(s). See also \code{\link{Uni2Entrez}} and \code{\link{Ref2Entrez}} for
preparing custom entrez files.}

\item{use_lowercase_aa}{Logical; if TRUE, modifications in amino acid residues
will be abbreviated with lower-case and/or \code{^_~}. See the table below
for details. The default is TRUE.}
}
\description{
Make DIANN peptide table
}
