% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{makeTrendRes}
\alias{makeTrendRes}
\title{Helper to make the outputs of trend analysis.}
\usage{
makeTrendRes(
  fn,
  choice = "cmeans",
  dots = NULL,
  id = "gene",
  df_mean_log2r = NULL,
  ext = "txt"
)
}
\arguments{
\item{fn}{A prefix of file name.}

\item{choice}{Character string; the clustering method in \code{c("cmeans",
"clara", "kmeans", "pam", "fanny")}. The default is "cmeans".}

\item{id}{An identifier.}

\item{df_mean_log2r}{Data of mean log2Ratio.}

\item{ext}{The extension of out filenames.}
}
\description{
Helper to make the outputs of trend analysis.
}
