% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{map_olr_pep}
\alias{map_olr_pep}
\title{Map outlier indexes to peptide sequences.}
\usage{
map_olr_pep(xs, ys)
}
\arguments{
\item{xs}{A vector of outlier indexes.}

\item{ys}{A vector of peptide sequences.}
}
\description{
Map outlier indexes to peptide sequences.
}
