% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{map_pepprot2}
\alias{map_pepprot2}
\title{Helper of \link{groupProts}.}
\usage{
map_pepprot2(df, out_path = NULL)
}
\arguments{
\item{df}{The data frame from upstream steps. It must contains the two
columns of \code{prot_acc} and \code{pep_seq}.}

\item{out_path}{An output path.}
}
\description{
Builds the logical map between peptide (in rows) and proteins (in columns).
}
\examples{
\donttest{
df <- data.frame(prot_acc = character(2000), pep_seq = character(2000))
set.seed(100)
df$prot_acc <- sample(LETTERS[1:20], 2000, replace = TRUE)
df$pep_seq <- sample(letters[1:26], 20, replace = TRUE)
df <- df[!duplicated(df), ] 

out <- map_pepprot2(df)
}
}
