% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{map_peps_prots}
\alias{map_peps_prots}
\title{Helper: maps peptides to possible proteins}
\usage{
map_peps_prots(df, group_psm_by = "pep_seq", group_pep_by = "prot_acc")
}
\arguments{
\item{df}{A data frame.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}
}
\description{
Helper: maps peptides to possible proteins
}
