% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{map_to_entrez_os_name}
\alias{map_to_entrez_os_name}
\title{Map \code{Uniprot} or \code{Refseq} to \code{entrez} (not currently used).}
\usage{
map_to_entrez_os_name(
  species = "human",
  abbr_species = NULL,
  os_name = "Homo sapiens",
  from = "UNIPROT",
  filename = NULL,
  db_path = "~/proteoQ/dbs/entrez",
  overwrite = FALSE
)
}
\arguments{
\item{species}{Character string; the name of a species.}

\item{abbr_species}{Two-letter character string; the abbreviated name of
 species used with
 \href{https://bioconductor.org/packages/release/data/annotation/html/org.Hs.eg.db.html}{org.Xx.eg.db}.
 The value of \code{abbr_species} will be determined automatically if the
 species is in one of \code{c("human", "mouse", "rat")}. Otherwise, for
 example, users need to provide \code{abbr_species = Ce} for fetching the
 \code{org.Ce.eg.db} package in the name space of proteoQ.

 For analysis against
 \href{http://current.geneontology.org/products/pages/downloads.html}{gene
 ontology} and \href{https://www.gsea-msigdb.org/gsea/index.jsp}{Molecular
 Signatures}, the argument is further applied to differentiate the same
 biological terms under different species; e.g., \code{GO~0072686 mitotic
 spindle} becomes \code{hs_GO~0072686 mitotic spindle} for human and
 \code{mm_GO~0072686 mitotic spindle} for mouse.}

\item{os_name}{An organism name by UniProt.}

\item{from}{the type of \code{keys}}

\item{filename}{Character string; An output file name. At the \code{NULL}
default, the name will be determined automatically at a given
\code{species}; i.e., \code{msig_hs.rds} for \code{human} data. The file is
saved as a \code{.rds} object for uses with \code{\link{prnGSPA}}.}

\item{db_path}{Character string; the local path for database(s). The default
is \code{"~/proteoQ/dbs/entrez"}.}

\item{overwrite}{Logical; if TRUE, overwrite the downloaded database(s). The
default is FALSE.}
}
\description{
Map \code{Uniprot} or \code{Refseq} to \code{entrez} (not currently used).
}
