% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mart_mm}
\alias{mart_mm}
\title{BioMart-Ensembl for mouse proteins.}
\format{
An object of class \code{Mart} of length 1.
}
\source{
\url{https://bioconductor.org/packages/release/bioc/html/biomaRt.html}
}
\usage{
mart_mm
}
\description{
A S4 object from \link[biomaRt]{useMart} at "dataset = mmusculus_gene_ensembl".
}
\keyword{datasets}
