% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulData.R
\name{mascot_refseq2uniprot}
\alias{mascot_refseq2uniprot}
\title{Helper function}
\usage{
mascot_refseq2uniprot(dat_dir)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Helper function
}
