% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_call_arg}
\alias{match_call_arg}
\title{Matches the arg to anal_prnGSPA}
\usage{
match_call_arg(call_rda = "foo", arg = "scale_log2r")
}
\arguments{
\item{call_rda}{the name of a rda}

\item{arg}{Argument to be matched.}
}
\description{
Matches the arg to anal_prnGSPA
}
