% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{match_calltime}
\alias{match_calltime}
\title{Finds the time stamp of a matched call from cached results.}
\usage{
match_calltime(
  path = "~/proteoQ/.MSearch/Cache/Calls",
  fun = "calc_pepmasses",
  nms = c("parallel", "out_path"),
  type = c(TRUE, FALSE)
)
}
\arguments{
\item{path}{A (system) file path.}

\item{fun}{The name of function being saved.}

\item{nms}{Names of arguments to be included in or excluded from matching.}

\item{type}{Logical; if TRUE, includes all arguments in \code{nms}. At FALSE,
excludes all \code{nms}.}
}
\value{
An empty object if no matches.
}
\description{
Finds the time stamp of a matched call from cached results.
}
