% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_expt}
\alias{match_expt}
\title{Matches the file name for expt_smry}
\usage{
match_expt(fn_pars = "load_expts.txt")
}
\arguments{
\item{fn_pars.}{}
}
\description{
Matches the file name for expt_smry
}
