% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_fasta}
\alias{match_fasta}
\title{Matches the name of fasta files to those in normPSM}
\usage{
match_fasta()
}
\arguments{
\item{fasta.}{}
}
\description{
Matches the name of fasta files to those in normPSM
}
