% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_fmls}
\alias{match_fmls}
\title{Matches formulas to those in calls to pepSig or prnSig}
\usage{
match_fmls(formulas)
}
\description{
Matches formulas to those in calls to pepSig or prnSig
}
