% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_frac}
\alias{match_frac}
\title{Matches the file name for frac_smry}
\usage{
match_frac(fn_pars = "load_expts.txt")
}
\arguments{
\item{fn_pars.}{}
}
\description{
Matches the file name for frac_smry
}
