% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_gset_nm}
\alias{match_gset_nm}
\title{Matches the current gene set names to those used in prnGSPA()}
\usage{
match_gset_nm(id = c("go_sets", "kegg_sets", "c2_msig"))
}
\arguments{
\item{id.}{}
}
\description{
Matches the current gene set names to those used in prnGSPA()
}
