% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_identifier}
\alias{match_identifier}
\title{Match the current id to the id in normPep or normPrn}
\usage{
match_identifier(id = c("pep_seq", "pep_seq_mod", "prot_acc", "gene"))
}
\arguments{
\item{id.}{}
}
\description{
Match the current id to the id in normPep or normPrn
}
\examples{
imputeNA(id = gene, m = 3, maxit = 3)

\dontrun{
}
}
