% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_logi_gv}
\alias{match_logi_gv}
\title{Match to a global logical variable}
\usage{
match_logi_gv(var, val)
}
\arguments{
\item{var}{Character string representation of a variable.}

\item{val}{The value of \code{var} before matching.}
}
\description{
Match to a global logical variable
}
\examples{
\donttest{
foo <- function(scale_log2r = FALSE) {
  match_logi_gv("scale_log2r", scale_log2r)
}

scale_log2r <- TRUE
foo()
}
}
