% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_normPep_id}
\alias{match_normPep_id}
\title{Matches the current id to the id in normPep}
\usage{
match_normPep_id(id = c("gene"))
}
\arguments{
\item{id.}{}
}
\description{
Matches the current id to the id in normPep
}
