% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_normPSM_id}
\alias{match_normPSM_id}
\title{Matches the current id to the id in normPSM}
\usage{
match_normPSM_id(id = c("pep_seq"))
}
\arguments{
\item{id.}{}
}
\description{
Matches the current id to the id in normPSM
}
