% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_normPSM_par}
\alias{match_normPSM_par}
\title{Find the setting of current par in normPSM}
\usage{
match_normPSM_par(par = "use_lowercase_aa")
}
\arguments{
\item{par}{}
}
\description{
Find the setting of current par in normPSM
}
