% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_normPSM_pepid}
\alias{match_normPSM_pepid}
\title{Matches the current pep_id to the pep_id in normPSM}
\usage{
match_normPSM_pepid(id = c("pep_seq"))
}
\arguments{
\item{id.}{}
}
\description{
Matches the current pep_id to the pep_id in normPSM
}
