% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_normPSM_protid}
\alias{match_normPSM_protid}
\title{Matches the current prot_id to the prot_id in normPSM}
\usage{
match_normPSM_protid(id = c("prot_acc"))
}
\arguments{
\item{id.}{}
}
\description{
Matches the current prot_id to the prot_id in normPSM
}
