% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_pepSig_scale_log2r}
\alias{match_pepSig_scale_log2r}
\title{Match scale_log2r}
\usage{
match_pepSig_scale_log2r(scale_log2r = TRUE, impute_na = FALSE)
}
\description{
\code{match_pepSig_scale_log2r} matches the value of \code{scale_log2r} to the value
in the most recent pepSig at a given impute_na status
}
