% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{match_secions}
\alias{match_secions}
\title{Matches against secondary ions}
\usage{
match_secions(theos, expts, type_ms2ions = "by", ppm_ms2 = 25, digits = 5)
}
\arguments{
\item{theos}{A list of theoretical values.}

\item{expts}{A list of experimental values}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Matches against secondary ions
}
