% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{match_valexpr}
\alias{match_valexpr}
\title{Matches the value of a function argument.}
\usage{
match_valexpr(f = NULL, arg = NULL, val = NULL, default = NULL)
}
\arguments{
\item{f}{The function where one of its argument value will be matched.}

\item{arg}{The argument where its value will be matched.}

\item{val}{The value to be matched.}

\item{default}{The default value of \code{arg}.}
}
\description{
Matches the value of a function argument.
}
\examples{
\donttest{
foo <- function (quant = c("none", "tmt6", "tmt10", "tmt11", "tmt16")) {
 val <- rlang::enexpr(quant)
 val <- match_valexpr(f = !!match.call()[[1]], arg = quant, val = !!val)
}

default <- foo()
custom <- foo(quant = tmt6)
}
}
