% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{matchMS}
\alias{matchMS}
\title{Searches MS ions.}
\usage{
matchMS(
  mgf_path = "~/proteoQ/mgfs",
  out_path = "~/proteoQ/outs",
  fasta = "~/proteoQ/dbs/fasta/uniprot/uniprot_hs_2020_05.fasta",
  fixedmods = c("TMT6plex (K)", "Carbamidomethyl (C)"),
  varmods = c("TMT6plex (N-term)", "Acetyl (Protein N-term)", "Oxidation (M)",
    "Deamidated (N)", "Gln->pyro-Glu (N-term = Q)"),
  include_insource_nl = FALSE,
  enzyme = c("trypsin"),
  maxn_fasta_seqs = 50000,
  maxn_vmods_setscombi = 64,
  maxn_vmods_per_pep = 5,
  maxn_sites_per_vmod = 3,
  maxn_vmods_sitescombi_per_pep = 32,
  min_len = 7,
  max_len = 100,
  max_miss = 2,
  type_ms2ions = "by",
  topn_ms2ions = 100,
  minn_ms2 = 6,
  ppm_ms1 = 20,
  ppm_ms2 = 25,
  pep_fdr = 0.01,
  digits = 5
)
}
\arguments{
\item{mgf_path}{The file path to a list of MGF files.}

\item{out_path}{The file name with prepending path for outputs.}

\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search. There is no default and users need to provide the
correct file path(s) and name(s).}

\item{fixedmods}{A character vector of fixed modifications. See also
\link{parse_unimod} for grammars.}

\item{varmods}{A character vector of variable modifications.}

\item{include_insource_nl}{Logical Logical; if TRUE, includes MS1 precursor
  masses with the losses of neutral species prior to MS2 fragmentation. The
  default is FALSE.

  Note that there is no combination of neutral losses for fixed modifications
  at the precursor levels. Changes from fixed to variable modifications for
  complete combinations.}

\item{enzyme}{A character string; the proteolytic specificity of the assumed
enzyme will be used to generate peptide sequences from proteins. The enzyme
is currently \code{trypsin}.}

\item{maxn_fasta_seqs}{Integer; the maximum number of protein sequences in
fasta files.}

\item{maxn_vmods_setscombi}{Integer; the maximum number of combinatorial variable
modifications and neutral losses.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{maxn_vmods_sitescombi_per_pep}{Integer; the maximum number of
combinatorial variable modifications per peptide sequence.}

\item{min_len}{Integer; the minimum length of peptides. Shorter peptides will
be excluded.}

\item{max_len}{Integer; the maximum length of peptides. Longer peptides will
be excluded.}

\item{max_miss}{The maximum number of mis-cleavages per peptide sequence.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz".}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches.}

\item{minn_ms2}{Integer; the minimum number of MS2 ions for consideration as
a hit.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{pep_fdr}{Numeric; the levels of peptide FDR.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Searches MS ions.
}
