% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-simulData.R
\name{maxquant_refseq2uniprot}
\alias{maxquant_refseq2uniprot}
\title{Helper function}
\usage{
maxquant_refseq2uniprot(dat_dir)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Helper function
}
