% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{mcalc_monopep}
\alias{mcalc_monopep}
\title{Calculates multiply the molecular weight of a polypeptides ([MH]+).}
\usage{
mcalc_monopep(aa_seqs, aa_masses, aa_masses_ref, digits = 5)
}
\arguments{
\item{aa_seqs}{Character string; a vector of peptide sequences with
one-letter representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
The calculations iterate through protein accessions and peptide sequences.
}
\examples{
\dontrun{
data(package = "proteoQ", aa_residues)
aa_masses <- aa_residues \%>\%
  dplyr::select(c("one_letter", "monoisotopic_da"))
aa_masses <- aa_masses$monoisotopic_da \%>\% `names<-`(aa_masses$one_letter)

x <- mcalc_monopep(list(protein_a = c("AAIDWFDGKEFSGNPIK", "MAAIDWFDGKEFSGNPIK"), 
                        protein_b = c("AAIDWFDGKEFSGNPIK")), 
                   aa_masses)
                   
aa_masses_all <- ext_aamasses()
aa_masses_ref <- aa_masses_all[[1]]

x <- mcalc_monopep(list(protein_a = c("AAIMDWFDMGKEFSGNPIK", "MAAIDWFDGKEFSGNPIK"), 
                        protein_b = c("AAIDWFDGKEFSGNPIK")), 
                   aa_masses_all[[11]], aa_masses_ref)
}
}
