% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{mcalcms2_a1_vnl0}
\alias{mcalcms2_a1_vnl0}
\title{Helper: masses of a series of MS2 ions.}
\usage{
mcalcms2_a1_vnl0(
  vmods_combi,
  ntmod,
  ctmod,
  aas,
  aa_masses,
  type_ms2ions,
  digits = 5
)
}
\arguments{
\item{vmods_combi}{Lists of variable modifications.}

\item{ntmod}{The parsed \emph{variable} \code{N-term} modifications from
\code{aa_masses}.}

\item{ctmod}{The parsed \emph{variable} \code{C-term} modifications from
\code{aa_masses}.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Updates \code{aas} (since `amods+`) and calculates masses for moduels (7)
"amods+ tmod- vnl- fnl-" and (8) "amods+ tmod+ vnl- fnl-".
}
