% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{mcalcpepmass_a1_fnl1}
\alias{mcalcpepmass_a1_fnl1}
\title{Helper: multiple mode of peptide-mass calculation.}
\usage{
mcalcpepmass_a1_fnl1(
  vmods_combi,
  nl_combi,
  amods,
  ntmod,
  ctmod,
  aas,
  aa_masses,
  digits
)
}
\arguments{
\item{vmods_combi}{Lists of variable modifications.}

\item{nl_combi}{Lists of combinations of neutral losses for corresponding
\code{vmods_combi}. Each list contains a table where each column
corresponds to a set of neutral loss. The first column corresponds to the
combination without NLs.}

\item{amods}{The parsed \emph{variable} \code{Anywhere} modifications from
\code{aa_masses}. The value can be \code{NULL}, e.g., at "amods- tmod- vnl-
fnl+".}

\item{ntmod}{The parsed \emph{variable} \code{N-term} modifications from
\code{aa_masses}.}

\item{ctmod}{The parsed \emph{variable} \code{C-term} modifications from
\code{aa_masses}.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Specific to "amods+" and "fnl+" (and vnl-).
}
\details{
\code{amods} is not empty; either \code{vnl} or \code{fnl} is not empty.
}
