% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{mdhyper}
\alias{mdhyper}
\title{The helper of \link[stats]{dhyper}.}
\usage{
mdhyper(x, m, n, k)
}
\arguments{
\item{x}{vector of quantiles representing the number of white balls
    drawn without replacement from an urn which contains both black and
    white balls.}

\item{m}{the number of white balls in the urn.}

\item{n}{the number of black balls in the urn.}

\item{k}{the number of balls drawn from the urn, hence must be in
    \eqn{0,1,\dots, m+n}.}
}
\description{
The helper of \link[stats]{dhyper}.
}
