% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{med_summarise_keys}
\alias{med_summarise_keys}
\title{Summary of peptide keys by mean or geomean}
\usage{
med_summarise_keys(df, ids)
}
\arguments{
\item{df}{A data frame.}

\item{ids}{A vector of column keys being the identifier of the level of
uniqueness. For example, it may be "pep_seq_mod" for non-SILAC or
c("pep_seq_mod", "pep_group") for SILAC with the conversion from PSMs to
peptides.}
}
\description{
Summary of peptide keys by mean or geomean
}
