% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlfq.R
\name{mergeSTY}
\alias{mergeSTY}
\title{Make MS1 S, T and Y matrices across adjacent scans}
\usage{
mergeSTY(
  xmat = NULL,
  ymat = NULL,
  tmat = NULL,
  tbars = NULL,
  rt_tol = 30,
  spc,
  are_refs,
  are_smpls
)
}
\arguments{
\item{xmat}{A matrix of X (retention time) values.}

\item{ymat}{A matrix of Y (intensity) values corresponding to \code{xs}.}

\item{tmat}{A matrix of retention-time values corresponding to \code{xs}.}

\item{tbars}{The mean of retention times.}

\item{rt_tol}{Retention time tolerance.}

\item{spc}{A species center(s).}

\item{are_refs}{Logical; are references or not.}

\item{are_smpls}{Logical; are samples or not.}
}
\description{
For adjusted step size.
}
