% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mfixBrukerMGF}
\alias{mfixBrukerMGF}
\title{Batch-fixing of Bruker MGF files.}
\usage{
mfixBrukerMGF(filepath, n_cores = 1L)
}
\arguments{
\item{filepath}{A file path to MGF.}

\item{n_cores}{The number of CPU cores.}
}
\description{
The first tilde in the \code{Title} lines needs to before \code{File}.
}
