% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{mmbr_psms}
\alias{mmbr_psms}
\title{Helper of \link{smbr_psms}.}
\usage{
mmbr_psms(dfqm, dfc, cols_mbr, cols_dfq, mbr_ret_tol = 60)
}
\arguments{
\item{dfqm}{psmQ subset with missing intensity values for MBR.}

\item{dfc}{Data from psmC.}

\item{cols_mbr}{Columns for MBR information borrowing.}

\item{cols_dfq}{Columns for copying from dfq.}

\item{mbr_ret_tol}{Retention time tolerance (in seconds) for LFQ-MBR.}
}
\description{
Batch processing of \link{smbr_psms}.
}
