% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-sigtests.R
\name{model_onechannel}
\alias{model_onechannel}
\title{Factorial model formula for interaction terms. All factors under one channel
in `label_scheme_sub`}
\usage{
model_onechannel(df, id, formula, label_scheme_sub, complete_cases, method,
  var_cutoff, pval_cutoff, logFC_cutoff, ...)
}
\description{
Factorial model formula for interaction terms. All factors under one channel
in `label_scheme_sub`
}
