% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mprocBrukerMGF}
\alias{mprocBrukerMGF}
\title{Batch-reprocessing of Bruker MGF files.}
\usage{
mprocBrukerMGF(filepath, n_cores = 1L)
}
\arguments{
\item{filepath}{A file path to MGF.}

\item{n_cores}{The number of CPU cores.}
}
\description{
Batch-reprocessing of Bruker MGF files.
}
