% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mq_mods}
\alias{mq_mods}
\title{A list of MaxQuant modifications.}
\format{
A data frame with 549 rows and 4 variables:
\describe{
  \item{title}{Modification}
  \item{position}{Position of a modification}
  \item{composition}{Composition of elements under a modification}
  \item{mass}{Monoisotopic mass of a modification}
}
}
\usage{
mq_mods
}
\description{
The dataset is based on "modificatins.xml" in MaxQuant.
}
\keyword{datasets}
