% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{ms1_a0_fnl1_byprot}
\alias{ms1_a0_fnl1_byprot}
\title{Helper of \link{calcms1mass_noterm}.}
\usage{
ms1_a0_fnl1_byprot(prot_peps, fnl_combi, aa_masses, digits = 4L)
}
\arguments{
\item{prot_peps}{Lists of named peptide masses under a protein.}

\item{fnl_combi}{A data.frame of combinations of neutral losses for fixed
modifications. Each row corresponds to a set of neutral loss. The first row
corresponds to the combination without NLs (all zeros).}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
By individual proteins.
}
