% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{ms1masses_bare}
\alias{ms1masses_bare}
\title{Calculates mono-isotopic masses of peptide sequences.}
\usage{
ms1masses_bare(
  seqs,
  aa_masses,
  ftmass = NULL,
  max_miss = 2L,
  min_len = 7L,
  max_len = 100L,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  is_fixed_protnt = FALSE,
  is_fixed_protct = FALSE,
  parallel = TRUE,
  digits = 4L
)
}
\arguments{
\item{seqs}{Sequences of peptides from FASTAs by protein accessions. Each
list contains two lists of sequences: (1) without and (2) with N-terminal
methionine.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{ftmass}{The sum of masses of \code{fixed} N-term and C-term
modifications.}

\item{max_miss}{The maximum number of mis-cleavages per peptide sequence.}

\item{min_len}{Integer; the minimum length of peptides. Shorter peptides will
be excluded.}

\item{max_len}{Integer; the maximum length of peptides. Longer peptides will
be excluded.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{is_fixed_protnt}{Logical; is protein N-terminal modification fixed.}

\item{is_fixed_protct}{Logical; is protein C-terminal modification fixed.}

\item{parallel}{Logical; if TRUE, performs parallel computation. The default
is TRUE.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Calculates mono-isotopic masses of peptide sequences.
}
