% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{ms1masses_noterm}
\alias{ms1masses_noterm}
\title{Helper of \link{ms1masses_bare}.}
\usage{
ms1masses_noterm(
  aa_seqs,
  aa_masses,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  parallel = TRUE,
  digits = 4L
)
}
\arguments{
\item{aa_seqs}{Character string; a vector of peptide sequences with
one-letter representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{parallel}{Logical; if TRUE, performs parallel computation. The default
is TRUE.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
For either forward or reversed sequences.
}
