% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ion_ladder.R
\name{ms2ions_by_type}
\alias{ms2ions_by_type}
\title{Helper: switches among ion types for calculating MS2 masses.}
\usage{
ms2ions_by_type(aas2, ntmass, ctmass, type_ms2ions, digits)
}
\arguments{
\item{aas2}{A sequence of amino-acid residues with \emph{masses}. Residues
are in names and masses in values (note that argument \code{aas}
corresponds to residues without masses).}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Helper: switches among ion types for calculating MS2 masses.
}
