% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{ms2ions_cols_fn11}
\alias{ms2ions_cols_fn11}
\title{Helper in calculating MS2 masses with "fnl+".}
\usage{
ms2ions_cols_fn11(aas, nl_combi, aa_masses, ntmod, ctmod, type_ms2ions, digits)
}
\arguments{
\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{nl_combi}{Lists of combinations of neutral losses for corresponding
\code{vmods_combi}. Each list contains a table where each column
corresponds to a set of neutral loss. The first column corresponds to the
combination without NLs.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{ntmod}{The parsed \emph{variable} \code{N-term} modifications from
\code{aa_masses}.}

\item{ctmod}{The parsed \emph{variable} \code{C-term} modifications from
\code{aa_masses}.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\value{
A numeric vector
}
\description{
For \code{fnl+} only (modules 5, 6, 11, 12).
}
\details{
The calculation runs column-wisely through the masses in \code{nl_combi}
(fixed). Note that the names of residues (\code{aas2}) have already been
updated before the utility was called.
}
