% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplots.R
\name{my_custom_cor}
\alias{my_custom_cor}
\title{Custom correlation plots}
\usage{
my_custom_cor(
  data,
  mapping,
  color = I("grey50"),
  sizeRange = c(1, 4),
  cor_method = "pearson",
  ...
)
}
\arguments{
\item{data}{A data frame}

\item{mapping}{A mapping aesthetics.}

\item{color}{A color.}

\item{sizeRange}{A range of sizes.}

\item{cor_method}{A correlation method.}
}
\description{
Custom correlation plots
}
\examples{
\donttest{
my_custom_cor(data, aes(x = col_nm_1, y = col_nm_1))
}
}
