% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{my_dist}
\alias{my_dist}
\title{A wrapper of \link[stats]{dist} with the handling of partial argument
matches.}
\usage{
my_dist(...)
}
\arguments{
\item{...}{Arguments for \link[stats]{dist}}
}
\description{
A wrapper of \link[stats]{dist} with the handling of partial argument
matches.
}
