% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-informatics.R
\name{my_pheatmap}
\alias{my_pheatmap}
\title{Wrapper around pheatmap}
\usage{
my_pheatmap(mat, filename, annotation_col, color, annotation_colors,
  breaks, ...)
}
\description{
Wrapper around pheatmap
}
