% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_split1}
\alias{my_split1}
\title{Split with NA values}
\usage{
my_split1(x, f, ty = "trans_matrix", replace_na_by = ".zzz")
}
\arguments{
\item{x}{Data for splitting.}

\item{f}{The factor to split against.}

\item{ty}{The type of data.}

\item{replace_na_by}{A character vector to replace NA values.}
}
\description{
Split with NA values
}
