% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_split2}
\alias{my_split2}
\title{Split with NA values}
\usage{
my_split2(x, f, ty = "trans_matrix")
}
\arguments{
\item{x}{Data for splitting.}

\item{f}{The factor to split against.}

\item{ty}{The type of data.}
}
\description{
Split with NA values
}
