% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{my_sum_n}
\alias{my_sum_n}
\title{Sums top-n.}
\usage{
my_sum_n(x, n = 3, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{An positive integer.}

\item{...}{Additional arguments for \code{sum}.}
}
\description{
Sums top-n.
}
