% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{my_tolower}
\alias{my_tolower}
\title{To lower cases}
\usage{
my_tolower(x = "", ch = "^")
}
\arguments{
\item{x}{A character string of amino acid sequence.}

\item{ch}{A tag before the letter to conversion to its lower case.}
}
\description{
To lower cases
}
