% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_union}
\alias{my_union}
\title{The union of named list}
\usage{
my_union(x, y)
}
\arguments{
\item{x}{A list of values.}

\item{y}{Another list of values.}
}
\description{
names will be kept after the unification
}
