% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{my_upper}
\alias{my_upper}
\title{To upper cases}
\usage{
my_upper(x = "", ch = "@")
}
\arguments{
\item{x}{A character string of amino acid sequence.}

\item{ch}{A tag before the letter to conversion to its upper case.}
}
\description{
To upper cases
}
