% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{my_which_max}
\alias{my_which_max}
\title{Finds the x at max density.}
\usage{
my_which_max(params, label_scheme)
}
\arguments{
\item{params}{Parameters from multiple Gaussians.}

\item{label_scheme}{The metadata of label scheme.}
}
\description{
Finds the x at max density.
}
