% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{n_LCMS}
\alias{n_LCMS}
\title{Finds the maximum number of LCMS injections under the same TMT_Set.}
\usage{
n_LCMS(label_scheme_full)
}
\arguments{
\item{label_scheme_full}{The label_scheme with the probable inclusion of
different LCMS_inj under the same TMT_Set.}
}
\description{
Finds the maximum number of LCMS injections under the same TMT_Set.
}
