% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{n_TMT_sets}
\alias{n_TMT_sets}
\title{The number of multiplex TMT experiments.}
\usage{
n_TMT_sets(label_scheme_full)
}
\description{
\code{n_TMT_sets} returns the number of multiplex TMT experiments.
}
