% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_fasta_name_by_prot_acc}
\alias{na_fasta_name_by_prot_acc}
\title{Helper of \link{annotPrn}.}
\usage{
na_fasta_name_by_prot_acc(df)
}
\arguments{
\item{df}{A data frame.}
}
\description{
Helper of \link{annotPrn}.
}
