% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_genes_by_acc}
\alias{na_genes_by_acc}
\title{Helper of \link{parse_fasta}.}
\usage{
na_genes_by_acc(df)

na_genes_by_acc(df)
}
\arguments{
\item{df}{A data frame.}

\item{acc_lookup}{The lookup of protein accession.}
}
\description{
Coerces NA values in genes by the value of accessions.
}
