% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{na_single_lfq}
\alias{na_single_lfq}
\title{Trivializes data rows with single LFQ intensity}
\usage{
na_single_lfq(df, pattern = "^I000 ")
}
\arguments{
\item{df}{A data frame.}

\item{pattern}{The pattern of intensity fields.}
}
\description{
Trivializes data rows with single LFQ intensity
}
