% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_species_by_org}
\alias{na_species_by_org}
\title{Helper of \link{parse_fasta}.}
\usage{
na_species_by_org(acc_lookup)
}
\arguments{
\item{acc_type}{The type of protein accession.}
}
\description{
Coerces NA values in species by the value of organism.
}
