% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{na_zeroIntensity}
\alias{na_zeroIntensity}
\title{replace zero intensity with NA}
\usage{
na_zeroIntensity(df)
}
\arguments{
\item{df}{A list of file names.}
}
\description{
\code{reorder_files} replace zero intensity with NA to avoid -Inf in log10 transformation.
}
