% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{names_pos<-}
\alias{names_pos<-}
\title{Prefix form of colnames(x)[c(2, 5, ...)] for use in pipes}
\usage{
names_pos(x, pos) <- value
}
\arguments{
\item{x}{A data frame.}

\item{pos}{Numeric.  The index of coloumns for name change.}

\item{value}{Characters.  The new column names.}
}
\value{
The data frame with new names: 
\code{refChannels}, reference channels(s); 
\code{emptyChannels}, empty channel(s) that were not used for sample labeling; 
\code{labeledChannels}, non-empty channels including reference(s) and sample(s).
}
\description{
\code{names_pos<-} rename the columns at indeces \code{pos}.
}
\details{
In the "label_scheme.csv" file,
}
\examples{
label_scheme <- readRDS("c:\\\\Results\\\\R\\\\proteoQ\\\\Data\\\\label_scheme_orig.rds")
channelInfo(label_scheme, 2)

\dontrun{
channelInfo(label_scheme, "a")
channelInfo(label_scheme, c(1, 2))
}
}
