% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{newColnames}
\alias{newColnames}
\title{Make new column names}
\usage{
newColnames(i, x, label_scheme, pattern = "[RI][0-9]{3}[NC]{0,1}")
}
\arguments{
\item{i}{Integer; the index of TMT experiment}

\item{x}{Data frame; log2FC data}

\item{label_scheme}{Experiment summary}
}
\description{
\code{newColnames} match names to Sample_ID in label_scheme
}
