% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-nmf.R
\name{nmfTest}
\alias{nmfTest}
\title{NMF analysis}
\usage{
nmfTest(df, id, r, nrun, col_group, label_scheme_sub, anal_type,
  scale_log2r, filepath, filename, complete_cases, ...)
}
\description{
NMF analysis
}
