% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-peptable.R
\name{normPep}
\alias{normPep}
\title{Reports peptide results}
\usage{
normPep(id = c("pep_seq", "pep_seq_mod"), method_psm_pep = c("median",
  "mean", "weighted.mean", "top.3", "mqpep"),
  group_pep_by = c("prot_acc", "gene"), method_align = c("MC",
  "MGKernel"), range_log2r = c(10, 90), range_int = c(5, 95),
  n_comp = NULL, seed = NULL, annot_kinases = FALSE,
  col_refit = NULL, cache = TRUE, plot_log2FC_cv = TRUE, ...)
}
\arguments{
\item{id}{Depreciated: character string; the variable for summarisation of
\code{PSMs} into peptides. The option \code{id = pep_seq} corresponds to the
summarisation by the primary sequences of peptides. The option \code{id =
pep_seq_mod} corresponds to the summarisation by both the primary sequences
and variable modifications of peptides. \code{PSMs} data with the same value
in \code{pep_seq} or \code{pep_seq_mod} will be summarised into a single
entry of peptide. NB: the value of \code{id} will match automatically to the
value of \code{group_psm_by} in \code{normPSM}.}

\item{method_psm_pep}{Character string; the method to summarise the
\code{log2FC} and the \code{intensity} of \code{PSMs} by peptide entries.
The descriptive statistics includes \code{c("mean", "median", "top.3",
"weighted.mean")}. The \code{log10-intensity} of reporter ions at the
\code{PSMs} levels will be the weight when summarising \code{log2FC} with
\code{"top.3"} or \code{"weighted.mean"}.}

\item{group_pep_by}{Depreciated: a character string for the grouping of
peptide entries. At the \code{prot_acc} default, descriptive statistics will
be calculated based on the same \code{prot_acc} groups. At
\code{group_pep_by = gene}, proteins with the same gene name but different
accession numbers will be treated as one group. NB: the value of
\code{group_pep_by} will match automatically to the value of
\code{group_pep_by} in \code{normPSM}.}

\item{method_align}{Character string or a list of gene symbols; the method to
align the \code{log2FC} of peptide/protein entries across samples.
\code{MC}: median-centering; \code{MGKernel}: the kernal density defined by
multiple Gaussian functions (\code{\link[mixtools]{normalmixEM}}). At
\code{method_align = "MC"}, the ratio profiles of each sample will be
aligned in that the medians of the \code{log2FC} are zero. At
\code{method_align = "MGKernel"}, the \code{log2FC} will be aligned in that
the maximums of kernel density are zero. It is also possible to align the
\code{log2FC} to the median of a list of user-supplied genes:
\code{method_align = c("ACTB", "GAPDH", ...)}.}

\item{range_log2r}{Numeric vector at length two; the range of the
\code{log2FC} of peptide/protein entries for use in the scaling
normalization of standard deviation across samples. The default is between
the 10th and the 90th quantiles.}

\item{range_int}{Numeric vector at length two; the range of the
\code{intensity} of reporter ions for use in the scaling normalization of
standard deviation across samples. The default is between the 5th and the
95th quantiles.}

\item{n_comp}{Integer; the number of Gaussian components to be used with
\code{method_align = "MGKernel"}. A typical value is 2 or 3. The variable
\code{n_comp} overwrites the augument \code{k} in
\code{\link[mixtools]{normalmixEM}}.}

\item{seed}{Integer; a seed for reproducible fitting at \code{method_align =
MGKernel}.}

\item{annot_kinases}{Logical; if TRUE, annotates kinase attributes of
proteins.}

\item{col_refit}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_refit} will be
used in the refit of \code{log2FC} using multiple Gaussian kernels. The
density estimates from an earlier analyis will be kept for the remaining
samples. At the \code{NULL} default, the column key of \code{Sample_ID} will
be used, which results in the refit of the \code{log2FC} for all samples.}

\item{cache}{Logical; if TRUE, use cache.}

\item{plot_log2FC_cv}{Logical; if TRUE, prepares the violin plots of the CV of
peptide \code{log2FC} for each sample.}

\item{...}{\code{filter_}: Logical expression(s) for the row filtration of
data; also see \code{\link{normPSM}}. \cr Additional parameters for
\code{\link[mixtools]{normalmixEM}}:\cr \code{maxit}, the maximum number of
iterations allowed; \cr \code{epsilon}, tolerance limit for declaring
algorithm convergence.}
}
\value{
The primary output is in \code{~\\dat_dir\\Peptide\\Peptide.txt}.
}
\description{
\code{normPep} summarises
\code{\href{https://www.ebi.ac.uk/pride/help/archive/search/tables}{PSMs}}
into peptides and normalizes the data across
\code{\href{https://en.wikipedia.org/wiki/Tandem_mass_tag}{TMT}} experiments
and \code{LC/MS} injections.
}
\details{
In the primary output file, "\code{Peptide.txt}", values under columns
\code{log2_R...} are logarithmic ratios at base 2 in relative to the
\code{reference(s)} within each multiplex TMT set, or to the row means if no
\code{reference(s)} are present. Values under columns \code{N_log2_R...} are
aligned \code{log2_R...} according to \code{method_align} without scaling
normalization. Values under columns \code{Z_log2_R...} are \code{N_log2_R...}
with additional scaling normalization. Values under columns \code{I...} are
\code{reporter-ion intensity} before normalization. Values under columns
\code{N_I...} are normalized \code{I...}. Values under columns
\code{sd_log2_R...} are the standard deviation of the \code{log2FC} of
proteins from ascribing peptides.

In general, median statistics is applied when summarising numeric values from
PSMs to peptides. One exception is \code{pep_expect} where geometric mean is
used.

Also see \code{\link{normPrn}} for more description of the column keys in the
output.

The peptide counts in individual peptide tables,
\code{TMTset1_LCMSinj1_Peptide_N.txt} et al., may be fewer than the entries
indicated under the \code{prot_n_pep} column after the peptide
removals/cleanups using \code{purgePSM}. Values under columns
\code{N_log2_R...} are intermediate reports by median-centering
\code{log2_R...} without scaling normalization.
}
\examples{
\dontrun{
# peptides results with exemplary `filter_...`
normPep(
  method_psm_pep = median,
  method_align = MGKernel,
  range_log2r = c(5, 95),
  range_int = c(5, 95),
  n_comp = 3,
  seed = 749662,
  maxit = 200,
  epsilon = 1e-05,

  # filter_by = exprs(pep_n_psm >= 2),
  # filter_by_sp = exprs(species == "human"),
)

# exemplary peptide purging; n: the number of peptides under a protein
purgePep(max_cv = .5, min_n = 2)

}
}
\seealso{
\code{\link{normPSM}} for PSM an \code{\link{normPrn}} for proteins.
}
