% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{normPep_Mplex}
\alias{normPep_Mplex}
\title{combined peptide reports across multiple TMT experiments}
\usage{
normPep_Mplex(
  group_psm_by = "pep_seq_mod",
  group_pep_by = "prot_acc",
  use_duppeps = TRUE,
  ...
)
}
\arguments{
\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{use_duppeps}{Logical; if TRUE, re-assigns double/multiple dipping
peptide sequences to the most likely proteins by majority votes.}

\item{...}{\code{filter_}: Variable argument statements for the filtration of
data rows. Each statement contains to a list of logical expression(s). The
\code{lhs} needs to start with \code{filter_}. The logical condition(s) at
the \code{rhs} needs to be enclosed in \code{exprs} with round parenthesis.
For example, \code{pep_expect} is a column key present in \code{Mascot} PSM
exports and \code{filter_psms_at = exprs(pep_expect <= 0.1)} will remove PSM
entries with \code{pep_expect > 0.1}.}
}
\description{
median summarization of data from the same TMT experiment at different LCMS injections
summed \code{pep_n_psm}, \code{prot_n_psm}, and \code{prot_n_pep} after data merging
no Z_log2_R yet available
  use \code{col_select = expr(Sample_ID)} not \code{col_select} to get all Z_log2_R
  why: users may specify \code{col_select} only partial to Sample_ID entries
}
