% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-peptable.R
\name{normPep_Splex}
\alias{normPep_Splex}
\title{Helper of calcPepide() for individual peptide reports}
\usage{
normPep_Splex(method_psm_pep = "median", cache = TRUE)
}
\arguments{
\item{method_psm_pep}{Character string; the method to summarise the
\code{log2FC} and the \code{intensity} of \code{PSMs} by peptide entries.
The descriptive statistics includes \code{c("mean", "median", "top.3",
"weighted.mean")}. The \code{log10-intensity} of reporter ions at the
\code{PSMs} levels will be the weight when summarising \code{log2FC} with
\code{"top.3"} or \code{"weighted.mean"}.}

\item{cache}{Logical; if TRUE, use cache.}
}
\value{
Results in \code{.txt} files for each of TMT experiments and LC/MS
  injections.
}
\description{
\code{normPep_Splex} prepares peptide data for each TMT experiment at one or
multiple LC/MS injections.
}
\examples{
	normPep_Splex(
		id = pep_seq_mod,
		method_psm_pep = median, 
		group_pep_by = prot_acc, 
	)

\dontrun{
}
}
