% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-prntable.R
\name{normPrn}
\alias{normPrn}
\title{Reports Protein Results}
\usage{
normPrn(id = c("prot_acc", "gene"), method_pep_prn = c("median",
  "mean", "weighted.mean", "top.3"), method_align = c("MC", "MGKernel"),
  range_log2r = c(20, 90), range_int = c(5, 95), n_comp = NULL,
  seed = NULL, fasta = NULL, ...)
}
\arguments{
\item{id}{The variable to summarise peptides into proteins. The option
 \code{"prot_acc"} corresponds to the summarisation by the accession numbers
 or entry names of proteins The option \code{"gene"} corresponds to the
 summarisation by the gene names of proteins.

 At \code{id = gene}, data under the same gene name but different acccesssion
 numbers or entry names will be summarised into one entry.}

\item{method_pep_prn}{The method to summarise the \code{log2-ratios} and the
 \code{intensity} of peptides by protein entries. The descriptive statistics
 includes \code{c("mean", "median", "top.3", "weighted.mean")}.

 The representative \code{log10-intensity} of reporter ions at the peptide
 levels (from \code{\link{normPep}}) will be the weigth when summarising
 log2-ratios with "top.3" or "weighted.mean".}

\item{method_align}{The method to align the \code{log2-ratios} of proteins
 entries across samples. \code{MC}: median-centering; \code{MGKernel}: the
 kernal density defined by multiple Gaussian functions
 (\code{\link[mixtools]{normalmixEM}}).

 At \code{method_align = "MC"}, the ratio profiles of each sample will be
 aligned in that the medians of the \code{log2-ratios} are zero. At
 \code{method_align = "MGKernel"}, the \code{log2-ratios} will be aligned in
 that the maximums of kernel density are zero. It is also possible to align
 the \code{log2-ratios} to the median of a list of user-supplied genes:
 \code{method_align = c("ACTB", "GAPDH", ...)}. At \code{method_align =
 NULL}, the protein summary by \code{method_pep_prn} will be taken without
 furthur normalization.}

\item{range_log2r}{The range of the \code{log2-ratios} of protein entries for
use in the scaling normalization of standard deviation across samples. The
default is between the 20th and the 90th quantiles.}

\item{range_int}{The range of the representative reporter-ion \code{intensity}
of proteins for use in the scaling normalization of standard deviation
across samples. The default is between the 5th and the 95th quantiles.}

\item{n_comp}{The number of Gaussian components to be used with
\code{method_align = "MGKernel"}. A typical value is 2 or 3. The variable
\code{n_comp} overwrites the augument \code{k} in
\code{\link[mixtools]{normalmixEM}}.}

\item{seed}{Integer; a seed for reproducible fitting at \code{method_align =
MGKernel}.}

\item{fasta}{The file name with prepended directory path to the fasta database
being used in MS/MS ion search. If not specified, a pre-compiled system file
will be used.}

\item{...}{Additional parameters from \code{\link[mixtools]{normalmixEM}}:\cr
\code{maxit}, the maximum number of iterations allowed; \cr \code{epsilon},
tolerance limit for declaring algorithm convergence.}
}
\value{
The primary output in "\code{C:\\my_directory\\Protein\\Protein.txt}".
}
\description{
\code{normPrn} summarises peptides into proteins and performs data
normalization.
}
\details{
The column keys are defined in
\code{\href{http://www.matrixscience.com/help/csv_headers.html}{Matrix
Science}} with the following additions:

\tabular{ll}{ \strong{Header}   \tab \strong{Descrption}\cr entrez \tab
\code{\href{https://www.ncbi.nlm.nih.gov/gene}{Protein Entrez ID}} \cr gene
\tab Protein gene name \cr kin_attr       \tab Optional; the attribute of
proteins being kinases \cr kin_class      \tab Optional; the classes of
kinases, e.g., TK, TKL... \cr kin_order      \tab Optional; the order of
"kin_class" using the \code{\href{http://kinase.com/human/kinome/}{kinase tree
diagram}} \cr n_pep          \tab The number of significant peptide sequences
matched to a proposed protein \cr n_psm \tab The number of significant MS/MS
queries matched to a peptide sequence \cr pep_seq_mod    \tab One-letter
representation of peptide sequences: amino acid residue(s) with variable
modifications are in the lower cases; the acetylations of protein N-terminals
indicated by '_' and the flanking residues on the N- or C-terminal side of
peptides separated by '.', e.g. "-._mAsGVAVSDGVIK.V"\cr prot_cover     \tab
Protein sequence coverage calculated from all available raw files   \cr
raw_file \tab MS file name(s) where peptides or proteins were identified \cr
refseq_acc \tab \code{\href{https://www.ncbi.nlm.nih.gov/refseq/}{Protein
RefSeq accession number}} \cr uniprot_acc    \tab
\code{\href{https://www.uniprot.org/help/accession_numbers}{Protein UniProt
accession number}} \cr uniprot_id     \tab
\code{\href{https://www.uniprot.org/help/entry_name}{Protein UniProt entry
name}} \cr I... (...)     \tab Reporter-ion intensity calculated from the
descriptive statistics in method_\code{psm_pep} for indicated samples \cr
N_I... (...)   \tab Normalized \code{I... (...)}. The calibration factors for
median centerring of \code{log2-ratios} are used to scale the reporter-ion
intensity \cr log2_R (...)   \tab \code{log2-ratios} relative to reference
materials for indicated samples \cr N_log2_R (...) \tab Aligned \code{log2_R
(...)} according to method_align without scaling normalization \cr Z_log2_R
(...) \tab \code{N_log2_R (...)} with scaling normalization \cr }
}
\examples{
	normPrn(
		id = gene, 
		method_pep_prn = median, 
		method_align = MGKernel, 
		range_log2r = c(5, 95), 
		range_int = c(5, 95), 
		n_comp = 2, 
		seed = 749662, 
		fasta = "C:\\\\Results\\\\DB\\\\Refseq\\\\RefSeq_HM_Frozen_20130727.fasta", 
		maxit = 200, 
		epsilon = 1e-05
	)
	
	normPrn(
		id = "prot_acc",
		fasta = "C:\\\\Results\\\\DB\\\\Refseq\\\\RefSeq_HM_Frozen_20130727.fasta"
	)

\dontrun{
}
}
\seealso{
\code{\link{normPSM}} for PSMs and \code{\link{normPep}} for
 peptides.
}
\concept{aggregate functions}
