% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{not_all_zero}
\alias{not_all_zero}
\title{Finds all-zero column(s)}
\usage{
not_all_zero(x)
}
\arguments{
\item{x}{A data frame of \code{log2-ratios} and \code{intensity}.}
}
\description{
\code{not_all_zero} identifies the column indeces with all NA values.
}
