% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ok_existing_params}
\alias{ok_existing_params}
\title{Checks the suitability of existing param files for MGKernel}
\usage{
ok_existing_params(filepath)
}
\arguments{
\item{filepath}{A file path to \code{"MGKernel_params_N.txt"}}
}
\description{
Checks the suitability of existing param files for MGKernel
}
