% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{ok_file_ncomp}
\alias{ok_file_ncomp}
\title{Compares to prior n_comp value.}
\usage{
ok_file_ncomp(filepath, filename, n_comp)
}
\arguments{
\item{filepath}{A file path.}

\item{filename}{A file name.}

\item{n_comp}{Integer; the number of Gaussian components to be used with
\code{method_align = MGKernel}. A typical value is 2 or 3. The variable
\code{n_comp} overwrites the argument \code{k} in
\code{\link[mixtools]{normalmixEM}}.}
}
\description{
Compares to prior n_comp value.
}
