% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{order_mascot_psm_cols}
\alias{order_mascot_psm_cols}
\title{Order Mascot PSM columns}
\usage{
order_mascot_psm_cols(df, psm_cols = NULL, rm_na_cols = FALSE)
}
\arguments{
\item{df}{A data frame of PSM.}

\item{psm_cols}{A character string of column names.}

\item{rm_na_cols}{Logical; if TRUE, remove columns of all NAs.}
}
\description{
Columns indicated by \code{psm_cols} will be moved to the front, and columns
of ratios and ratios and intensities to the end.
}
