% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{order_psm_cols}
\alias{order_psm_cols}
\title{Order PSM columns.}
\usage{
order_psm_cols(df, cols)
}
\arguments{
\item{df}{A PSM data frame.}

\item{cols}{A vector; the names of columns to be moved to the front of
\code{df}.}
}
\description{
Only for certain columns.
}
