% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{p_dixon}
\alias{p_dixon}
\title{Outlier removals with Dixon's method}
\usage{
p_dixon(q, n, type = 10)
}
\arguments{
\item{q}{Vector of quantiles.}

\item{n}{Length of sample Length of sample.}

\item{type}{The type of Dixon method. See also outliers::dixon.test.}
}
\description{
Outlier removals with Dixon's method
}
