% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{p_grubbs}
\alias{p_grubbs}
\title{Outlier removals with Grubbs' method}
\usage{
p_grubbs(q, n, type = 10)
}
\arguments{
\item{q}{Vector of quantiles.}

\item{n}{Length of sample Length of sample.}

\item{type}{The type of Dixon method. See also \link[outliers]{dixon.test}.}
}
\description{
Outlier removals with Grubbs' method
}
