% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{pad_grp_samples}
\alias{pad_grp_samples}
\title{Pads sample groups.}
\usage{
pad_grp_samples(
  df,
  sids,
  tmt_levels = NULL,
  type_levels = c("I", "N_I", "sd_log2_R", "log2_R", "N_log2_R")
)
}
\arguments{
\item{df}{A data frame.}

\item{sids}{The Sample_IDs from label_scheme.}

\item{tmt_levels}{The levels of TMT: 000, 126, 127 etc. without the
\code{TMT-} prefix.}

\item{type_levels}{The levels of five types.}
}
\description{
For uses with group searches (with non-trivial values under
\code{pep_group}). For example, heavy may be missing in complete under one
sample and light missing in another.
}
\examples{
\donttest{
# 10-plex
tmt_levels <- TMT_levels(10)
tmt_levels <- gsub("^TMT-", "", tmt_levels)

group_ids <- c("light", "heavy")

sids <- LETTERS[1:10]
len <- length(sids)
sids <- rep(sids, each = length(group_ids))

group_ids <- rep(group_ids, len)
sids <- paste0(sids, " [", group_ids, "]")

rm(list = c("len", "group_ids"))

# 1-plex
#' tmt_levels <- TMT_levels(1)
tmt_levels <- gsub("^TMT-", "", tmt_levels)

group_ids <- c("light", "heavy")

sids <- LETTERS[1]
len <- length(sids)
sids <- rep(sids, each = length(group_ids))

group_ids <- rep(group_ids, len)
sids <- paste0(sids, " [", group_ids, "]")

rm(list = c("len", "group_ids"))
}
}
