% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{pad_mascot_channels}
\alias{pad_mascot_channels}
\title{Pad Mascot TMT channels to the highest plex}
\usage{
pad_mascot_channels(file, ...)
}
\arguments{
\item{file}{An intermediate PSM table.}

\item{...}{filter_dots.}
}
\description{
Pad Mascot TMT channels to the highest plex
}
