% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{pad_psm_fields}
\alias{pad_psm_fields}
\title{Pads PSM exports.}
\usage{
pad_psm_fields(dfs = NULL)
}
\arguments{
\item{dfs}{Intermediate PSM tables.}
}
\description{
Intensity or ratio columns are handled during channel padding and thus
excluded.
}
