% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{pad_sm_channels}
\alias{pad_sm_channels}
\title{Pad Spectrum Mill TMT channels to the highest plex}
\usage{
pad_sm_channels(file, ...)
}
\arguments{
\item{file}{A file name of PSM table from MaxQuant export}
}
\description{
Pad Spectrum Mill TMT channels to the highest plex
}
