% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{pad_tmt_channels}
\alias{pad_tmt_channels}
\title{Pads TMT channels to the highest multiplex.}
\usage{
pad_tmt_channels(file = NULL, ...)
}
\arguments{
\item{file}{An intermediate PSM table.}

\item{...}{filter_dots.}
}
\description{
For mzion.
}
