% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{par_dist}
\alias{par_dist}
\title{Helper of \link{parDist}.}
\usage{
par_dist(cols, mat)
}
\arguments{
\item{cols}{The column indexes of \code{mat}.}

\item{mat}{A bool matrix.}
}
\description{
R version of parallel distance. Not currently used.
}
