% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{par_dist1}
\alias{par_dist1}
\title{Helper of \link{parDist}.}
\usage{
par_dist1(cols, mat)
}
\arguments{
\item{cols}{The column indexes of \code{mat}.}

\item{mat}{A bool matrix.}
}
\description{
R version of parallel distance. Uses a subset of mat for memory efficiency.
Not currently used.
}
\examples{
\donttest{
x <- par_dist1(idxes[[i]], mat[idxes[[i]][1]:len])
}
}
