% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{par_dist2}
\alias{par_dist2}
\title{Helper of \link{parDist}.}
\usage{
par_dist2(cols, mat)
}
\arguments{
\item{cols}{The column indexes of \code{mat}.}

\item{mat}{A bool matrix.}
}
\description{
For coupling to par_distC; not currently used.
}
