% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{parse_aamasses}
\alias{parse_aamasses}
\title{Parses \code{aa_masses}.}
\usage{
parse_aamasses(aa_masses)
}
\arguments{
\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}
}
\description{
Parses \code{aa_masses}.
}
