% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{parse_acc_temp}
\alias{parse_acc_temp}
\title{Determine the protein accession type from PSM tables}
\usage{
parse_acc_temp(df, prx)
}
\description{
Find the protein accession from a non-cRAP entry and parse it.
}
\details{
\code{parse_acc} parse the protein accession.
}
