% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_col_select}
\alias{parse_col_select}
\title{Parses \code{col_select}.}
\usage{
parse_col_select(col_select, label_scheme)
}
\arguments{
\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
At the \code{NULL} default, the column key of \code{Select} in
\code{expt_smry.xlsx} will be used. In the case of no samples being
specified under \code{Select}, the column key of \code{Sample_ID} will be
used. The non-empty entries under the ascribing column will be used in
indicated analysis.}

\item{label_scheme}{Experiment summary}
}
\description{
Parses \code{col_select}.
}
