% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_filename}
\alias{parse_filename}
\title{Parses file name.}
\usage{
parse_filename(filename, dat_dir, must_exists = FALSE)
}
\arguments{
\item{filename}{A file name.}

\item{dat_dir}{A working directory.}

\item{must_exists}{Logical; if TRUE, the file must be present.}
}
\description{
Parses file name.
}
