% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{parse_unimod}
\alias{parse_unimod}
\title{Parse the name of a Unimod}
\usage{
parse_unimod(unimod)
}
\arguments{
\item{unimod}{The name of a \href{https://www.unimod.org/}{Unimod} modification.}
}
\description{
The general format: \code{parse_unimod("title (position = site)")}.
}
\examples{
\donttest{
# "dot" for anywhere (either position or site)
x1 <- parse_unimod("Carbamidomethyl (. = C)")
x2 <- parse_unimod("Carbamidomethyl (Anywhere = C)")
x3 <- parse_unimod("Carbamidomethyl (C)")

identical(x1, x2); identical(x2, x3)

# Any residue on protein N-term 
x1 <- parse_unimod("Acetyl (Protein N-term = .)")
x2 <- parse_unimod("Acetyl (Protein N-term)")

identical(x1, x2)

# Any N-term residue
x1 <- parse_unimod("Acetyl (N-term)")
x2 <- parse_unimod("Acetyl (N-term = .)")

identical(x1, x2)

# N-term Q
x1 <- parse_unimod("Gln->pryo-Glu (N-term = Q)")
x2 <- parse_unimod("Gln->pryo-Glu (N-term Q)")

identical(x1, x2)

# ok with parenthesis in the 'title' 
x <- parse_unimod("Hex(5)HexNAc(2) (N)")

# ok with spaces in the 'title' 
x <- parse_unimod("Met-loss (Protein N-term = M)")
}

\dontrun{
# No modification of anywhere and anything 
# (to every position and site)
x <- parse_unimod("Carbamidomethyl (Anywhere = .)")
x <- parse_unimod("Carbamidomethyl")
x <- parse_unimod("Carbamidomethyl (. = .)")

# Prefer an "=" sign between 'N-term' and 'Q'
x <- parse_unimod("Gln->pyro-Glu (N-term Q)")
}
}
