% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_uniprot_fasta}
\alias{parse_uniprot_fasta}
\title{Parse UniProt FASTA for accession lookups}
\usage{
parse_uniprot_fasta(df, fasta, entrez)
}
\arguments{
\item{df}{An input data frame.}

\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search. There is no default and users need to provide the
correct file path(s) and name(s).}

\item{entrez}{Character string(s) to the name(s) of entrez file(s) with
prepended directory path. At the \code{NULL} default, a convenience lookup
is available for species among \code{c("human", "mouse", "rat")}. For other
species, users need to provide the file path(s) and name(s) for the lookup
table(s). See also \code{\link{Uni2Entrez}} and \code{\link{Ref2Entrez}}
for preparing custom entrez files.}
}
\description{
Parse UniProt FASTA for accession lookups
}
