% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{pep_mf_lfq}
\alias{pep_mf_lfq}
\title{Handling of MSFragger peptide table}
\usage{
pep_mf_lfq(
  label_scheme,
  omit_single_lfq,
  group_psm_by,
  use_lowercase_aa = FALSE
)
}
\arguments{
\item{label_scheme}{Experiment summary}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{use_lowercase_aa}{Logical; if TRUE, modifications in amino acid residues
will be abbreviated with lower-case and/or \code{^_~}. See the table below
for details. The default is TRUE.}
}
\description{
Fill back intensity values.
}
