% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{pep_mq_lfq}
\alias{pep_mq_lfq}
\title{Handling of MaxQuant peptide.txt}
\usage{
pep_mq_lfq(
  label_scheme,
  omit_single_lfq = FALSE,
  use_spec_counts = FALSE,
  use_maxlfq = TRUE
)
}
\arguments{
\item{label_scheme}{Experiment summary.}

\item{omit_single_lfq}{Omit entries with single LFQ value or not.}

\item{use_spec_counts}{Not yet used. Logical; use spectrum counts or not.}

\item{use_maxlfq}{Logical; use MaxLFQ values or not.}
}
\description{
Fill back intensity values that are not filled in LFQ msms.txt.
}
