% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-mds.R
\name{pepEucDist}
\alias{pepEucDist}
\title{Distance plots}
\usage{
pepEucDist(...)
}
\description{
Euclidean distance of peptide data
}
\examples{
pepEucDist(
  scale_log2r = TRUE,
  adjEucDist = FALSE,
  show_ids = FALSE,
  annot_cols = c("Peptide_Yield", "Group"),
  display_numbers = TRUE,
  number_color = "grey30",
  number_format = "\%.2f",
  clustering_distance_rows = "euclidean",
  clustering_distance_cols = "euclidean",
  fontsize = 16,
  fontsize_row = 20,
  fontsize_col = 20,
  fontsize_number = 8,
  cluster_rows = TRUE,
  show_rownames = TRUE,
  show_colnames = TRUE,
  border_color = "grey60",
  cellwidth = 24,
  cellheight = 24
)


\dontrun{
pepEucDist(
  col_color = "column_key_not_existed",
  col_shape = "also_a_missing_column_key",
  annot_cols = c("bad_column_key", "yet_another_bad_column_key")
)
}

}
\seealso{
\code{\link{proteoEucDist}} for parameters
}
