% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-histo.R
\name{pepHist}
\alias{pepHist}
\title{Histograms of Peptide Data}
\usage{
pepHist(...)
}
\description{
Histograms of Peptide Data
}
\examples{
# no scaling normalization
pepHist(
  scale_log2r = FALSE, 
  show_curves = TRUE,
  show_vline = TRUE,
  ncol = 4
)

# scaling normalization
pepHist(
  scale_log2r = TRUE, 
  show_curves = TRUE,
  show_vline = TRUE,
  ncol = 4
)

# direct call to proteoHist()
proteoHist(
  id = pep_seq_mod, 
  scale_log2r = TRUE, 
  show_curves = TRUE,
  show_vline = TRUE,
  ncol = 4
)


\dontrun{
pepHist(
  col_select = "a_non_existed_column_key"
)
}

}
\seealso{
\code{\link{proteoHist}} for parameters
}
