% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-mds.R
\name{pepMDS}
\alias{pepMDS}
\title{MDS plots}
\usage{
pepMDS(...)
}
\description{
MDS plots of peptide data
}
\examples{
pepMDS(
  scale_log2r = TRUE,
  col_color = Color,
  col_shape = Shape,
  show_ids = TRUE,
)

\dontrun{
pepMDS(
  col_color = "column_key_not_existed",
  col_shape = "also_a_missing_column_key"
)
}

}
\seealso{
\code{\link{proteoMDS}} for parameters
}
