% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-sigtests.R
\name{pepSig}
\alias{pepSig}
\title{Significance Tests of Peptide \code{log2-ratios}}
\usage{
pepSig(...)
}
\description{
Significance Tests of Peptide \code{log2-ratios}
}
\seealso{
\code{\link{proteoSigtest}} for parameters
}
