% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{peptideMBR}
\alias{peptideMBR}
\title{MBR of peptides}
\usage{
peptideMBR(
  base_name,
  ms1files,
  xvs,
  yvs,
  tvs,
  svs,
  mbr_mzs,
  mbr_ys,
  mbr_rets,
  mbr_peps,
  dat_dir,
  mbr_ret_tol = 25,
  min_y = 2e+06,
  yco = 100,
  step = 1e-05,
  max_mbr_fold = 20,
  gap = mbr_ret_tol + 35
)
}
\arguments{
\item{base_name}{The base name of a peptide table file
(\code{TMTset[i]_LCMSinj[j]_Peptide_N.txt}).}

\item{ms1files}{A list of file names of retention-time calibrated RAW MS1
data corresponding to the peptide table at \code{base_name} (More than one
file if with analyte pre-fractionation for a sample).}

\item{xvs}{A vector of X (m-over-z) values under the peptide table at the
\code{base_name}. The length is the number of peptides in the universe. NA
values correspond to peptides (\code{pep_seq_modz}) that are not present in
the current peptide table but found in others.}

\item{yvs}{A vector of apex intensities.}

\item{tvs}{A vector of apex retention times.}

\item{svs}{A vector of apex scan numbers.}

\item{mbr_mzs}{A vector of m-over-z values in the universe.}

\item{mbr_ys}{A vector of intensity values in the universe (sample average).}

\item{mbr_rets}{A vector of retention times in the universe.}

\item{mbr_peps}{A vector of \code{pep_seq_modz} sequences in the universe.}

\item{dat_dir}{The working data directory.}

\item{mbr_ret_tol}{The tolerance in MBR retention time in seconds.}

\item{min_y}{The cut-off of intensity values in MBR.}

\item{step}{The mass error in \code{ppm / 1e6}.}

\item{max_mbr_fold}{The maximum absolute fold change in MBR.}

\item{gap}{A gap in retention-time window in seconds for finding gates in the
Y values of LC. The gap value serves as a margin to trace out a whole peak
with its apex within the \code{mbr_ret_tol}.}
}
\description{
MBR of peptides
}
