% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{peptideMBR2}
\alias{peptideMBR2}
\title{MBR of peptides}
\usage{
peptideMBR2(
  ms1files,
  base_name,
  xvs,
  yvs,
  tvs,
  svs,
  mbr_mzs,
  mbr_ys,
  mbr_rets,
  mbr_peps,
  dat_dir,
  mbr_ret_tol = 25,
  max_mbr_fold = 20,
  step = 1e-05,
  gap = mbr_ret_tol + 45
)
}
\arguments{
\item{ms1files}{The names of MS1 data files corresponding to a peptide table,
e.g., \code{TMTset1_LCMSinj1_Peptide_N.txt}.}

\item{base_name}{The basename of a peptide table.}

\item{xvs}{A vector of moverzs.}

\item{yvs}{A vector of apex intensities.}

\item{tvs}{A vector of apex retention times.}

\item{svs}{A vector of apex scan numbers.}

\item{mbr_mzs}{Values of m-over-zs for MBR.}

\item{mbr_ys}{Intensity values for MBR.}

\item{mbr_rets}{Values of retention times for MBR.}

\item{mbr_peps}{Sequences of peptides for MBR.}

\item{dat_dir}{The working data directory.}

\item{mbr_ret_tol}{The tolerance in MBR retention time in seconds.}

\item{max_mbr_fold}{The maximum absolute fold change in MBR.}

\item{step}{The mass error in \code{ppm / 1e6}.}

\item{gap}{A gap in retention-time window in seconds for finding gates in the
Y values of LC. The gap value need to be broad enough.}
}
\description{
MBR of peptides
}
