% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplots.R
\name{plot_corr_sub}
\alias{plot_corr_sub}
\title{Make correlation plots}
\usage{
plot_corr_sub(
  df,
  cor_method = "pearson",
  xlab,
  ylab,
  filename,
  filepath,
  xmin,
  xmax,
  xbreaks,
  width,
  height,
  digits = 2L,
  ...
)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{filename}{A representative file name to outputs. By default, the
name(s) will be determined automatically. For text files, a typical file
extension is \code{.txt}. For image files, they are typically saved via
\code{\link[ggplot2]{ggsave}} or \code{\link[pheatmap]{pheatmap}} where the
image type will be determined by the extension of the file name.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{xmin}{minimal x.}

\item{xmax}{maximal x.}

\item{xbreaks}{breaks on x-axis.}

\item{width}{plot width}

\item{height}{plot height}

\item{...}{additional arguments for ggsave.}
}
\description{
Make correlation plots
}
