% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanos.R
\name{plot_gspa_volcano}
\alias{plot_gspa_volcano}
\title{Volcano plots of protein \code{log2FC} under given gene sets}
\usage{
plot_gspa_volcano(
  gt,
  gsets,
  gsea_res,
  dfw,
  id = "gene",
  id_gspa = "entrez",
  show_sig = "none",
  topn_labels = 20L,
  xco = 1.2,
  yco = 0.05,
  pval_cutoff = 0.05,
  logFC_cutoff = log2(1.2),
  x_label = NULL,
  y_label = NULL,
  theme = NULL,
  myPalette = c("#377EB8", "#E41A1C"),
  filepath,
  fml_nm,
  custom_prefix,
  fn_suffix,
  dots
)
}
\arguments{
\item{gsets}{The gene sets.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{id_gspa}{Character string; the ID type used in GSPA. The default is
\code{entrez}.}

\item{show_sig}{Character string indicating the type of significance values to
be shown with \code{\link{prnGSPAMap}}. The default is \code{"none"}.
Additional choices are from \code{c("pVal", "qVal")} where \code{pVal} or
\code{qVal} will be shown, respectively, in the facet grid of the plots.}

\item{topn_labels}{A non-negative integer; the top-n species for labeling in a
plot. At \code{topn_labels = 0}, no labels of proteins/peptides will be
shown. The default is to label the top-20 species with the lowest p-values.}

\item{pval_cutoff}{Numeric value or vector; the cut-off in protein
significance \code{pVal}. Entries with \code{pVals} less significant than
the threshold will be excluded from enrichment analysis. The default is 0.05
for all formulas matched to or specified in argument \code{fml_nms}.
Formula-specific threshold is allowed by supplying a vector of cut-off
values.}

\item{logFC_cutoff}{Numeric value or vector; the cut-off in protein
\code{log2FC}. Entries with absolute \code{log2FC} smaller than the
threshold will be excluded from enrichment analysis. The default magnitude
is \code{log2(1.2)} for all formulas matched to or specified in argument
\code{fml_nms}. Formula-specific threshold is allowed by supplying a vector
of absolute values in \code{log2FC}.}

\item{theme}{A
\href{https://ggplot2.tidyverse.org/reference/ggtheme.html}{ggplot2} theme,
i.e., theme_bw(), or a custom theme. At the NULL default, a system theme
will be applied.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{fml_nm}{A character string; the name of \code{fml}.}
}
\description{
Volcano plots of protein \code{log2FC} under given gene sets
}
