% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanos.R
\name{plot_ibaq}
\alias{plot_ibaq}
\title{Rank plot of protein iBAQ.}
\usage{
plot_ibaq(outname = NULL, highlights = NULL, ...)
}
\arguments{
\item{outname}{The output file name.}

\item{highlights}{The protein names to be highlighted.}

\item{...}{Additional arguments passed to \code{ggsave}.}
}
\description{
Rank plot of protein iBAQ.
}
