% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-nmf.R
\name{plot_metaNMF}
\alias{plot_metaNMF}
\title{NMF coefficients}
\usage{
plot_metaNMF(...)
}
\description{
Visualizes the coefficient heat map
}
\examples{
plot_metaNMF(
  scale_log2r = TRUE,
  col_order = Order,
  n_clust = 6
)

}
\seealso{
\code{\link{proteoTrend}} for parameters
}
