% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-nmf.R
\name{plot_prnNMFCoef}
\alias{plot_prnNMFCoef}
\title{NMF coefficients}
\usage{
plot_prnNMFCoef(...)
}
\description{
Visualizes the coefficient heat map
}
\examples{
plot_prnNMFCoef(
  scale_log2r = TRUE,
  col_order = Order,
  n_clust = 6
)

}
\seealso{
\code{\link{proteoTrend}} for parameters
}
