% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-nmf.R
\name{plot_prnNMFCon}
\alias{plot_prnNMFCon}
\title{NMF consensus}
\usage{
plot_prnNMFCon(...)
}
\description{
Visualizes the consensus heat map
}
\examples{
plot_prnNMFCon(
  scale_log2r = TRUE,
  col_order = Order,
  n_clust = 6
)

}
\seealso{
\code{\link{proteoTrend}} for parameters
}
