% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular.R
\name{plot_prnSubcell_UMAP}
\alias{plot_prnSubcell_UMAP}
\title{Plot UMAP of sub-cellular scores.}
\usage{
plot_prnSubcell_UMAP(
  dat_dir = NULL,
  df_expr = NULL,
  id = "gene",
  n_clust = 16:36,
  levels_subcell_compartments = c("Cytosol", "Nucleus", "ER", "PM", "Mitochondrion",
    "Golgi", "Lysosome", "Ribosome", "Stress.granule", "Proteasome", "Peroxisome",
    "Centrosome", "Endosome", "Unknown"),
  col_subtype = NULL,
  df2 = NULL,
  filepath = NULL,
  filename = NULL,
  ext = "png",
  wrap_facets = TRUE,
  seed = 123,
  ...
)
}
\arguments{
\item{dat_dir}{The current working directory.}

\item{df_expr}{A matrix of expression data.}

\item{id}{An identifier.}

\item{n_clust}{Not used. The number of clusters.}

\item{levels_subcell_compartments}{The factor levels of subcellular
compartments. Set \code{levels_subcell_compartments = NULL} for automatic
determination.}

\item{df2}{Not used. The names of secondary input files.}

\item{filepath}{A file path of inputs and outputs.}

\item{filename}{An output filename.}

\item{ext}{An extension of output filename.}

\item{wrap_facets}{Wrap facets or not.}

\item{seed}{A seed for umap.}
}
\description{
Inputs are expression data of normalized intensities.
}
\examples{
\dontrun{
plot_prnSubcell_UMAP(
  dat_dir = dat_dir, df_expr = NULL, id = "gene",
  df2 = NULL, filepath = NULL, filename = NULL, ext = "pdf",
  wrap_facets = TRUE, seed = 123, nrow = 2, ncol = 7, width = 16, height = 6)
}
}
