% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular.R
\name{plot_prnSubcellular}
\alias{plot_prnSubcellular}
\title{Visualization of subcellular results.}
\usage{
plot_prnSubcellular(
  dat_dir = NULL,
  col_select = NULL,
  col_order = NULL,
  levels_subcellular = NULL,
  levels_subtype = NULL,
  tie_method = "none",
  qt = 0.5,
  n_clust = NULL,
  panel_ids = NULL,
  pat = "Trend_[ONZ]_.*nclust\\\\d+\\\\.txt$",
  ext = "txt",
  data_type = "Trend",
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  df2 = NULL,
  filepath = NULL,
  filename = NULL,
  make_plot = TRUE,
  theme = NULL,
  ...
)
}
\arguments{
\item{dat_dir}{The working directory.}

\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
At the \code{NULL} default, the column key of \code{Select} in
\code{expt_smry.xlsx} will be used. In the case of no samples being
specified under \code{Select}, the column key of \code{Sample_ID} will be
used. The non-empty entries under the ascribing column will be used in
indicated analysis.}

\item{col_order}{Character string to a column key in \code{expt_smry.xlsx}.
Numeric values under which will be used for the left-to-right arrangement of
samples in graphic outputs or top-to-bottom arrangement in text outputs. At
the NULL default, the column key \code{Order} will be used. If values under
column \code{Order} are left blank, samples will be ordered by their names.}

\item{levels_subcellular}{The levels of subcellular fractions.}

\item{levels_subtype}{The levels of sample subtypes, e.g., Control, Treated.}

\item{qt}{Not used. The quantile of localization score for thresholding
subcellular compartment specificity. The default is to use median.}

\item{n_clust}{Numeric vector; the number(s) of clusters that data will be
divided into. At the NULL default, it will be determined by the gap method
in \code{\link[cluster]{clusGap}}. The \code{n_clust} overwrites the
argument \code{centers} in \code{\link[e1071]{cmeans}}.}

\item{panel_ids}{A subset of panel IDs for visualization.}

\item{pat}{The pattern of input filenames.}

\item{ext}{The extension of input filenames.}

\item{data_type}{The type of input data.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE. At
\code{scale_log2r = NA}, the raw \code{log2FC} without normalization will
be used.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used. The default is FALSE.}

\item{df2}{Character vector or string; the name(s) of secondary data file(s).
An informatic task, i.e. \code{anal_prnTrend(...)} against a primary
\code{df} generates secondary files such as
\code{Protein_Trend_Z_nclust6.txt} etc. See also \code{\link{prnHist}} for
the description of a primary \code{df}; \code{\link{normPSM}} for the lists
of \code{df} and \code{df2}.}

\item{filepath}{Use system default.}

\item{filename}{A representative file name to outputs. By default, it will be
determined automatically by the name of the current call.}

\item{make_plot}{Logical; to bypass the plotting at FALSE.}

\item{theme}{A
\href{https://ggplot2.tidyverse.org/reference/ggtheme.html}{ggplot2} theme,
i.e., theme_bw(), or a custom theme. At the NULL default, a system theme
will be applied.}

\item{...}{\code{filter_}: Variable argument statements for the row
filtration against data in a primary file linked to \code{df}. See also
\code{\link{normPSM}} for the format of \code{filter_} statements. \cr \cr
\code{arrange_}: Variable argument statements for the row ordering against
data in a primary file linked to \code{df}. See also \code{\link{prnHM}}
for the format of \code{arrange_} statements. \cr \cr Additional arguments
for \link[e1071]{cmeans}, \link[stats]{kmeans}, \link[cluster]{clara},
\link[cluster]{pam}. Note that \code{centers} in \link[e1071]{cmeans} or
\link[stats]{kmeans} is replaced with \code{n_clust}. The same applies to
\code{k} in \link[cluster]{clara} or \link[cluster]{pam}. \cr With
\link[e1071]{cmeans}, \code{m} is according to Schwaemmle and Jensen if not
provided; \cr \code{x} is disabled with input data being determined
automatically.}
}
\description{
\code{plot_prnSubcellular} plots the subcellular purity of protein
expressions from \code{\link{anal_prnTrend}}.
}
\examples{
if (FALSE) {
  cl <- 16; width = 10; height = 7.5;
  plot_prnSubcellular(
    df2 = paste0("Protein_Trend_Z_nclust", cl, ".txt"),
    col_order = Order,
    n_clust = cl,
    levels_subcellular = c("CP", "NP", "ChA"),
    ncol = 4,
    width = !!width,
    height = !!height,
  )
}
}
