% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular.R
\name{plot_prnSubcellular_UMAP}
\alias{plot_prnSubcellular_UMAP}
\title{Plot UMAP of subcellular scores.}
\usage{
plot_prnSubcellular_UMAP(
  df2 = NULL,
  levels_fraction = NULL,
  levels_subtype = NULL,
  key_subcellular = "location_slim3",
  key_col = "loc_score",
  filename_ref = NULL,
  scale_log2r = TRUE,
  impute_na = FALSE,
  filename = NULL,
  seed = 42,
  ...
)
}
\arguments{
\item{df2}{Character vector or string; the name(s) of secondary data file(s).
An informatic task, i.e. \code{anal_prnTrend(...)} against a primary
\code{df} generates secondary files such as
\code{Protein_Trend_Z_nclust6.txt} etc. See also \code{\link{prnHist}} for
the description of a primary \code{df}; \code{\link{normPSM}} for the lists
of \code{df} and \code{df2}.}

\item{levels_fraction}{The levels of subcellular fractions.}

\item{levels_subtype}{The levels of sample subtypes, e.g., Control, Treated.}

\item{key_subcellular}{The key of subcellular compartments from the reference
data.}

\item{key_col}{The column key where its values are used for UMAP.}

\item{filename_ref}{Optional; a file name with prepending path to a reference
subcellular annotation file, e.g., UniProt and GO intersection.}

\item{scale_log2r}{Logical; at the TRUE default, input files with
\code{_Z[...].txt} in name will be used. Otherwise, files with
\code{_N[...].txt} in name will be taken. An error will be thrown if no
files are matched under given conditions.}

\item{impute_na}{Logical; at the TRUE default, input files with
\code{_impNA[...].txt} in name will be loaded. Otherwise, files without
\code{_impNA} in name will be taken. An error will be thrown if no files are
matched under given conditions.}

\item{filename}{A representative file name to outputs. By default, it will be
determined automatically by the name of the current call.}

\item{seed}{A seed for umap.}

\item{...}{\code{filter2_}: Variable argument statements for the row
filtration against data in secondary file(s) of
\code{[...]Protein_Trend_[...].txt}. See also \code{\link{prnGSPAHM}} for
the format of \code{filter2_} statements. \cr \cr \code{arrange2_}: Variable
argument statements for the row ordering against data in secondary file(s)
of \code{[...]Protein_Trend_[...].txt}. See also \code{\link{prnGSPAHM}} for
the format of \code{arrange2_} statements. \cr \cr Additional parameters for
use in \code{plot_} functions: \cr \code{ymin}, the minimum y at \code{log2}
scale; \cr \code{ymax}, the maximum y at \code{log2} scale; \cr
\code{ybreaks}, the breaks in y-axis at \code{log2} scale; \cr \code{nrow},
the number of rows; \cr \code{width}, the width of plot; \cr \code{height},
the height of plot; \cr \code{color}, the color of lines; \cr \code{alpha},
the transparency of lines.}
}
\description{
Against localization scores.
}
