% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm.R
\name{plot_sub_hms}
\alias{plot_sub_hms}
\title{Plot subtrees of heat maps.}
\usage{
plot_sub_hms(
  df,
  id = "gene",
  label_scheme_sub,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  clustering_distance_rows,
  clustering_distance_cols,
  p_dist_rows,
  p_dist_cols,
  hc_method_rows,
  hc_method_cols,
  h_cols,
  p,
  annotation_col = TRUE,
  annotation_row = FALSE,
  mypalette,
  color_breaks,
  annotation_colors,
  filepath,
  fn_prefix,
  fn_suffix,
  sample_ids,
  complete_cases = FALSE,
  ok_group_renorm = TRUE,
  dots
)
}
\arguments{
\item{df}{A data frame of both intensities and log2Ratios.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{label_scheme_sub}{A subset of metadata.}

\item{cluster_rows}{Logical; cluster rows or not.}

\item{cluster_cols}{Logical; cluster columns or not.}

\item{clustering_distance_rows}{A method for clustering rows by distances.}

\item{clustering_distance_cols}{A method for clustering columns by distances.}

\item{p_dist_rows}{Numeric; the power of the Minkowski distance in the
measures of row \code{\link[stats]{dist}} at \code{clustering_distance_rows
= "minkowski"}. The default is 2.}

\item{p_dist_cols}{Numeric; similar to \code{p_dist_rows} but for column data.}

\item{hc_method_rows}{A character string; the same agglomeration method for
\code{\link[stats]{hclust}} of data rows. The default is \code{complete}.}

\item{hc_method_cols}{A character string; similar to \code{hc_method_rows} but
for column data.}

\item{h_cols}{The column hierarchy from the full data.}

\item{p}{A plot object from the full data.}

\item{annotation_col}{Dummy argument to avoid incurring the corresponding
argument in \link[pheatmap]{pheatmap}.}

\item{annotation_row}{Dummy argument to avoid incurring the corresponding
argument in \link[pheatmap]{pheatmap}.}

\item{mypalette}{A color palette.}

\item{color_breaks}{Color breaks.}

\item{annotation_colors}{Annotation colors.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{fn_prefix}{The prefix of an output file.}

\item{fn_suffix}{The suffix of an output file.}

\item{sample_ids}{Sample_ids from metadata.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{ok_group_renorm}{Logical; OK to renormalize samples with groups or not.}

\item{dots}{Additional arguments.}
}
\description{
Plot subtrees of heat maps.
}
