% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanos.R
\name{plot_venn}
\alias{plot_venn}
\title{Plots Venn.}
\usage{
plot_venn(Counts, filepath, direction, fml_nm)
}
\arguments{
\item{Counts}{Counts from \link{summ_venn}.}

\item{filepath}{A file path.}

\item{direction}{A direction of up or down.}

\item{fml_nm}{Formula name.}
}
\description{
Plots Venn.
}
