% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-nmf.R
\name{plotNMF_temp}
\alias{plotNMF_temp}
\title{Plots results from NMF analysis}
\usage{
plotNMF_temp(df, id, r, nrun, col_group, label_scheme_sub, filepath,
  filename, complete_cases, xmin = -1, xmax = 1, x_margin = 0.1,
  annot_cols = NULL, annot_colnames = NULL, width_consensus = 6,
  height_consensus = 6, width_coefmap = 6, height_coefmap = 6, ...)
}
\description{
Plots results from NMF analysis
}
