% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-nmf.R
\name{plotNMFCoef}
\alias{plotNMFCoef}
\title{Plots coef results from NMF analysis}
\usage{
plotNMFCoef(
  id,
  rank,
  label_scheme_sub,
  scale_log2r,
  complete_cases,
  impute_na,
  df2,
  filepath,
  filename,
  ...
)
}
\description{
Plots coef results from NMF analysis
}
