% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-volcanos.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Plot volcanos}
\usage{
plotVolcano(
  df = NULL,
  df2 = NULL,
  id = "gene",
  adjP = FALSE,
  show_labels = TRUE,
  anal_type = "Volcano",
  gspval_cutoff = 0.05,
  gslogFC_cutoff = log2(1.2),
  topn = Inf,
  show_sig = "none",
  fml_nms = NULL,
  gset_nms = "go_sets",
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  filepath = NULL,
  filename = NULL,
  theme = NULL,
  ...
)
}
\description{
Plot volcanos
}
