% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-volcanos.R
\name{plotVolcano_sub}
\alias{plotVolcano_sub}
\title{Perform significance tests}
\usage{
plotVolcano_sub(df = NULL, id = "gene", filepath = NULL,
  filename = NULL, adjP = FALSE, show_labels = TRUE, anal_type)
}
\description{
Perform significance tests
}
