% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{pmatch_bymgfs_i}
\alias{pmatch_bymgfs_i}
\title{Helper of \link{pmatch_bymgfs}.}
\usage{
pmatch_bymgfs_i(
  i,
  aa_masses,
  mgf_path,
  n_cores,
  out_path,
  mod_indexes,
  type_ms2ions,
  maxn_vmods_per_pep,
  maxn_sites_per_vmod,
  maxn_vmods_sitescombi_per_pep,
  minn_ms2,
  ppm_ms1,
  ppm_ms2,
  digits
)
}
\arguments{
\item{i}{Integer; the index for a set of corresponding aa_masses and
theoretical peptides.}

\item{mgf_path}{The file path to a list of MGF files. There is no default and
the experimenters need to supply the files.}

\item{n_cores}{Integer; the number of computer cores.}

\item{out_path}{The output path.}

\item{mod_indexes}{Integer; the indexes of fixed and/or variable
modifications.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{maxn_vmods_sitescombi_per_pep}{Integer; the maximum number of
combinatorial variable modifications per peptide sequence.}

\item{minn_ms2}{Integer; the minimum number of MS2 ions for consideration as
a hit.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{digits}{Integer; the number of decimal places to be used.}

\item{is_target}{Logical; indicator of target or decoy peptides for
searching.}
}
\description{
Helper of \link{pmatch_bymgfs}.
}
