% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{point_in_polygon}
\alias{point_in_polygon}
\title{Find points within a contour}
\usage{
point_in_polygon(x, y, xp, yp)
}
\arguments{
\item{x}{A vector of X values.}

\item{y}{A vector of Y values.}

\item{xp}{A vector of polygon X values.}

\item{yp}{A vector of polygon Y values.}
}
\description{
Find points within a contour
}
