% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{post_frame_adv}
\alias{post_frame_adv}
\title{Post frame advancing.}
\usage{
post_frame_adv(res, mgf_frames)
}
\arguments{
\item{res}{Results from frame-advanced searches.}

\item{mgf_frames}{Data of MGF frames.}
}
\description{
Post frame advancing.
}
