% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{post_ms2match}
\alias{post_ms2match}
\title{Post processing after ms2match.}
\usage{
post_ms2match(out, i, aa_masses, out_path)
}
\arguments{
\item{out}{An output from various ms2match(es).}

\item{i}{Integer; the index for a set of corresponding aa_masses and
theoretical peptides.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{out_path}{An output path.}
}
\description{
Post processing after ms2match.
}
