% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{post_pepscores}
\alias{post_pepscores}
\title{Cleanups post \link{calc_pepscores}.}
\usage{
post_pepscores(df)
}
\arguments{
\item{df}{A results after pep_scores.}
}
\description{
Cleanups post \link{calc_pepscores}.
}
