% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{postMZpepLFQ}
\alias{postMZpepLFQ}
\title{Post-processing of Mzion PSM tables.}
\usage{
postMZpepLFQ(df, group_pep_by = "gene", type_sd = "log2_R", set_idx = 1L)
}
\arguments{
\item{df}{A data frame of peptide table.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{type_sd}{Character string; the type of log2Ratios for SD calculations.
The value is one \code{log2_R}, \code{N_log2_R} or \code{Z_log2_R}.}

\item{set_idx}{A set index of TMT experiment.}
}
\description{
Post-processing of Mzion PSM tables.
}
