% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pre_mq_meta}
\alias{pre_mq_meta}
\title{Updates MaxQuant experimentalDesignTemplate.txt}
\usage{
pre_mq_meta(
  dat_dir = proteoQ:::get_gl_dat_dir(),
  expt_smry = "expt_smry.xlsx",
  filename = "combined/experimentalDesignTemplate.txt"
)
}
\arguments{
\item{dat_dir}{The working directory.}

\item{expt_smry}{The filename of proteoQ metadata.}

\item{filename}{The filename of MaxQuant metadata.}
}
\description{
Updates MaxQuant experimentalDesignTemplate.txt
}
