% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{prep_folded_tdata}
\alias{prep_folded_tdata}
\title{Prepares folded, transposed data}
\usage{
prep_folded_tdata(df, folds, label_scheme_sub, col_group)
}
\arguments{
\item{df}{Input data frame}

\item{folds}{Not currently used. Integer; the degree of folding data into
subsets. The default is one without data folding.}

\item{label_scheme_sub}{Metadata for data subset}

\item{col_group}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_group} will be
used for sample grouping in the indicated analysis. At the NULL default, the
column key \code{Group} will be used. No data annotation by groups will be
performed if the fields under the indicated group column is empty.}
}
\description{
Prepares folded, transposed data
}
