% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{prep_fraction_scheme}
\alias{prep_fraction_scheme}
\title{Loads the information of analyte prefractionation}
\usage{
prep_fraction_scheme(dat_dir, filename)
}
\description{
Loads the information of analyte prefractionation
}
