% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspa.R
\name{prep_gspa}
\alias{prep_gspa}
\title{Helper of GSPA}
\usage{
prep_gspa(
  df = NULL,
  id = NULL,
  df_trend = NULL,
  is_subcellular = FALSE,
  col_subcellular = NULL,
  col_subtype = NULL,
  col_group = NULL,
  fml_nm = NULL,
  dat_dir = NULL,
  label_scheme_sub = NULL,
  col_ind = 0L,
  pval_cutoff = 0.05,
  logFC_cutoff = log2(1.2),
  use_adjP = FALSE,
  scale_log2r = TRUE,
  impute_na = FALSE
)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{id}{Currently only "entrez".}

\item{df_trend}{Trend data.}

\item{is_subcellular}{Logical; does data contain subcellular specificity.}

\item{col_subtype}{Temporarily applied only to subcellular data analysis.
Character string to a column key in \code{expt_smry.xlsx} indicating sample
subtypes linked to subcellular fractions. This is needed, e.g., when each
subcellular fraction contain multiple subtypes of control, treated etc.}

\item{col_group}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_group} will be
used for sample grouping in the indicated analysis. At the NULL default, the
column key \code{Group} will be used. No data annotation by groups will be
performed if the fields under the indicated group column is empty.}

\item{fml_nm}{A character string; the name of \code{fml}.}

\item{dat_dir}{The working directory.}

\item{label_scheme_sub}{A data frame. Subset entries from \code{label_scheme}
for selected samples.}

\item{col_ind}{Numeric vector; the indexes of columns for the ascribed
\code{fml_nm}.}

\item{pval_cutoff}{Numeric value or vector; the cut-off in protein
significance \code{pVal}. Entries with \code{pVals} less significant than
the threshold will be excluded from enrichment analysis. The default is 0.05
for all formulas matched to or specified in argument \code{fml_nms}.
Formula-specific threshold is allowed by supplying a vector of cut-off
values.}

\item{logFC_cutoff}{Numeric value or vector; the cut-off in protein
\code{log2FC}. Entries with absolute \code{log2FC} smaller than the
threshold will be excluded from enrichment analysis. The default magnitude
is \code{log2(1.2)} for all formulas matched to or specified in argument
\code{fml_nms}. Formula-specific threshold is allowed by supplying a vector
of absolute values in \code{log2FC}.}

\item{use_adjP}{Logical; if TRUE, use Benjamini-Hochberg pVals. The default is
FALSE.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE. At
\code{scale_log2r = NA}, the raw \code{log2FC} without normalization will
be used.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used. The default is FALSE.}
}
\description{
Helper of GSPA
}
