% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{prep_label_scheme}
\alias{prep_label_scheme}
\title{Processes the metadata of TMT experiments}
\usage{
prep_label_scheme(
  dat_dir = NULL,
  expt_smry = "expt_smry.xlsx",
  frac_smry = "frac_smry.xlsx"
)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{expt_smry}{A character string to a \code{.xlsx} file containing the
metadata of TMT or LFQ experiments. The default is \code{expt_smry.xlsx}.}

\item{frac_smry}{A character string to a \code{.xlsx} file containing peptide
fractionation summary. The default is \code{frac_smry.xlsx}.}
}
\description{
Processes the metadata of TMT experiments
}
\section{\code{Empty samples}}{
 Under the same TMT_Set and TMT_Channel, the
  Sample_ID can be non-trivial under one LCMS_Injection but Empty under
  another. During the \link{PSM2Pep} roll-ups, the values under Empty samples
  will be all NA and thus has no contributions whereas the values of the
  non-trivial samples will be used normally.
}

