% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_range}
\alias{prep_range}
\title{Standardizes range values.}
\usage{
prep_range(from_to = c(0, 1))
}
\arguments{
\item{from_to}{A numeric vector of length 2.}
}
\description{
Standardizes range values.
}
