% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{prepString}
\alias{prepString}
\title{Downloads and prepares STRING databases}
\usage{
prepString(
  species = "human",
  links_url = NULL,
  aliases_url = NULL,
  info_url = NULL,
  db_path = "~/proteoQ/dbs/string",
  filename = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{species}{Character string; the name of a species for the
\emph{conveninent} preparation of STRING databases. The species available
for the convenience feature is in one of \code{c("human", "mouse", "rat")}
with \code{"human"} being the default. The argument is not required for
other species; instead, users will provide values under arguments
\code{links_url}, \code{aliases_url} and \code{info_url}.}

\item{links_url}{A URL to \href{http://string-db.org/}{STRING} interaction
data. A valid web address is required for species other than
\code{c("human", "mouse", "rat")}. At the NULL default and the species in
one of \code{c("human", "mouse", "rat")}, the link will be determined
automatically; note that users can overwrite the default interaction data by
providing their own URL.}

\item{aliases_url}{A URL to \href{http://string-db.org/}{STRING} aliases
between \code{string_protein_ids} and \code{protein accessions}. A valid web
address is required for species other than \code{c("human", "mouse",
"rat")}. At the NULL default and the species in one of \code{c("human",
"mouse", "rat")}, the link will be determined automatically; note that users
can overwrite the default alias data by providing their own URL.}

\item{info_url}{A URL to \href{http://string-db.org/}{STRING} display names
between \code{string_protein_ids} and \code{gene names}. A valid web address
is required for species other than \code{c("human", "mouse", "rat")}. At the
NULL default and the species in one of \code{c("human", "mouse", "rat")},
the link will be determined automatically; note that users can overwrite the
default \code{info} data by providing their own URL.}

\item{db_path}{Character string; the local path for database(s). The default
is \code{"~/proteoQ/dbs/string"}.}

\item{filename}{Use system default. Otherwise, the user-provided basename will
be prepended with \code{_ppi.tsv} for network data and \code{_expr.tsv} for
expression data.}

\item{overwrite}{Logical; if TRUE, overwrite the downloaded database(s). The
default is FALSE.}
}
\description{
\code{prepString} downloads and prepares the data sets of
\href{https://string-db.org/}{STRING} interactions, accession lookups and gene
name lookups.
}
\examples{
prepString(
 # species = fly, 
 links_url = "https://stringdb-static.org/download/protein.links.full.v11.0/7227.protein.links.full.v11.0.txt.gz",
 aliases_url = "https://stringdb-static.org/download/protein.aliases.v11.0/7227.protein.aliases.v11.0.txt.gz",
 info_url = "https://stringdb-static.org/download/protein.info.v11.0/7227.protein.info.v11.0.txt.gz", 
 filename = string_dm.rds,
)
}
\seealso{
\code{\link{anal_prnString}} for protein-protein interaction
 networks.
}
