% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{prn_annot_crap}
\alias{prn_annot_crap}
\title{Lookups of cRAP proteins.}
\format{
A data frame with 195 rows and 9 variables:
\describe{
  \item{uniprot_acc}{UniProt accession number}
  \item{uniprot_id}{UniProt ID}
  \item{status}{the review status according to UniProt}
  \item{prot_desc}{protein description}
  \item{gene}{gene names}
  \item{organism}{the organism attribute according to UniProt}
  \item{length}{the number of amino acid residues under a proposed protein}
  \item{refseq_acc}{RefSeq accession number}
  \item{entrez}{Entrez ID (not currently used)}
}
}
\source{
\url{https://www.thegpm.org/crap/}
}
\usage{
prn_annot_crap
}
\description{
A dataset containing the IDs of cRAP proteins.
}
\keyword{datasets}
