% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-corrplots.R
\name{prnCorr}
\alias{prnCorr}
\title{Correlation plots of protein data}
\usage{
prnCorr(...)
}
\description{
Correlation plots of protein data
}
\examples{
prnCorr(
  use_log10 = TRUE,
  scale_log2r = TRUE,
  min_int = 3.5,
  max_int = 6.5,
  min_log2r = -2,
  max_log2r = 2
)

}
\seealso{
\code{\link{proteoCorrplot}} for parameters
}
