% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{prnImp}
\alias{prnImp}
\title{Impute NA values for protein data}
\usage{
prnImp(...)
}
\description{
Impute NA values for protein data
}
\seealso{
\code{\link{imputeNA}} for parameters
}
