% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-nmf.R
\name{prnNMF}
\alias{prnNMF}
\title{NMF Analysis of Protein \code{log2-ratios}}
\usage{
prnNMF(...)
}
\description{
NMF Analysis of Protein \code{log2-ratios}
}
\seealso{
\code{\link{proteoNMF}} for parameters
}
