% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-trends.R
\name{prnTrend}
\alias{prnTrend}
\title{Trend Analysis of Protein \code{log2-ratios}}
\usage{
prnTrend(...)
}
\description{
Trend Analysis of Protein \code{log2-ratios}
}
\seealso{
\code{\link{proteoTrend}} for parameters
}
