% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanos.R
\name{pepVol}
\alias{pepVol}
\alias{prnVol}
\title{Volcano plots}
\usage{
pepVol(
  dat_dir = NULL,
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  adjP = FALSE,
  topn_labels = 20,
  df = NULL,
  filepath = NULL,
  filename = NULL,
  fml_nms = NULL,
  theme = NULL,
  highlights = NULL,
  grids = NULL,
  is_subcellular = FALSE,
  ...
)

prnVol(
  dat_dir = NULL,
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  adjP = FALSE,
  topn_labels = 20,
  df = NULL,
  filepath = NULL,
  filename = NULL,
  fml_nms = NULL,
  theme = NULL,
  highlights = NULL,
  grids = NULL,
  is_subcellular = FALSE,
  ...
)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE. At
\code{scale_log2r = NA}, the raw \code{log2FC} without normalization will
be used.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used. The default is FALSE.}

\item{adjP}{Logical; if TRUE, use Benjamini-Hochberg pVals in volcano plots.
The default is FALSE.}

\item{topn_labels}{A non-negative integer; the top-n species for labeling in a
plot. At \code{topn_labels = 0}, no labels of proteins/peptides will be
shown. The default is to label the top-20 species with the lowest p-values.}

\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{filepath}{Use system default.}

\item{filename}{A representative file name to outputs. By default, it will be
determined automatically by the name of the current call.}

\item{fml_nms}{Character string or vector; the formula name(s). By default,
the formula(s) will match those used in \code{\link{pepSig}} or
\code{\link{prnSig}}.}

\item{theme}{A
\href{https://ggplot2.tidyverse.org/reference/ggtheme.html}{ggplot2} theme,
i.e., theme_bw(), or a custom theme. At the NULL default, a system theme
will be applied.}

\item{highlights}{A list of entries for highlighting. See also \code{filter_}
varargs for the format.}

\item{grids}{An integer or integer vector for subset visualization of
contrasts within a group. For example with a group of three contrasts,
\code{grids = 2:3} will hide the first grid from displaying. At the
NULL default, all available grids will be shown.}

\item{is_subcellular}{Logical; does data contain subcellular specificity.}

\item{...}{\code{filter_}: Variable argument statements for the row filtration
against data in a primary file linked to \code{df}. See also
\code{\link{normPSM}} for the format of \code{filter_} statements. \cr \cr
Additional parameters for plotting: \cr \code{xco}, the cut-off lines of
fold changes at position \code{x}; the default is at \eqn{-1.2} and
\eqn{+1.2}. \cr \code{yco}, the cut-off line of \code{pVal} at position
\code{y}; the default is \eqn{0.05}. \cr \code{width}, the width of plot;
\cr \code{height}, the height of plot. \cr \code{nrow}, the number of rows
in a plot. \cr \code{xmin}, the minimum \code{x}. \cr \code{xmax}, the
maximum \code{x}. \cr \code{ymin}, the minimum \code{y}. \cr \code{ymax},
the maximum \code{y}. \cr \code{x_label}, the label on \code{x}. \cr
\code{y_label}, the label on \code{y}.}
}
\description{
\code{pepVol} visualizes the volcano plots of peptide data.

\code{prnVol} visualizes the volcano plots of protein data.
}
\examples{
\donttest{
# ===================================
# Volcano plots
# ===================================

## !!!Require the brief working example in `?load_expts`

## Global option
scale_log2r <- TRUE

## For all peptides or proteins
# all peptides
pepVol()

# All proteins
prnVol(
  xco = 1.2,
  yco = 0.01,
)

# Hide `xco` and/or `yco` lines
# (xco = 0 -> log2(xco) = - Inf)
prnVol(
  xco = 0,
  yco = Inf,
  filename = no_xylines.png,
)

# Shows vertical center line at log2(1)
# (xco = 1 -> log2(xco) = 0)
prnVol(
  xco = 1,
  yco = Inf,
  filename = no_xylines.png,
)

# Kinases and prot_n_pep >= 2
prnVol(
  xco = 1.2,
  yco = 0.01,
  filter_prots_by_kin = exprs(kin_attr, prot_n_pep >= 2),
  filename = "kin_npep2.png"
)

# Selected formula and/or customization
prnVol(
  fml_nms = "W2_bat",
  xmin = -5,
  xmax = 5, 
  ymin = 0, 
  ymax = 30,
  x_label = "Ratio ("*log[2]*")",
  y_label = "pVal ("*-log[10]*")", 
  height = 6,
  width = 6*2.7,
  filename = custom.png,
)

# Custom theme
library(ggplot2)
my_theme <- theme_bw() +
  theme(
    axis.text.x = element_text(angle = 0, vjust = 0.5, size = 24),
    axis.ticks.x = element_blank(),
    axis.text.y = element_text(angle = 0, vjust = 0.5, size = 24),
    axis.title.x = element_text(colour = "black", size = 24),
    axis.title.y = element_text(colour="black", size = 24),
    plot.title = element_text(face = "bold", colour = "black", size = 14, 
                              hjust = .5, vjust = .5),
    
    panel.grid.major.x = element_blank(),
    panel.grid.minor.x = element_blank(),
    panel.grid.major.y = element_blank(),
    panel.grid.minor.y = element_blank(),
    
    strip.text.x = element_text(size = 16, colour = "black", angle = 0),
    strip.text.y = element_text(size = 16, colour = "black", angle = 90),
    
    legend.key = element_rect(colour = NA, fill = 'transparent'),
    legend.background = element_rect(colour = NA,  fill = "transparent"),
    legend.position = "none",
    legend.title = element_text(colour="black", size = 18),
    legend.text = element_text(colour="black", size = 18),
    legend.text.align = 0,
    legend.box = NULL
  )

prnVol(theme = my_theme, filename = my_theme.png)

# Custom plot
# ("W2_bat" etc. are contrast names in `pepSig`)
prnVol(fml_nms = c("W2_bat", "W2_loc"), filename = foo.png)

res <- readRDS(file.path(dat_dir, "Protein/Volcano/W2_bat/foo.rds"))
# names(res)

p <- ggplot() +
  geom_point(data = res$data, mapping = aes(x = log2Ratio, y = -log10(pVal)), 
             size = 3, colour = "#f0f0f0", shape = 20, alpha = .5) +
  geom_point(data = res$greater, mapping = aes(x = log2Ratio, y = -log10(pVal)), 
             size = 3, color = res$palette[2], shape = 20, alpha = .8) +
  geom_point(data = res$less, mapping = aes(x = log2Ratio, y = -log10(pVal)), 
             size = 3, color = res$palette[1], shape = 20, alpha = .8) +
  geom_hline(yintercept = -log10(res$yco), linetype = "longdash", linewidth = .5) +
  geom_vline(xintercept = -log2(res$xco), linetype = "longdash", linewidth = .5) +
  geom_vline(xintercept = log2(res$xco), linetype = "longdash", linewidth = .5) +
  scale_x_continuous(limits = c(res$xmin, res$xmax)) +
  scale_y_continuous(limits = c(res$ymin, res$ymax)) +
  labs(title = res$title, x = res$x_label, y = res$y_label) +
  res$theme

p <- p + geom_text(data = res$topns, 
                   mapping = aes(x = log2Ratio, 
                                 y = -log10(pVal), 
                                 label = Index, 
                                 color = Index),
                   size = 3, 
                   alpha = .5, 
                   hjust = 0, 
                   nudge_x = 0.05, 
                   vjust = 0, 
                   nudge_y = 0.05, 
                   na.rm = TRUE)

p <- p + facet_wrap(~ Contrast, nrow = 1, labeller = label_value)

p <- p + geom_table(data = res$topn_labels, aes(table = gene), 
                    x = -res$xmax*.85, y = res$ymax/2)

# Highlight
prnVol(
  highlights = rlang::exprs(gene \%in\% c("ACTB", "GAPDH")), 
  filename = highlights.png
)


## protein subgroups by gene sets
# prerequisite analysis of GSPA
prnGSPA(
  impute_na = FALSE,
  pval_cutoff = 1E-2, # protein pVal threshold
  logFC_cutoff = log2(1.1), # protein log2FC threshold
  gspval_cutoff = 5E-2, # gene-set pVal threshold
  gslogFC_cutoff = log2(1.2), # gene-set log2FC threshold
  gset_nms = c("go_sets"),
)

# Mapping gene sets to volcano-plot visualization
# (1) forced lines of `pval_cutoff` and `logFC_cutoff`  
#   according to the corresponding `prnGSPA` in red; 
# (2) optional lines of `xco` and `yco` in grey
prnGSPAMap(
  show_sig = pVal, 
)

# Disable the lines of `xco` and `yco`, 
prnGSPAMap(
  impute_na = FALSE,
  topn_gsets = 20, 
  show_sig = pVal, 
  xco = 0, 
  yco = Inf, 
)

# Customized thresholds for visualization
prnGSPAMap(
  fml_nms = c("W2_bat", "W2_loc", "W16_vs_W2"),
  gspval_cutoff = c(5E-2, 5E-2, 1E-10),
  gslogFC_cutoff = log2(1.2),
  topn_gsets = 20, 
  topn_labels = 0,
  show_sig = pVal,
  xco = 0, 
  yco = Inf, 
)

## prnGSPAMap(...) maps secondary results of `[...]Protein_GSPA_{NZ}[_impNA].txt` 
#  from prnGSPA(...) onto a primary `df` of `Protein[_impNA]_pVal.txt` 
#  
#  See also ?prnGSPA for additional examples involving both `df` and `df2`, 
#  as well as `filter_` and `filter2_`
}
}
\seealso{
\emph{Metadata} \cr \code{\link{load_expts}} for metadata preparation
 and a reduced working example in data normalization \cr

 \emph{Data normalization} \cr \code{\link{normPSM}} for extended examples in
 PSM data normalization \cr \code{\link{PSM2Pep}} for extended examples in
 PSM to peptide summarization \cr \code{\link{mergePep}} for extended
 examples in peptide data merging \cr \code{\link{standPep}} for extended
 examples in peptide data normalization \cr \code{\link{Pep2Prn}} for
 extended examples in peptide to protein summarization \cr
 \code{\link{standPrn}} for extended examples in protein data normalization.
 \cr \code{\link{purgePSM}} and \code{\link{purgePep}} for extended examples
 in data purging \cr \code{\link{pepHist}} and \code{\link{prnHist}} for
 extended examples in histogram visualization. \cr \code{\link{extract_raws}}
 and \code{\link{extract_psm_raws}} for extracting MS file names \cr

 \emph{Variable arguments of `filter_...`} \cr \code{\link{contain_str}},
 \code{\link{contain_chars_in}}, \code{\link{not_contain_str}},
 \code{\link{not_contain_chars_in}}, \code{\link{start_with_str}},
 \code{\link{end_with_str}}, \code{\link{start_with_chars_in}} and
 \code{\link{ends_with_chars_in}} for data subsetting by character strings
 \cr

 \emph{Missing values} \cr \code{\link{pepImp}} and \code{\link{prnImp}} for
 missing value imputation \cr

 \emph{Informatics} \cr \code{\link{pepSig}} and \code{\link{prnSig}} for
 significance tests \cr \code{\link{pepVol}} and \code{\link{prnVol}} for
 volcano plot visualization \cr \code{\link{prnGSPA}} for gene set enrichment
 analysis by protein significance pVals \cr \code{\link{prnGSPAMap}} for mapping
 GSPA to volcano plot visualization \cr \code{\link{prnGSPAHM}} for heat map
 and network visualization of GSPA results \cr \code{\link{prnGSVA}} for gene
 set variance analysis \cr \code{\link{prnGSEA}} for data preparation for
 online GSEA. \cr \code{\link{pepMDS}} and \code{\link{prnMDS}} for MDS
 visualization \cr \code{\link{pepPCA}} and \code{\link{prnPCA}} for PCA
 visualization \cr \code{\link{pepLDA}} and \code{\link{prnLDA}} for LDA
 visualization \cr \code{\link{pepHM}} and \code{\link{prnHM}} for heat map
 visualization \cr \code{\link{pepCorr_logFC}}, \code{\link{prnCorr_logFC}},
 \code{\link{pepCorr_logInt}} and \code{\link{prnCorr_logInt}}  for
 correlation plots \cr \code{\link{anal_prnTrend}} and
 \code{\link{plot_prnTrend}} for trend analysis and visualization \cr
 \code{\link{anal_pepNMF}}, \code{\link{anal_prnNMF}},
 \code{\link{plot_pepNMFCon}}, \code{\link{plot_prnNMFCon}},
 \code{\link{plot_pepNMFCoef}}, \code{\link{plot_prnNMFCoef}} and
 \code{\link{plot_metaNMF}} for NMF analysis and visualization \cr

 \emph{Custom databases} \cr \code{\link{Uni2Entrez}} for lookups between
 UniProt accessions and Entrez IDs \cr \code{\link{Ref2Entrez}} for lookups
 among RefSeq accessions, gene names and Entrez IDs \cr \code{\link{prepGO}}
 for
 \code{\href{http://current.geneontology.org/products/pages/downloads.html}{gene
  ontology}} \cr \code{\link{prepMSig}} for
 \href{https://data.broadinstitute.org/gsea-msigdb/msigdb/release/7.0/}{molecular
  signatures} \cr \code{\link{prepString}} and \code{\link{anal_prnString}}
 for STRING-DB \cr

 \emph{Column keys in PSM, peptide and protein outputs} \cr
 system.file("extdata", "psm_keys.txt", package = "proteoQ") \cr
 system.file("extdata", "peptide_keys.txt", package = "proteoQ") \cr
 system.file("extdata", "protein_keys.txt", package = "proteoQ") \cr
}
