% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-go.R
\name{proc_gaf}
\alias{proc_gaf}
\title{Helper to save `gaf` without header}
\usage{
proc_gaf(db_path, fn_gaf)
}
\arguments{
\item{db_path}{Character string; the local path for database(s). The default
is \code{"~\\proteoQ\\dbs\\go"}.}

\item{fn_gaf}{filename according to \code{gaf_url}}
}
\description{
Helper to save `gaf` without header
}
