% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{proc_mgf_chunks}
\alias{proc_mgf_chunks}
\title{Processes MGF entries in chunks.}
\usage{
proc_mgf_chunks(
  lines,
  topn_ms2ions = 100,
  ret_range = c(0, Inf),
  filepath = file.path("~/proteoQ/mgfs/temp")
)
}
\arguments{
\item{lines}{MGF lines.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{ret_range}{The range of retention time in seconds.}

\item{filepath}{The file path to a list of mgf files.}
}
\description{
Processes MGF entries in chunks.
}
