% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{proc_mgf_chunks_i}
\alias{proc_mgf_chunks_i}
\title{Helper of \link{proc_mgf_chunks}.}
\usage{
proc_mgf_chunks_i(
  file,
  topn_ms2ions = 100L,
  ret_range = c(0L, Inf),
  ppm_ms2 = 25L,
  min_ms2mass = 110L,
  index_ms2 = FALSE
)
}
\arguments{
\item{file}{A mgf chunk (chunk_1.mgf etc.) with prepending file path.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{ret_range}{The range of retention time in seconds.}
}
\description{
Helper of \link{proc_mgf_chunks}.
}
