% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proc_mgf_timstof}
\alias{proc_mgf_timstof}
\title{Add the MS file name to timsTOF mgf files.}
\usage{
proc_mgf_timstof(path, n = 1000)
}
\arguments{
\item{path}{The file path to mgf files.}

\item{n}{Integer; the first \code{n} lines for quick checks of the MS file
name.}
}
\description{
Add the MS file name to timsTOF mgf files.
}
