% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{proc_mgfs}
\alias{proc_mgfs}
\title{Helper in processing MGF entries in chunks.}
\usage{
proc_mgfs(lines, topn_ms2ions = 100, ret_range = c(0, Inf))
}
\arguments{
\item{lines}{MGF lines.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches.}

\item{ret_range}{The range of retention time in seconds.}
}
\description{
Helper in processing MGF entries in chunks.
}
