% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{proc_obo}
\alias{proc_obo}
\title{Helper to save `obo` without header}
\usage{
proc_obo(
  db_path,
  fn_obo,
  type = c("biological_process", "cellular_component", "molecular_function")
)
}
\arguments{
\item{db_path}{Character string; the local path for database(s). The default
is \code{"~\\proteoQ\\dbs\\go"}.}

\item{fn_obo}{filename according to \code{obo_url}}

\item{type}{Character vector. The name space in gene ontology to be included.
The default is to include all in \code{c("biological_process",
"cellular_component", "molecular_function")}. In the example of \code{type =
c("biological_process", "cellular_component")}, terms of
\code{molecular_function} will be excluded.}
}
\description{
Helper to save `obo` without header
}
