% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{proc_psms}
\alias{proc_psms}
\title{Process PSMs.}
\usage{
proc_psms(
  df = NULL,
  scale_rptr_int = FALSE,
  rptr_intco = 0,
  rm_craps = FALSE,
  rm_krts = FALSE,
  annot_kinases = FALSE,
  plot_rptr_int = TRUE,
  parallel = TRUE
)
}
\arguments{
\item{df}{PSM data.}

\item{scale_rptr_int}{Logical; if TRUE, scales (up) MS2 reporter-ion
intensities by MS1 precursor intensity: \eqn{I_{MS1}*(I_{x}/\sum I_{MS2})}.
\eqn{I_{MS1}}, MS1 precursor intensity; \eqn{I_{MS2}}, MS2 reporter-ion
intensity; \eqn{I_{x}}, MS2 reporter-ion intensity under TMT channel
\eqn{x}. Note that the scaling will not affect \code{log2FC}.}

\item{rptr_intco}{Numeric; the threshold of reporter-ion intensity being
considered non-trivial. The default is 0 without cut-offs. The filtration
will not be applied synchronously to the precursor intensity under the same
PSM query. To guard against odds such as higher MS2 reporter-ion
intensities than their contributing MS1 precursor intensity, employs for
example \code{filter_... = rlang::exprs(pep_tot_int >= your_ms1_cutoff)}
during \link{PSM2Pep}.}

\item{rm_craps}{Logical; if TRUE,
\href{https://www.thegpm.org/crap/}{cRAP} proteins will be removed.
The default is FALSE.}

\item{rm_krts}{Logical; if TRUE, keratin entries will be removed. The default
is FALSE.}

\item{annot_kinases}{Logical; if TRUE, proteins of human or mouse origins
will be annotated with their kinase attributes. The default is FALSE.}

\item{plot_rptr_int}{Logical; if TRUE, the distributions of reporter-ion
intensities will be plotted. The default is TRUE. The argument is also
applicable to the precursor intensity with MaxQuant LFQ.}

\item{parallel}{Logical; if TRUE, performs parallel computation. The default
is TRUE.}
}
\description{
Common routines in PSM processing across search engines.
}
