% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{proc_remap}
\alias{proc_remap}
\title{Helper to save ReMap}
\usage{
proc_remap(
  db_path = "~/proteoQ/dbs/remap",
  filename = "remap2022_nr_macs2_mm39_v1_0.bed.gz"
)
}
\arguments{
\item{db_path}{A file path.}

\item{filename}{A file name.}
}
\description{
Helper to save ReMap
}
