% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{proc_rpt}
\alias{proc_rpt}
\title{Helper to save rpt}
\usage{
proc_rpt(
  db_path = "~/proteoQ/dbs/gene_homology",
  filename = "HOM_AllOrganism.rpt"
)
}
\arguments{
\item{db_path}{A file path.}

\item{filename}{A file name.}
}
\description{
Helper to save rpt
}
