% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{procBrukerMGF}
\alias{procBrukerMGF}
\title{Reprocesses of Bruker MGF files.}
\usage{
procBrukerMGF(file)
}
\arguments{
\item{file}{A file name with prepending path.}
}
\description{
Reprocesses of Bruker MGF files.
}
\examples{
\donttest{
filepath = "F:/timsTOF/mgf"
mprocBrukerMGF(filepath)
}
}
