% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{procBrukerMGF_v1}
\alias{procBrukerMGF_v1}
\title{Reprocesses of Bruker MGF files.}
\usage{
procBrukerMGF_v1(file)
}
\arguments{
\item{file}{A file name with prepending path.}
}
\description{
Reprocesses of Bruker MGF files.
}
\examples{
\donttest{
filepath = "F:/timsTOF/mgf"
mprocBrukerMGF(filepath)
}
}
