% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{procMSGFmods}
\alias{procMSGFmods}
\title{Processes MSGF modifications.}
\usage{
procMSGFmods(filename = "MSFG_mods.txt", dat_dir = NULL)
}
\arguments{
\item{filename}{The file name that contains MSGF modifications.}

\item{dat_dir}{The working directory.}
}
\description{
Processes MSGF modifications.
}
