% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-specialty.R
\name{proteo_hm}
\alias{proteo_hm}
\title{Makes heat maps}
\usage{
proteo_hm(
  df = NULL,
  id = NULL,
  df_meta = NULL,
  sample_ids = NULL,
  filepath = NULL,
  filename = NULL,
  complete_cases = FALSE,
  annot_cols = NULL,
  annot_colnames = NULL,
  annot_rows = NULL,
  xmin = -1,
  xmax = 1,
  xmargin = 0.1,
  ...
)
}
\arguments{
\item{df_meta}{A file name of meta data.}
}
\description{
Makes heat maps
}
\examples{
proteo_hm(
  df = Protein_delta.txt, 
  id = gene, 
  df_meta = hm_meta.xlsx, 
  filepath = file.path(dat_dir, "Protein\\\\Heatmap"), 
  filename = "kin_delta.png",
  complete_cases = FALSE, 
  annot_cols = NULL, 
  annot_colnames = NULL, 
  annot_rows = c("kin_class"), 
  cluster_rows = FALSE, 
  xmin = -1, 
  xmax = 1, 
  xmargin = .1, 
  width = 5, 
  height = 12,
  arrange2_by = exprs(kin_class, gene), 
)
}
