% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gsea.R
\name{proteoGSEA}
\alias{proteoGSEA}
\alias{prnGSEA}
\title{GSEA of protein data}
\usage{
proteoGSEA(id = gene, scale_log2r = TRUE, df = NULL,
  filepath = NULL, filename = NULL, impute_na = TRUE,
  complete_cases = FALSE, gset_nms = "go_sets", var_cutoff = 0.5,
  pval_cutoff = 0.05, logFC_cutoff = log2(1.2), gspval_cutoff = 0.05,
  min_size = 10, max_size = Inf, min_greedy_size = 1,
  fml_nms = NULL, task = "anal", ...)

prnGSEA(...)
}
\arguments{
\item{id}{Character string indicating the type of data. The value will be
determined automatically.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE.}

\item{df}{The name of input data file. By default, it will be determined
automatically by the value of \code{id}.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{filename}{A file name to output results. By default, it will be
determined automatically by the name of the calling function and the value
of id in the call.}

\item{impute_na}{Logical. At the NULL default, the TRUE or FALSE will match
the choice in \code{\link{pepSig}} for peptide and \code{\link{prnSig}} for
protein data.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{gset_nms}{Not currently used. Character string or vector containing the
name(s) of gene sets for enrichment analysis. The possible values are in
\code{c("go_sets", "kegg_sets")}. Note that the currently supported species
are human, mouse and rat.}

\item{var_cutoff}{Numeric value or vector; the cut-off in the variance of
protein \code{log2FC} across samples. Entries with \code{variances} less
than the threshold will be removed.}

\item{pval_cutoff}{Numeric value or vector; the cut-off in protein
significance \code{pVal}. Entries with \code{pVals} less significant than
the threshold will be ignored during enrichment analysis.}

\item{logFC_cutoff}{Numeric value or vector; the cut-off in protein
\code{log2FC}. Entries with absolute \code{log2FC} smaller than the
threshold will be ignored during enrichment analysis.}

\item{gspval_cutoff}{Not currently used. Numeric value or vector; the cut-off in
gene-set significance \code{pVal}. Only enrichment terms with \code{pVals}
more significant than the threshold will be reported.}

\item{min_size}{Not currently used. Numeric value or vector; minimum number of
protein entries for consideration in gene set tests. The number is the sum
of up or down-expressed proteins after data filtration by
\code{pval_cutoff}, \code{logFC_cutoff} or varargs expressions under
\code{filter_}.}

\item{max_size}{Not currently used. Numeric value or vector; maximum number of
protein entries for consideration in gene set tests. The number is the sum
of up or down-expressed proteins after data filtration by
\code{pval_cutoff}, \code{logFC_cutoff} or varargs expressions under
\code{filter_}. The default in infinite. For samples that are vastly
different between contrasts, essential gene sets may be blackholed with
trivial terms such as cell parts, molecular functions et al. In cases like
this, a smaller \code{max_size} may be considered.}

\item{min_greedy_size}{Not currently used. Numeric value or vector; minimum
number of unique protein entries for a gene set to be considered essential.}

\item{fml_nms}{Character string or vector; the formula name(s). By default,
the names will match those used in \code{\link{pepSig}} or
\code{\link{prnSig}}.}

\item{...}{\code{filter_}: Logical expression(s) for the row filtration of
data; also see \code{\link{normPSM}}.}
}
\description{
\code{prnGSEA} prepares data for the analysis of
\code{\href{http://software.broadinstitute.org/gsea/index.jsp}{GSEA}} aganist
protein \code{log2FC} data. Users should avoid calling the method directly,
but instead use the following wrappers.

\code{prnGSEA} is a wrapper of \code{\link{proteoGSEA}} for gene set
enrichment analysis.
}
\details{
The arguments \code{var_cutoff}, \code{pval_cutoff} and \code{logFC_cutoff}
are used to filter out low influence genes. Additional subsetting of data via
the \code{vararg} approach of \code{filter_} is feasible.

The outputs include \code{Protein_GSEA.gct} and \code{protein_GSEA.cls} for
samples indicated in file \code{Protein_pVals.txt} or
\code{Protein_impNA_pVals.txt}. These outputs can be used with online
\code{\href{http://software.broadinstitute.org/gsea/index.jsp}{GSEA}}.

The current GSEA may not yet support the comparisons between two grouped
conditions, i.e.,  (grpA + grpB) versus (grpC + grpD). The \code{prnGSEA}
utility further breaks the input data into pairs of groups according to the
formulae and contrasts defined in \code{pepSig} or \code{prnSig}. The
phenotype labels are then reformed in reflection of the original group names,
weights and directions, i.e., \code{0.5xgrpA&0.5xgrpB	-0.5xgrpC&-0.5xgrpD}.
The corresponding \code{.gct} and \code{.cls} files can be used with the
online or the github version of R-GSEA.
}
