% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{proteoGSPA}
\alias{proteoGSPA}
\alias{prnGSPA}
\alias{prnGSPAHM}
\title{GSPA of protein data}
\usage{
proteoGSPA(id = gene, scale_log2r = TRUE, df = NULL,
  filepath = NULL, filename = NULL, impute_na = TRUE,
  complete_cases = FALSE, gset_nms = "go_sets", var_cutoff = 0.5,
  pval_cutoff = 0.05, logFC_cutoff = log2(1.2), gspval_cutoff = 0.01,
  min_size = 10, task = "anal", ...)

prnGSPA(...)

prnGSPAHM(annot_cols = NULL, annot_colnames = NULL, ...)
}
\arguments{
\item{id}{Character string to indicate the type of data. The value will be
determined automatically. Peptide data will be used at \code{id = pep_seq}
or \code{pep_seq_mod}, and protein data will be used at \code{id = prot_acc}
or \code{gene}.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples.}

\item{df}{The name of input data file. By default, it will be determined
automatically by the value of \code{id}.}

\item{filepath}{Use system default.}

\item{filename}{Use system default.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used.}

\item{gset_nms}{Character vector containing the name(s) of gene sets for
enrichment analysis. The possible values are in \code{c("go_sets",
"kegg_sets", "c2_msig")}.}

\item{var_cutoff}{Not currently used.}

\item{pval_cutoff}{The cut-off in protein significance \code{pVal}. Entries
with \code{pVals} smaller than the threshold will be ignored.}

\item{logFC_cutoff}{The cut-off in protein \code{log2FC}. Entries with
\code{log2FC} smaller than the threshold will be ignored.}

\item{gspval_cutoff}{The cut-off in significance \code{pVal} of gene sets.
Gene sets with \code{pVals} smaller than the threshold will be ignored.}

\item{min_size}{Minimum number of protein entries for consideration of a gene
set test. The number is the sum of both up and down-expressed proteins after
data filtration by `pval_cutoff`, `logFC_cutoff` or varargs `filter_`.}

\item{...}{\code{filter_}: Logical expression(s) for the row filtration of
data; also see \code{\link{normPSM}}.}

\item{annot_cols}{A character vector of column keys that can be found in
\code{essmap_Protein_GSPA_...csv} output files. The values under the
selected keys will be used to color-code sample IDs on the top of heat maps
by \code{prnGSPAHM}.}

\item{annot_colnames}{A character vector of replacement name(s) to
\code{annot_cols}.}
}
\description{
\code{prnGSPA} performs the analysis of Gene Set Probability Asymmetricity
(GSPA) aganist protein \code{log2FC} data. Users should avoid call the method
directly, but instead use the following wrappers.

\code{prnGSPA} is a wrapper of \code{\link{proteoGSPA}} for enrichment analysis

\code{prnGSPAHM} is a wrapper of \code{\link{proteoGSPA}} for heat map and
network visualization
}
\details{
The significance \code{pVals} of proteins are first obtained from
\code{\link{prnSig}}, which involves linear modelings using
\code{\link[limma]{eBayes}} or \code{\link[lmerTest]{lmer}}. The geometric
mean of \code{pVals} are then each calculated for the groups of up or down
regulated proteins. The quotient of the two \code{pVals} is used to prepsent
the significance of gene set enrichment. The arguments \code{pval_cutoff} and
\code{logFC_cutoff} are used to filter out low influence genes.

The formula(s) of contrast(s) used in \code{\link{prnSig}} will be taken by
default.

\code{prnGSPAHM} visualizes the distance heat map between essential and all
gene sets. The distance is \eqn{1 - f} where \eqn{f} is the fraction overlap
for ids under a gene set.
}
\section{\code{Protein_GSPA_...csv}}{


 \tabular{ll}{ \strong{Key}   \tab \strong{Descrption}\cr term \tab the name
 of a gene set \cr is_essential \tab a logicial indicator of gene set
 essentiality \cr count \tab the number of entrez IDs with significance
 \code{pVal} <= \code{pval_cutoff} under a \code{term} \cr contrast \tab a
 contrast of sample groups \cr p_val \tab significance p values \cr q_val
 \tab \code{p_val} with \code{BH} adjustment for multiple tests \cr log2fc
 \tab the fold change of a gene set at logarithmic base of 2 \cr }
}

\section{\code{essmap_Protein_GSPA_...csv}}{


 \tabular{ll}{ \strong{Key}   \tab \strong{Descrption}\cr ess_term \tab the
 name of an essential gene set \cr term \tab the name of a gene set \cr
 ess_count \tab the number of entrez IDs under a \code{term} with matches to
 a correspoding \code{ess_term} \cr fraction \tab a fraction of matches in
 entrez IDs between a \code{term} and a \code{ess_term}; for example at
 \code{fraction = 1} between a \code{term}, \eqn{x}, and an \code{ess_term},
 \eqn{y}, all of the gene IDs under \eqn{x} are present in \eqn{y} \cr
 distance \tab 1 - \code{fraction} \cr source \tab a numeric index of
 essential gene sets \cr target \tab a numeric index of all gene sets \cr }
}

\examples{
# enrichment analysis
prnGSPA(
  scale_log2r = TRUE,
  impute_na = FALSE,
  pval_cutoff = 5E-2,
  gspval_cutoff = 5E-3,
  gset_nms = c("go_sets", "kegg_sets"),

  filter_by_npep = exprs(prot_n_pep >= 2),
  # `filename(s)` will be automated,
)


# distance heat map and network of GSPA terms
# a `term` is subset to an `ess_term` if the distance is zero
# `source` is a column key in `essmap_Protein_GSPA...csv`
prnGSPAHM(
  filter_by = exprs(distance <= .2),
  annot_cols = "source",
  annot_colnames = "Essential set index",

  fontsize = 16,
  fontsize_row = 20,
  fontsize_col = 20,
  fontsize_number = 8,
  border_color = "grey60",
  cellwidth = 24,
  cellheight = 24,
  filename = show_redundancy_at_small_dist.png,
)

# human terms only
prnGSPAHM(
  filter_num = exprs(distance <= .8),
  filter_sp = exprs(start_with_str("hs", term)),
  annot_cols = "source",
  annot_colnames = "Essential set index",

  fontsize = 16,
  fontsize_row = 20,
  fontsize_col = 20,
  fontsize_number = 8,
  border_color = "grey60",
  cellwidth = 24,
  cellheight = 24,
  filename = show_connectivity_at_large_dist.png,
)

# custom color palette
prnGSPAHM(
  annot_cols = c("source", "ess_count"),
  annot_colnames = c("Essential set index", "Count"),
  filter_by = exprs(ess_count >= 8, distance <= .80),

  fontsize = 16,
  fontsize_row = 20,
  fontsize_col = 20,
  fontsize_number = 8,
  border_color = "grey60",
  cellwidth = 24,
  cellheight = 24,
  color = colorRampPalette(c("blue", "white", "red"))(100),
  filename = "custom_colors.png"
)

\dontrun{
}

}
