% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-histo.R
\name{proteoHist}
\alias{proteoHist}
\alias{pepHist}
\alias{prnHist}
\title{Visualizes histograms}
\usage{
proteoHist(id = c("pep_seq", "pep_seq_mod", "prot_acc", "gene"),
  col_select = NULL, scale_log2r = FALSE, show_curves = TRUE,
  show_vline = TRUE, new_fit = FALSE, df = NULL, filepath = NULL,
  filename = NULL, ...)

pepHist(...)

prnHist(...)
}
\arguments{
\item{id}{Character string to indicate the type of data. Peptide data will be
used at \code{id = pep_seq} or \code{pep_seq_mod}, and protein data at
\code{id = prot_acc} or \code{gene}.}

\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
The default key is \code{Select}. Samples corresponding to non-empty entries
under \code{col_select} will be included in the indicated analysis.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2-ratios} to the same
scale of standard deviation for all samples.}

\item{show_curves}{Logical; if TRUE, shows the fitted curves.}

\item{show_vline}{Logical; if TRUE, shows the vertical lines at \code{x = 0}.}

\item{new_fit}{Not currently used.}

\item{df}{The file name of input data. By default, it will be determined by
the value of \code{id}.}

\item{filepath}{The file path to output results. By default, it will be
determined by the name of the current function \code{call} and the value of
\code{id}.}

\item{filename}{A representative file name to output images. By default, it
will be determined by the names of the current \code{call}. The images are
saved via \code{\link[ggplot2]{ggsave}} and the image type will be
determined by the extension of file names with the default being
\code{.png}.}

\item{...}{Additional parameters for plotting: \cr \code{xmin}, the minimum
\eqn{x} at a log2 scale; the default is -2. \cr \code{xmax}, the maximum
\eqn{x} at a log2 scale; the default is +2. \cr \code{x_breaks}, the breaks
in \eqn{x}-axis at a log2 scale; the default is 1. \cr \code{binwidth}, the
binwidth of \code{log2-ratios}; the default is \eqn{(xmax - xmin)/80}. \cr
\code{ncol}, the number of columns; the default is 1. \cr \code{width}, the
width of plot; \cr \code{height}, the height of plot.}
}
\value{
The histograms of \code{log2-ratios} under
 \code{~\\dat_dir\\Peptide\\Histogram} or
 \code{~\\dat_dir\\Protein\\Histogram}.
}
\description{
\code{proteoHist} prepares the histogram visualization of \code{log2-ratios}
for proteins or peptides data.

\code{pepHist} is a wrapper function of \code{\link{proteoHist}}

\code{prnHist} is a wrapper function of \code{\link{proteoHist}}
}
\details{
In the histograms, the \code{log2-ratios} under each TMT channel are
color-coded by their contributing reporter-ion intensity.

The function matches the current \code{id} to those in the latest \code{call}
to \code{\link{normPep}} or \code{\link{normPrn}}.  For example, if
\code{pep_seq} was used in \code{\link{normPep()}}, the current \code{id =
pep_seq_mod} will be matched to \code{id = pep_seq}.
}
\examples{
# without scaling normalization
pepHist(
  scale_log2r = FALSE,
  xmin = -1,
  xmax = 1,
  ncol = 5
)

# with scaling normalization
pepHist(
  scale_log2r = TRUE,
  xmin = -1,
  xmax = 1,
  ncol = 5
)

\dontrun{
pepHist(
  col_select = "a_non_existed_column_key"
)
}

# no scaling normalization
prnHist(
  ncol = 10
)

# scaling normalization
prnHist(
  scale_log2r = TRUE,
  ncol = 10
)


}
