% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-hm.R
\name{proteoKinHM}
\alias{proteoKinHM}
\title{Visualizes kinase heat maps}
\usage{
proteoKinHM(id = gene, col_select = NULL, col_benchmark = NULL,
  scale_log2r = TRUE, df = NULL, filepath = NULL, filename = NULL,
  complete_cases = FALSE, impute_na = FALSE, anal_type = "Heatmap",
  xmin = -1, xmax = 1, x_margin = 0.1, annot_cols = NULL, ...)
}
\description{
Specialized for kinase heat maps
}
