% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-ma.R
\name{proteoMA}
\alias{proteoMA}
\title{MA Plots}
\usage{
proteoMA(id = gene, col_select = NULL, col_group = NULL,
  scale_log2r = FALSE, df = NULL, filepath = NULL, filename = NULL,
  ...)
}
\arguments{
\item{id}{The name of a unique identifier (see \code{\link[proteoQ]{MDS}}).}

\item{scale_log2r}{Logical; if TRUE, rescales \code{log2-ratios} to the same
scale of standard deviation for all samples.}

\item{annot_kinases}{Logical; if TRUE, annotates proteins being kinases or
not.}
}
\value{
Images stored under the file folders that are associated to
  \code{id}, \code{anal_type} and \code{annot_kinases}.
}
\description{
\code{proteoMA} produces MA plots.
}
\details{
reads the data from either "\code{~\\Direcotry\\Peptide\\Peptide All.txt}" at
\code{id = pep_seq_mod}, or "\code{~\\Direcotry\\Protein\\Protein All by
Accession.txt}" at \code{id = prot_acc} or
"\code{~\\Direcotry\\Protein\\Protein All by gene.txt}" at \code{id = gene}.
}
\examples{
MA(
	id = gene,
	scale_log2r = scale_log2r,
	annot_kinases = annot_kinases,
)

\dontrun{
}
}
