% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-nmf.R
\name{proteoNMF}
\alias{proteoNMF}
\title{NMF}
\usage{
proteoNMF(id = c("pep_seq", "pep_seq_mod", "prot_acc", "gene"),
  col_select = NULL, scale_log2r = FALSE, impute_na = TRUE,
  complete_cases = FALSE, df = NULL, filepath = NULL,
  filename = NULL, r = 4, nrun = 200, xmin = -1, xmax = 1,
  x_margin = 0.1, annot_cols = NULL, width_consensus = 6,
  height_consensus = 6, width_coefmap = 6, height_coefmap = 6, ...)
}
\arguments{
\item{id}{The name of a unique identifier (see \code{\link[proteoQ]{MDS}}).}

\item{scale_log2r}{Logical; if TRUE, rescales \code{log2-ratios} to the same
scale of standard deviation for all samples.}

\item{annot_kinases}{Logical; if TRUE, annotates proteins being kinases or
not.}
}
\value{
Images stored under the file folders that are associated to
  \code{id}, \code{anal_type} and \code{annot_kinases}.
}
\description{
\code{proteoNMF} produces heat maps.
}
\details{
reads the data from either "\code{~\\Direcotry\\Peptide\\Peptide All.txt}" at
\code{id = pep_seq_mod}, or "\code{~\\Direcotry\\Protein\\Protein All by
Accession.txt}" at \code{id = prot_acc} or
"\code{~\\Direcotry\\Protein\\Protein All by gene.txt}" at \code{id = gene}.
}
\examples{
MA(
	id = gene,
	scale_log2r = scale_log2r,
	annot_kinases = annot_kinases,
)

# or use the form of functional factories
	my_prnMA <- info_anal(df = NULL, id = prot_acc, scale_log2r, filepath = NULL, filename = NULL, annot_kinases, anal_type = "MA")
	my_prnMA()

	my_pepMA <- info_anal(df = NULL, id = pep_seq_mod, scale_log2r, filepath = NULL, filename = NULL, annot_kinases, anal_type = "MA")
	my_pepMA()

\dontrun{
}
}
