% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-mds.R
\name{proteoPCA}
\alias{proteoPCA}
\title{Visualizes PCA plots}
\usage{
proteoPCA(id = gene, col_select = NULL, col_group = NULL,
  col_color = NULL, col_fill = NULL, col_shape = NULL,
  col_size = NULL, col_alpha = NULL, scale_log2r = TRUE,
  show_ids = TRUE, annot_cols = NULL, df = NULL, filepath = NULL,
  filename = NULL, ...)
}
\arguments{
\item{id}{Character string to indicate the type of data. Peptide data will be
used at column \code{id = "pep_seq"} or \code{"pep_seq_mod"}, and protein
data at \code{id = "prot_acc"} or \code{"gene"}.}

\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
The default key is \code{Select}. Samples corresponding to non-empty entries
under \code{col_select} will be included in the indicated analysis.}

\item{col_group}{Not currently used.}

\item{col_color}{Character string to a column key in \code{expt_smry.xlsx}.
The default key is \code{Color}. Values under which will be used for the
\code{color} aesthetics in \code{ggplot2}.}

\item{col_fill}{Character string to a column key in \code{expt_smry.xlsx}.
The default key is \code{Fill}. Values under which will be used for the
\code{fill} aesthetics in \code{ggplot2}.}

\item{col_shape}{Character string to a column key in \code{expt_smry.xlsx}.
The default key is \code{Shape}. Values under which will be used for the
\code{shape} aesthetics in \code{ggplot2}.}

\item{col_size}{Character string to a column key in \code{expt_smry.xlsx}.
The default key is \code{Size}. Values under which will be used for the
\code{size} aesthetics in \code{ggplot2}.}

\item{col_alpha}{Character string to a column key in \code{expt_smry.xlsx}.
The default key is \code{Alpha}. Values under which will be used for the
\code{alpha} aesthetics in \code{ggplot2}.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2-ratios} to the same
scale of standard deviation for all samples.}

\item{show_ids}{Logical; if TRUE, shows the sample IDs in \code{MDS/PCA}
plots.}

\item{annot_cols}{Not currently used.}

\item{df}{The filename of input data. By default, it will be determined by the
value of \code{id}.}

\item{filepath}{The filepath to output results. By default, it will be
determined by the name of the current function \code{call} and the value of
\code{id}.}

\item{filename}{A representative filename to output images. By default, it
will be determined by the names of the current \code{call}.}

\item{...}{Parameters inherited from \code{ggsave}: \cr \code{width}, the
width of plot; \cr \code{height}, the height of plot \cr \code{...}}
}
\value{
PCA plots.
}
\description{
\code{protePCA} visualizes the results from principal component analysis
(PCA).
}
\details{
\code{log2-ratios} are used in PCA (\code{\link[stats]{prcomp}}).
}
