% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-purge.R
\name{proteoPurge}
\alias{proteoPurge}
\alias{purPep}
\alias{purPrn}
\title{Purge ata}
\usage{
proteoPurge(id = c("pep_seq", "pep_seq_mod", "prot_acc", "gene"),
  cv_cutoff = NULL, nseq_cutoff = 1, df = NULL, filepath = NULL,
  filename = NULL, ...)

purPep(...)

purPrn(...)
}
\arguments{
\item{id}{Character string to indicate the type of data. Peptide data will be
used at \code{id = pep_seq} or \code{pep_seq_mod}, and protein data at
\code{id = prot_acc} or \code{gene}.}

\item{cv_cutoff}{Numeric; the cut-off of CV. The CVs are from the ascribing
PSMs for peptide data or ascribing peptides for protein data.}

\item{nseq_cutoff}{Positive integer; for peptide data, peptide entries with the
number of identifying PSMs smaller than the cutoff will be removed; for
protein data, protein entries with the number of identifying peptides
smaller than the cutoff will be removed.}

\item{df}{The name of input data file. By default, it will be determined by
the value of \code{id}.}

\item{filepath}{A file path to output results. By default, it will be
determined by the name of the calling \code{function} and the value of
\code{id} in the \code{call}.}

\item{filename}{A representative file name to output image(s). By default, it
will be determined by the name of the current \code{call}. The image(s) are
saved via \code{\link[ggplot2]{ggsave}} where the image type will be
determined by the extension of the file name. A \code{.png} format will be
used at default or an unrecognized file extension.}

\item{...}{Additional parameters for plotting: \cr \code{xmin}, the minimum
\eqn{x} at a log2 scale; the default is -2. \cr \code{xmax}, the maximum
\eqn{x} at a log2 scale; the default is +2. \cr \code{x_breaks}, the breaks
in \eqn{x}-axis at a log2 scale; the default is 1. \cr \code{binwidth}, the
binwidth of \code{log2FC}; the default is \eqn{(xmax - xmin)/80}. \cr
\code{ncol}, the number of columns; the default is 1. \cr \code{width}, the
width of plot; \cr \code{height}, the height of plot.}
}
\description{
\code{proteoPurge} provides additional cleanup of protein or peptide data.

\code{purPep} is a wrapper of \code{\link{proteoPurge}} for peptide data

\code{purPrn} is a wrapper of \code{\link{proteoPurge}} for protein data
}
\details{
The function matches the current \code{id} to those in the latest \code{call}
to \code{\link{normPep}} or \code{\link{normPrn}}.  For example, if
\code{pep_seq} was used in \code{\link{normPep()}}, the current \code{id =
pep_seq_mod} will be matched to \code{id = pep_seq}.
}
\examples{
\dontrun{
purPep(
  cv_cutoff = 0.4,
  nseq_cutoff = 1,
)
}

\dontrun{
purPrn(
  cv_cutoff = 0.4,
  nseq_cutoff = 1,
)
}

}
