% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-string.R
\name{proteoString}
\alias{proteoString}
\alias{anal_prnString}
\title{STRING outputs}
\usage{
proteoString(id = c("pep_seq", "pep_seq_mod", "prot_acc", "gene"),
  task = "anal", col_select = NULL, col_group = NULL,
  col_order = NULL, impute_na = FALSE, complete_cases = FALSE,
  db_path = "~\\\\proteoQ\\\\dbs\\\\string", score_cutoff = 0.7,
  scale_log2r = TRUE, df = NULL, filepath = NULL, filename = NULL,
  ...)

anal_prnString(...)
}
\arguments{
\item{id}{Character string to indicate the type of data. The value will be
determined automatically.}

\item{col_select}{Not currently used.}

\item{col_group}{Not currently used.}

\item{col_order}{Not currently used.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used. The default is FALSE.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{db_path}{Character string; the local path for database(s). The default
is \code{"~\\proteoQ\\dbs\\string"}.}

\item{score_cutoff}{Numeric; the threshold in the \code{combined_score} of
protein-protein interaction. The default is 0.7.}

\item{scale_log2r}{Not currently used.}

\item{filename}{Use system default. Otherwise, the basename will be prepended
to \code{_[species]_ppi.tsv} for network data and \code{_[species]_expr.tsv}
for expression data.}

\item{...}{\code{filter_}: Logical expression(s) for the row filtration of
data; also see \code{\link{normPSM}}.}
}
\description{
\code{proteoString} prepares the data of both
\code{\href{https://string-db.org/}{STRING}} protein-protein interactions
(ppi) and companion protein expressions. Users should avoid calling the method
directly, but instead use the following wrappers.

\code{anal_prnString} is a wrapper of \code{\link{proteoString}} for the
STRING analysis of protein data
}
\details{
Convenience features, such as data row filtration via \code{filter_} varargs,
are available. The ppi file, \code{..._ppi.tsv}, and the expression file,
\code{..._expr.tsv}, are also compatible with a third-party
\code{\href{https://cytoscape.org/}{Cytoscape}}.
}
\examples{
# ===================================
# String DB
# ===================================

## !!!require the brief working example in `?load_expts`

string_dir <- "~\\\\proteoQ\\\\dbs\\\\string"
dir.create(string_dir, recursive = TRUE, showWarnings = FALSE)

## download
dl_stringdbs(
  species = human,
  db_path = string_dir,
)

## ppi and expression
# human and mouse and probably 'unknown' species
anal_prnString(
  db_path = string_dir,
  score_cutoff = .9,
  filter_prots_by = exprs(prot_n_pep >= 2),
)

# cleaner outputs by specifying species
# ('unknown' will be removed)
anal_prnString(
  db_path = string_dir,
  score_cutoff = .9,
  filter_by_sp = exprs(species == "human"),
  filter_prots_by = exprs(prot_n_pep >= 2),
)

anal_prnString(
  db_path = string_dir,
  score_cutoff = .9,
  filter_by_sp = exprs(species == "mouse"),
  filter_prots_by = exprs(prot_n_pep >= 2),
)

# custom filename with outputs in: 
# prot_npep2 W16_vs_W2.pVal (W16-W2)_human_expr.tsv
# prot_npep2 W16_vs_W2.pVal (W16-W2)_human_ppi.tsv
# prot_npep2 W16_vs_W2.pVal (W16-W2)_mouse_expr.tsv
# prot_npep2 W16_vs_W2.pVal (W16-W2)_mouse_ppi.tsv
anal_prnString(
  db_path = string_dir,
  score_cutoff = .9,
  complete_cases = TRUE, 
  filter_by = exprs(species \%in\% c("human", "mouse"), 
                    prot_n_pep >= 2, 
                    `W16_vs_W2.pVal (W16-W2)` <= 1e-6),
  filename = "prot_npep2 W16_vs_W2.pVal (W16-W2).txt",
)
}
\seealso{
\code{\link{dl_stringdbs}} for database downloads. \cr
 \code{\link{prnSig}} for significance tests \cr
}
