% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{psm_mcleanup}
\alias{psm_mcleanup}
\title{Parallel PSM cleanup}
\usage{
psm_mcleanup(file, rm_outliers, group_psm_by, dat_dir, TMT_plex, rm_allna)
}
\arguments{
\item{file}{A PSM file (name).}

\item{rm_outliers}{Logical; if TRUE, PSM outlier removals will be performed
for peptides with more than two identifying PSMs. Dixon's method will be
used when \eqn{2 < n \le 25} and Rosner's method will be used when \eqn{n >
25}. The default is FALSE.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{TMT_plex}{Numeric; the multiplexity of TMT, i.e., 10, 11 etc.}
}
\description{
Parallel PSM cleanup
}
