% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{psm_msplit}
\alias{psm_msplit}
\title{Parallel split of PSM tables}
\usage{
psm_msplit(df, nm, fn_lookup, dat_dir, plot_rptr_int, TMT_plex)
}
\arguments{
\item{df}{A list of data frames.}

\item{nm}{Names of the \code{df}'s.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{plot_rptr_int}{Logical; if TRUE, the distributions of reporter-ion
intensities will be plotted. The default is TRUE.}

\item{TMT_plex}{Numeric; the multiplexity of TMT, i.e., 10, 11 etc.}
}
\description{
Parallel split of PSM tables
}
