% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{psm_to_pep}
\alias{psm_to_pep}
\title{Helper of PSM2Pep}
\usage{
psm_to_pep(
  file = NULL,
  dat_dir = NULL,
  label_scheme_full = NULL,
  group_psm_by = "pep_seq",
  group_pep_by = "prot_acc",
  method_psm_pep = "median",
  lfq_ret_tol = 60L,
  rm_allna = FALSE,
  type_sd = "log2_R",
  engine = "mz",
  lfq_mbr = FALSE,
  ...
)
}
\arguments{
\item{file}{The name of a PSM file.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{label_scheme_full}{The label_scheme with the probable inclusion of
different LCMS_inj under the same TMT_Set.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{method_psm_pep}{Character string; the method to summarize the
\code{log2FC} and the \code{intensity} of \code{PSMs} by peptide entries.
The \code{log10-intensity} of reporter ions at the \code{PSMs} levels will 
be the weight when summarizing \code{log2FC} with various \code{"top_n"} 
statistics or \code{"weighted_mean"}. The choices of summary statistics for
 LFQ are depreciated at the users' levels.}

\item{lfq_ret_tol}{The tolerance of retention time (in seconds) for the
aggregation of LFQ data.}

\item{rm_allna}{Logical; if TRUE, removes data rows that are exclusively NA
across ratio columns of \code{log2_R126} etc. The setting also applies to
\code{log2_R000} in LFQ.}

\item{type_sd}{Character string; the type of log2Ratios for SD calculations.
The value is one \code{log2_R}, \code{N_log2_R} or \code{Z_log2_R}.}

\item{lfq_mbr}{Logical; performs MBR for LFQ or not.}

\item{...}{filter_dots.}
}
\description{
Helper of PSM2Pep
}
