% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{purge_search_space}
\alias{purge_search_space}
\title{Subset the search space.}
\usage{
purge_search_space(i, aa_masses, mgf_path, n_cores, ppm_ms1, fmods_nl = NULL)
}
\arguments{
\item{i}{Integer; the index for a set of corresponding aa_masses and
theoretical peptides.}

\item{aa_masses}{Amino-acid lookup at a given combination of fixed and
variable.}

\item{mgf_path}{The file path to a list of MGF files. There is no default and
the experimenters need to supply the files.}

\item{n_cores}{The number of CPU cores.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}

\item{fmods_nl}{The attribute of \code{fmods_nl} from an \code{aa_masses}.}
}
\description{
Subset the search space.
}
