% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-purge.R
\name{purgePep}
\alias{purgePep}
\title{Purge peptide data}
\usage{
purgePep(dat_dir = NULL, max_cv = NULL, min_n = 1, adjSD = FALSE,
  keep_ohw = TRUE, col_select = NULL, col_order = NULL,
  filename = NULL, ...)
}
\arguments{
\item{dat_dir}{A character string to the working directory.}

\item{max_cv}{Numeric; the cut-off in CV. The CV are from the ascribing PSMs
for the filtration of peptide entries or ascribing peptides for the
filtration of protein entries.}

\item{min_n}{Positive integer. When calling from \code{purgePSM}, peptide
entries in PSM tables with the number of identifying PSMs smaller than
\code{min_n} will be replaced with NA. When calling from \code{purgePep},
protein entries in peptide tables with the number of identifying peptides
smaller than \code{min_n} will be replaced with NA.}

\item{adjSD}{Logical; if TRUE, adjust the standard deviation in relative to
the width of ratio profiles.}

\item{keep_ohw}{Logical; if TRUE, keep one-hit-wonders with unknown CV. The
default is TRUE.}

\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under the column key will be
included in the indicated analysis. At the \code{NULL} default, the column
key will be \code{Select}.}

\item{filename}{A representative file name to output image(s). By default, it
will be determined automatically by the name of the current \code{call}. The
image(s) are saved via \code{\link[ggplot2]{ggsave}} where the image type
will be determined by the extension of the file name. A \code{.png} format
will be used at default or at an unrecognized file extension.}

\item{...}{\code{filter_}: Logical expression(s) for the row filtration of
data; also see \code{\link{normPSM}}. Additional parameters for plotting:
\cr \code{ymax}, the maximum \eqn{y} at a log2 scale; the default is +0.6.
\cr \code{ybreaks}, the breaks in \eqn{y}-axis at a log2 scale; the default
is 0.2. Note: \code{y_breaks} depreciated. \cr \code{width}, the width of
plot. \cr \code{height}, the height of plot. \cr \code{flip_coord}, logical;
if TRUE, flip \code{x} and \code{y} axis.}
}
\description{
\code{purgePep} removes \code{protein} entries from peptide table(s) by
selection criteria.
}
\details{
The CV of proteins are calculated from contributing peptides at the basis of
per TMT experiment per LCMS run.
}
