% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-purge.R
\name{purgePSM}
\alias{purgePSM}
\title{Purge PSM data}
\usage{
purgePSM(dat_dir = NULL, pt_cv = NULL, max_cv = NULL,
  adjSD = FALSE, keep_ohw = TRUE, ...)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{pt_cv}{Numeric between 0 and 1; the percentile of CV. Values above the
percentile threshold will be replaced with NA. The default is NULL with no
data trimming by CV percentile.}

\item{max_cv}{Numeric; the cut-off in maximum CV. Values above the threshold
will be replaced with NA. The default is NULL with no data trimming by max
CV.}

\item{adjSD}{Logical; if TRUE, adjust the standard deviation in relative to
the width of ratio profiles.}

\item{keep_ohw}{Logical; if TRUE, keep one-hit-wonders with unknown CV. The
default is TRUE.}

\item{...}{Additional parameters for plotting: \cr \code{ymax}, the maximum
\eqn{y} at a log2 scale. \cr \code{ybreaks}, the breaks in \eqn{y}-axis at a
log2 scale. \cr \code{width}, the width of plot. \cr \code{height}, the
height of plot. \cr \code{flip_coord}, logical; if TRUE, flip \code{x} and
\code{y} axis.}
}
\description{
\code{purgePSM} removes \code{peptide} entries from PSM tables by selection
criteria. It further plots the distributions of \code{log2FC} by TMT
experiments and LC/MS series.
}
\details{
The CV of peptides are calculated from contributing PSMs at the basis of per
TMT experiment per series of LC/MS. Note that greater CV may be encountered
for samples that are more different to reference material(s).
}
\examples{
# ===================================
# PSM cleanup by CV
# ===================================

## !!!require the brief working example in `?load_expts`

## additional examples
# visualization only
purgePSM()

# by percent CV
purgePSM(pt_cv = .95)

# by max CV
purgePSM(max_cv = .5)

# by `max_cv` then by `pt_cv`
purgePSM(max_cv = .5)
purgePSM(pt_cv = .95)

# actually 90\% CV 
purgePSM(pt_cv = .95)
purgePSM(pt_cv = .95)

}
\seealso{
\code{\link{load_expts}} for a minimal working example in data normalization \cr
 \code{\link{normPSM}} for extended examples in PSM data normalization \cr
 \code{\link{purgePSM}} for extended examples in PSM data purging \cr
 \code{\link{PSM2Pep}} for extended examples in PSM to peptide summarization \cr 
 \code{\link{mergePep}} for extended examples in peptide data merging \cr 
 \code{\link{standPep}} for extended examples in peptide data normalization \cr
 \code{\link{pepHist}} for extended examples in peptide data histogram visualization. \cr 
 \code{\link{purgePep}} for extended examples in peptide data purging \cr
 \code{\link{Pep2Prn}} for extended examples in peptide to protein summarization \cr
 \code{\link{standPrn}} for extended examples in protein data normalization. \cr 
 \code{\link{prnHist}} for extended examples in protein data histogram visualization.
}
