% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{q_grubbs}
\alias{q_grubbs}
\title{Distribution function}
\usage{
q_grubbs(p, n, type = 10, rev = FALSE)
}
\arguments{
\item{p}{Vector of probabilities. See also \link[outliers]{dixon.test}.}

\item{n}{Length of sample Length of sample.}

\item{type}{The type of Dixon method. See also \link[outliers]{dixon.test}.}

\item{rev}{Logical; if TRUE, acts as p_dixon.}
}
\description{
Distribution function
}
