% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{q_table}
\alias{q_table}
\title{Distribution table}
\usage{
q_table(p, probs, quants)
}
\arguments{
\item{p}{Vector of probabilities. See also \link[outliers]{dixon.test}.}

\item{probs}{Vector of given probabilities.}

\item{quants}{Vector of given corresponding quantiles.}
}
\description{
Distribution table
}
