% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{quick_leftjoin}
\alias{quick_leftjoin}
\title{Left-joining of two data frames.}
\usage{
quick_leftjoin(x, y, by = NULL)
}
\arguments{
\item{x}{The left data frame.}

\item{y}{The right data frame.}

\item{by}{The key.}
}
\description{
Left-joining of two data frames.
}
