% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{quick_rightjoin}
\alias{quick_rightjoin}
\title{Right-joining of two data frames.}
\usage{
quick_rightjoin(x, y, by = NULL)
}
\arguments{
\item{x}{The left data frame (to be proliferated by rows).}

\item{y}{The right data frame (the dominated one).}

\item{by}{The key.}
}
\description{
Rows ordered by \code{y}, which is different to \link[dplyr]{right_join}.
}
\examples{
\donttest{
df1 <- data.frame(A = c("a", "b", "c"), B = c(1, 1, 1))
df2 <- data.frame(A = c("a", "c", "d"), C = c(2, 2, "3"))

x1 <- quick_rightjoin(df1, df2, by = "A")
x1 <- x1[, c("A", "B", "C")]
rownames(x1) <- seq_len(nrow(x1))

x2 <- dplyr::right_join(df1, df2, by = "A")
x2 <- x2[, c("A", "B", "C")]

stopifnot(identical(x1, x2))

# row order may be different
x1 <- quick_rightjoin(df2, df1, by = "A")
x1 <- x1[, c("A", "B", "C")]
rownames(x1) <- seq_len(nrow(x1))

x2 <- dplyr::right_join(df2, df1, by = "A")
x2 <- x2[, c("A", "B", "C")]

# FALSE
identical(x1, x2)
}
}
