% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{ratio_toCtrl}
\alias{ratio_toCtrl}
\title{Calculate the log2-ratio to the control group of samples}
\usage{
ratio_toCtrl(df, id, label_scheme_sub, nm_ctrl)
}
\description{
Calculate the log2-ratio to the control group of samples
}
\examples{
ratio_toCtrl(df, "gene", label_scheme_sub, Heatmap_Group)
}
