% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Reads the metadata of expt_smry or frac_smry.}
\usage{
read_metadata(
  filename = "expt_smry.xlsx",
  dat_dir = NULL,
  ext = "xlsx",
  metatype = "expt"
)
}
\arguments{
\item{filename}{The filename of metadata.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{ext}{The extension of a file name.}

\item{metatype}{The type of metadata in one of "expt" or "frac".}
}
\description{
Reads the metadata of expt_smry or frac_smry.
}
