% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{read_mgf_chunks}
\alias{read_mgf_chunks}
\title{Reads mgfs in chunks.}
\usage{
read_mgf_chunks(
  filepath = "~/proteoQ/mgfs",
  topn_ms2ions = 100,
  ret_range = c(0, Inf)
)
}
\arguments{
\item{filepath}{The file path to a list of mgf files.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{ret_range}{The range of retention time in seconds.}
}
\description{
Reads mgfs in chunks.
}
