% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{readMGF}
\alias{readMGF}
\title{Reads MGF files in chunks.}
\usage{
readMGF(
  filepath = "~/proteoQ/mgfs",
  min_mass = 516.2405,
  topn_ms2ions = 100,
  ret_range = c(0, Inf),
  ppm_ms1 = 20,
  out_path = file.path(filepath, "mgf_queries.rds")
)
}
\arguments{
\item{filepath}{The file path to a list of mgf files.}

\item{min_mass}{Numeric; the minimum mass of MS1 species. The value needs to
match the one in  \link{binTheoPeps}.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches.}

\item{ret_range}{The range of retention time in seconds.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}
}
\description{
Reads MGF files in chunks.
}
\examples{
\donttest{
mgf_queries <- readMGF()
}
}
