% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{readMGF_br}
\alias{readMGF_br}
\title{Add the MS file name to timsTOF mgf files.}
\usage{
readMGF_br(
  filepath = "~/proteoQ/mgf_test",
  min_mass = 516.2405,
  ret_range = c(0, Inf)
)
}
\arguments{
\item{filepath}{The file path to a list of mgf files.}

\item{min_mass}{Numeric; the minimum mass of MS1 species. The value needs to
match the one in  \link{binTheoPeps}.}

\item{ret_range}{The range of retention time in seconds.}
}
\description{
Uses base R readLines (to avoid memory failure with stringi).
}
\examples{
\donttest{
mgf_queries <- readMGF()
}
}
