% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated.R
\name{recalc_ms1mass}
\alias{recalc_ms1mass}
\title{Recalculates the MS1 precursor masses.}
\usage{
recalc_ms1mass(
  df,
  maxn_vmods_setscombi = 64,
  mod_indexes = NULL,
  include_insource_nl = FALSE,
  maxn_vmods_per_pep = 5,
  maxn_sites_per_vmod = 3,
  parallel = TRUE,
  digits = 5
)
}
\arguments{
\item{df}{A data frame.}

\item{maxn_vmods_setscombi}{Integer; the maximum number of combinatorial variable
modifications and neutral losses.}

\item{mod_indexes}{Integer; the indexes of fixed and/or variable
modifications.}

\item{include_insource_nl}{Logical Logical; if TRUE, includes MS1 precursor
masses with the losses of neutral species prior to MS2 fragmentation. The
default is FALSE. The setting at TRUE remains experimenting by allowing
additional masses in the universe of MS1 precursors. The offsets in NLs (by
adding back precursor masses) have not yet taken into account in MS2 ion
searches. A more systemically approach such as \code{open} MS1 masses might
be developed in the future.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{parallel}{Logical; if TRUE, performs parallel computation. The default
is TRUE.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Incorrect without permutation.
}
