% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{recheck_tmt_channels}
\alias{recheck_tmt_channels}
\title{Rechecks the completeness of TMT_Channel.}
\usage{
recheck_tmt_channels(df, TMT_plex)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}

\item{TMT_plex}{Numeric; the multiplexity of TMT.}
}
\description{
Rows may be deleted later. Checks the completeness of TMT_Channel (RAW_File
is either all NA or all filled)
}
