% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlfq.R
\name{reduce_lfq_rtstep}
\alias{reduce_lfq_rtstep}
\title{Reduce the step size of retention times for binning}
\usage{
reduce_lfq_rtstep(
  xs,
  ixs,
  lwr,
  step,
  fcts = c(0.8, 0.6, 0.4, 0.2, 0.1, 0.05, 0.01, 0.005, 0.001)
)
}
\arguments{
\item{xs}{Retention times.}

\item{ixs}{Indexes of retention times.}

\item{lwr}{The lower limit for indexing \code{xs}.}

\item{step}{The step size of binning.}

\item{fcts}{The scaling factor for reduced step size.}
}
\description{
Handling the case of multiple MS1 features in the same bin
}
