% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{refseq_entrez_hs}
\alias{refseq_entrez_hs}
\title{Lookups among RefSeq accessions, Entrez IDs and gene names for human
proteins.}
\format{A data frame with 31833 rows and 4 variables:
\describe{
  \item{entrez}{Entrez ID}
  \item{refseq_acc}{RefSeq accession number}
  \item{gene}{gene names}
}}
\usage{
refseq_entrez_hs
}
\description{
A dataset containing human \code{refseq} accessions, \code{entrez} IDs and
\code{gene} names.
}
\keyword{datasets}
