% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-datasets.R
\docType{data}
\name{refseq_mm_crossref}
\alias{refseq_mm_crossref}
\title{Lookups among RefSeq accessions, Entrez IDs and gene names for mouse
proteins.}
\format{A data frame with 24906 rows and 4 variables: \describe{
  \item{entrez}{Entrez ID} \item{refseq_acc}{RefSeq accession number}
  \item{gene}{gene names} \item{organism}{the organism attribute according to
  UniProt} }}
\usage{
refseq_mm_crossref
}
\description{
A dataset containing mouse \code{refseq} accessions, \code{entrez} IDs and
\code{gene} names.
}
\keyword{datasets}
