% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{reloc_col}
\alias{reloc_col}
\title{Relocates column "m/z" to be immediately before column "Mass".}
\usage{
reloc_col(df = NULL, from = "m/z", to = "Mass")
}
\arguments{
\item{df}{The original data frame.}

\item{from}{The column to be moved from.}

\item{to}{The column to which the \code{from} will be moved before.}
}
\description{
Not currently used.
}
