% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{reloc_col_after_first}
\alias{reloc_col_after_first}
\title{Relocates column "to_move" immediately after the first column.}
\usage{
reloc_col_after_first(df, to_move = "after_anchor")
}
\arguments{
\item{df}{The original data frame.}

\item{to_move}{The column to be moved.}
}
\description{
Relocates column "to_move" immediately after the first column.
}
