% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{reloc_col_before_first}
\alias{reloc_col_before_first}
\title{Relocate column "to_move" immediately before the first column.}
\usage{
reloc_col_before_first(df, to_move = "after_anchor")
}
\arguments{
\item{df}{The original data frame.}

\item{to_move}{The column to be moved.}
}
\description{
Relocate column "to_move" immediately before the first column.
}
