% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{reorder_files}
\alias{reorder_files}
\title{Re-order file names}
\usage{
reorder_files(filelist, n_TMT_sets)
}
\arguments{
\item{filelist}{A list of file names.}

\item{n_TMT_sets}{the number of multiplex TMT experiments.}
}
\description{
\code{reorder_files} re-orders file names by TMT set numbers then by LCMS
injection numbers.
}
