% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reorder_files2}
\alias{reorder_files2}
\title{Re-order file names}
\usage{
reorder_files2(filelist)
}
\arguments{
\item{filelist}{A list of file names (with probable prepending paths).}
}
\description{
By TMT set numbers then by LCMS injection numbers.
}
