% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{reorderCols}
\alias{reorderCols}
\title{Re-order columns in a data frame}
\usage{
reorderCols(df, endColIndex, col_to_rn)
}
\arguments{
\item{endColIndex}{the indeces of columns to be moved to the end of \code{df}.}

\item{col_to_rn}{the column identifier where the values under that column will be used as row names.}

\item{filelist}{A list of file names.}
}
\description{
\code{reorder_files} re-ordered file names by TMT set numbers then by injection numbers.
}
