% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{reorderCols2}
\alias{reorderCols2}
\title{Re-order columns in a data frame}
\usage{
reorderCols2(df = NULL, pattern = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{pattern}{columns matched the pattern will be moved to the end of
\code{df}.}
}
\description{
\code{reorderCols2} re-orders columns in a data frame.
}
