% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{rep_ls_groups}
\alias{rep_ls_groups}
\title{Replicates new label_scheme with new Sample_IDs by group_ids.}
\usage{
rep_ls_groups(label_scheme, group_ids)
}
\arguments{
\item{label_scheme}{Experiment summary.}

\item{group_ids}{A character vector; for example, c("heavy", "light") that
can be found from column \code{pep_group} in \code{df}; \code{group_ids} is
NULL if column \code{pep_group} is not in \code{df}.}
}
\description{
The new Sample_ID: "Sample [heavy]", "Sample [light]" etc.
}
