% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{replace_by_rowone}
\alias{replace_by_rowone}
\title{Replaces values by the first row.}
\usage{
replace_by_rowone(df, col_nms)
}
\arguments{
\item{df}{A data frame.}

\item{col_nms}{The column names under which the values in \code{df} will be
replaced with those in the first row.}
}
\description{
Replaces values by the first row.
}
