% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{replace_cols_at}
\alias{replace_cols_at}
\title{Replace columns in the original PSM table.}
\usage{
replace_cols_at(df, df2, idxs)
}
\arguments{
\item{df}{The original data frame.}

\item{df2}{The data columns to replace those in \code{df}.}

\item{idxs}{The sequences of column indexes in \code{df}. Note that
\code{idxs} need to be a continuous sequences.}
}
\description{
The column index(es) need to be continuous.
}
