% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{replace_na_with_empty}
\alias{replace_na_with_empty}
\title{Replaces NA sample IDs with Empty.1 etc.}
\usage{
replace_na_with_empty(x, prefix = "Empty")
}
\arguments{
\item{x}{A vector of Sample_ID.}

\item{prefix}{A character string prefixing \code{Empty} samples.}
}
\description{
Replaces NA sample IDs with Empty.1 etc.
}
