% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_NorZ_names}
\alias{replace_NorZ_names}
\title{Replaces column names to get Sample_ID(s).}
\usage{
replace_NorZ_names(NorZ_ratios, nms)
}
\arguments{
\item{NorZ_ratios}{A string.}

\item{nms}{A vector of names.}
}
\description{
Replaces column names to get Sample_ID(s).
}
