% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{reproc_psmC}
\alias{reproc_psmC}
\title{Reprocessing of \code{psmC.txt}.}
\usage{
reproc_psmC(out_path = NULL, fdr_type = "protein", combine_tier_three = FALSE)
}
\arguments{
\item{out_path}{The output path.}

\item{fdr_type}{Character string; the type of FDR control. The value is in
one of c("psm", "peptide", "protein"). Note that \code{fdr_type = protein}
is equivalent to \code{fdr_type = peptide} with the additional filtration
of data at \code{prot_tier == 1}. A variant is to set \code{fdr_type =
psm}, followed by a data filtration at \code{prot_tier == 1}.}

\item{combine_tier_three}{Logical; if TRUE, combines all protein results to
the output of \code{psmQ.txt}. Outputs under the option TRUE are often
comparable to Mascot outputs with FDR controls at the levels of PSMs or
peptides.}
}
\description{
Protein grouping from \code{psmC.txt} to \code{psmQ.txt}.
}
\details{
May solve some memory shortage issues for large data sets (e.g., over a
million peptide sequences * 35000 proteins from \code{psmC.txt}).
}
