% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs2.R
\name{rev_pepseqs}
\alias{rev_pepseqs}
\title{Reverses peptide sequences.}
\usage{
rev_pepseqs(xs)
}
\arguments{
\item{xs}{Lists of proteins with named peptide masses under each entry.}
}
\description{
Reverses peptide sequences.
}
