% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{rm_2RToutliers}
\alias{rm_2RToutliers}
\title{Remove retention-time outliers between two}
\usage{
rm_2RToutliers(mx, my, mt, ms, mf, mn, zero = NA_real_)
}
\arguments{
\item{mx}{A two-row matrix of m-over-z values with large standard deviations
in retention times.}

\item{my}{A two-row matrix of apex intensity values.}

\item{mt}{A two-row matrix of apex retention times.}

\item{ms}{A two-row matrix of apex scan numbers.}

\item{mf}{A two-row matrix of apex FWHM values.}

\item{mn}{A two-row matrix of apex lengths.}

\item{zero}{A zero value for outlier replacements.}
}
\description{
Remove retention-time outliers between two
}
