% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{rm_apexrt_outliers}
\alias{rm_apexrt_outliers}
\title{Remove PSM retention outliers}
\usage{
rm_apexrt_outliers(df = NULL, key = "pep_seq_mod", rt_tol = 240, y_tol = 0.01)
}
\arguments{
\item{df}{A PSM table at a TMTset[i]LCMSinj[j].}

\item{key}{A column key for data grouping.}

\item{rt_tol}{Tolerance in retention times between \code{pep_ret_range}  and
\code{pep_apex_ret} in seconds.}

\item{y_tol}{Tolerance in intensity in relative to the based apex intensity.}
}
\description{
Nullify the apexes with large distance between \code{pep_ret_range} and
\code{pep_apex_ret}. Also remove the entries of very low intensity entries
(in multi-PSM groups of \code{pep_seq_mod}).
}
