% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs2.R
\name{rm_char_in_nfirst2}
\alias{rm_char_in_nfirst2}
\title{Remove a starting character from the first \code{n} entries.}
\usage{
rm_char_in_nfirst2(x, char = "^-", n = (max_miss + 1L) * 2L)
}
\arguments{
\item{x}{A list of character strings.}

\item{char}{A starting character to be removed.}

\item{n}{The number of beginning entries to be considered.}
}
\description{
Remove a starting character from the first \code{n} entries.
}
