% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{rm_diann_empties}
\alias{rm_diann_empties}
\title{Helper in removing DIANN samples not found in metadata.}
\usage{
rm_diann_empties(df, label_scheme_full)
}
\arguments{
\item{df}{A data frame of DIANN outputs.}

\item{label_scheme_full}{Metadata.}
}
\description{
Helper in removing DIANN samples not found in metadata.
}
