% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{rm_fully_empty_tmt_sets}
\alias{rm_fully_empty_tmt_sets}
\title{Removes fully empty TMT sets.}
\usage{
rm_fully_empty_tmt_sets(
  df,
  frac_smry = NULL,
  TMT_plex = 0L,
  dat_dir = NULL,
  ext = NULL
)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}

\item{frac_smry}{A character string to a \code{.xlsx} file containing peptide
fractionation summary. The default is \code{frac_smry.xlsx}.}

\item{TMT_plex}{The multiplexity of TMT.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{ext}{The extension of a file name.}
}
\description{
If all \code{Sample_ID}'s are \code{Empty} within a \code{TMT_Set}. The whole
set of experiment, namely all \code{TMT_Channel}'s, will be removed.
Otherwise, \code{Empty} sample IDs will still be kept (in that the number of
Sample_IDs within a \code{TMT_Set} will be always equal to the multiplexity
of a TMT experiment).
}
\details{
The utility is also applicable to LFQ with a special handling at TMT_plex = 0
(not 1).

Sample IDs started with \code{Empty} are at first replaced with NA.
}
