% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{rm_nRToutliers}
\alias{rm_nRToutliers}
\title{Remove retention-time outliers for more than two values}
\usage{
rm_nRToutliers(mx, my, mt, ms, mf, mn, err, zero = NA_real_)
}
\arguments{
\item{mx}{A multi-row matrix of m-over-z values.}

\item{my}{A multi-row matrix of apex intensity values.}

\item{mt}{A multi-row matrix of apex retention times.}

\item{ms}{A multi-row matrix of apex scan numbers.}

\item{mf}{A multi-row matrix of apex FWHM values.}

\item{mn}{A multi-row matrix of apex lengths.}

\item{err}{The tolerance in retention times.}

\item{zero}{A zero value for outlier replacements.}
}
\description{
Remove retention-time outliers for more than two values
}
