% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{rm_psm_spikes}
\alias{rm_psm_spikes}
\title{Remove PSM spikes by a group of \code{pep_seq_modz}}
\usage{
rm_psm_spikes(df, yfrac = 0.02, sdco = 3, add_sd = TRUE)
}
\arguments{
\item{df}{A data frame for a \code{pep_seq_modz} at a given charge state.}

\item{yfrac}{A factor of intensity fraction. Peaks with intensities \eqn{le}
the fraction of the base peak will be ignored.}

\item{sdco}{The cut-off in standard deviations.}

\item{add_sd}{Logical; if TRUE, add the attribute of \code{sd}.}
}
\description{
May be deleted.
}
