% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{rm_spn_empties}
\alias{rm_spn_empties}
\title{Helper in removing Spectranaut samples not found in metadata.}
\usage{
rm_spn_empties(df, label_scheme_full)
}
\arguments{
\item{df}{A data frame of Spectranaut outputs.}

\item{label_scheme_full}{Metadata.}
}
\description{
Helper in removing Spectranaut samples not found in metadata.
}
