% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{rmPSMHeaders}
\alias{rmPSMHeaders}
\title{Removes PSM Headers}
\usage{
rmPSMHeaders()
}
\value{
Header file(s) and data file(s) without the header.
}
\description{
\code{rmPSMHeaders} removes the header of PSM outputs from
\code{\href{https://http://www.matrixscience.com/}{Mascot}}. It also removes
the spacer columns in the fields of ratio and intensity values.
}
