% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{roll_sum}
\alias{roll_sum}
\title{Concatenates adjacent peptides in a list (with mass).}
\usage{
roll_sum(peps, n = 2L, include_cts = TRUE)
}
\arguments{
\item{peps}{A list of peptide sequences with a one-letter representation of
amino acid residues.}

\item{n}{The number of mis-cleavages for consideration.}

\item{include_cts}{Logical; the list, \code{peps}, includes the protein
C-terminal sequence or not. At the default of TRUE, mis-cleaved peptides at
the end of the protein C-terms will be added as they should. The arguments
would be typically at FALSE, for example, when used for generating
mis-cleaved peptides from the N-terminal of peptides with the removal of a
starting residue \code{M}.}
}
\description{
Concatenates adjacent peptides in a list (with mass).
}
\examples{
\donttest{
peps <- c(a = 1, b = 2, c = 3)
res <- roll_sum(peps, 2)
}
}
