% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{Rosner_outliers}
\alias{Rosner_outliers}
\title{Outlier removals with Rosner's method}
\usage{
Rosner_outliers(x, k = 1L, alpha = 0.05, warn = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{k}{The number of suspected outliers.}

\item{alpha}{The level of type-I error.}

\item{warn}{Logical; show warning message or not.}
}
\description{
Outlier removals with Rosner's method
}
