% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{round_msfmod_decimals}
\alias{round_msfmod_decimals}
\title{For MSFragger}
\usage{
round_msfmod_decimals(
  mods,
  pat = ".*\\\\((\\\\d+\\\\.?\\\\d*)\\\\)$",
  pat2 = "\\\\(\\\\d+\\\\.?\\\\d*\\\\)$"
)
}
\arguments{
\item{mods}{Character strings under column Assigned Modifications.}

\item{pat}{A regular expression.}

\item{pat2}{A regular expression.}
}
\description{
...
}
