% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlfq.R
\name{saddMBR}
\alias{saddMBR}
\title{MBR of peptides}
\usage{
saddMBR(
  base_name,
  ms1files,
  row_id = 1L,
  xsmat = NULL,
  ysmat = NULL,
  tsmat = NULL,
  ssmat = NULL,
  fsmat = NULL,
  nsmat = NULL,
  mbr_mzs,
  mbr_rets,
  mbr_peps,
  dat_dir,
  mbr_ret_tol = 25,
  gap = mbr_ret_tol + 90,
  min_y = 2e+06,
  yco = 100,
  step = 1e-05,
  fwhm_co = 0.5
)
}
\arguments{
\item{base_name}{The base name of a peptide table file
(\code{TMTset[i]_LCMSinj[j]_Peptide_N.txt}).}

\item{ms1files}{A list of file names of retention-time calibrated RAW MS1
data corresponding to the peptide table at \code{base_name} (More than one
file if with analyte pre-fractionation for a sample).}

\item{row_id}{An integer corresponding to \code{base_name}, as well as the
row number in \code{xmat} etc.}

\item{xsmat}{A matrix of X (m-over-z) values under the peptide table at the
\code{base_name}. The length is the number of peptides in the universe. NA
values correspond to peptides (\code{pep_seq_modz}) that are not present in
the current peptide table but found in others. Each cell in the matrix
contains a list of all possible m-over-z candidates.}

\item{ysmat}{A matrix of apex intensities.}

\item{tsmat}{A matrix of apex retention times.}

\item{ssmat}{A matrix of apex scan numbers.}

\item{fsmat}{A matrix of FWHMs.}

\item{nsmat}{A matrix of apex width in counts of MS2 scans.}

\item{mbr_mzs}{A vector of m-over-z values in the universe (sample average).}

\item{mbr_rets}{A vector of retention times in the universe (sample average).}

\item{dat_dir}{The working data directory.}

\item{mbr_ret_tol}{The tolerance in MBR retention time in seconds.}

\item{gap}{A gap in retention-time window in seconds for finding gates in the
Y values of LC. The gap value serves as a margin to trace out a whole peak
with its apex within the \code{mbr_ret_tol}.}

\item{min_y}{The cut-off of intensity values in MBR.}

\item{yco}{The cut-off in Y-values.}

\item{step}{The mass error in \code{ppm / 1e6}.}

\item{fwhm_co}{The cut-off in FWHM values.}
}
\description{
MBR of peptides
}
