% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{save_call}
\alias{save_call}
\title{Save the arguments in a function call}
\usage{
save_call(call_pars, fn)
}
\arguments{
\item{call_pars}{Language.}

\item{fn}{The name of function being called.}
}
\description{
Save the arguments in a function call
}
