% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{save_call2}
\alias{save_call2}
\title{Saves the arguments in a function call}
\usage{
save_call2(path, fun, time = NULL)
}
\arguments{
\item{path}{A (system) file path.}

\item{fun}{The name of function being saved.}

\item{time}{The time stamp.}
}
\description{
Saves the arguments in a function call
}
