% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{scalc_pepprobs}
\alias{scalc_pepprobs}
\title{Calculates peptide scores by single MGF entries (scan numbers).}
\usage{
scalc_pepprobs(
  entry,
  topn_ms2ions = 100L,
  type_ms2ions = "by",
  penalize_sions = FALSE,
  ppm_ms2 = 25L,
  digits = 4L
)
}
\arguments{
\item{entry}{A row of data from \link{pmatch_bymgfs}.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Each entry corresponds to a row in \code{ion_matches.rds}.
}
