% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{scalc_pepscores}
\alias{scalc_pepscores}
\title{Calculates peptide scores by single MGF entries (scan numbers).}
\usage{
scalc_pepscores(entry, topn_ms2ions = 100)
}
\arguments{
\item{entry}{A row of data from \link{pmatch_bymgfs}.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches.}
}
\description{
Each entry corresponds to a row.
}
