% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-mds.R
\name{scoreEucDist}
\alias{scoreEucDist}
\title{Scores Euclidean distance}
\usage{
scoreEucDist(
  df,
  id,
  label_scheme_sub,
  anal_type,
  scale_log2r,
  adjEucDist = FALSE,
  ...
)
}
\description{
Scores Euclidean distance
}
