% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{search_mgf}
\alias{search_mgf}
\title{Searches a single MGF.}
\usage{
search_mgf(
  expt_mass_ms1,
  expt_moverz_ms2,
  theomasses_bf_ms1,
  theomasses_cr_ms1,
  theomasses_af_ms1,
  theos_bf_ms2,
  theos_cr_ms2,
  theos_af_ms2,
  minn_ms2 = 7L,
  ppm_ms1 = 20L,
  ppm_ms2 = 25L
)
}
\arguments{
\item{expt_mass_ms1}{Numeric; the experimental MS1 mass.}

\item{expt_moverz_ms2}{A numeric list; the experimental MS2 m/z's.}

\item{theomasses_bf_ms1}{Numeric vector; the theoretical MS1 masses at the
preceding \code{-1} frame.}

\item{theomasses_cr_ms1}{Numeric vector; the theoretical MS1 masses at the
current frame.}

\item{theomasses_af_ms1}{Numeric vector; the theoretical MS1 masses at the
following \code{+1} frame.}

\item{theos_bf_ms2}{Numeric vector; the theoretical MS2 m/z's at the
preceding \code{-1} frame.}

\item{theos_cr_ms2}{Numeric vector; the theoretical MS2 m/z's at the
current frame.}

\item{theos_af_ms2}{Numeric vector; the theoretical MS2 m/z's at the
following \code{+1} frame.}

\item{minn_ms2}{Integer; the minimum number of MS2 ions for consideration as
a hit.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}
}
\value{
Lists of tibbles.
}
\description{
Searches a single MGF.
}
