% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{search_mgf_frames_i}
\alias{search_mgf_frames_i}
\title{Helper of \link{search_mgf_frames}}
\usage{
search_mgf_frames_i(
  mgf_frames,
  i,
  aa_masses,
  out_path,
  mod_indexes,
  type_ms2ions = "by",
  maxn_vmods_per_pep = 5,
  maxn_sites_per_vmod = 3,
  maxn_vmods_sitescombi_per_pep = 32,
  minn_ms2 = 7,
  ppm_ms1 = 20,
  ppm_ms2 = 25,
  digits = 5
)
}
\arguments{
\item{mgf_frames}{MGFs in frames. Each frame contains one to multiple MGFs
whose MS1 masses are in the same interval.}

\item{i}{Integer; the index of the i-th \code{aa_masses} table.}

\item{aa_masses}{Amino-acid lookup at a given combination of fixed and
variable.}

\item{out_path}{The output path from \link{binTheoPeps}.}

\item{mod_indexes}{Integer; the indexes of fixed and/or variable
modifications.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{maxn_vmods_sitescombi_per_pep}{Integer; the maximum number of
combinatorial variable modifications per peptide sequence.}

\item{minn_ms2}{Integer; the minimum number of MS2 ions for consideration as
a hit.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Searches MGFs in a frame at a given combination of fixed and variable
modifications.
}
