% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtests.R
\name{sep_data_by_group}
\alias{sep_data_by_group}
\title{Helper: separate data by contrast groups.}
\usage{
sep_data_by_group(elements, dfR, dfI, sids, key_col, label_scheme)
}
\arguments{
\item{elements}{Elements of contrasts.}

\item{dfR}{Data frame of log2Ratios.}

\item{dfI}{Data frame of intensities.}

\item{sids}{Sample IDs.}

\item{key_col}{The name of a key column.}

\item{label_scheme}{Metadata.}
}
\description{
Helper: separate data by contrast groups.
}
