% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_cutpoints}
\alias{set_cutpoints}
\title{Generate cut points.}
\usage{
set_cutpoints(cut_points = NULL, data = NULL)
}
\arguments{
\item{cut_points}{A numeric vector defines the cut points (knots) in
histograms. The values of the knots indicate the summarized
\code{log10(Intentisy)} of ions. For example, at \code{cut_points = seq(4,
7, .5)}, values of \code{log2FC} will be binned from \eqn{-Inf} to \eqn{Inf}
according to the cut points at the reporter-ion (or LFQ) intensity of
\eqn{10^4, 10^4.5, ... 10^7}. The default is \code{cut_points = NULL} where
the cut points will be at the quantiles of 25%, 50%, 75%. To disable data
binning, set \code{cut_points} to \code{Inf} or \code{-Inf}. See also
\code{\link{mergePep}} for data alignment with binning.}

\item{data}{A sequence of numeric values with probable NA value.}
}
\description{
Generate cut points.
}
