% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{set_dat_dir}
\alias{set_dat_dir}
\title{Set global \code{dat_dir}}
\usage{
set_dat_dir(dat_dir = NULL)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Set global \code{dat_dir}
}
