% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_ggsave_dots}
\alias{set_ggsave_dots}
\title{Set dot-dot-dot for ggsave.}
\usage{
set_ggsave_dots(dots, dummies = c("filename", "plot", "device", "path"))
}
\arguments{
\item{dots}{Arguments passed on.}

\item{dummies}{Arguments automated for ggsave.}
}
\description{
Set dot-dot-dot for ggsave.
}
