% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{set_mascot_colnms}
\alias{set_mascot_colnms}
\title{Set names for Mascot ratio and intensity columns.}
\usage{
set_mascot_colnms(TMT_plex)
}
\arguments{
\item{TMT_plex}{Numeric; the multiplexity of TMT, i.e., 10, 11 etc.}
}
\description{
Space columns are included.
}
