% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{setHMColor_temp}
\alias{setHMColor_temp}
\title{Customize the annotation colors for columns}
\usage{
setHMColor_temp(annotation_col)
}
\description{
Customize the annotation colors for columns
}
\examples{
anno_colors <- annotation_col \%>\% dplyr::select(-c("Peptide_Yield")) \%>\% setHMColor(.) 

\dontrun{
}

}
