% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{setHMlims}
\alias{setHMlims}
\title{Sets the upper and the lower limits in the range of heat maps}
\usage{
setHMlims(x, xmin, xmax)
}
\arguments{
\item{x}{A data frame of \code{log2FC}.}

\item{xmin}{the lower limit.}

\item{xmax}{the upper limit.}
}
\description{
\code{setHMlims} imputes values beyond the limits to the corresponding
limits.
}
\examples{
setHMlims(df, -1, 1)

}
