% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{setup_expts}
\alias{setup_expts}
\title{Set up Experiments}
\usage{
setup_expts(dat_dir = NULL, expt_smry = "expt_smry.xlsx",
  frac_smry = "frac_smry.xlsx")
}
\arguments{
\item{dat_dir}{A character string to the working directory.}

\item{expt_smry}{A character string to the file of TMT experimental summary.
The default is "expt_smry.csv".}

\item{frac_smry}{A character string to the file of peptide fractionation
summary. The default is "frac_smry.csv".}
}
\description{
A function processes \code{.csv} files containing the metadata of TMT
experiments
}
\section{\code{expt_smry.csv}}{
 The \code{.csv} files should be located
 immediately under the file folder defined by the character vector of
 \code{dat_dir}.

 In the \code{expt_smry.csv} file, values under the \code{Sample_ID} column
 should be unique for entries at different indeces of \code{TMT_Channel} and
 \code{TMT_Set}, or left blank for used entries. Samples with the same
 indeces of \code{TMT_Channel} and \code{TMT_Set} but different indeces of
 \code{LCMS_Injection} should have the same value in \code{Sample_ID}.

 Under the \code{Reference} column, entries should be indicated with non-void
 string(s) for reference(s) and left blank for non-references.

 Users can add or modify the \code{optional column keys} in the \code{.csv}
 file for suitable analysis (see \code{\link{prnSig}}... for the
 customization of column keys).

 \tabular{ll}{ \strong{Reserved column key}   \tab \strong{Descrption}\cr
 TMT_Channel \tab TMT channel names: \code{126}, \code{127N}, \code{127C} et
 al. \cr TMT_Set \tab TMT experiment indeces  \cr LCMS_Injection   \tab LC/MS
 injection indeces under a \code{TMT_Set} \cr RAW_File \tab MS data filenames
 originated by \code{Xcalibur} with or without the \code{.raw} extension
 (\strong{Notes}: (1) \code{OS} filenames can be different to those recorded
 in \code{Xcalibur} and the later should be used. (2) For analysis with
 offline fractionations of peptides before LC/MS, the \code{RAW_File} column
 should be left blank. The correspondence between the fractions and
 \code{RAW_File} names should be specified in a separate file named
 \code{fraction_scheme.csv}.)\cr Accession_Type \tab Protein accession types
 (The currently supported names are "uniprot_id" for
 \code{\href{https://www.uniprot.org/help/entry_name}{UniProt entry name}}
 and "refseq_acc" for
 \code{\href{https://www.ncbi.nlm.nih.gov/refseq/}{RefSeq accession
 number}}.)\cr Species \tab Species names (The currently supported names are
 "human", "mouse", "rat". The name is "human and mouse" or "pdx" for
 \code{\href{https://en.wikipedia.org/wiki/Patient_derived_xenograft}{"patient
  derived xenografts"}}.)\cr Sample_ID \tab Unique sample IDs (\strong{Note}:
 no space in name; leave the sample IDs blank for unused TMT channels.) \cr
 Reference \tab Labels indicating reference samples in TMT experiments
 (\strong{Note}: leave the labels blank for non-reference TMT channels.) \cr
 \cr }

 \tabular{ll}{ \strong{Optional column key}   \tab \strong{Descrption}\cr
 Select \tab Selected \code{Sample_IDs} for indicated analysis of "MDS",
 "Heat map"... \cr Group \tab Aesthetic labels annotating the prior knowledge
 of sample groups, e.g., Ctrl_T1, Ctrl_T2, Disease_T1, Disease_T2, ... \cr
 Order \tab Numeric labels specifying the order of sample \code{groups}
 (\strong{Note}: leave the labels blank for unsupervised analysis.)\cr Color
 \tab Aesthetic labels for sample annotation by color\cr Shape \tab Aesthetic
 labels for sample annotation by shape\cr Size \tab Aesthetic labels for
 sample annotation by size \cr Alpha \tab Aesthetic labels for sample
 annotation by transparency \cr Term \tab Categorial terms for statistical
 modeling \cr Duplicate \tab Indicators of duplicated samples for corrections
 in statistical significance \cr ... \tab Customed columns for mapping to
 indicated analysis}
}

\section{\code{frac_smry.csv}}{
 There will be no need to prepare the
 "frac_smry.csv" file if no peptide fractionations were performed in TMT
 experiments.

 \tabular{ll}{ \strong{Column key}   \tab \strong{Descrption}\cr TMT_Set \tab
 TMT experiment indeces  \cr LCMS_Injection   \tab LC/MS injection indeces
 under a \code{TMT_Set} \cr Fraction \tab Fraction indeces under a
 \code{TMT_Set} \cr RAW_File \tab MS data filenames originated by
 \code{Xcalibur} with or without the \code{.raw} extension }
}

\examples{
# An examplary "expt_smry.csv"
system.file("extdata", "expt_smry.csv", package = "proteoQ")

# An examplary "frac_smry.csv"
system.file("extdata", "frac_smry.csv", package = "proteoQ")

# Set up the working directory and experiments
dat_dir <- c("C:\\\\my_direcotry")
setup_expts()

}
