% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-sigtests.R
\name{sigTest}
\alias{sigTest}
\title{Perform significance tests}
\usage{
sigTest(df, id, label_scheme_sub, filepath, filename, complete_cases,
  method, var_cutoff, pval_cutoff, logFC_cutoff, ...)
}
\description{
Perform significance tests
}
