% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulData.R
\name{simulUniprotPSM}
\alias{simulUniprotPSM}
\title{Generate simulated UniProt results from RefSeq PSM data}
\usage{
simulUniprotPSM(type = "Mascot", dat_dir = NULL)
}
\arguments{
\item{type}{The type of PSM data in \code{c("Masoct", "MaxQuant")}.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Generate simulated UniProt results from RefSeq PSM data
}
