% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{single_mq_prntable}
\alias{single_mq_prntable}
\title{Check the single presence of MaxQuant proteinGroups[...].txt.}
\usage{
single_mq_prntable(dat_dir)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Check the single presence of MaxQuant proteinGroups[...].txt.
}
