% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_tmt_lcms}
\alias{sort_tmt_lcms}
\title{Sorts by the indexes of TMT_Set and LCMS_Injection}
\usage{
sort_tmt_lcms(nms)
}
\arguments{
\item{nms}{A list of names of "TMT_Set" and "LCMS_Injection" separated by dot.}
}
\description{
"8.3"    "8.5"   "11.1"   "11.3"   "99.2"  "99.10" 
Not "8.3"    "8.5"   "11.1"   "11.3"   "99.10"  "99.2" 
Nor "11.1"   "11.3"  "8.3"    "8.5"   "99.10"  "99.2"
}
