% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{sp_lookup_go}
\alias{sp_lookup_go}
\title{Helper to look up GO species}
\usage{
sp_lookup_go(species)
}
\arguments{
\item{species}{Character string; the name of a species for the
\emph{conveninent} preparation of \code{MSig} data bases. The species
available for the convenience feature is in one of \code{c("human", "mouse",
"rat")} with \code{"human"} being the default. The argument is not required
for other species; instead, users will provide values under arguments
\code{ortho_mart} for the lookup of orthologs to human.}
}
\description{
Helper to look up GO species
}
