% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sp_lookup_Ul}
\alias{sp_lookup_Ul}
\title{Species lookup UpperLower (Ul)}
\usage{
sp_lookup_Ul(species)
}
\arguments{
\item{species}{Character string; the name of a species.}
}
\description{
Species lookup UpperLower (Ul)
}
