% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{splitPSM}
\alias{splitPSM}
\title{Splits PSM tables}
\usage{
splitPSM(
  group_psm_by = "pep_seq",
  group_pep_by = "prot_acc",
  fasta = NULL,
  entrez = NULL,
  pep_unique_by = "group",
  scale_rptr_int = FALSE,
  rm_craps = FALSE,
  rm_krts = FALSE,
  purge_phosphodata = TRUE,
  annot_kinases = FALSE,
  plot_rptr_int = TRUE,
  rptr_intco = 0,
  rptr_intrange = c(0, 100),
  use_lowercase_aa = TRUE,
  parallel = TRUE,
  ...
)
}
\arguments{
\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search. There is no default and users need to provide the
correct file path(s) and name(s).}

\item{entrez}{Character string(s) to the name(s) of entrez file(s) with
prepended directory path. At the \code{NULL} default, a convenience lookup
is available for species among \code{c("human", "mouse", "rat")}. For other
species, users need to provide the file path(s) and name(s) for the lookup
table(s). See also \code{\link{Uni2Entrez}} and \code{\link{Ref2Entrez}}
for preparing custom entrez files.}

\item{rm_craps}{Logical; if TRUE,
\href{https://www.thegpm.org/crap/}{cRAP} proteins will be removed.
The default is FALSE.}

\item{rm_krts}{Logical; if TRUE, keratin entries will be removed. The default
is FALSE.}

\item{purge_phosphodata}{Logical; if TRUE and phosphorylation present as
variable modification(s), entries without phosphorylation will be removed.
The default is TRUE.}

\item{annot_kinases}{Logical; if TRUE, proteins of human or mouse origins
will be annotated with their kinase attributes. The default is FALSE.}

\item{plot_rptr_int}{Logical; if TRUE, the distributions of reporter-ion
intensities will be plotted. The default is TRUE. The argument is also
applicable to the precursor intensity with MaxQuant LFQ.}

\item{rptr_intco}{Numeric; the threshold of reporter-ion intensity (TMT:
\code{I126} etc.; LFQ: \code{I000}) being considered non-trivial. The
default is 0 without cut-offs. The data nullification will not be applied
synchronously to the precursor intensity (\code{pep_tot_int}) under the
same PSM query. To guard against odds such as higher MS2 reporter-ion
intensities than their contributing MS1 precursor intensity, employs for
example \code{filter_... = rlang::exprs(pep_tot_int >= your_ms1_cutoff)}
during \link{PSM2Pep}.}

\item{rptr_intrange}{Numeric vector at length two. The argument specifies the
range of reporter-ion intensities (TMT: \code{I126} etc.; LFQ: \code{I000})
being considered non-trivial. The default is between 0 and 100 percentile
without cut-offs. While argument \code{rptr_intco} employs a universal
cut-off across samples by absolute values, \code{range_int} provides an
alternative means of sample-specific thresholding of intensities by
percentiles. The data nullification will not be applied synchronously to
the precursor intensity under the same PSM query.}

\item{use_lowercase_aa}{Logical; if TRUE, modifications in amino acid
residues will be abbreviated with lower-case and/or \code{^_~}. See the
table below for details. The default is TRUE.}

\item{parallel}{Logical; if TRUE, performs parallel computation. The default
is TRUE.}

\item{...}{Not currently used.}
}
\description{
\code{splitPSM} splits the PSM outputs after \code{rmPSMHeaders()}. It
separates PSM data by TMT/LFQ experiment and LC/MS injection.
}
\details{
Arguments \code{group_psm_by} and \code{group_pep_by} are used to update
\code{prot_matches_sig} and \code{prot_sequences_sig} after data merge.
}
