% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{splitPSM_mq}
\alias{splitPSM_mq}
\title{Splits PSM tables}
\usage{
splitPSM_mq(fasta = NULL, corrected_int = TRUE, rptr_intco = 1000,
  rm_craps = FALSE, rm_reverses = TRUE, plot_violins = TRUE)
}
\arguments{
\item{fasta}{Character string(s); the file name(s) with prepended directory
path to the \code{fasta} database being used in MS/MS ion search.}

\item{corrected_int}{Logical; if TRUE, use values under columns "Reporter
intensity corrected" in MaxQuant PSM results. Otherwise, use "Report
intensity" without corrections.}

\item{rptr_intco}{Numeric; the threshold of reporter ion intensity.}

\item{rm_craps}{Logical; if TRUE, removes
\code{\href{https://www.thegpm.org/crap/}{cRAP}} proteins.}

\item{rm_reverses}{Logical; if TRUE, removes \code{Reverse} entries from
MaxQuant peptide results.}

\item{plot_violins}{Logical; if TRUE, prepares the violin plots of
reporter-ion intensities.}
}
\description{
\code{splitPSM} splits the PSM outputs after \code{rmPSMHeaders()}. It
separates PSM data by TMT experiment and LC/MS injection.
}
\examples{
splitPSM(rptr_intco = 1000, rm_craps = TRUE, plot_violins = TRUE)

}
