% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{str_exclude_count}
\alias{str_exclude_count}
\title{Excludes a character in string counting}
\usage{
str_exclude_count(x, char = "-")
}
\arguments{
\item{x}{A character string}

\item{char}{A character to be excluded for counting.}
}
\description{
Excludes a character in string counting
}
