% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{sub_by_bestz}
\alias{sub_by_bestz}
\title{Subset PSMs by the best charge state under a \code{pep_seq_mod}}
\usage{
sub_by_bestz(df, chs)
}
\arguments{
\item{df}{A subset of PSM table at the same \code{pep_seq_mod}.}

\item{chs}{A vector of unique charge states under the \code{pep_seq_mod}.}
}
\description{
Subset PSMs by the best charge state under a \code{pep_seq_mod}
}
