% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{sub_modmasses}
\alias{sub_modmasses}
\title{Calculates of the mono-isotopic masses for \code{Anywhere} variable
modifications.}
\usage{
sub_modmasses(aa_seq, monomasses, site, mass_delta, digits = 5)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{monomasses}{A vector of mono-isotopic masses for peptides.}

\item{site}{An amino-acide site.}

\item{mass_delta}{The mass delta between current aa_masses table and the
reference ass_masses table.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
At all counts of variable modifications for a given \code{site}. For example,
calculations of masses at 1 to 3 oxidations of M. Only counts, not
permutations.
}
