% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{sub_neuloss_peps}
\alias{sub_neuloss_peps}
\title{Subsets theoretical peptides.}
\usage{
sub_neuloss_peps(pattern, theopeps)
}
\arguments{
\item{pattern}{A regex of amino-acid residue(s).}

\item{theopeps}{Lists of theoretical peptides. A column of \code{pep_seq} is
assumed.}
}
\description{
Only entries containing the site of neuloss will be kept.
}
