% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ions.R
\name{subpeps_by_vmods}
\alias{subpeps_by_vmods}
\title{Subset peptides by variable modifications.}
\usage{
subpeps_by_vmods(aa_masses, peps)
}
\arguments{
\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{peps}{A list of peptide sequences with a one-letter representation of
amino acid residues.}
}
\description{
From N-term to C-term.
}
