% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_keepattr}
\alias{subset_keepattr}
\title{Subsets elements with attributes kept.}
\usage{
subset_keepattr(data, nm = c("M", "N"))
}
\arguments{
\item{data}{The data vector.}

\item{nm}{The names of elements in the vector for removals.}
}
\description{
Subsets elements with attributes kept.
}
