% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ions.R
\name{subset_protntsite}
\alias{subset_protntsite}
\alias{subset_protntany}
\alias{subset_anyntsite}
\alias{subset_anyntany}
\alias{subset_anysite}
\alias{subset_protctsite}
\alias{subset_anyctsite}
\alias{subset_anyctany}
\title{Subsets proteins by variable modifications.}
\usage{
subset_protntsite(peps, sites = "M")

subset_protntany(peps)

subset_anyntsite(peps, sites = "Q")

subset_anyntany(peps)

subset_anysite(peps, sites = "N")

subset_protctsite(peps, sites = "V")

subset_anyctsite(peps, sites = "Q")

subset_anyctany(peps)
}
\arguments{
\item{peps}{A list of peptide sequences with a one-letter representation of
amino acid residues.}

\item{sites}{A list of sites. No need to specify for any terminal sites.}
}
\description{
Flowchart (1-nt): `Dimethyl (Protein N-term = P)`

Flowchart (2-nt): `Acetyl (Protein N-term)`

Flowchart (3-nt): `Gln->pyro Glu (N-term = Q)`

Flowchart f(4-nt): `Acetyl (N-term)`

Flowchart (5): `Oxidation (M)`

Flowchart (1-ct): `Dehydrated (Protein C-term = N)`

Flowchart (3-ct): `Oxidation (C-term = G)`

Flowchart (4-ct): `Amidated (C-term)`
}
