% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{subset_theoframes}
\alias{subset_theoframes}
\title{Subsets the frames of theoretical peptides.}
\usage{
subset_theoframes(mgf_frames = NULL, theopeps = NULL)
}
\arguments{
\item{mgf_frames}{MGFs in frames. Each frame contains one to multiple MGFs
whose MS1 masses are in the same interval.}

\item{theopeps}{Binned theoretical peptides at a given combination of fixed
and variable.}
}
\description{
Subsets the frames of theoretical peptides.
}
