% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanos.R
\name{subset_volcano_df}
\alias{subset_volcano_df}
\title{Sets data for volcan plots.}
\usage{
subset_volcano_df(x, df, keys = c("pVal", "adjP", "log2Ratio"))
}
\arguments{
\item{x}{A contrast}

\item{df}{A data frame}

\item{keys}{Column keys.}
}
\description{
Sets data for volcan plots.
}
