% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsetBrukerMGF}
\alias{subsetBrukerMGF}
\title{Subsets Bruker's MGF data}
\usage{
subsetBrukerMGF(
  file,
  begin_offset = 5L,
  charge_offset = 5L,
  topn_ms2ions = Inf
)
}
\arguments{
\item{file}{A file name}

\item{begin_offset}{The number of lines before a BEGIN line.}

\item{charge_offset}{The number lines after a BEGIN line to a following
CHARGE line.}

\item{topn_ms2ions}{Top-n MS2 ions to be retained.}
}
\description{
Applied after the fixes of TITLE lines.
}
