% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{sumdnorm}
\alias{sumdnorm}
\title{Data normalization}
\usage{
sumdnorm(x, xmin = -4, xmax = 4, by = xmax/200)
}
\arguments{
\item{x}{A numeric vector.}

\item{xmin}{the miminal x values.}

\item{xmax}{the maximal x values.}

\item{by}{the step length.}
}
\value{
A data frame.
}
\description{
\code{sumdnorm} calculates summed density from \code{normMulGau}.
}
