% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular.R
\name{svmSubcell}
\alias{svmSubcell}
\title{SVM classification of subcellular locations.}
\usage{
svmSubcell(
  species = "human",
  id = "gene",
  col_subtype = NULL,
  col_compartment = "compartment",
  dat_dir = NULL,
  filepath = NULL,
  n_clust = NULL,
  levels_subcellular = c("1K", "3K", "5K", "9K", "12K", "15K", "30K", "79K", "120K",
    "SN"),
  seed = 42,
  ...
)
}
\arguments{
\item{species}{Species in one of human, mouse and rat.}

\item{id}{Identifier.}

\item{col_subtype}{Temporarily applied only to subcellular data analysis.
Character string to a column key in \code{expt_smry.xlsx} indicating sample
subtypes linked to subcellular fractions. This is needed, e.g., when each
subcellular fraction contain multiple subtypes of control, treated etc.}

\item{col_compartment}{The identifier of subcellular compartment.}

\item{dat_dir}{The current working directory.}

\item{filepath}{The filepath of inputs and outputs.}

\item{n_clust}{The number of clusters.}

\item{levels_subcellular}{The levels of subcellular fractions.}

\item{seed}{A seed for SVm.}
}
\description{
SVM classification of subcellular locations.
}
