% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{syn_empty_channels}
\alias{syn_empty_channels}
\title{Synchronizes Sample_ID(s) for Empty samples.}
\usage{
syn_empty_channels(df)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}
}
\description{
Synchronizes Sample_ID(s) for Empty samples.
}
