% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{taxid_lookup_rev}
\alias{taxid_lookup_rev}
\title{Reversed toxonomy lookup}
\usage{
taxid_lookup_rev(species)
}
\arguments{
\item{species}{Character string; the name of a species.}
}
\description{
Reversed toxonomy lookup
}
