% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{TMT_levels}
\alias{TMT_levels}
\title{Finds the factor levels of TMT labels}
\usage{
TMT_levels(TMT_plex)
}
\arguments{
\item{TMT_plex}{Numeric; the multiplexity of TMT, i.e., 10, 11 etc.}
}
\description{
\code{TMT_levels} returns the factor levels of TMT labels.
}
