% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{TMT_plex}
\alias{TMT_plex}
\title{Finds the multiplexity of TMT labels}
\usage{
TMT_plex(label_scheme_full)
}
\arguments{
\item{label_scheme_full}{The label_scheme with the probable inclusion of
different LCMS_inj under the same TMT_Set.}
}
\description{
\code{TMT_plex} returns the multiplexity of TMT labels.
}
