% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{TMT_plex2}
\alias{TMT_plex2}
\title{Finds the multiplexity of TMT labels from previously made
label_scheme_full.rda}
\usage{
TMT_plex2()
}
\description{
\code{TMT_plex} returns the multiplexity of TMT labels.
}
