% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TMT_top_n}
\alias{TMT_top_n}
\title{Average top_n}
\usage{
TMT_top_n(x, id, ...)
}
\arguments{
\item{x}{A data frame of \code{log2FC} and \code{intensity}.}

\item{id}{The variable to summarize \code{log2FC}.}

\item{...}{Additional arguments for \code{mean}.}
}
\description{
\code{TMT_wt_mean} calculates the weighted mean of \code{log2FC} and
\code{intensity}. Note the difference to \link{aggrTopn}, which uses sum
statistics.
}
\examples{
\donttest{df_num <- TMT_top_n(df, prot_acc, na.rm = TRUE)}
}
