% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{TMT_wt_mean}
\alias{TMT_wt_mean}
\title{Calculates weighted mean}
\usage{
TMT_wt_mean(x, id, ...)
}
\arguments{
\item{x}{A data frame of \code{log2FC} and \code{intensity}.}

\item{id}{The variable to summarise \code{log2FC}.}
}
\description{
\code{TMT_wt_mean} calculates the weighted mean of \code{log2FC} and
\code{intensity}.
}
