% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tmt_wtmean}
\alias{tmt_wtmean}
\title{Calculates weighted mean}
\usage{
tmt_wtmean(x, id, na.rm = TRUE, ...)
}
\arguments{
\item{x}{A data frame of \code{log2FC} and \code{intensity}.}

\item{id}{The variable to summarize \code{log2FC}.}

\item{na.rm}{Logical; if TRUE, removes NA values.}

\item{...}{Additional arguments for \code{weighted.mean}.}
}
\description{
\code{tmt_wtmean} calculates the weighted mean of \code{log2FC} based on
\code{intensity}.
}
