% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{to_complete_cases}
\alias{to_complete_cases}
\title{force 'complete_cases = TRUE' at 'impute_na = FALSE'}
\usage{
to_complete_cases(complete_cases = FALSE, impute_na = FALSE)
}
\description{
force 'complete_cases = TRUE' at 'impute_na = FALSE'
}
