% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{topx}
\alias{topx}
\title{Finds the top-n entries without re-ordering.}
\usage{
topx(x, n = 50L, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{The number of top entries to keep.}
}
\value{
The top-n entries.
}
\description{
Finds the top-n entries without re-ordering.
}
