% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular.R
\name{train_loc_svm}
\alias{train_loc_svm}
\title{Train localization SVM.}
\usage{
train_loc_svm(
  df_expr,
  sub_type = NULL,
  df_ref,
  id = "gene",
  col_y = "location",
  col_subtype = NULL,
  col_compartment = "compartment",
  filepath = NULL,
  seed = 42L,
  ...
)
}
\arguments{
\item{df_expr}{A matrix of expression data, typically normalized intensities
that summed to one across fractions for each protein.}

\item{sub_type}{A label of data subtype, e.g., "Control".}

\item{df_ref}{A reference data frame.}

\item{id}{An identifier.}

\item{col_y}{The column key in \code{df_ref} for training labels.}

\item{col_compartment}{The identifier of subcellular compartment.}

\item{seed}{A seed for SVM.}
}
\description{
Train localization SVM.
}
