% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-trends.R
\name{trendTest}
\alias{trendTest}
\title{Trend analysis}
\usage{
trendTest(df, id, col_group, col_order, label_scheme_sub, n_clust,
  complete_cases, scale_log2r, filepath, filename, ...)
}
\description{
Trend analysis
}
