% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{triv_optcols_at_empty}
\alias{triv_optcols_at_empty}
\title{Trivializes optional columns for Empty samples.}
\usage{
triv_optcols_at_empty(df)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}
}
\description{
Trivializes optional columns for Empty samples.
}
