% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{try_calc_peploc}
\alias{try_calc_peploc}
\title{Helper of \link{calc_peploc}.}
\usage{
try_calc_peploc(out = NULL)
}
\arguments{
\item{out}{A result of \code{psmC.txt}.}
}
\description{
Not yet currently used.
}
