% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{uniprot_entrez_mm}
\alias{uniprot_entrez_mm}
\title{Lookups between UniProt accessions and Entrez IDs for mouse proteins.}
\format{A data frame with 33231 rows and 2 variables:
\describe{
  \item{entrez}{Entrez ID}
  \item{uniprot_acc}{UniProt accession number}
}}
\usage{
uniprot_entrez_mm
}
\description{
A dataset containing mouse \code{entrez} IDs and \code{UniProt} accessions.
}
\keyword{datasets}
