% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{uniprot_entrez_rn}
\alias{uniprot_entrez_rn}
\title{Lookups between UniProt accessions and Entrez IDs for rat proteins.}
\format{A data frame with 21197 rows and 2 variables:
\describe{
  \item{entrez}{Entrez ID}
  \item{uniprot_acc}{UniProt accession number}
}}
\usage{
uniprot_entrez_rn
}
\description{
A dataset containing rat \code{entrez} IDs and \code{UniProt} accessions.
}
\keyword{datasets}
