% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{update_aas_anywhere}
\alias{update_aas_anywhere}
\title{Updates the labels of amino-acid residues.}
\usage{
update_aas_anywhere(vmods_combi, aas, amods)
}
\arguments{
\item{vmods_combi}{Lists of variable modifications.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{amods}{The parsed \emph{variable} \code{Anywhere} modifications from
\code{aa_masses}. The value can be \code{NULL}, e.g., at "amods- tmod- vnl-
fnl+".}
}
\description{
The mappings are according to fixed and variable modifications. The
\code{aas} in a \emph{single} sequence and \code{vmods_combi} is also a
\emph{single} vector.
}
\details{
The mappings are for the count combination of \code{Anywhere} sites. It does
not apply to terminal sites (as there is only one N- or C-term).
}
