% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{update_df}
\alias{update_df}
\title{Update df after normalization.}
\usage{
update_df(df, label_scheme_fit, cf_x_fit, sd_coefs_fit)
}
\arguments{
\item{df}{A data frame of peptide or protein table.}

\item{label_scheme_fit}{Experiment summary for samples being selected for
fitting.}

\item{cf_x_fit}{A data frame containing the \code{x} positions for each
samples indicated in label_scheme_fit.}

\item{sd_coefs_fit}{The standard deviations for each samples indicated in
label_scheme_fit.}
}
\description{
Update df after normalization.
}
