% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{update_frac_smry}
\alias{update_frac_smry}
\title{Updates pre-existed frac_smry.xlsx.}
\usage{
update_frac_smry(
  frac_smry = "frac_smry.xlsx",
  expt_smry = "expt_smry",
  dat_dir = NULL,
  label_scheme_full = NULL,
  ext = "xlsx",
  tbl_mascot = NULL,
  tbl_lfq = NULL,
  is_tmt = TRUE
)
}
\arguments{
\item{frac_smry}{A character string to a \code{.xlsx} file containing peptide
fractionation summary. The default is \code{frac_smry.xlsx}.}

\item{expt_smry}{A character string to a \code{.xlsx} file containing the
metadata of TMT or LFQ experiments. The default is \code{expt_smry.xlsx}.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{label_scheme_full}{The metadata of expt_smry.}

\item{ext}{The extension of a file name.}

\item{tbl_mascot}{A table of message for Mascot workflow.}

\item{tbl_lfq}{A table of message for LFQ workflow.}

\item{is_tmt}{Logical; is TMT experiments or not.}
}
\description{
Updates pre-existed frac_smry.xlsx.
}
