% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{update_ls_by_col_order}
\alias{update_ls_by_col_order}
\title{Helper: update metadata by column keys of \code{col_group} and
\code{col_order}.}
\usage{
update_ls_by_col_order(label_scheme_sub, col_group, col_order)
}
\arguments{
\item{label_scheme_sub}{Metadata.}
}
\description{
Helper: update metadata by column keys of \code{col_group} and
\code{col_order}.
}
