% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_nodes}
\alias{update_nodes}
\title{Helper of \link{make_mq_meta2}}
\usage{
update_nodes(mqxml, field = "filePaths", type = "string", vals)
}
\arguments{
\item{mqxml}{The XML from mqpar.xml.}

\item{field}{The field to be updated.}

\item{type}{The type of data.}

\item{vals}{The Values of data.}
}
\description{
Helper of \link{make_mq_meta2}
}
