% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{updatePSMsd}
\alias{updatePSMsd}
\title{Update cross-sample PSM standard deviation}
\usage{
updatePSMsd(dfs, filelist, key = "pep_seq_mod", clean_up = FALSE)
}
\arguments{
\item{dfs}{A list of PSM tables.}

\item{filelist}{The file names corresponding to \code{dfs}.}

\item{key}{The key for data grouping}

\item{clean_up}{Logical; nullify data of retention outliers or not.}
}
\description{
Update cross-sample PSM standard deviation
}
