% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{use_mf_peptable}
\alias{use_mf_peptable}
\title{Check the single presence of MSFragger peptide table}
\usage{
use_mf_peptable(dat_dir, label_scheme_full, group_psm_by = "pep_seq_mod")
}
\arguments{
\item{dat_dir}{The working directory.}

\item{label_scheme_full}{The metadata}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}
}
\description{
Check the single presence of MSFragger peptide table
}
