% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{use_mq_prottable}
\alias{use_mq_prottable}
\title{Check the single presence of MaxQuant combined_protein[...].tsv}
\usage{
use_mq_prottable(dat_dir)
}
\arguments{
\item{dat_dir}{The working directory.}
}
\description{
Check the single presence of MaxQuant combined_protein[...].tsv
}
