% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informatics.R
\name{vararg_secmsg}
\alias{vararg_secmsg}
\title{helper for finding input `df`}
\usage{
vararg_secmsg(id = NULL, anal_type = NULL, ...)
}
\arguments{
\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{anal_type}{Character string; the type of analysis that are preset for
method dispatch in function factories. The value will be determined
automatically. Exemplary values include \code{anal_type = c("PCA",
"Corrplot", "EucDist", "GSPA", "Heatmap", "Histogram", "MDS", "Model",
"NMF", "Purge", "Trend", "LDA", ...)}.}

\item{...}{Not currently used.}
}
\description{
helper for finding input `df`
}
