% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{vmods_elements}
\alias{vmods_elements}
\title{Find the sets of variable modifications.}
\usage{
vmods_elements(
  aas,
  residue_mods,
  ntmod,
  ctmod,
  aa_masses,
  maxn_vmods_per_pep = 5,
  maxn_sites_per_vmod = 3,
  digits = 5
)
}
\arguments{
\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{residue_mods}{Amino-acid residues with Unimod names. For example
rownames of \code{Carbamidomethyl (M)} and \code{Oxidation (M)} and a
column residues of \code{M, M}.}

\item{ntmod}{The parsed \emph{variable} \code{N-term} modifications from
\code{aa_masses}.}

\item{ctmod}{The parsed \emph{variable} \code{C-term} modifications from
\code{aa_masses}.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Excluding position differences, i.e., \code{A, B} and \code{B, A} is the
same set.
}
