% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{which_topx}
\alias{which_topx}
\title{Finds the indexes of top-n entries without re-ordering.}
\usage{
which_topx(x, n = 50L, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{The number of top entries to keep.}
}
\value{
The indexes of the top-n entries.
}
\description{
Finds the indexes of top-n entries without re-ordering.
}
