% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_excel_wb}
\alias{write_excel_wb}
\title{Wrapper of writing Excel}
\usage{
write_excel_wb(df, sheetName = "Setup", dat_dir = NULL, filename)
}
\arguments{
\item{df}{A data frame.}

\item{sheetName}{An Excel sheet name.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{filename}{An output file name.}
}
\description{
Wrapper of writing Excel
}
