% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{write_metadata}
\alias{write_metadata}
\title{Writes the metadata of label_scheme or fraction_scheme.}
\usage{
write_metadata(
  df,
  dat_dir = NULL,
  filename = "expt_smry.xlsx",
  metatype = "expt"
)
}
\arguments{
\item{df}{A data frame of metadata.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{filename}{The filename of metadata.}

\item{metatype}{The type of metadata in one of "expt" or "frac".}
}
\description{
Writes the metadata of label_scheme or fraction_scheme.
}
