% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{add_mean_dev}
\alias{add_mean_dev}
\title{Add mean-deviation}
\usage{
add_mean_dev(df, label_scheme_fit, filepath)
}
\arguments{
\item{df}{A data frame.}

\item{filepath}{A file path.}

\item{label_schem_fit}{A subset of lable_scheme.}
}
\description{
Add mean-deviation
}
