% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{add_pep_retsd}
\alias{add_pep_retsd}
\title{Helper of \link{calc_pep_retsd}}
\usage{
add_pep_retsd(
  df,
  group_psm_by = "pep_seq_mod",
  use_unique = TRUE,
  col_rt = "pep_apex_ret"
)
}
\arguments{
\item{df}{A data frame.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{use_unique}{Logical; if TRUE, filter data by uniqueness.}

\item{col_rt}{The column key of retention time.}
}
\description{
Helper of \link{calc_pep_retsd}
}
