% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{aggrNums}
\alias{aggrNums}
\title{Aggregate numeric values}
\usage{
aggrNums(f)
}
\arguments{
\item{f}{A function for data aggregation.}
}
\description{
\code{aggrNums} aggregate \code{log2-ratios} and \code{intensity} by \code{c("mean", "median", "weighted.mean", "top.3")}
}
\examples{
df_num <- aggrNums(median)(df, prot_acc, na.rm = TRUE)

\dontrun{
df_num <- aggrNums("median")(df, prot_acc, na.rm = TRUE)
}
}
