% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{alignPSMApexs}
\alias{alignPSMApexs}
\title{Align PSM apex retention times}
\usage{
alignPSMApexs(
  dfs,
  ms1full_files = NULL,
  path_ms1 = NULL,
  key = "pep_seq_modz",
  temp_dir = NULL,
  min_n = NULL
)
}
\arguments{
\item{dfs}{A list of PSM tables by TMTSet[i]LCMSinj[j].}

\item{ms1full_files}{A list of \code{ms1full_} files.}

\item{path_ms1}{The file path to \code{ms1full_} files.}

\item{key}{The key for data grouping.}

\item{temp_dir}{A path for temporary files.}

\item{min_n}{The minimum number of samples for RT alignments.}
}
\description{
Also nullify retention-time outliers.
}
