% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{anal_pepTrend}
\alias{anal_pepTrend}
\alias{anal_prnTrend}
\alias{plot_pepTrend}
\title{Trend analysis of peptide data.}
\usage{
anal_pepTrend(
  dat_dir = NULL,
  col_select = NULL,
  col_group = NULL,
  col_order = NULL,
  col_fraction = NULL,
  col_subtype = NULL,
  cluster_data = TRUE,
  choice = c("cmeans", "clara", "kmeans", "pam", "fanny"),
  n_clust = NULL,
  group_data_by = c("ratio", "intensity"),
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  impute_group_na = TRUE,
  p_outlier = 0.05,
  df = NULL,
  filepath = NULL,
  filename = NULL,
  ...
)

anal_prnTrend(
  dat_dir = NULL,
  col_select = NULL,
  col_group = NULL,
  col_order = NULL,
  col_fraction = NULL,
  col_subtype = NULL,
  cluster_data = TRUE,
  choice = c("cmeans", "clara", "kmeans", "pam", "fanny"),
  n_clust = NULL,
  group_data_by = c("ratio", "intensity"),
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  impute_group_na = TRUE,
  p_outlier = 0.05,
  df = NULL,
  filepath = NULL,
  filename = NULL,
  ...
)

plot_pepTrend(
  dat_dir = NULL,
  col_select = NULL,
  col_order = NULL,
  levels_group = NULL,
  n_clust = NULL,
  panel_ids = NULL,
  show_panel_ids = TRUE,
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  df2 = NULL,
  filepath = NULL,
  filename = NULL,
  theme = NULL,
  ...
)
}
\arguments{
\item{dat_dir}{The current working directory.}

\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
At the \code{NULL} default, the column key of \code{Select} in
\code{expt_smry.xlsx} will be used. In the case of no samples being
specified under \code{Select}, the column key of \code{Sample_ID} will be
used. The non-empty entries under the ascribing column will be used in
indicated analysis.}

\item{col_group}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_group} will be
used for sample grouping in the indicated analysis. At the NULL default, the
column key \code{Group} will be used. No data annotation by groups will be
performed if the fields under the indicated group column is empty.}

\item{col_order}{Character string to a column key in \code{expt_smry.xlsx}.
Numeric values under which will be used for the left-to-right arrangement of
samples in graphic outputs or top-to-bottom arrangement in text outputs. At
the NULL default, the column key \code{Order} will be used. If values under
column \code{Order} are left blank, samples will be ordered by their names.}

\item{col_fraction}{Only applicable to subcellular data analysis for
  assessing fraction purity etc. Character string to a column key in
  \code{expt_smry.xlsx} indicating subcellular fractions. No calculation of
  subcellular location score at the \code{NULL} default.

  Suggest, e.g., the corresponding column key \code{Fraction} in metadata.}

\item{col_subtype}{Only applicable to subcellular data analysis for assessing
sub-group fraction purity etc. Character string to a column key in
\code{expt_smry.xlsx} indicating sample subtypes linked to subcellular
fractions. This is needed, e.g., when each subcellular fraction contain
multiple subtypes of control, treated etc. and purity will be asssessed
independently.}

\item{cluster_data}{Logical; to cluster data or not. The default is
\code{TRUE}. Set \code{cluster_data = FALSE} To return only aggregated
log2Ratio and intensity data without further clustering.}

\item{choice}{Character string; the clustering method in \code{c("cmeans",
"clara", "kmeans", "pam", "fanny")}. The default is "cmeans".}

\item{n_clust}{Numeric vector; the number(s) of clusters that data will be
divided into. At the NULL default, it will be determined by the gap method
in \code{\link[cluster]{clusGap}}. The \code{n_clust} overwrites the
argument \code{centers} in \code{\link[e1071]{cmeans}}.}

\item{group_data_by}{The type of values (\code{"ratio"} or
\code{"intensity"})for clustering. \code{log2Ratio} are used at
\code{group_data_by = "ratio"} and apex-normalized intensity at
\code{group_data_by = "intensity"}.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE. At
\code{scale_log2r = NA}, the raw \code{log2FC} without normalization will
be used.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used. The default is FALSE.}

\item{impute_group_na}{Logical; if TRUE, impute (intensity) values that are
exclusively NA under a sample group. The sample grouping is defined under
\code{label_scheme}.}

\item{p_outlier}{Type-I error in outlier assessment. Set p = 1.0 to skip
outlier removals.}

\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{filepath}{Use system default.}

\item{filename}{A representative file name to outputs. By default, it will be
determined automatically by the name of the current call.}

\item{...}{\code{filter_}: Variable argument statements for the row
filtration against data in a primary file linked to \code{df}. See also
\code{\link{normPSM}} for the format of \code{filter_} statements. \cr \cr
\code{arrange_}: Variable argument statements for the row ordering against
data in a primary file linked to \code{df}. See also \code{\link{prnHM}}
for the format of \code{arrange_} statements. \cr \cr Additional arguments
for \link[e1071]{cmeans}, \link[stats]{kmeans}, \link[cluster]{clara},
\link[cluster]{pam}. Note that \code{centers} in \link[e1071]{cmeans} or
\link[stats]{kmeans} is replaced with \code{n_clust}. The same applies to
\code{k} in \link[cluster]{clara} or \link[cluster]{pam}. \cr With
\link[e1071]{cmeans}, \code{m} is according to Schwaemmle and Jensen if not
provided; \cr \code{x} is disabled with input data being determined
automatically.}

\item{type}{The type of data: either \code{Protein} or \code{Peptide}.}
}
\value{
Classified \code{log2FC}.
}
\description{
Unsupervised clustering of \code{log2FC} or \code{Intensity}.

Unsupervised clustering of \code{log2FC} or \code{Intensity}.

#'Plots the trends of protein expressions from
\code{\link{anal_pepTrend}}.
}
\details{
The option of \code{complete_cases} will be forced to \code{TRUE} at
\code{impute_na = FALSE}
}
\examples{
\donttest{
# ===================================
# Trend analysis
# ===================================

## !!!require the brief working example in `?load_expts`

## global option
scale_log2r <- TRUE


# ===================================
# Analysis
# ===================================
## base (proteins, with sample order supervision)
anal_prnTrend(
  impute_na = FALSE,
  col_order = Order,
  n_clust = c(5:6), 
)

## against selected samples
anal_prnTrend(
  col_select = BI,
  impute_na = FALSE,
  col_order = Order,
  n_clust = c(5:6), 
  filename = sel.txt,
)

## row filtration (proteins)
anal_prnTrend(
  impute_na = FALSE,
  col_order = Order,
  n_clust = c(5:6), 
  filter_prots_by = exprs(prot_n_pep >= 2),
)

## manual m degree of fuzziness (proteins)
anal_prnTrend(
  impute_na = FALSE,
  col_order = Order,
  n_clust = c(5:6), 
  filter_prots = exprs(prot_n_pep >= 2),
  m = 1.5,
  filename = my_m.txt,
)

## additional row filtration by pVals (proteins, impute_na = FALSE)
# if not yet, run prerequisitive significance tests at `impute_na = FALSE`
pepSig(
  impute_na = FALSE, 
  W2_bat = ~ Term["(W2.BI.TMT2-W2.BI.TMT1)", 
                  "(W2.JHU.TMT2-W2.JHU.TMT1)", 
                  "(W2.PNNL.TMT2-W2.PNNL.TMT1)"],
  W2_loc = ~ Term_2["W2.BI-W2.JHU", 
                    "W2.BI-W2.PNNL", 
                    "W2.JHU-W2.PNNL"],
  W16_vs_W2 = ~ Term_3["W16-W2"], 
)

prnSig(impute_na = FALSE)

# (`W16_vs_W2.pVal (W16-W2)` now a column key)
anal_prnTrend(
  impute_na = FALSE,
  col_order = Order,
  n_clust = c(5:6), 
  filter_prots_by_npep = exprs(prot_n_pep >= 3), 
  filter_prots_by_pval = exprs(`W16_vs_W2.pVal (W16-W2)` <= 1e-6), 
)

## additional row filtration by pVals (impute_na = TRUE)
# if not yet, run prerequisitive NA imputation and corresponding 
# significance tests at `impute_na = TRUE`
pepImp(m = 2, maxit = 2)
prnImp(m = 5, maxit = 5)

pepSig(
  impute_na = TRUE, 
  W2_bat = ~ Term["(W2.BI.TMT2-W2.BI.TMT1)", 
                  "(W2.JHU.TMT2-W2.JHU.TMT1)", 
                  "(W2.PNNL.TMT2-W2.PNNL.TMT1)"],
  W2_loc = ~ Term_2["W2.BI-W2.JHU", 
                    "W2.BI-W2.PNNL", 
                    "W2.JHU-W2.PNNL"],
  W16_vs_W2 = ~ Term_3["W16-W2"], 
)

prnSig(impute_na = TRUE)

anal_prnTrend(
  impute_na = TRUE,
  col_order = Order,
  n_clust = c(5:6), 
  filter_prots_by_npep = exprs(prot_n_pep >= 3), 
  filter_prots_by_pval = exprs(`W16_vs_W2.pVal (W16-W2)` <= 1e-6), 
)


# ===================================
# Visualization
# ===================================
## base (proteins, no NA imputation) 
plot_prnTrend(
  col_order = Order, 
)

# at specific cluster ID(s)
# (`cluster` is a column key in `Protein_Trend_[...].txt`)
plot_prnTrend(
  impute_na = FALSE, 
  col_order = Order,
  filter2_by_clusters = exprs(cluster == 5),
  width = 8, 
  height = 10,
  filename = cl5.png,
)

# manual selection of secondary input data file(s)
# may be used for optimizing individual plots
plot_prnTrend(
  df2 = c("Protein_Trend_Z_nclust5.txt"),
  col_order = Order, 
  filename = n5.png,
)

# manual secondary input(s) at specific rank(s)
plot_prnTrend(
  df2 = c("Protein_Trend_Z_nclust5.txt"),
  impute_na = FALSE, 
  col_order = Order,
  filter2_by_clusters = exprs(cluster == 5),
  width = 8, 
  height = 10,
  filename = n5_cl5.png,
)

## NA imputation
# also save as pdf
plot_prnTrend(
  impute_na = TRUE,
  col_order = Order,
  filename = my.pdf,
)

## against selected samples
plot_prnTrend(
  col_order = Order, 
  col_select = BI,
  filename = bi.png,
)

## custom theme
library(ggplot2)
my_trend_theme <- theme_bw() + theme(
  axis.text.x  = element_text(angle=60, vjust=0.5, size=24),
  axis.ticks.x  = element_blank(), 
  axis.text.y  = element_text(angle=0, vjust=0.5, size=24),
  axis.title.x = element_text(colour="black", size=24),
  axis.title.y = element_text(colour="black", size=24),
  plot.title = element_text(face="bold", colour="black",
                            size=20, hjust=.5, vjust=.5),
  panel.grid.major.x = element_blank(),
  panel.grid.minor.x = element_blank(),
  panel.grid.major.y = element_blank(),
  panel.grid.minor.y = element_blank(),
  panel.background = element_rect(fill = '#0868ac', colour = 'red'),
  
  strip.text.x = element_text(size = 24, colour = "black", angle = 0),
  strip.text.y = element_text(size = 24, colour = "black", angle = 90),
  
  plot.margin = unit(c(5.5, 55, 5.5, 5.5), "points"), 
  
  legend.key = element_rect(colour = NA, fill = 'transparent'),
  legend.background = element_rect(colour = NA,  fill = "transparent"),
  legend.position = "none",
  legend.title = element_text(colour="black", size=18),
  legend.text = element_text(colour="black", size=18),
  legend.text.align = 0,
  legend.box = NULL
)

plot_prnTrend(
  col_order = Order, 
  col_select = BI,
  theme = my_trend_theme,
  filename = my_theme.png,
)

## no grouping 
# each sample under column `Select` forms its own group
anal_prnTrend(
  col_group = Select,
  col_order = Order,
  n_clust = 6, 
  filter_prots = exprs(prot_n_pep >= 2),
  filename = sample_ids_as_groups.txt,
)

plot_prnTrend(
  df2 = "sample_ids_as_groups_Protein_Trend_Z_nclust6.txt",
  filter2_by_clusters = exprs(cluster == 4),
  width = 24,
  height = 16,
)

## grouped by column `Term_2` in metadata
anal_prnTrend(
  col_group = Term_2,
  col_order = Order,
  n_clust = 6, 
  filter_prots = exprs(prot_n_pep >= 2),
  filename = term_2_grouping.txt,
)

plot_prnTrend(
  df2 = "term_2_grouping_Protein_Trend_Z_nclust6.txt",
  filter2_by_clusters = exprs(cluster == 3),
  width = 6,
  height = 6,
)

## Cytoscape visualization
# (Make sure that Cytoscape is open.)
# Human
cluego(
  df2 = "Protein_Trend_Z_nclust5.txt", 
  species = c(human = "Homo Sapiens"), 
  n_clust = c(3, 5)
)

# Mouse
cluego(
  df2 = "Protein_Trend_Z_nclust5.txt", 
  species = c(mouse = "Mus Musculus"), 
  n_clust = c(3:4)
)

}
}
\seealso{
\emph{Metadata} \cr \code{\link{load_expts}} for metadata
  preparation and a reduced working example in data normalization \cr

  \emph{Data normalization} \cr \code{\link{normPSM}} for extended examples
  in PSM data normalization \cr \code{\link{PSM2Pep}} for extended examples
  in PSM to peptide summarization \cr \code{\link{mergePep}} for extended
  examples in peptide data merging \cr \code{\link{standPep}} for extended
  examples in peptide data normalization \cr \code{\link{Pep2Prn}} for
  extended examples in peptide to protein summarization \cr
  \code{\link{standPrn}} for extended examples in protein data normalization.
  \cr \code{\link{purgePSM}} and \code{\link{purgePep}} for extended examples
  in data purging \cr \code{\link{pepHist}} and \code{\link{prnHist}} for
  extended examples in histogram visualization. \cr
  \code{\link{extract_raws}} and \code{\link{extract_psm_raws}} for
  extracting MS file names \cr

  \emph{Variable arguments of `filter_...`} \cr \code{\link{contain_str}},
  \code{\link{contain_chars_in}}, \code{\link{not_contain_str}},
  \code{\link{not_contain_chars_in}}, \code{\link{start_with_str}},
  \code{\link{end_with_str}}, \code{\link{start_with_chars_in}} and
  \code{\link{ends_with_chars_in}} for data subsetting by character strings
  \cr

  \emph{Missing values} \cr \code{\link{pepImp}} and \code{\link{prnImp}} for
  missing value imputation \cr

  \emph{Informatics} \cr \code{\link{pepSig}} and \code{\link{prnSig}} for
  significance tests \cr \code{\link{pepVol}} and \code{\link{prnVol}} for
  volcano plot visualization \cr \code{\link{prnGSPA}} for gene set
  enrichment analysis by protein significance pVals \cr \code{\link{gspaMap}}
  for mapping GSPA to volcano plot visualization \cr \code{\link{prnGSPAHM}}
  for heat map and network visualization of GSPA results \cr
  \code{\link{prnGSVA}} for gene set variance analysis \cr
  \code{\link{prnGSEA}} for data preparation for online GSEA. \cr
  \code{\link{pepMDS}} and \code{\link{prnMDS}} for MDS visualization \cr
  \code{\link{pepPCA}} and \code{\link{prnPCA}} for PCA visualization \cr
  \code{\link{pepLDA}} and \code{\link{prnLDA}} for LDA visualization \cr
  \code{\link{pepHM}} and \code{\link{prnHM}} for heat map visualization \cr
  \code{\link{pepCorr_logFC}}, \code{\link{prnCorr_logFC}},
  \code{\link{pepCorr_logInt}} and \code{\link{prnCorr_logInt}}  for
  correlation plots \cr \code{\link{anal_prnTrend}} and
  \code{\link{plot_prnTrend}} for trend analysis and visualization \cr
  \code{\link{cluego}} for the visualization of \code{\link{anal_prnTrend}}
  and \code{\link{plot_prnTrend}} via \code{Cytoscape/ClueGO} \cr
  \code{\link{anal_pepNMF}}, \code{\link{anal_prnNMF}},
  \code{\link{plot_pepNMFCon}}, \code{\link{plot_prnNMFCon}},
  \code{\link{plot_pepNMFCoef}}, \code{\link{plot_prnNMFCoef}} and
  \code{\link{plot_metaNMF}} for NMF analysis and visualization \cr

  \emph{Custom databases} \cr \code{\link{Uni2Entrez}} for lookups between
  UniProt accessions and Entrez IDs \cr \code{\link{Ref2Entrez}} for lookups
  among RefSeq accessions, gene names and Entrez IDs \cr
 \code{\link{prepGO}} for \code{\href{http://current.geneontology.org/products/pages/downloads.html}{gene
 ontology}} \cr
 \code{\link{prepMSig}} for \href{https://data.broadinstitute.org/gsea-msigdb/msigdb/release/7.0/}{molecular
 signatures} \cr
  \code{\link{prepString}} and \code{\link{anal_prnString}} for STRING-DB \cr

  \emph{Column keys in PSM, peptide and protein outputs} \cr
  system.file("extdata", "psm_keys.txt", package = "proteoQ") \cr
  system.file("extdata", "peptide_keys.txt", package = "proteoQ") \cr
  system.file("extdata", "protein_keys.txt", package = "proteoQ") \cr
}
