% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-trends.R
\name{anal_prnTrend}
\alias{anal_prnTrend}
\title{Trend analysis}
\usage{
anal_prnTrend(...)
}
\description{
Trend analysis of protein \code{log2-ratios}
}
\examples{
anal_prnTrend(
  scale_log2r = TRUE,
  col_order = Order,
  n_clust = 6
)

}
\seealso{
\code{\link{proteoTrend}} for parameters
}
