% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-trends.R
\name{analTrend}
\alias{analTrend}
\title{Trend analysis}
\usage{
analTrend(
  df,
  id,
  col_group,
  col_order,
  label_scheme_sub,
  n_clust,
  scale_log2r,
  complete_cases,
  impute_na,
  filepath,
  filename,
  anal_type,
  ...
)
}
\description{
Trend analysis
}
