% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-go.R
\name{annot_from_to}
\alias{annot_from_to}
\title{Helper to map `SYMBOL` to `ENTREZID`}
\usage{
annot_from_to(
  abbr_species = "Hs",
  keys = NULL,
  from = "SYMBOL",
  to = "ENTREZID"
)
}
\description{
Helper to map `SYMBOL` to `ENTREZID`
}
