% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-string.R
\name{annot_stringdb}
\alias{annot_stringdb}
\title{Annotates protein STRING ids by species}
\usage{
annot_stringdb(species, df, db_path, id, score_cutoff)
}
\arguments{
\item{db_path}{Character string; the local path for database(s).}

\item{score_cutoff}{Numeric; the threshold of the \code{combined_score} of
protein-protein interaction.}
}
\description{
Annotates protein STRING ids by species
}
