% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annotPeppos}
\alias{annotPeppos}
\title{Annotation of peptide positions and adjacent amino acid residues}
\usage{
annotPeppos(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
\code{annotPeppos} annotates the start and the end positions of peptides in
ascribed proteins description based on the \code{fasta}. It also annotates the
preceding and the following AA residues.
}
