% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{annotPrn}
\alias{annotPrn}
\title{Add annotation based on the "prot_acc"}
\usage{
annotPrn(df, acc_type)
}
\description{
Add annotation based on the "prot_acc"
}
\examples{
annotPrn(id = gene, m = 3, maxit = 3)

\dontrun{
}
}
