% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{annotPrn}
\alias{annotPrn}
\title{Adds protein annotation}
\usage{
annotPrn(df, fasta)
}
\description{
\code{annotPrn} cross-referencing proteins among \code{uniprot_acc},
\code{uniprot_id}, \code{refseq} and \code{entrez}.
}
