% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{annotPSM_mq}
\alias{annotPSM_mq}
\title{Annotates PSM results}
\usage{
annotPSM_mq(expt_smry = "expt_smry.xlsx", fasta = NULL,
  rm_krts = FALSE, plot_rptr_int = TRUE)
}
\arguments{
\item{expt_smry}{A character string to the \code{.xlsx} file containing the
metadata of TMT experiments. The default is "expt_smry.xlsx".}

\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search.}

\item{plot_rptr_int}{Logical; if TRUE, prepares the violin plots of
reporter-ion intensities.}
}
\description{
\code{annotPSM_mq} adds fields of annotation to PSM tables.
}
