% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{annotPSM_mq}
\alias{annotPSM_mq}
\title{Annotates MaxQuant PSM results}
\usage{
annotPSM_mq(
  group_psm_by = "pep_seq",
  group_pep_by = "prot_acc",
  mc_psm_by = "peptide",
  fasta = NULL,
  expt_smry = "expt_smry.xlsx",
  rm_krts = FALSE,
  plot_rptr_int = TRUE,
  plot_log2FC_cv = TRUE,
  ...
)
}
\arguments{
\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{mc_psm_by}{A character string specifying the method in the median
centering of PSM \code{log2FC} across samples. At the \code{peptide}
default, the median description of PSMs (grouped by \code{pep_seq} or
\code{pep_seq_mod} according to \code{group_psm_by}) will be first
calculated and the offsets to zero (of logarithmic 2) will be used for the
centering of PSMs across samples. At \code{mc_psm_by = protein}, the median
description of PSMs (grouped by \code{prot_acc} or \code{gene} according to
\code{group_pep_by}) will be calculated and the corresponding offsets to
zero will be applied. At the \code{mc_psm_by = psm}, PSMs will be median
centered without grouping.}

\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search. There is no default and users need to provide the
correct file path(s) and name(s).}

\item{expt_smry}{A character string to a \code{.xlsx} file containing the
metadata of TMT or LFQ experiments. The default is \code{expt_smry.xlsx}.}

\item{rm_krts}{Logical; if TRUE, keratin entries will be removed. The default
is FALSE.}

\item{plot_rptr_int}{Logical; if TRUE, the distributions of reporter-ion
intensities will be plotted. The default is TRUE.}

\item{plot_log2FC_cv}{Logical; if TRUE, the distributions of the CV of peptide
\code{log2FC} will be plotted. The default is TRUE.}

\item{...}{Not currently used.}
}
\description{
\code{annotPSM_mq} adds fields of annotation to MaxQuant PSM tables.
}
