% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{assign_duppeps}
\alias{assign_duppeps}
\title{Assign duplicated peptides to a leading protein}
\usage{
assign_duppeps(df, group_psm_by, group_pep_by)
}
\arguments{
\item{df}{A PSM data frame}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}
}
\description{
Assign duplicated peptides to a leading protein
}
