% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtests.R
\name{base_sigtest_y}
\alias{base_sigtest_y}
\title{Calculate baseline intensities.}
\usage{
base_sigtest_y(
  dfR,
  dfI,
  elements = NULL,
  key_col = NULL,
  label_scheme_sub_sub,
  seed = NULL
)
}
\arguments{
\item{dfR}{The complete data frame of log2Ratios.}

\item{dfI}{The corresponding intensity data frame.}

\item{elements}{The elements of groups in contrast fits.}

\item{key_col}{The key column in contrast fits, e.g., column \code{Term}.}

\item{label_scheme_sub_sub}{The metadata corresponding \code{dfR}.}

\item{seed}{A seed for reproducible random number generations.}

\item{ys_base}{The baseline intensities. The length is equal to the number of
columns in \code{dfR}.}
}
\description{
Calculate baseline intensities.
}
