% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{binTheoPeps}
\alias{binTheoPeps}
\title{Separates theoretical peptides into mass groups.}
\usage{
binTheoPeps(
  res,
  min_mass = 516.24046,
  max_mass = 10000,
  ppm = 20,
  out_path = file.path("E:/proteoQ/outs", "pepmasses/binned_theopeps.rds")
)
}
\arguments{
\item{res}{Lists of data containing theoretical peptides and masses from
\link{readRDS}.}

\item{min_mass}{Numeric; the minimum MS1 mass.}

\item{max_mass}{Numeric; the maximum MS1 mass.}

\item{ppm}{Numeric; the eror tolerance of MS1 mass in ppm.}

\item{out_path}{The output path.}
}
\value{
Lists of theoretical peptides binned by MS1 masses. The lists
  correspond to the lists of \code{res}.
}
\description{
Separates theoretical peptides into mass groups.
}
\examples{
\donttest{
res <- readRDS("~/proteoQ/dbs/fasta/uniprot/pepmass/uniprot_hs_2020_05_2miss.rds")
theopeps <- binTheoPeps(res)
}
}
