% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{calc_lfq_log2r}
\alias{calc_lfq_log2r}
\title{Helper: calculates LFQ log2FC.}
\usage{
calc_lfq_log2r(df, type, refChannels)
}
\arguments{
\item{df}{A data frame.}

\item{type}{The type of intensity data.}

\item{refChannels}{The reference channels.}
}
\description{
For rows with single intensity values: original intensities kept but
\code{log2_R000} coerced from 0 to NA. This will keep single-value rows away
from data alignment that are based on \code{log2_R000}. Otherwise, the
alignment may be trapped to the spike at \code{log2_R000 = 0}.
}
