% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{calc_lfq_pepnums2}
\alias{calc_lfq_pepnums2}
\title{Calculates \code{pep_tot_int}, \code{pep_razor_int} and
\code{pep_unique_int} in LFQ}
\usage{
calc_lfq_pepnums2(df, filelist, group_psm_by)
}
\arguments{
\item{df}{A data frame.}
}
\description{
Calculates \code{pep_tot_int}, \code{pep_razor_int} and
\code{pep_unique_int} in LFQ
}
