% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calc_monopep}
\alias{calc_monopep}
\title{Helper: calculates the mono-isotopic mass of a peptide sequence.}
\usage{
calc_monopep(
  aa_seq,
  aa_masses,
  mod_indexes = NULL,
  max_nvmods_per_site = 3,
  max_ncombvmods_per_pep = 64,
  digits = 5
)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{mod_indexes}{Integer; the indexes of fixed and/or variable
modifications.}

\item{max_nvmods_per_site}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{max_ncombvmods_per_pep}{Integer; the maximum number of combinatorial
variable modifications per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Helper: calculates the mono-isotopic mass of a peptide sequence.
}
\examples{
\dontrun{
## No modifications
data(package = "proteoQ", aa_residues)
aa_masses <- aa_residues \%>\%
  dplyr::select(c("one_letter", "monoisotopic_da"))
aa_masses <- aa_masses$monoisotopic_da \%>\% `names<-`(aa_masses$one_letter)
aa_masses["-"] <- 0

x <- calc_monopep("AAIDWFDGKEFSGNPIK", aa_masses)
x <- calc_monopep("-AAAAAAAGDSDSWDADAFSVEDPVR", aa_masses)

## With possible modifications
library(purrr)

aa_masses_all <- calc_aamasses()

calc_monopep("GFGFVSNFER", aa_masses_all[[1]])
calc_monopep("GFGMFVSNFER", aa_masses_all[[1]])
calc_monopep("GFGMFVSNMFER", aa_masses_all[[1]])
calc_monopep("GFGMFVSNMFER", aa_masses_all[[11]])

# Error if modifications not in 'aa_masses_all'
# (`subpeps_by_vmods()` subsets peptides by 
#  variable modifications in an 'aa_masses')
map(aa_masses_all, ~ calc_monopep("GFGFVTFSNSMAEVDAAMAAR", .x))

calc_monopep("GFGFNNTFSSMAEVDAMMANAR", aa_masses_all[[11]])
map(aa_masses_all, ~ calc_monopep("GFGFNNTFSSMAEVDAMMANAR", .x))
}
}
