% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calc_monopep}
\alias{calc_monopep}
\title{Helper: calculates the mono-isotopic mass of a peptide sequence.}
\usage{
calc_monopep(
  aa_seq,
  aa_masses,
  vmods_ps,
  amods,
  tmod,
  ntmod,
  ctmod,
  maxn_vmods_per_pep = 5,
  maxn_sites_per_vmod = 3,
  digits = 5
)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{vmods_ps}{A named list of positions and sites with positions being names
and sites being values.}

\item{amods}{Anywhere modifications subset from \code{vmods_ps}.}

\item{tmod}{A terminal modification subset from \code{vmod_ps}.}

\item{ntmod}{A N-term modification subset from \code{tmod}.}

\item{ctmod}{A C-term modification subset from \code{tmod}.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Helper: calculates the mono-isotopic mass of a peptide sequence.
}
