% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calc_monopeptide}
\alias{calc_monopeptide}
\title{Calculates the mono-isotopic mass of a peptide sequence.}
\usage{
calc_monopeptide(
  aa_seq,
  aa_masses,
  maxn_vmods_per_pep = 5,
  maxn_sites_per_vmod = 3,
  digits = 5
)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
For direct uses from an R console (with trade-offs in speed).
}
\examples{
\dontrun{
## No modifications
data(package = "proteoQ", aa_residues)
aa_masses <- aa_residues \%>\%
  dplyr::select(c("one_letter", "monoisotopic_da"))
aa_masses <- aa_masses$monoisotopic_da \%>\% `names<-`(aa_masses$one_letter)
aa_masses["-"] <- 0

x <- calc_monopeptide("AAIDWFDGKEFSGNPIK", aa_masses)
x <- calc_monopeptide("-AAAAAAAGDSDSWDADAFSVEDPVR", aa_masses)

## With possible modifications
library(purrr)

aa_masses_all <- calc_aamasses()

calc_monopeptide("GFGFVSNFER", aa_masses_all[[1]])
calc_monopeptide("GFGMFVSNFER", aa_masses_all[[1]])
calc_monopeptide("GFGMFVSNMFER", aa_masses_all[[1]])
calc_monopeptide("GFGMFVSNMFER", aa_masses_all[[11]])

# Error if modifications not in 'aa_masses_all'
# (`subpeps_by_vmods()` subsets peptides by 
#  variable modifications in an 'aa_masses')
map(aa_masses_all, ~ calc_monopeptide("GFGFVTFSNSMAEVDAAMAAR", .x))

calc_monopeptide("GFGFNNTFSSMAEVDAMMANAR", aa_masses_all[[11]])
map(aa_masses_all, ~ calc_monopeptide("GFGFNNTFSSMAEVDAMMANAR", .x))
}
}
