% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calc_monopeptide}
\alias{calc_monopeptide}
\title{Calculates the mono-isotopic mass of a peptide sequence.}
\usage{
calc_monopeptide(
  aa_seq,
  fixedmods,
  varmods,
  maxn_vmods_setscombi = 64,
  maxn_vmods_per_pep = Inf,
  maxn_sites_per_vmod = Inf,
  digits = 5
)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{fixedmods}{A character vector of fixed modifications. See also
\link{parse_unimod} for grammars.}

\item{varmods}{A character vector of variable modifications.}

\item{maxn_vmods_setscombi}{Integer; the maximum number of combinatorial variable
modifications and neutral losses.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
For direct uses from an R console (with trade-offs in speed).
}
\examples{
\donttest{
## No variable modifications
# (1)
x <- calc_monopeptide("MAKEMASSPECFUN", 
                      fixedmods = NULL, 
                      varmods = NULL)

x$mass

# (2)
x <- calc_monopeptide("MAKEMASSPECFUN", 
                      fixedmods = "Oxidation (M)", 
                      varmods = NULL)

x$mass

## With variable modifications
# (3)
x <- calc_monopeptide("MAKEMASSPECFUN", 
                      fixedmods = NULL, 
                      varmods = "Oxidation (M)")

x$mass
# x$vmods_ps

# (4)
x <- calc_monopeptide("MAKEMASSPECFUN",
                      c("TMT6plex (N-term)", 
                        "TMT6plex (K)", 
                        "Carbamidomethyl (C)"),
                      c("Acetyl (N-term)", 
                        "Gln->pyro-Glu (N-term = Q)", 
                        "Oxidation (M)"))
                      
x$mass

# The N-term M realizes with acetylation
x$vmods_ps[[5]]

}
}
