% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{calc_ms2ionseries}
\alias{calc_ms2ionseries}
\title{Calculates the masses of MS2 ion series.}
\usage{
calc_ms2ionseries(
  aa_seq = NA,
  aa_masses = NA,
  type = "b",
  maxn_vmods_per_pep = 5,
  maxn_sites_per_vmod = 3,
  maxn_vmods_sitescombi_per_pep = 32,
  z = 1,
  digits = 5
)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{type}{Character; the type of fragment ions.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{maxn_vmods_sitescombi_per_pep}{Integer; the maximum number of
combinatorial variable modifications per peptide sequence.}

\item{z}{Integer; the charge state of an MS1 ion. Currently only for postive
charges.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
For a given type of fragmentation. Minimal error handling for speeds.
}
