% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{calc_mvmods_ms2masses}
\alias{calc_mvmods_ms2masses}
\title{Helper: masses of a series of MS2 ions.}
\usage{
calc_mvmods_ms2masses(
  vmods,
  ntmod,
  ctmod,
  aa_seq,
  aas,
  aa_masses,
  type = "by",
  digits = 5
)
}
\arguments{
\item{vmods}{A list of variable modifications. Site indexes in names;
modifications in values.}

\item{ntmod}{The \code{N-term} modification.}

\item{ctmod}{The \code{C-term} modification.}

\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{type}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Helper: masses of a series of MS2 ions.
}
