% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{calc_mvmods_ms2masses}
\alias{calc_mvmods_ms2masses}
\title{Helper: masses of a series of MS2 ions.}
\usage{
calc_mvmods_ms2masses(
  vmods_combi,
  ntmod,
  ctmod,
  aas,
  mass,
  aa_masses,
  type_ms2ions = "by",
  digits = 5
)
}
\arguments{
\item{vmods_combi}{Lists of variable modifications.}

\item{ntmod}{The parsed \emph{variable} \code{N-term} modifications from
\code{aa_masses}.}

\item{ctmod}{The parsed \emph{variable} \code{C-term} modifications from
\code{aa_masses}.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{mass}{The mass of a theoretical MS1.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz".}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Helper: masses of a series of MS2 ions.
}
