% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_pepfdr}
\alias{calc_pepfdr}
\title{Calculates the cut-off score at a peptide FDR.}
\usage{
calc_pepfdr(out, nms, pep_fdr = 0.01)
}
\arguments{
\item{out}{The output from \link{calc_pepscores}.}

\item{nms}{The name(s) of \code{out} that correspond(s) to decoy results.}

\item{pep_fdr}{Numeric; the levels of peptide FDR.}
}
\description{
Calculates the cut-off score at a peptide FDR.
}
