% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calc_pepmasses}
\alias{calc_pepmasses}
\title{Calculates the masses of tryptic peptides from a fasta database.}
\usage{
calc_pepmasses(
  fasta = "~/proteoQ/dbs/fasta/uniprot/uniprot_hs_2020_05.fasta",
  out_path = "~/proteoQ/dbs/fasta/uniprot/pepmasses/pepmasses.rds",
  min_len = 7,
  max_len = 100,
  max_miss = 2
)
}
\arguments{
\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search. There is no default and users need to provide the
correct file path(s) and name(s).}

\item{out_path}{The file path to outputs.}

\item{min_len}{The minimum length of peptides. Shorter peptides will be
excluded.}

\item{max_miss}{The maximum length of peptides. Longer peptides will be
excluded.}

\item{out_name}{The file name of outputs.}
}
\description{
Calculates the masses of tryptic peptides from a fasta database.
}
\examples{
\dontrun{
res <- calc_pepmasses()
}
}
