% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_pepscores}
\alias{calc_pepscores}
\title{Calculates the scores of peptides.}
\usage{
calc_pepscores(res = NULL, topn_ms2ions = 100, out_path = "~/proteoQ/outs")
}
\arguments{
\item{res}{The result from \link{pmatch_bymgfs}.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches.}

\item{out_path}{The file name with prepending path for outputs.}
}
\description{
Calculates the scores of peptides.
}
