% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_pepscores}
\alias{calc_pepscores}
\title{Calculates the scores of peptides.}
\usage{
calc_pepscores(
  topn_ms2ions = 100L,
  type_ms2ions = "by",
  target_fdr = 0.01,
  fdr_type = "psm",
  min_len = 7L,
  max_len = 100L,
  penalize_sions = FALSE,
  ppm_ms2 = 25L,
  out_path = "~/proteoQ/outs",
  digits = 5L
)
}
\arguments{
\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{target_fdr}{Numeric; a targeted false-discovery rate (FDR) at the
levels of PSM, peptide or protein. See also argument \code{fdr_type}.}

\item{fdr_type}{Character string; the type of FDR control. The value is in
one of c("psm", "peptide", "protein"). Note that \code{fdr_type = protein}
is equivalent to \code{fdr_type = peptide} with the additional filtration
of data at \code{prot_tier == 1}. A variant is to set \code{fdr_type =
psm}, followed by a data filtration at \code{prot_tier == 1}.}

\item{min_len}{Integer; the minimum length of peptides. Shorter peptides will
be excluded.}

\item{max_len}{Integer; the maximum length of peptides. Longer peptides will
be excluded.}

\item{penalize_sions}{Logical; if TRUE, penalizes secondary ions of b0, y0
etc. with lower weights in peptide scoring.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{out_path}{The output path.}

\item{digits}{Integer; the number of decimal places to be used.}

\item{res}{The result from \link{pmatch_bymgfs}. At the \code{NULL} default,
the file \code{ion_matches.rds} will be used.}
}
\description{
Calculates the scores of peptides.
}
