% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_probi_byvmods}
\alias{calc_probi_byvmods}
\title{Helper for score calculations}
\usage{
calc_probi_byvmods(
  df,
  nms,
  expt_moverzs,
  expt_ints,
  N,
  type_ms2ions = "by",
  topn_ms2ions = 100L,
  penalize_sions = TRUE,
  ppm_ms2 = 25L,
  digits = 5L
)
}
\arguments{
\item{df}{Two lists of \code{theo} and matched \code{expt} m-over-z.}

\item{nms}{The names (character strings indicating the names and position of
variable modifications).}

\item{expt_moverzs}{Nested list of match and unmatched experimental m-over-z.}

\item{expt_ints}{Nested list of match and unmatched experimental intensity.}

\item{N}{Numeric; the number of MS2 features in an MGF query.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{penalize_sions}{Logical; if TRUE, penalizes secondary ions of b0, y0
etc. with lower weights in peptide scoring.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
By the positions of variable modifications.
}
\examples{
\donttest{
df <- list(theo = c(390.2009, 550.2315, 710.2622, 809.3306, 880.3677, 
                    995.3946, 1151.4957, 175.1190, 290.1459, 361.1830, 
                    460.2514, 620.2821, 780.3127, 940.3434), 
           expt = c(390.2008, 550.2323, 710.2624, 809.3301, 880.3662, 
                    995.3970, NA, 175.1191, 290.1458, 361.1832, 
                    460.2517, 620.2880, 780.3126, 940.3438))

expt_moverzs <- c(110.0717, 112.0509, 112.0873, 113.0713, 115.0869, 
                  116.0167, 116.0709, 126.1280, 127.1250, 127.1313, 
                  128.1284, 128.1346, 129.1317, 129.1380, 130.0978, 
                  130.1350, 130.1413, 131.1384, 133.0432, 136.0619, 
                  139.0504, 157.1083, 158.0563, 158.0925, 159.0763, 
                  173.1498, 175.1191, 176.1223, 176.1600, 178.0646, 
                  186.1531, 188.1597, 212.1031, 230.1144, 230.1702, 
                  232.1115, 248.1808, 255.1082, 261.6220, 264.6251, 
                  273.1193, 275.6198, 284.1326, 284.6348, 290.1458, 
                  310.1302, 321.0684, 329.1280, 344.1566, 359.6652, 
                  361.1832, 362.2062, 390.2008, 407.2268, 420.1369, 
                  459.2220, 460.2517, 481.0993, 491.1726, 522.2371, 
                  539.7526, 540.2550, 550.2323, 551.2338, 567.2598, 
                  576.7457, 584.7561, 585.2570, 585.7582, 586.2587, 
                  619.2524, 620.2880, 682.2661, 683.2708, 710.2624, 
                  711.2637, 718.3199, 780.3126, 781.3342, 782.3379, 
                  809.3301, 810.3351, 880.3662, 881.3693, 921.3688, 
                  922.3726, 923.3927, 924.3849, 940.3438, 941.3491, 
                  995.3970, 996.3967, 997.3690, 998.3657, 1011.3803, 
                  1012.3803, 1013.3842, 1014.3911, 1015.3893, 1016.3904)

expt_ints <- c(12810.80, 14142.40, 58754.70, 12451.00, 29055.70, 
               45291.00, 63865.00, 250674.00, 261949.00, 179089.00, 
               253049.00, 190448.00, 240766.00, 275813.00, 28354.30, 
               219360.00, 189991.00, 229268.00, 60450.10, 12415.10, 
               11351.30, 17766.50, 29119.50, 105925.00, 10832.10, 
               15792.60, 707208.00, 29073.60, 12632.30, 18499.40, 
               18826.60, 33715.50, 12418.70, 18046.80, 164112.00, 
               15920.80, 13090.70, 24475.10, 14995.90, 49102.20, 
               56960.70, 17143.40, 93462.60, 15536.50, 23416.20, 
               15584.90, 30465.80, 12715.50, 31551.40, 12031.20, 
               22367.40, 55729.80, 77277.60, 19092.70, 29280.50, 
               17574.20, 21419.20, 10681.80, 8850.03, 27071.20, 
               25317.90, 13518.70, 55593.70, 14856.20, 30853.00, 
               11551.30, 22966.50, 515863.00, 243235.00, 11709.10, 
               47070.90, 19336.10, 57334.40, 11747.80, 147943.00, 
               42007.30, 18287.80, 51689.70, 62069.30, 13403.00, 
               84499.10, 24180.00, 47260.80, 13985.00, 14132.90, 
               10097.10, 12578.40, 13326.10, 49003.80, 12951.10, 
               98873.10, 38704.30, 12971.00, 8924.51, 89986.00, 
               162963.00, 119860.00, 114223.00, 125885.00, 32305.60)

calc_probi_byvmods(df, nms = "0000000", expt_moverzs, expt_ints, N = 190)

# 
df2 <- df
df2$expt[8] <- NA
calc_probi_byvmods(df2, nms = "0000000", expt_moverzs, expt_ints, N = 190)
}
}
