% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calc_prot_pepmasses}
\alias{calc_prot_pepmasses}
\title{Helper function for parallel calculations peptide masses by proteins.}
\usage{
calc_prot_pepmasses(
  prot_peps,
  aa_masses,
  include_insource_nl,
  maxn_vmods_per_pep,
  maxn_sites_per_vmod,
  digits
)
}
\arguments{
\item{prot_peps}{Lists of peptides under a proteins.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{include_insource_nl}{Logical Logical; if TRUE, includes MS1 precursor
  masses with the losses of neutral species prior to MS2 fragmentation. The
  default is FALSE.

  Note that there is no combination of neutral losses for fixed modifications
  at the precursor levels. Changes from fixed to variable modifications for
  complete combinations.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
For single protein.
}
