% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calc_prot_pepmasses}
\alias{calc_prot_pepmasses}
\title{Helper function for parallel calculations peptide masses by proteins.}
\usage{
calc_prot_pepmasses(
  prot_peps,
  aa_masses,
  include_insource_nl,
  maxn_vmods_per_pep,
  maxn_sites_per_vmod,
  digits
)
}
\arguments{
\item{prot_peps}{Lists of peptides under a proteins.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{include_insource_nl}{Logical Logical; if TRUE, includes MS1 precursor
masses with the losses of neutral species prior to MS2 fragmentation. The
default is FALSE. The setting at TRUE remains experimenting by allowing
additional masses in the universe of MS1 precursors. The offsets in NLs (by
adding back precursor masses) have not yet taken into account in MS2 ion
searches. A more systemically approach such as \code{open} MS1 masses might
be developed in the future.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
For single protein.
}
