% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calc_prots_pepmasses}
\alias{calc_prots_pepmasses}
\title{Helper function for parallel calculations peptide masses by proteins.}
\usage{
calc_prots_pepmasses(
  aa_seqs,
  aa_masses,
  maxn_vmods_per_pep,
  maxn_sites_per_vmod,
  digits
)
}
\arguments{
\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}

\item{prot_peps}{Lists of peptides under a proteins.}
}
\description{
For each split of multiple proteins.
}
