% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{calc_prots_pepmasses}
\alias{calc_prots_pepmasses}
\title{Helper function for parallel calculations peptide masses by proteins.}
\usage{
calc_prots_pepmasses(
  aa_seqs,
  aa_masses,
  vmods_ps,
  amods,
  tmod,
  ntmod,
  ctmod,
  mod_indexes,
  maxn_vmods_per_pep,
  maxn_sites_per_vmod,
  digits
)
}
\arguments{
\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{vmods_ps}{A named list of positions and sites with positions being names
and sites being values.}

\item{amods}{Anywhere modifications subset from \code{vmods_ps}.}

\item{tmod}{A terminal modification subset from \code{vmod_ps}.}

\item{ntmod}{A N-term modification subset from \code{tmod}.}

\item{ctmod}{A C-term modification subset from \code{tmod}.}

\item{mod_indexes}{Integer; the indexes of fixed and/or variable
modifications.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}

\item{prot_peps}{Lists of peptides under a proteins.}
}
\description{
For each split of multiple proteins.
}
