% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{calc_sd_fcts}
\alias{calc_sd_fcts}
\title{SD for all non-trival samples.}
\usage{
calc_sd_fcts(df, range_log2r, range_int, label_scheme)
}
\arguments{
\item{df}{A data frame.}

\item{range_log2r}{Numeric vector at length two. The argument specifies the
range of the \code{log2FC} for use in the scaling normalization of standard
deviation across samples. The default is between the 10th and the 90th
quantiles.}

\item{range_int}{Numeric vector at length two. The argument specifies the
range of the \code{intensity} of reporter ions for use in the scaling
normalization of standard deviation across samples. The default is between
the 5th and the 95th quantiles.}

\item{label_scheme}{Experiment summary}
}
\description{
SD for all non-trival samples.
}
