% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{calc_subcell_frac}
\alias{calc_subcell_frac}
\title{Calculate subcellular intensity fractions.}
\usage{
calc_subcell_frac(
  dfR = NULL,
  dfI = NULL,
  label_scheme_sub = NULL,
  col_group = "Group",
  col_fraction = "Subcellular",
  col_subtype = NULL
)
}
\arguments{
\item{dfR}{A data frame of log2Ratios.}

\item{dfI}{A data frame of intensities.}

\item{label_scheme_sub}{Metadata.}

\item{col_group}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_group} will be
used for sample grouping in the indicated analysis. At the NULL default, the
column key \code{Group} will be used. No data annotation by groups will be
performed if the fields under the indicated group column is empty.}

\item{col_fraction}{Only applicable to subcellular data analysis for
  assessing fraction purity etc. Character string to a column key in
  \code{expt_smry.xlsx} indicating subcellular fractions. No calculation of
  subcellular location score at the \code{NULL} default.

  Suggest, e.g., the corresponding column key \code{Fraction} in metadata.}

\item{col_subtype}{Only applicable to subcellular data analysis for assessing
sub-group fraction purity etc. Character string to a column key in
\code{expt_smry.xlsx} indicating sample subtypes linked to subcellular
fractions. This is needed, e.g., when each subcellular fraction contain
multiple subtypes of control, treated etc. and purity will be asssessed
independently.}
}
\value{
Classified \code{log2FC}.
}
\description{
Unsupervised clustering of \code{log2FC} or \code{Intensity}.

Unsupervised clustering of \code{log2FC} or \code{Intensity}.

#'Plots the trends of protein expressions from
\code{\link{anal_pepTrend}}.
}
\details{
The option of \code{complete_cases} will be forced to \code{TRUE} at
\code{impute_na = FALSE}
}
\examples{
\donttest{
# ===================================
# Trend analysis
# ===================================

## !!!require the brief working example in `?load_expts`

## global option
scale_log2r <- TRUE


# ===================================
# Analysis
# ===================================
## base (proteins, with sample order supervision)
anal_prnTrend(
  impute_na = FALSE,
  col_order = Order,
  n_clust = c(5:6), 
)

## against selected samples
anal_prnTrend(
  col_select = BI,
  impute_na = FALSE,
  col_order = Order,
  n_clust = c(5:6), 
  filename = sel.txt,
)

## row filtration (proteins)
anal_prnTrend(
  impute_na = FALSE,
  col_order = Order,
  n_clust = c(5:6), 
  filter_prots_by = exprs(prot_n_pep >= 2),
)

## manual m degree of fuzziness (proteins)
anal_prnTrend(
  impute_na = FALSE,
  col_order = Order,
  n_clust = c(5:6), 
  filter_prots = exprs(prot_n_pep >= 2),
  m = 1.5,
  filename = my_m.txt,
)

## additional row filtration by pVals (proteins, impute_na = FALSE)
# if not yet, run prerequisitive significance tests at `impute_na = FALSE`
pepSig(
  impute_na = FALSE, 
  W2_bat = ~ Term["(W2.BI.TMT2-W2.BI.TMT1)", 
                  "(W2.JHU.TMT2-W2.JHU.TMT1)", 
                  "(W2.PNNL.TMT2-W2.PNNL.TMT1)"],
  W2_loc = ~ Term_2["W2.BI-W2.JHU", 
                    "W2.BI-W2.PNNL", 
                    "W2.JHU-W2.PNNL"],
  W16_vs_W2 = ~ Term_3["W16-W2"], 
)

prnSig(impute_na = FALSE)

# (`W16_vs_W2.pVal (W16-W2)` now a column key)
anal_prnTrend(
  impute_na = FALSE,
  col_order = Order,
  n_clust = c(5:6), 
  filter_prots_by_npep = exprs(prot_n_pep >= 3), 
  filter_prots_by_pval = exprs(`W16_vs_W2.pVal (W16-W2)` <= 1e-6), 
)

## additional row filtration by pVals (impute_na = TRUE)
# if not yet, run prerequisitive NA imputation and corresponding 
# significance tests at `impute_na = TRUE`
pepImp(m = 2, maxit = 2)
prnImp(m = 5, maxit = 5)

pepSig(
  impute_na = TRUE, 
  W2_bat = ~ Term["(W2.BI.TMT2-W2.BI.TMT1)", 
                  "(W2.JHU.TMT2-W2.JHU.TMT1)", 
                  "(W2.PNNL.TMT2-W2.PNNL.TMT1)"],
  W2_loc = ~ Term_2["W2.BI-W2.JHU", 
                    "W2.BI-W2.PNNL", 
                    "W2.JHU-W2.PNNL"],
  W16_vs_W2 = ~ Term_3["W16-W2"], 
)

prnSig(impute_na = TRUE)

anal_prnTrend(
  impute_na = TRUE,
  col_order = Order,
  n_clust = c(5:6), 
  filter_prots_by_npep = exprs(prot_n_pep >= 3), 
  filter_prots_by_pval = exprs(`W16_vs_W2.pVal (W16-W2)` <= 1e-6), 
)


# ===================================
# Visualization
# ===================================
## base (proteins, no NA imputation) 
plot_prnTrend(
  col_order = Order, 
)

# at specific cluster ID(s)
# (`cluster` is a column key in `Protein_Trend_[...].txt`)
plot_prnTrend(
  impute_na = FALSE, 
  col_order = Order,
  filter2_by_clusters = exprs(cluster == 5),
  width = 8, 
  height = 10,
  filename = cl5.png,
)

# manual selection of secondary input data file(s)
# may be used for optimizing individual plots
plot_prnTrend(
  df2 = c("Protein_Trend_Z_nclust5.txt"),
  col_order = Order, 
  filename = n5.png,
)

# manual secondary input(s) at specific rank(s)
plot_prnTrend(
  df2 = c("Protein_Trend_Z_nclust5.txt"),
  impute_na = FALSE, 
  col_order = Order,
  filter2_by_clusters = exprs(cluster == 5),
  width = 8, 
  height = 10,
  filename = n5_cl5.png,
)

## NA imputation
# also save as pdf
plot_prnTrend(
  impute_na = TRUE,
  col_order = Order,
  filename = my.pdf,
)

## against selected samples
plot_prnTrend(
  col_order = Order, 
  col_select = BI,
  filename = bi.png,
)

## custom theme
library(ggplot2)
my_trend_theme <- theme_bw() + theme(
  axis.text.x  = element_text(angle=60, vjust=0.5, size=24),
  axis.ticks.x  = element_blank(), 
  axis.text.y  = element_text(angle=0, vjust=0.5, size=24),
  axis.title.x = element_text(colour="black", size=24),
  axis.title.y = element_text(colour="black", size=24),
  plot.title = element_text(face="bold", colour="black",
                            size=20, hjust=.5, vjust=.5),
  panel.grid.major.x = element_blank(),
  panel.grid.minor.x = element_blank(),
  panel.grid.major.y = element_blank(),
  panel.grid.minor.y = element_blank(),
  panel.background = element_rect(fill = '#0868ac', colour = 'red'),
  
  strip.text.x = element_text(size = 24, colour = "black", angle = 0),
  strip.text.y = element_text(size = 24, colour = "black", angle = 90),
  
  plot.margin = unit(c(5.5, 55, 5.5, 5.5), "points"), 
  
  legend.key = element_rect(colour = NA, fill = 'transparent'),
  legend.background = element_rect(colour = NA,  fill = "transparent"),
  legend.position = "none",
  legend.title = element_text(colour="black", size=18),
  legend.text = element_text(colour="black", size=18),
  legend.text.align = 0,
  legend.box = NULL
)

plot_prnTrend(
  col_order = Order, 
  col_select = BI,
  theme = my_trend_theme,
  filename = my_theme.png,
)

## no grouping 
# each sample under column `Select` forms its own group
anal_prnTrend(
  col_group = Select,
  col_order = Order,
  n_clust = 6, 
  filter_prots = exprs(prot_n_pep >= 2),
  filename = sample_ids_as_groups.txt,
)

plot_prnTrend(
  df2 = "sample_ids_as_groups_Protein_Trend_Z_nclust6.txt",
  filter2_by_clusters = exprs(cluster == 4),
  width = 24,
  height = 16,
)

## grouped by column `Term_2` in metadata
anal_prnTrend(
  col_group = Term_2,
  col_order = Order,
  n_clust = 6, 
  filter_prots = exprs(prot_n_pep >= 2),
  filename = term_2_grouping.txt,
)

plot_prnTrend(
  df2 = "term_2_grouping_Protein_Trend_Z_nclust6.txt",
  filter2_by_clusters = exprs(cluster == 3),
  width = 6,
  height = 6,
)

## Cytoscape visualization
# (Make sure that Cytoscape is open.)
# Human
cluego(
  df2 = "Protein_Trend_Z_nclust5.txt", 
  species = c(human = "Homo Sapiens"), 
  n_clust = c(3, 5)
)

# Mouse
cluego(
  df2 = "Protein_Trend_Z_nclust5.txt", 
  species = c(mouse = "Mus Musculus"), 
  n_clust = c(3:4)
)

}
}
\seealso{
\emph{Metadata} \cr \code{\link{load_expts}} for metadata
  preparation and a reduced working example in data normalization \cr

  \emph{Data normalization} \cr \code{\link{normPSM}} for extended examples
  in PSM data normalization \cr \code{\link{PSM2Pep}} for extended examples
  in PSM to peptide summarization \cr \code{\link{mergePep}} for extended
  examples in peptide data merging \cr \code{\link{standPep}} for extended
  examples in peptide data normalization \cr \code{\link{Pep2Prn}} for
  extended examples in peptide to protein summarization \cr
  \code{\link{standPrn}} for extended examples in protein data normalization.
  \cr \code{\link{purgePSM}} and \code{\link{purgePep}} for extended examples
  in data purging \cr \code{\link{pepHist}} and \code{\link{prnHist}} for
  extended examples in histogram visualization. \cr
  \code{\link{extract_raws}} and \code{\link{extract_psm_raws}} for
  extracting MS file names \cr

  \emph{Variable arguments of `filter_...`} \cr \code{\link{contain_str}},
  \code{\link{contain_chars_in}}, \code{\link{not_contain_str}},
  \code{\link{not_contain_chars_in}}, \code{\link{start_with_str}},
  \code{\link{end_with_str}}, \code{\link{start_with_chars_in}} and
  \code{\link{ends_with_chars_in}} for data subsetting by character strings
  \cr

  \emph{Missing values} \cr \code{\link{pepImp}} and \code{\link{prnImp}} for
  missing value imputation \cr

  \emph{Informatics} \cr \code{\link{pepSig}} and \code{\link{prnSig}} for
  significance tests \cr \code{\link{pepVol}} and \code{\link{prnVol}} for
  volcano plot visualization \cr \code{\link{prnGSPA}} for gene set
  enrichment analysis by protein significance pVals \cr \code{\link{gspaMap}}
  for mapping GSPA to volcano plot visualization \cr \code{\link{prnGSPAHM}}
  for heat map and network visualization of GSPA results \cr
  \code{\link{prnGSVA}} for gene set variance analysis \cr
  \code{\link{prnGSEA}} for data preparation for online GSEA. \cr
  \code{\link{pepMDS}} and \code{\link{prnMDS}} for MDS visualization \cr
  \code{\link{pepPCA}} and \code{\link{prnPCA}} for PCA visualization \cr
  \code{\link{pepLDA}} and \code{\link{prnLDA}} for LDA visualization \cr
  \code{\link{pepHM}} and \code{\link{prnHM}} for heat map visualization \cr
  \code{\link{pepCorr_logFC}}, \code{\link{prnCorr_logFC}},
  \code{\link{pepCorr_logInt}} and \code{\link{prnCorr_logInt}}  for
  correlation plots \cr \code{\link{anal_prnTrend}} and
  \code{\link{plot_prnTrend}} for trend analysis and visualization \cr
  \code{\link{cluego}} for the visualization of \code{\link{anal_prnTrend}}
  and \code{\link{plot_prnTrend}} via \code{Cytoscape/ClueGO} \cr
  \code{\link{anal_pepNMF}}, \code{\link{anal_prnNMF}},
  \code{\link{plot_pepNMFCon}}, \code{\link{plot_prnNMFCon}},
  \code{\link{plot_pepNMFCoef}}, \code{\link{plot_prnNMFCoef}} and
  \code{\link{plot_metaNMF}} for NMF analysis and visualization \cr

  \emph{Custom databases} \cr \code{\link{Uni2Entrez}} for lookups between
  UniProt accessions and Entrez IDs \cr \code{\link{Ref2Entrez}} for lookups
  among RefSeq accessions, gene names and Entrez IDs \cr
 \code{\link{prepGO}} for \code{\href{http://current.geneontology.org/products/pages/downloads.html}{gene
 ontology}} \cr
 \code{\link{prepMSig}} for \href{https://data.broadinstitute.org/gsea-msigdb/msigdb/release/7.0/}{molecular
 signatures} \cr
  \code{\link{prepString}} and \code{\link{anal_prnString}} for STRING-DB \cr

  \emph{Column keys in PSM, peptide and protein outputs} \cr
  system.file("extdata", "psm_keys.txt", package = "proteoQ") \cr
  system.file("extdata", "peptide_keys.txt", package = "proteoQ") \cr
  system.file("extdata", "protein_keys.txt", package = "proteoQ") \cr
}
