% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{calc_tmt_nums}
\alias{calc_tmt_nums}
\title{Calculates numeric values}
\usage{
calc_tmt_nums(df, filelist, group_psm_by, parallel)
}
\arguments{
\item{df}{A data frame of PSM.}

\item{filelist}{A list of PSM files.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{parallel}{Logical; if TRUE, performs parallel computation. The default
is TRUE.}
}
\description{
Fields of numeric values: sd_log2_R, log2_R, log2_R, I, N_I.
}
