% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{calc_tmt_prnnums}
\alias{calc_tmt_prnnums}
\title{Helper: calculates the TMT log2FC and reporter-ion intensity of proteins}
\usage{
calc_tmt_prnnums(
  df,
  use_unique_pep = TRUE,
  id = "prot_acc",
  method_pep_prn = "median"
)
}
\arguments{
\item{df}{A data frame.}

\item{use_unique_pep}{Logical. If TRUE, only entries that are \code{TRUE} or
equal to \code{1} under the column \code{pep_isunique} in
\code{Peptide.txt} will be used, for summarizing the \code{log2FC} and the
\code{intensity} of peptides into protein values. The default is to use
unique peptides only. For \code{MaxQuant} data, the levels of uniqueness
are according to the \code{pep_unique_by} in \code{\link{normPSM}}. The
argument currently do nothing to \code{Spectrum Mill} data where both
unique and shared peptides will be kept.}

\item{id}{Always "prot_acc".}

\item{method_pep_prn}{Character string; the method to summarize the
\code{log2FC} and the \code{intensity} of peptides by protein entries. The
descriptive statistics includes \code{c("mean", "median", "weighted_mean",
"top_3_mean", "lfq_max", "lfq_top_2_sum", "lfq_top_3_sum", "lfq_all")} with
\code{median} being the default for TMT and \code{lfq_top_3_sum} for LFQ.
The representative \code{log10-intensity} of reporter (or LFQ) ions at the
peptide levels will be the weight when summarizing \code{log2FC} with
various \code{"top_n"} statistics or \code{"weighted_mean"}.}
}
\description{
Helper: calculates the TMT log2FC and reporter-ion intensity of proteins
}
