% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{calc_tmtint}
\alias{calc_tmtint}
\title{Reporter-ion quantitation.}
\usage{
calc_tmtint(
  data = NULL,
  quant = c("none", "tmt6", "tmt10", "tmt11", "tmt16"),
  ppm_reporters = 10
)
}
\arguments{
\item{data}{An upstream result from \link{matchMS}.}

\item{quant}{A quantitation method. The default is "none".}

\item{ppm_reporters}{The mass tolerance of MS2 reporter ions.}
}
\description{
Reporter-ion quantitation.
}
