% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{calc_tmtint}
\alias{calc_tmtint}
\title{Reporter-ion quantitation.}
\usage{
calc_tmtint(
  data = NULL,
  quant = c("none", "tmt6", "tmt10", "tmt11", "tmt16"),
  ppm_reporters = 10
)
}
\arguments{
\item{data}{An upstream result from \link{matchMS}.}

\item{quant}{A quantitation method. The default is "none". Additional choices
include \code{tmt6} etc. For other multiplicities of \code{tmt}, use the
compatible higher plexes, for example, \code{tmt16} for \code{tmt12} etc.
and \code{tmt10} for \code{tmt8} etc.}

\item{ppm_reporters}{The mass tolerance of MS2 reporter ions.}
}
\description{
Reporter-ion quantitation.
}
