% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{calcLFQprnnums}
\alias{calcLFQprnnums}
\title{Adds the \code{total}, \code{razor} and \code{unique} intensities of
proteins.}
\usage{
calcLFQprnnums(
  df,
  label_scheme,
  use_unique_pep = TRUE,
  group_psm_by = "pep_seq",
  group_pep_by = "gene",
  method_pep_prn = "lfq_top_3_sum",
  impute_prot_na = TRUE,
  use_spec_counts = FALSE
)
}
\arguments{
\item{df}{A data frame.}

\item{label_scheme}{Meta data.}

\item{use_unique_pep}{Logical. If TRUE, only entries that are \code{TRUE} or
equal to \code{1} under the column \code{pep_isunique} in \code{Peptide.txt}
will be used, for summarizing the \code{log2FC} and the \code{intensity} of
peptides into protein values. The default is to use unique peptides only.
For \code{MaxQuant} data, the levels of uniqueness are according to the
\code{pep_unique_by} in \code{\link{normPSM}}. The argument currently do
nothing to \code{Spectrum Mill} data where both unique and shared peptides
will be kept.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{method_pep_prn}{Character string; the method to summarize the
the \code{intensity} of peptides by protein entries. The
descriptive statistics includes \code{c("mean", "median", "weighted_mean",
"top_3_mean", "lfq_max", "lfq_top_2_sum", "lfq_top_3_sum", "lfq_all")} with
\code{median} being the default for TMT and \code{lfq_top_3_sum} for LFQ.
The representative \code{log10-intensity} of reporter (or LFQ) ions at the
peptide levels will be the weight when summarizing \code{log2FC} with
various \code{"top_n"} statistics or \code{"weighted_mean"}.

The method to summarize \code{log2FC} is \code{median}.}

\item{impute_prot_na}{Logical; impute NA values of protein log2FC or not.}

\item{use_spec_counts}{Logical; If TRUE, uses spectrum counts for
quantitation with Mascot or Mzion outputs.}
}
\description{
Adds the \code{total}, \code{razor} and \code{unique} intensities of
proteins.
}
