% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{calcms2_a0_t1_nl0}
\alias{calcms2_a0_t1_nl0}
\title{(2) "amods- tmod+ vnl- fnl-".}
\usage{
calcms2_a0_t1_nl0(aas, aa_masses, ntmod, ctmod, type_ms2ions, digits)
}
\description{
(2) "amods- tmod+ vnl- fnl-".
}
