% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{calcPeptide}
\alias{calcPeptide}
\title{Calculate peptide data for individual TMT experiments}
\usage{
calcPeptide(
  df,
  group_psm_by,
  method_psm_pep,
  group_pep_by,
  dat_dir,
  set_idx,
  injn_idx
)
}
\arguments{
\item{df}{A data frame containing the PSM table from database searches.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{method_psm_pep}{Character string; the method to summarize the
\code{log2FC} and the \code{intensity} of \code{PSMs} by peptide entries.
The descriptive statistics includes \code{c("mean", "median", "top.3",
"weighted.mean")} with \code{median} being the default. The
\code{log10-intensity} of reporter (or LFQ) ions at the \code{PSMs} levels
will be the weight when summarizing \code{log2FC} with \code{"top.3"} or
\code{"weighted.mean"}.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{set_idx}{Numeric.  The index of a multiplex TMT experiment in metadata
files such as \code{label_scheme.xlsx} and \code{frac_scheme.xlsx}.}

\item{injn_idx}{Numeric. The index of \code{LCMS_Inj} in metadata files such
as \code{label_scheme.xlsx} and \code{frac_scheme.xlsx}.}
}
\description{
Argument \code{injn_idx} not currently used.
}
