% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{channelInfo}
\alias{channelInfo}
\title{Extracts the channel information in TMT experiments}
\usage{
channelInfo(label_scheme, set_idx)
}
\arguments{
\item{label_scheme}{The data frame returned by \code{\link{load_expts}}.}

\item{set_idx}{Numeric.  The index of a multiplex TMT experiment in metadata
files such as \code{label_scheme.xlsx} and \code{frac_scheme.xlsx}.}
}
\value{
Three lists of indexes: \code{refChannels}, reference channels(s);
  \code{emptyChannels}, empty channel(s) that were not used for sample
  labeling; \code{labeledChannels}, non-empty channels including both
  reference(s) and sample(s).
}
\description{
A function returns the indexes of TMT channels that are associated to
reference(s), sample(s) and probable unused void(s).
}
