% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{channelInfo}
\alias{channelInfo}
\title{Extract the information of reference and unused TMT channels.}
\usage{
channelInfo(label_scheme, set_idx)
}
\arguments{
\item{label_scheme}{The data frame returned by \code{\link{setup_expts}}.}

\item{set_idx}{Numeric.  The index of a multiplex TMT experiment.}
}
\value{
Three lists of indeces: 
\code{refChannels}, reference channels(s); 
\code{emptyChannels}, empty channel(s) that were not used for sample labeling; 
\code{labeledChannels}, non-empty channels including both reference(s) and sample(s).
}
\description{
A function returns the indeces of TMT channels that are associated to reference(s), sample(s) and probable unused void(s).
}
