% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_depreciated_args}
\alias{check_depreciated_args}
\title{Checks depreciated arguments}
\usage{
check_depreciated_args(blacklist = NULL, ...)
}
\arguments{
\item{blacklist}{A list of paired character vectors. In each pair, the first
element is the old argument name if incurred and the second is the new name
for reminding.}

\item{...}{A list of arguments for checking. The depreciated argument(s) are
no longer in the formalArgs of a function. If present, they will be in
\code{...}.}
}
\description{
Checks depreciated arguments
}
