% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_ggplot_aes}
\alias{check_ggplot_aes}
\title{Checks missing values in a ggplot2 object.}
\usage{
check_ggplot_aes(p)
}
\arguments{
\item{p}{A ggplot2 object.}
}
\description{
Checks missing values in a ggplot2 object.
}
