% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gset_nms}
\alias{check_gset_nms}
\title{check 'gset_nms'}
\usage{
check_gset_nms(gset_nms)
}
\arguments{
\item{gset_nms}{Character string or vector containing the shorthanded name(s),
full file path(s) or both to gene sets for enrichment analysis. For species
among \code{"human", "mouse", "rat"}, the default of \code{c("go_sets",
"c2_msig")} will utilize terms from both gene ontology (\code{GO}) and
molecular signatures (\code{MSig}). Custom data bases of \code{GO} and
curated \code{MSig}, and/or additional species are also supported. See also
\code{\link{prepGO}} for the preparation of custom \code{GO} and
\code{\link{prepMSig}} for the preparation of custom \code{MSig}.}
}
\description{
It is OK that `gset_nms` not in `c("go_sets", "c2_msig", "kegg_sets")` as
custom gene sets may be applied.
}
