% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{check_label_scheme}
\alias{check_label_scheme}
\title{Checks the uniqueness of sample IDs in \code{label_scheme_full}}
\usage{
check_label_scheme(label_scheme_full)
}
\arguments{
\item{label_scheme_full}{The data frame returned by \code{\link{load_expts}},
including multiple LCMS series.}
}
\description{
\code{check_label_scheme} will stop the analysis if the number of unique
samples are less than expected.
}
