% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_optional_cols}
\alias{check_optional_cols}
\title{Checks the default columns in metadata.}
\usage{
check_optional_cols(df)
}
\arguments{
\item{df}{A data frame of metadata (e.g., label_scheme_full).}
}
\description{
Checks the default columns in metadata.
}
