% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular.R
\name{classify_subcellular}
\alias{classify_subcellular}
\title{A wrapper of classifying subcellular locations.}
\usage{
classify_subcellular(
  dat_dir = NULL,
  filepath = NULL,
  id = "gene",
  species = "human",
  n_clust = 16:36,
  levels_subcellular = c("1K", "3K", "5K", "9K", "12K", "15K", "30K", "79K", "120K",
    "SN"),
  col_group = NULL,
  col_order = NULL,
  col_fraction = NULL,
  col_subtype = NULL,
  min_prot_n_peps = 2L,
  plot_prob_clusters = TRUE,
  seed = 42L,
  ...
)
}
\arguments{
\item{dat_dir}{The working directory.}

\item{filepath}{A filepath for inputs and outputs.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{species}{A species in one of \code{"human", "mouse", "rat"} for
identifying the database of reference markers.}

\item{n_clust}{The number of clusters.}

\item{min_prot_n_peps}{The minimum number of detected peptides under a
protein for consideration.}

\item{plot_prob_clusters}{Logical; an option to plot clusters of
probabilities.}

\item{seed}{A random seed.}

\item{choice}{A clustering method.}
}
\description{
A wrapper of classifying subcellular locations.
}
\examples{
\dontrun{
classify_subcellular(
  dat_dir = dat_dir, id = "gene", species = "mouse", n_clust = 16:36, 
  col_group = "Group", col_order = "Order", col_fraction = "Fraction", 
  levels_subcellular = 
    c("1K", "3K", "5K", "9K", "12K", "15K", "30K", "79K", "120K", "SN"))
}

}
