% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular.R
\name{classify_trends}
\alias{classify_trends}
\title{Classify trends.}
\usage{
classify_trends(
  df_med = NULL,
  tie_method = "none",
  col_y = "score",
  col_cluster = "cluster",
  col_fraction = NULL,
  col_subtype = NULL
)
}
\arguments{
\item{df_med}{A median description of localization scores by modules.}

\item{tie_method}{A method handling ties in localization scores.}

\item{col_y}{The column name of Y values for summarization.}

\item{col_cluster}{The column key of cluster nodes.}

\item{col_fraction}{Only applied to subcellular data analysis. Character
  string to a column key in \code{expt_smry.xlsx} indicating subcellular
  fractions. No calculation of subcellular location score at the \code{NULL}
  default.

  Suggest the corresponding column key \code{Subcellular} in metadata.}

\item{col_subtype}{Temporarily applied only to subcellular data analysis.
Character string to a column key in \code{expt_smry.xlsx} indicating sample
subtypes linked to subcellular fractions. This is needed, e.g., when each
subcellular fraction contain multiple subtypes of control, treated etc.}
}
\description{
Classify trends.
}
