% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{cleanupPSM}
\alias{cleanupPSM}
\title{Cleans Up PSM results}
\usage{
cleanupPSM(
  rm_outliers = FALSE,
  group_psm_by = "pep_seq",
  rm_allna = FALSE,
  parallel = TRUE
)
}
\arguments{
\item{rm_outliers}{Logical; if TRUE, PSM outlier removals will be performed
for peptides with more than two identifying PSMs. Dixon's method will be
used when \eqn{2 < n \le 25} and Rosner's method will be used when \eqn{n >
25}. The default is FALSE.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{rm_allna}{Logical; if TRUE, removes data rows that are exclusively NA
across ratio columns of \code{log2_R126} etc. The setting also applies to
\code{log2_R000} in LFQ.}
}
\description{
\code{cleanupPSM} removes PSM outliers after \code{splitPSM},
\code{splitPSM_mq} or \code{splitPSM_sm}. The outlier removals will be
assessed at the basis of per peptide per TMT channel.
}
\details{
Dixon's method will be used when \eqn{2 < n \le 25}; Rosner's method will be
used when \eqn{n > 25}.
}
