% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlfq.R
\name{collapseSTY}
\alias{collapseSTY}
\title{Make MS1 S, T and Y matrices across adjacent scans}
\usage{
collapseSTY(
  xs = NULL,
  ys = NULL,
  zs = NULL,
  lwr = 10,
  step = 0.002,
  spc = 0,
  coll = TRUE,
  are_refs,
  are_smpls
)
}
\arguments{
\item{xs}{Vectors of X (retention time) values in an ASCENDING order.}

\item{ys}{Vectors of Y (intensity) values corresponding to \code{xs}.}

\item{zs}{Vectors of Z (scan numbers) values corresponding to \code{xs}.}

\item{lwr}{The lower (retention time) limit.}

\item{step}{The bin size in converting numeric X values to integers.}

\item{spc}{A species center.}

\item{coll}{Logical; collapse adjacent columns or not.}

\item{are_refs}{Logical; are references or not.}

\item{are_smpls}{Logical; are samples or not.}
}
\description{
Near identical to \code{mzion:::collapse_mms1ints} with three fields instead
of two.
}
