% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecombi.R
\name{combi_mvmods}
\alias{combi_mvmods}
\title{The combinations of variable modifications (multiple sites)}
\usage{
combi_mvmods(amods, aa_seq, aa_masses, maxn_vmods_per_resid, digits)
}
\arguments{
\item{amods}{\code{Anywhere} modifications.}

\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\value{
Lists by residues in \code{amods}.
}
\description{
For all the \code{Anywhere} modifications specified in \code{amods}.
}
