% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_phosphopeps}
\alias{count_phosphopeps}
\title{Counts phospho.}
\usage{
count_phosphopeps(collapses = c("mn"), rm_allna = FALSE)
}
\arguments{
\item{collapses}{The modified residues to be collapsed with unmodified
counterparts.}

\item{rm_allna}{Logical; if TRUE, removes data rows that are exclusively NA
across ratio columns of \code{log2_R126} etc. The setting also applies to
\code{log2_R000} in LFQ.}
}
\description{
Not currently used.
}
