% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-normalization.R
\name{dblTrim}
\alias{dblTrim}
\title{Data normalization}
\usage{
dblTrim(df, range_log2r, range_int, type_r = "N_log2_R", type_int = "N_I")
}
\value{
A data frame.
}
\description{
\code{dblTrim} doubly trims the \code{log2FC} and reporter-ion intensity by
the given ranges.
}
