% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{Dixon_outliers}
\alias{Dixon_outliers}
\title{Outlier removals with Dixon's method}
\usage{
Dixon_outliers(x, p = 0.05)
}
\arguments{
\item{x}{A numeric vector.}

\item{p}{P-value cut-off.}
}
\description{
Outlier removals with Dixon's method
}
