% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{Dixon_outliers}
\alias{Dixon_outliers}
\title{Outlier removals with Dixon's method}
\usage{
Dixon_outliers(x)
}
\arguments{
\item{x}{A matrix or data.frame.}
}
\description{
Outlier removals with Dixon's method
}
