% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{extract_psm_raws}
\alias{extract_psm_raws}
\title{Extract RAW file names from Mascot PSM outputs}
\usage{
extract_psm_raws(dat_dir)
}
\description{
\code{extract_psm_raws} extracts the RAW file names from the PSM data under
the current working directory.
}
\examples{
extract_psm_raws()

}
