% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{extract_raws}
\alias{extract_raws}
\title{Extract RAW MS file names}
\usage{
extract_raws(raw_dir = NULL)
}
\arguments{
\item{raw_dir}{A character string to the directory of MS data.}
}
\description{
Extract a list of \code{RAW} file names that can be passed to \code{frac_smry.xlsx}
}
\examples{
\dontrun{
# Supposed that RAW MS files are stored under "~\my_raw"
extract_raws("~\\\\my_raw")
}

}
