% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlfq.R
\name{fillMBRnaRTs}
\alias{fillMBRnaRTs}
\title{Fill NA retention times with the MS2 triggering retention times}
\usage{
fillMBRnaRTs(mbr_rets, mbr_peps, dat_dir)
}
\arguments{
\item{mbr_rets}{A vector of retention times.}

\item{mbr_peps}{A vector of \code{pep_seq_modz}.}

\item{dat_dir}{The working directory.}
}
\description{
To obtain a reference point for reassess apexes and MBR.
}
