% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{fillMFprnnums}
\alias{fillMFprnnums}
\title{Adds MSFragger intensity and ratio columns}
\usage{
fillMFprnnums(dat_dir)
}
\arguments{
\item{dat_dir}{The working directory.}
}
\description{
Adds MSFragger intensity and ratio columns
}
