% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-go.R
\name{find_abbr_species}
\alias{find_abbr_species}
\title{Helper to find the two-letter character string of abbreviated species}
\usage{
find_abbr_species(species = "human", abbr_species = NULL)
}
\description{
Helper to find the two-letter character string of abbreviated species
}
