% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{find_fit_nms}
\alias{find_fit_nms}
\title{Find sample IDS for used in the current fitting.}
\usage{
find_fit_nms(nm_a, nm_b)
}
\arguments{
\item{nm_a}{A name list from data table.}

\item{nm_b}{A name list from metadata.}
}
\description{
Find sample IDS for used in the current fitting.
}
