% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_fn_exts}
\alias{find_fn_exts}
\title{find the extensions of files}
\usage{
find_fn_exts(filename, type = "text")
}
\arguments{
\item{filename}{A character string of filename}

\item{type}{the type of filename extension}
}
\description{
find the extensions of files
}
