% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-go.R
\name{find_human_orthologs}
\alias{find_human_orthologs}
\title{Helper to find human orthologs
#' @examples
res <- find_human_orthologs(mouse)}
\usage{
find_human_orthologs(species, ortho_mart)
}
\description{
Helper to find human orthologs
#' @examples
res <- find_human_orthologs(mouse)
}
