% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated.R
\name{find_mbr_int}
\alias{find_mbr_int}
\title{Find the MBR intensity}
\usage{
find_mbr_int(
  ys,
  ts,
  ss,
  mbr_ret,
  mbr_y,
  mbr_ret_tol = 25,
  sp_cent = 0,
  max_mbr_fold = 20L,
  n_dia_scans = 6L,
  min_y = 2e+06,
  yco = 100,
  fwhm_co = 0.5,
  zero = NA_real_
)
}
\arguments{
\item{ys}{A vector of intensity values.}

\item{ts}{A vector of time values.}

\item{ss}{A vector of scan numbers.}

\item{mbr_ret}{The reference MBR retention time.}

\item{mbr_y}{The reference MBR intensity.}

\item{mbr_ret_tol}{The tolerance in MBR retention time in seconds.}

\item{sp_cent}{A species center (at log2 scale).}

\item{max_mbr_fold}{The maximum absolute fold change in MBR.}

\item{n_dia_scans}{The maximum number of zero-intensity scans for gap filling
in peak tracing. Not adjustable by users but synchronized with mzion.}

\item{min_y}{The cut-off of intensity values in MBR. Change to a smaller
value with PASEF.}

\item{yco}{The cut-off of intensity values.}

\item{fwhm_co}{The cut-off in FWHM values.}

\item{zero}{A zero value for replacements.}
}
\description{
Not used.
}
