% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{find_ms1_cutpoints}
\alias{find_ms1_cutpoints}
\title{Finds the cut-points of MS1 masses for binning.}
\usage{
find_ms1_cutpoints(from = 350, to = 1700, ppm = 20)
}
\arguments{
\item{from}{Numeric; the starting MS1 mass.}

\item{to}{Numeric; the ending MS1 mass.}

\item{ppm}{Numeric; the ppm for data binning.}
}
\description{
The cut-points will be used as the boundary in data binning. Note that the
upper bound is open.
}
