% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{find_ms1_cutpoints}
\alias{find_ms1_cutpoints}
\title{Finds the cut-points of MS1 masses for binning.}
\usage{
find_ms1_cutpoints(from = 500L, to = 10000L, ppm = 20L)
}
\arguments{
\item{from}{Numeric; the starting MS1 mass.}

\item{to}{Numeric; the ending MS1 mass.}

\item{ppm}{Numeric; the ppm for data binning.}
}
\value{
Cut points.
}
\description{
The cut-points will be used as the boundary in data binning. Note that the
upper bound is open.
}
\seealso{
find_ms1_interval
}
