% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{find_ms1_interval}
\alias{find_ms1_interval}
\title{Calculates the frame numbers for a list of experimental MS1 mass by
intervals.}
\usage{
find_ms1_interval(mass = 1714.81876, from = 350L, ppm = 20L)
}
\arguments{
\item{mass}{Numeric; a list of MS1 masses.}

\item{from}{Numeric; the starting MS1 mass.}

\item{ppm}{Numeric; the ppm for data binning.}
}
\value{
Frame numbers.
}
\description{
Needs correct \code{from}.
}
\examples{
\donttest{
find_ms1_interval(c(500, 800.1))
}
}
\seealso{
find_ms1_cutpoints
}
