% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{find_padding_pos}
\alias{find_padding_pos}
\title{Finds the padding positions of TMT channels.}
\usage{
find_padding_pos(this_plex = 10L, TMT_plex = 10L)
}
\arguments{
\item{this_plex}{Numeric; the multiplexity of TMT, i.e., 10, 11 etc. before
padding.}

\item{Numeric;}{the maximum multiplexity of TMT indicated in metadata of
\code{label_scheme.xlsx}.}
}
\description{
Finds the padding positions of TMT channels.
}
