% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mzion.R
\name{find_path_ms1}
\alias{find_path_ms1}
\title{Find the path of MS1 data}
\usage{
find_path_ms1(dat_dir = NULL, n_files = 0L)
}
\arguments{
\item{dat_dir}{A working directory.}

\item{n_files}{The number of RAW files.}
}
\description{
Find the path of MS1 data
}
