% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{find_pep_pos}
\alias{find_pep_pos}
\title{Find peptide start and end positions}
\usage{
find_pep_pos(prot_acc, pep_seq, fasta)
}
\arguments{
\item{prot_acc}{Protein accession}

\item{pep_seq}{Peptide sequence}

\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search. There is no default and users need to provide the
correct file path(s) and name(s).}
}
\description{
\code{find_pep_pos} finds the start and the end positions of peptides in
ascribed proteins description based on the \code{fasta}.
}
