% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{find_ppm_outer_bycombi}
\alias{find_ppm_outer_bycombi}
\title{Finds the the outer products for a vector of MS2 ions at a given ion series.}
\usage{
find_ppm_outer_bycombi(theos, expts, ppm_ms2)
}
\arguments{
\item{theos}{Numeric vector; one series of theoretical MS2s.}

\item{expts}{Numeric vector; one series experimental MS2s.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}
}
\description{
A theoretical peptide at a given position permutation.
}
