% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{find_reporter_ints}
\alias{find_reporter_ints}
\title{Finds the intensities of reporter-ions.}
\usage{
find_reporter_ints(
  ms2_moverzs,
  ms2_ints,
  theos,
  ul,
  ppm_reporters = 10,
  len,
  nms
)
}
\arguments{
\item{ms2_moverzs}{Numeric vector; a series of experimental MS2 m-over-z's
(in the region of reporter ions).}

\item{ms2_ints}{Numeric vector; a series of experimental MS2 intensities (in
the region of reporter ions).}

\item{theos}{The theoretical m-over-z of reporter ions.}

\item{ul}{The upper and lower bound for reporter-ion m-over-z's.}

\item{ppm_reporters}{The mass tolerance of MS2 reporter ions.}

\item{len}{The length of reporter-ion plexes.}

\item{nms}{The names of reporter-ion channels.}
}
\description{
Finds the intensities of reporter-ions.
}
\examples{
\donttest{
ms2_moverzs <- c(112.0873, 126.1280, 127.1251, 127.1313, 128.1250,
                 128.1284, 128.1347, 129.1317, 129.1380, 130.0654,
                 130.1351, 130.1413, 131.1384)

ms2_ints <- c(5113.79, 135569.00, 120048.00, 122599.00, 3397.98,
              140551.00, 144712.00, 103166.00, 145452.00, 3851.82,
              148218.00, 135393.00, 131215.00)

theos <- c(126.1277, 127.1248, 127.1311, 128.1281, 128.1344,
           129.1315, 129.1378, 130.1348, 130.1411, 131.1382)
names(theos) <- c("126", "127N", "127C", "128N", "128C",
                  "129N", "129C", "130N", "130C", "131N")

ppm_reporters <- 10
ul <- c(126.1, 131.2)
len <- 10
nms <- names(theos)

x <- find_reporter_ints(ms2_moverzs, ms2_ints, theos, ul, ppm_reporters = 10,
                        len , nms)

x <- find_reporter_ints(ms2_moverzs, ms2_ints, theos, ul, ppm_reporters = 25,
                        len , nms)

# Two `129C`, no `127N` etc.
ms2_moverzs <- c(105.1503, 107.0428, 111.7716, 120.0811, 126.1281, 127.1312,
                 128.1282, 128.1349, 129.1317, 129.1365, 129.1382, 230.1694,
                 233.4857, 233.4964, 337.3533, 352.1844, 376.2764, 463.3083,
                 525.2150, 562.3732, 569.3899, 591.2545, 596.0308, 632.3300,
                 636.3959, 703.3637, 789.0423, 816.4487, 817.4516, 839.9531,
                 864.3056, 914.7645, 921.5302, 1479.9816)

ms2_ints <- c(1201.79, 1319.32, 1603.45, 1595.34, 2148.66, 1785.74, 1254.24,
              1986.43, 10127.40, 1522.60, 1562.71, 2926.01, 1590.48, 1692.17,
              1347.88, 1412.64, 3050.10, 3231.10, 1355.21, 2424.18, 1783.26,
              1365.32, 1727.12, 2661.72, 1660.05, 5525.95, 1399.96, 4654.03,
              1990.57, 1758.72, 1655.09, 1460.68, 1641.39, 1721.33)

x <- find_reporter_ints(ms2_moverzs, ms2_ints, theos, ul, ppm_reporters = 25,
                        len , nms)
}
}
