% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{find_trend_df2}
\alias{find_trend_df2}
\title{Finds the input file for Trend plots.}
\usage{
find_trend_df2(
  df2 = NULL,
  n_clust = NULL,
  scale_log2r = TRUE,
  impute_na = FALSE,
  filepath = NULL
)
}
\arguments{
\item{df2}{Filename(s) of trend inputs.}

\item{n_clust}{The number of clusters.}

\item{filepath}{Input and output filepath.}
}
\description{
Finds the input file for Trend plots.
}
