% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-normalization.R
\name{fitKernelDensity}
\alias{fitKernelDensity}
\title{Data normalization}
\usage{
fitKernelDensity(df, n_comp = 3, ...)
}
\value{
A data frame.
}
\description{
\code{fitKernelDensity} calculates the fitted density of \code{log2-ratios}.
}
