% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{fml_gspa}
\alias{fml_gspa}
\title{A helper function for GSPA}
\usage{
fml_gspa(df, formula, col_ind, id, gsets, pval_cutoff, logFC_cutoff,
  gspval_cutoff, min_size, max_size, min_greedy_size, filepath, filename,
  ...)
}
\description{
A helper function for GSPA
}
