% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-volcanos.R
\name{fml_volcano}
\alias{fml_volcano}
\title{Formula specific volcano plots}
\usage{
fml_volcano(fml_nm, pval_cutoff, logFC_cutoff, df, col_ind, id, filepath,
  filename, adjP, show_labels, anal_type, show_sig, gset_nms, scale_log2r,
  complete_cases, impute_na, theme = NULL, ...)
}
\description{
Formula specific volcano plots
}
