% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-go.R
\name{get_full_entrez}
\alias{get_full_entrez}
\title{Helper to get a complete list of `ENTREZID` from `egUNIPROT`}
\usage{
get_full_entrez(species, from = "egUNIPROT")
}
\description{
Helper to get a complete list of `ENTREZID` from `egUNIPROT`
}
