% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-string.R
\name{getStringDB}
\alias{getStringDB}
\title{STRING outputs}
\usage{
getStringDB(db_path = "~\\\\proteoQ\\\\dbs\\\\string",
  score_cutoff = 0.7, adjP = FALSE, ...)
}
\arguments{
\item{db_path}{Character string; the local path for database(s).}

\item{score_cutoff}{Numeric; the threshold of the \code{combined_score} of
protein-protein interaction.}

\item{adjP}{Logical; if TRUE, use Benjamini-Hochberg pVals in volcano plots.
The default is FALSE.}

\item{...}{\code{filter_}: Logical expression(s) for the row filtration of
data; also see \code{\link{normPSM}}.}
}
\description{
\code{getStringDB} produces locally the
\code{\href{https://string-db.org/}{STRING}} results of protein-protein
interaction and protein expressions. The interaction file,
\code{Protein_STRING_ppi.tsv}, and the expression file,
\code{Protein_STRING_expr.tsv}, are compatible with
\code{\href{https://cytoscape.org/}{Cytoscape}}.
}
\seealso{
\code{\link{dl_stringdbs}} for database downloads.
}
