% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-string.R
\name{getStringDB}
\alias{getStringDB}
\title{STRING outputs}
\usage{
getStringDB(db_path = "~\\\\proteoQ\\\\dbs\\\\string",
  score_cutoff = 0.7, adjP = FALSE, impute_na = FALSE, ...)
}
\arguments{
\item{db_path}{Character string; the local path for database(s). The default
is \code{"~\\proteoQ\\dbs\\string"}.}

\item{score_cutoff}{Numeric; the threshold in the \code{combined_score} of
protein-protein interaction. The default is 0.7.}

\item{adjP}{Logical. At the FALSE default, unadjusted pVals from
\code{\link{prnSig}} will be used in PPI reports; otherwise, the
Benjamini-Hochberg pVals will be used in the reports.}

\item{impute_na}{Logical. At the NULL default, the TRUE or FALSE will match
the choice in \code{\link{pepSig}} for peptide and \code{\link{prnSig}} for
protein data.}

\item{...}{\code{filter_}: Logical expression(s) for the row filtration of
data; also see \code{\link{normPSM}}.}
}
\description{
\code{getStringDB} prepares locally the
\code{\href{https://string-db.org/}{STRING}} results of protein-protein
interaction and protein expressions. The interaction file,
\code{Protein_STRING_ppi.tsv}, and the expression file,
\code{Protein_STRING_expr.tsv}, are compatible with
\code{\href{https://cytoscape.org/}{Cytoscape}}.
}
\examples{
# ===================================
# String DB
# ===================================

## !!!require the brief working example in `?load_expts`

string_dir <- "~\\\\proteoQ\\\\dbs\\\\string"
dir.create(string_dir, recursive = TRUE, showWarnings = FALSE)

## download
dl_stringdbs(
  species = human,
  db_path = string_dir,
)

## ppi and expression
# human and mouse and probably 'unknown' species
anal_prnString(
  db_path = string_dir,
  score_cutoff = .9,
  filter_prots_by = exprs(prot_n_pep >= 2),
)

# human only
anal_prnString(
  db_path = string_dir,
  score_cutoff = .9,
  filter_by_sp = exprs(species == "human"),
  filter_prots_by = exprs(prot_n_pep >= 2),
)

# custom filename
# outputs in 'my_human_ppi.tsv' and 'my_human_expr.tsv' 
anal_prnString(
  db_path = string_dir,
  score_cutoff = .9,
  complete_cases = TRUE, 
  filter_by = exprs(species == "human", 
                    prot_n_pep >= 2, 
                    `W16_vs_W2.pVal (W16-W2)` <= 1e-6),
  filename = my_human.txt,
)

# outputs in: 
# prot_npep2 W16_vs_W2.pVal (W16-W2)_human_expr.tsv
# prot_npep2 W16_vs_W2.pVal (W16-W2)_human_ppi.tsv
# prot_npep2 W16_vs_W2.pVal (W16-W2)_mouse_expr.tsv
# prot_npep2 W16_vs_W2.pVal (W16-W2)_mouse_ppi.tsv
anal_prnString(
  db_path = string_dir,
  score_cutoff = .9,
  complete_cases = TRUE, 
  filter_by = exprs(prot_n_pep >= 2, 
                    `W16_vs_W2.pVal (W16-W2)` <= 1e-6),
  filename = "prot_npep2 W16_vs_W2.pVal (W16-W2).txt",
)
}
\seealso{
\code{\link{dl_stringdbs}} for database downloads. \cr 
\code{\link{prnSig}} for significance tests \cr
}
