% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{groupMZPepZ2}
\alias{groupMZPepZ2}
\title{Group Mzion peptides by charge states.}
\usage{
groupMZPepZ2(df)
}
\arguments{
\item{df_num}{A peptide table.}
}
\description{
Uses the charge state that has least number of missing values.
}
