% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{groupMZPSM1}
\alias{groupMZPSM1}
\title{Group PSMs}
\usage{
groupMZPSM1(df, yco = 1e+07, sdco = 3)
}
\arguments{
\item{df}{A Mzion PSM table.}

\item{yco}{An intensity cut-off.}

\item{sdco}{A standard deviation cut-off.}
}
\description{
For \code{lfq_mbr = TRUE}. Delayed \code{pep_seq_modz} to \code{pep_seq_mod}
conversion.
}
