% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{Grubbs_outliers}
\alias{Grubbs_outliers}
\title{Outlier removals with Grubbs's method}
\usage{
Grubbs_outliers(x, type = 10)
}
\arguments{
\item{x}{A matrix or data.frame.}

\item{type}{Type for grubbs.test.}
}
\description{
Outlier removals with Grubbs's method
}
