% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{gspa_colAnnot}
\alias{gspa_colAnnot}
\title{Sets up the column annotation in heat maps}
\usage{
gspa_colAnnot(annot_cols = NULL, df, sample_ids, annot_colnames = NULL)
}
\description{
Sets up the column annotation in heat maps
}
