% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{gspaTest}
\alias{gspaTest}
\title{Perform GSPA tests}
\usage{
gspaTest(
  df = NULL,
  id = "entrez",
  label_scheme_sub = NULL,
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  filepath = NULL,
  filename = NULL,
  gset_nms = "go_sets",
  var_cutoff = 0.5,
  pval_cutoff = 0.05,
  logFC_cutoff = log2(1.2),
  gspval_cutoff = 0.05,
  gslogFC_cutoff = log2(1),
  min_size = 6,
  max_size = Inf,
  min_delta = 4,
  min_greedy_size = 1,
  method = "mean",
  anal_type = "GSPA",
  ...
)
}
\description{
logFC_cutoff setset of data before the calculation of adjusted pvals
}
