% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{gspaTest}
\alias{gspaTest}
\title{Perform GSPA tests}
\usage{
gspaTest(df, id = "entrez", label_scheme_sub, filepath, filename,
  complete_cases = FALSE, gset_nm = "go_sets", var_cutoff = 0.5,
  pval_cutoff = 0.05, logFC_cutoff = log2(1.2), min_size = 6, ...)
}
\description{
logFC_cutoff subsets data for adjusted pvals
}
