% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspa.R
\name{gspaTest}
\alias{gspaTest}
\title{Perform GSPA tests}
\usage{
gspaTest(
  df = NULL,
  id = "entrez",
  id_gspa = "entrez",
  label_scheme_sub = NULL,
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  filepath = NULL,
  filename = NULL,
  gset_nms = "go_sets",
  var_cutoff = 0.5,
  pval_cutoff = 0.05,
  logFC_cutoff = log2(1.2),
  gsscore_cutoff = 10,
  gslogFC_cutoff = log2(1),
  gspval_cutoff = 0.01,
  min_size = 6L,
  max_size = .Machine$integer.max,
  min_delta = 5L,
  min_greedy_size = 1L,
  use_adjP = FALSE,
  method = "mean",
  anal_type = "GSPA",
  ...
)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{id}{Currently only "entrez".}

\item{id_gspa}{Character string; the ID type used in GSPA. The default is
\code{entrez}.}

\item{label_scheme_sub}{A data frame. Subset entries from \code{label_scheme}
for selected samples.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE. At
\code{scale_log2r = NA}, the raw \code{log2FC} without normalization will
be used.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used. The default is FALSE.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{filename}{A representative file name to outputs. By default, the
name(s) will be determined automatically. For text files, a typical file
extension is \code{.txt}. For image files, they are typically saved via
\code{\link[ggplot2]{ggsave}} or \code{\link[pheatmap]{pheatmap}} where the
image type will be determined by the extension of the file name.}

\item{gset_nms}{Character string or vector containing the shorthanded name(s),
full file path(s), or both, to gene sets for enrichment analysis. For
species among \code{"human", "mouse", "rat"}, the default of
\code{c("go_sets", "c2_msig", "kinsub")} will utilize terms from gene
ontology (\code{GO}), molecular signatures (\code{MSig}) and
kinase-substrate network (\code{PSP Kinase-Substrate}). Custom \code{GO},
\code{MSig} and other data bases at given species are also supported. See
also: \code{\link{prepGO}} for the preparation of custom \code{GO};
\code{\link{prepMSig}} for the preparation of custom \code{MSig}. For other
custom data bases, follow the same format of list as \code{GO} or
\code{MSig}.}

\item{var_cutoff}{Numeric; the cut-off in the variances of protein log2FC.
Entries with variances smaller than the threshold will be removed from GSVA.
The default is 0.5.}

\item{pval_cutoff}{Numeric value or vector; the cut-off in protein
significance \code{pVal}. Entries with \code{pVals} less significant than
the threshold will be excluded from enrichment analysis. The default is 0.05
for all formulas matched to or specified in argument \code{fml_nms}.
Formula-specific threshold is allowed by supplying a vector of cut-off
values.}

\item{logFC_cutoff}{Numeric value or vector; the cut-off in protein
\code{log2FC}. Entries with absolute \code{log2FC} smaller than the
threshold will be excluded from enrichment analysis. The default magnitude
is \code{log2(1.2)} for all formulas matched to or specified in argument
\code{fml_nms}. Formula-specific threshold is allowed by supplying a vector
of absolute values in \code{log2FC}.}

\item{gslogFC_cutoff}{Numeric value or vector; the cut-off in gene-set
enrichment fold change. Only enrichment terms with absolute fold change
greater than the threshold will be reported. The default magnitude is
\code{log2(1.2)} for all formulas matched to or specified in argument
\code{fml_nms}. Formula-specific threshold is allowed by supplying a vector
of absolute values in \code{log2FC}.}

\item{gspval_cutoff}{Depreciated. Numeric value or vector; the cut-off in
gene-set significance \code{pVal}. Only enrichment terms with \code{pVals}
more significant than the threshold will be reported. The default is 0.05
for all formulas matched to or specified in argument \code{fml_nms}.
Formula-specific threshold is allowed by supplying a vector of cut-off
values.}

\item{min_size}{Numeric value or vector; minimum number of protein entries for
consideration in gene set tests. The number is after data filtration by
\code{pval_cutoff}, \code{logFC_cutoff} or varargs expressions under
\code{filter_}. The default is 10 for all formulas matched to or specified
in argument \code{fml_nms}. Formula-specific threshold is allowed by
supplying a vector of sizes.}

\item{max_size}{Numeric value or vector; maximum number of protein entries for
consideration in gene set tests. The number is after data filtration by
\code{pval_cutoff}, \code{logFC_cutoff} or varargs expressions under
\code{filter_}. The default in infinite for all formulas matched to or
specified in argument \code{fml_nms}. Formula-specific threshold is allowed
by supplying a vector of sizes.}

\item{min_delta}{Numeric value or vector; the minimum count difference between
the up- and the down-expressed group of proteins for consideration in gene
set tests. For example at \code{min_delta = 4}, a gene set will 6
upregulated proteins and 2 down-expressed proteins, or vice versa, will be
assessed. The number is after data filtration by \code{pval_cutoff},
\code{logFC_cutoff} or varargs expressions under \code{filter_}. The default
is 4 for all formulas matched to or specified in argument \code{fml_nms}.
Formula-specific threshold is allowed by supplying a vector of sizes.}

\item{min_greedy_size}{Numeric value or vector; minimum number of unique
protein entries for a gene set to be considered essential. The default in
\code{1} for all formulas matched to or specified in argument
\code{fml_nms}. Formula-specific threshold is allowed by supplying a vector
of sizes.}

\item{use_adjP}{Logical; if TRUE, use Benjamini-Hochberg pVals. The default is
FALSE.}

\item{method}{Dummy argument to avoid incurring the corresponding argument in
\link[stats]{dist} by partial argument matches.}

\item{anal_type}{Character string; the type of analysis that are preset for
method dispatch in function factories. The value will be determined
automatically. Exemplary values include \code{anal_type = c("PCA",
"Corrplot", "EucDist", "GSPA", "Heatmap", "Histogram", "MDS", "Model",
"NMF", "Purge", "Trend", "LDA", ...)}.}

\item{...}{\code{filter_}: Variable argument statements for the row
filtration of data against the column keys in \code{Peptide.txt} for
peptides or \code{Protein.txt} for proteins. Each statement contains to a
list of logical expression(s). The \code{lhs} needs to start with
\code{filter_}. The logical condition(s) at the \code{rhs} needs to be
enclosed in \code{exprs} with round parenthesis. \cr \cr For example,
\code{pep_len} is a column key in \code{Peptide.txt}. The statement
\code{filter_peps_at = exprs(pep_len <= 50)} will remove peptide entries
with \code{pep_len > 50}. See also \code{\link{normPSM}}. \cr \cr
Additional parameters for plotting with \code{ggplot2}: \cr \code{xmin},
the minimum \eqn{x} at a log2 scale; the default is -2. \cr \code{xmax},
the maximum \eqn{x} at a log2 scale; the default is +2. \cr \code{xbreaks},
the breaks in \eqn{x}-axis at a log2 scale; the default is 1. \cr
\code{binwidth}, the binwidth of \code{log2FC}; the default is \eqn{(xmax -
xmin)/80}. \cr \code{ncol}, the number of columns; the default is 1. \cr
\code{width}, the width of plot; \cr \code{height}, the height of plot. \cr
\code{scales}, should the scales be fixed across panels; the default is
"fixed" and the alternative is "free".}
}
\description{
logFC_cutoff Numeric A threshold for the subset of data before the calculation
of adjusted pvals
}
