% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-volcanos.R
\name{gsvaMap}
\alias{gsvaMap}
\title{Volcano plots of protein \code{log2FC} under gene sets}
\usage{
gsvaMap(...)
}
\description{
\code{gsvaMap} is a wrapper function of \code{\link{proteoVolcano}} for
mapping of protein data by gene sets.
}
