% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gsva.R
\name{gsvaTest}
\alias{gsvaTest}
\title{Perform GSVA tests}
\usage{
gsvaTest(df, id = "entrez", label_scheme_sub, filepath, filename,
  complete_cases = FALSE, method = "limma", gset_nm = "go_sets",
  var_cutoff = 0.5, pval_cutoff = 1e-04, logFC_cutoff = log2(1.1),
  mx.diff = TRUE, ...)
}
\description{
Perform GSVA tests
}
