% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-volcanos.R
\name{gsVolcano}
\alias{gsVolcano}
\title{Volcano plots of protein \code{log2FC} under given gene sets}
\usage{
gsVolcano(in_name = NULL, df = NULL, contrast_groups = NULL,
  gsea_key = "term", gsets = NULL, theme = NULL, fml_nm = NULL,
  filepath = NULL, filename = NULL, adjP = FALSE,
  show_labels = TRUE, show_sig = "none", pval_cutoff = 1e-06,
  logFC_cutoff = log2(1.2), ...)
}
\description{
Volcano plots of protein \code{log2FC} under given gene sets
}
