% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-volcanos.R
\name{gsVolcano}
\alias{gsVolcano}
\title{Volcano Plots of Protein \code{log2-ratios} under Given Gene Sets}
\usage{
gsVolcano(df, contrast_groups, gsea_res, gsea_key = "term", gsets,
  volcano_theme, filepath = NULL, filename = NULL, adjP = FALSE,
  show_labels = TRUE, show_sig = NULL, pval_cutoff = 1e-06, ...)
}
\description{
Volcano Plots of Protein \code{log2-ratios} under Given Gene Sets
}
