% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-volcanos.R
\name{gsVolcano}
\alias{gsVolcano}
\title{Volcano plots of protein \code{log2FC} under given gene sets}
\usage{
gsVolcano(
  df2 = NULL,
  df = NULL,
  contrast_groups = NULL,
  gsea_key = "term",
  gsets = NULL,
  theme = NULL,
  fml_nm = NULL,
  filepath = NULL,
  filename = NULL,
  adjP = FALSE,
  show_labels = TRUE,
  show_sig = "none",
  gspval_cutoff = 1e-06,
  gslogFC_cutoff = log2(1.2),
  topn = Inf,
  ...
)
}
\arguments{
\item{df2}{Character vector or string; the name(s) of secondary data file(s).
An informatic task, i.e. \code{anal_prnTrend(...)} against a primary
\code{df} generates secondary files such as
\code{Protein_Trend_Z_nclust6.txt} etc. See also the \code{Auguments}
section or \code{\link{prnHist}} for the description of a primary \code{df}.}

\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{contrast_groups}{The contrast groups defined under a formula at \code{fml_nm}.}

\item{gsea_key}{Character string; the column key indicating the terms of gene sets.}

\item{gsets}{The gene sets.}

\item{theme}{A
\href{https://ggplot2.tidyverse.org/reference/ggtheme.html}{ggplot2}
theme, i.e., theme_bw(), or a custom theme. At the NULL default, a system
theme will be applied.}

\item{fml_nm}{A character string; the name of \code{fml}.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{filename}{A representative file name to outputs. By default, the name(s)
will be determined automatically. For text files, a typical file extension
is \code{.txt}. For image files, they are typically saved via
\code{\link[ggplot2]{ggsave}} or \code{\link[pheatmap]{pheatmap}} where the
image type will be determined by the extension of the file name.}

\item{adjP}{Logical; if TRUE, use Benjamini-Hochberg pVals in volcano plots.
The default is FALSE.}

\item{show_labels}{Logical; if TRUE, shows the labels of top twenty entries.
The default is TRUE.}

\item{show_sig}{Character string indicating the type of significance values to
be shown with \code{\link{gspaMap}}. The default is \code{"none"}.
Additional choices are from \code{c("pVal", "qVal")} where \code{pVal} or
\code{qVal} will be shown, respectively, in the facet grid of the plots.}

\item{gspval_cutoff}{Numeric value or vector for uses with
\code{\link{gspaMap}}. \code{Gene sets} with enrichment \code{pVals} less
significant than the threshold will be excluded from volcano plot
visualization. The default significance is 0.05 for all formulas matched to
or specified in argument \code{fml_nms}. Formula-specific threshold is
allowed by supplying a vector of cut-off values.}

\item{gslogFC_cutoff}{Numeric value or vector for uses with
\code{\link{gspaMap}}. \code{Gene sets} with absolute enrichment
\code{log2FC} less than the threshold will be excluded from volcano plot
visualization. The default magnitude is \code{log2(1.2) } for all formulas
matched to or specified in argument \code{fml_nms}. Formula-specific
threshold is allowed by supplying a vector of absolute values in
\code{log2FC}.}

\item{topn}{Numeric value or vector; top entries in gene sets ordered by
increasing \code{pVal} for visualization. The default is to use all
available entries.}

\item{...}{\code{filter_}: Variable argument statements for the row filtration
against data in a primary file linked to \code{df}. See also
\code{\link{normPSM}} for the format of \code{filter_} statements. \cr \cr
Additional parameters for plotting: \cr \code{xco}, the cut-off lines of
fold changes at position \code{x}; the default is at \eqn{-1.2} and
\eqn{+1.2}. \cr \code{yco}, the cut-off line of \code{pVal} at position
\code{y}; the default is \eqn{0.05}. \cr \code{width}, the width of plot;
\cr \code{height}, the height of plot. \cr \code{nrow}, the number of rows
in a plot.}
}
\description{
Volcano plots of protein \code{log2FC} under given gene sets
}
