% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hparse_fasta}
\alias{hparse_fasta}
\title{Helper of \link{parse_fasta} by accession type.}
\usage{
hparse_fasta(
  df_sub,
  fasta_db,
  pat_uni_acc,
  pat_uni_id,
  pat_ref_acc,
  fasta,
  warns
)
}
\arguments{
\item{df_sub}{An input data frame with a single type of protein accession.}

\item{fasta_db}{FASTA databases}

\item{pat_uni_acc}{Patterns of Uniprot accessions.}

\item{pat_uni_id}{Pattern of Uniprot IDs.}

\item{pat_ref_acc}{Patterns of RefSeq accessions.}

\item{fasta}{Names of FASTA database.}

\item{warns}{The type of warning.}
}
\description{
NA genes are replaced with prot_acc.
}
