% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{hpeptideMBR}
\alias{hpeptideMBR}
\title{Helper of \link{peptideMBR}}
\usage{
hpeptideMBR(
  ms1files,
  filelist,
  dat_dir,
  path_ms1,
  mbr_ret_tol = 25,
  max_mbr_fold = 20,
  step = 1e-05
)
}
\arguments{
\item{ms1files}{The names of MS1 data files.}

\item{filelist}{The name of peptide files.}

\item{dat_dir}{The working directory.}

\item{path_ms1}{The path to MS1 data files.}

\item{mbr_ret_tol}{The tolerance in MBR retention time in seconds.}

\item{max_mbr_fold}{The maximum absolute fold change in MBR.}

\item{step}{The mass error in \code{ppm / 1e6}.}
}
\description{
Helper of \link{peptideMBR}
}
