% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{impRefNA}
\alias{impRefNA}
\title{Impute NA for references}
\usage{
impRefNA(
  df,
  cols_ref,
  cols_smpl,
  refChannels,
  group_psm_by = "pep_seq_mod",
  sp_centers = NULL,
  imp_refs = FALSE,
  add_errs = FALSE
)
}
\arguments{
\item{df}{A data frame.}

\item{cols_ref}{The colume keys of reference IDs.}

\item{cols_smpl}{The column keys of sample IDs.}

\item{refChannels}{The reference channels.}

\item{group_psm_by}{A key for PSM grouping.}

\item{sp_centers}{A vector of log2FC centers for each species.}

\item{imp_refs}{Logical; impute missing references or not.}

\item{add_errs}{Logical; add normal errors or not.}
}
\description{
Impute NA for references
}
