% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{imputeNA}
\alias{imputeNA}
\alias{pepImp}
\alias{prnImp}
\title{Imputation of NA values}
\usage{
imputeNA(id, overwrite = FALSE, ...)

pepImp(...)

prnImp(...)
}
\arguments{
\item{id}{Character string to indicate the type of data. The value will be
determined automatically. Peptide data will be used at \code{id = pep_seq}
or \code{pep_seq_mod}, and protein data will be used at \code{id = prot_acc}
or \code{gene}.}

\item{overwrite}{Logical. If true, overwrite the previous results. The default
is FALSE.}

\item{...}{Parameters for \code{\link[mice]{mice}}}
}
\value{
\code{Peptide_impNA.txt} for peptide data and \code{Protein_impNA.txt}
 for protein data.
}
\description{
\code{imputeNA} imputes NA \code{log2FC} values in protein or peptide data.
Users should avoid call the method directly, but instead use the following
wrappers.

\code{pepImp} is a wrapper of \code{\link{imputeNA}} for peptide data

\code{prnImp} is a wrapper of \code{\link{imputeNA}} for protein data
}
\examples{
# ===================================
# Peptide NA imputation
# ===================================
pepImp(
  m = 3,
  maxit = 3,
)

# ===================================
# Protein NA imputation
# ===================================
prnImp(
  m = 5,
  maxit = 5,
)

}
