% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-informatics.R
\name{info_anal}
\alias{info_anal}
\title{Function factories for informatic analysis}
\usage{
info_anal(
  id = gene,
  col_select = NULL,
  col_group = NULL,
  col_order = NULL,
  col_color = NULL,
  col_fill = NULL,
  col_shape = NULL,
  col_size = NULL,
  col_alpha = NULL,
  color_brewer = NULL,
  fill_brewer = NULL,
  size_manual = NULL,
  shape_manual = NULL,
  alpha_manual = NULL,
  col_benchmark = NULL,
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  df = NULL,
  df2 = NULL,
  filepath = NULL,
  filename = NULL,
  anal_type = c("Corrplot", "Heatmap", "Histogram", "MA", "MDS", "Model", "NMF",
    "Trend")
)
}
\arguments{
\item{anal_type}{Character string; the type of analysis that are preset for
method dispatch in function factories. The value will be determined
automatically. Examplary values include \code{anal_type = c("PCA",
"Corrplot", "EucDist", "GSPA", "Heatmap", "Histogram", "MDS", "Model",
"NMF", "Purge", "Trend", ...)}.}
}
\value{
a function to the given \code{anal_type}.
}
\description{
\code{info_anal} produces functions for selected informatic analysis.
}
