% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-informatics.R
\name{info_anal}
\alias{info_anal}
\title{Functional factories for informatic analysis}
\usage{
info_anal(id = gene, col_select = NULL, col_group = NULL,
  col_order = NULL, col_color = NULL, col_fill = NULL,
  col_shape = NULL, col_size = NULL, col_alpha = NULL,
  col_benchmark = NULL, scale_log2r = FALSE, impute_na = FALSE,
  df = NULL, filepath = NULL, filename = NULL,
  anal_type = c("Corrplot", "Heatmap", "Histogram", "MA", "MDS", "Model",
  "NMF", "Trend"))
}
\arguments{
\item{anal_type}{The type of analysis; \code{anal_type = c("MDS", "PCA",
"EucDist", "MA", "Heatmap", "Histogram", "Corrplot", "Trend", "NMF",
"Model")}.}
}
\value{
a function to the given \code{anal_type}.
}
\description{
\code{info_anal} produces functions for selected informatic analysis.
}
