% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-specialty.R
\name{labEffPSM}
\alias{labEffPSM}
\title{TMT labeling efficiency}
\usage{
labEffPSM(
  group_psm_by = c("pep_seq", "pep_seq_mod"),
  group_pep_by = c("prot_acc", "gene"),
  dat_dir = NULL,
  expt_smry = "expt_smry.xlsx",
  frac_smry = "frac_smry.xlsx",
  fasta = NULL,
  pep_unique_by = "group",
  corrected_int = TRUE,
  rm_reverses = TRUE,
  rptr_intco = 1000,
  rm_craps = FALSE,
  rm_krts = FALSE,
  rm_outliers = FALSE,
  annot_kinases = FALSE,
  plot_rptr_int = TRUE,
  plot_log2FC_cv = TRUE,
  use_lowercase_aa = TRUE,
  ...
)
}
\arguments{
\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{expt_smry}{A character string to the \code{.xlsx} file containing the
metadata of TMT experiments. The default is \code{expt_smry.xlsx}.}

\item{frac_smry}{A character string to the \code{.xlsx} file containing
peptide fractionation summary. The default is \code{frac_smry.xlsx}.}

\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search. There is no default and users need to provide the
correct file path(s) and name(s).}

\item{pep_unique_by}{A character string for annotating the uniqueness of
peptides in \code{MaxQuant} PSMs. At the \code{group} default, the
uniqueness of peptides is by protein groups. At a more strigent criterion of
\code{protein}, the uniqueness of peptides is by protein entries. A new
column of \code{pep_isunique} with corresponding logical TRUE or FALSE will
be added to the PSM reports.}

\item{corrected_int}{A logical argument for uses with \code{MaxQuant} data. At
the TRUE default, values under columns "Reporter intensity corrected..." in
\code{MaxQuant} PSM results (\code{msms.txt}) will be used. Otherwise,
"Reporter intensity" values without corrections will be used.}

\item{rm_reverses}{A logical argument for uses with \code{MaxQuant} data. At
the TRUE default, \code{Reverse} entries will be removed.}

\item{rptr_intco}{Numeric; the threshold of reporter ion intensity. The
default is 1,000.}

\item{rm_craps}{Logical; if TRUE,
\code{\href{https://www.thegpm.org/crap/}{cRAP}} proteins will be removed.
The default is FALSE.}

\item{rm_krts}{Logical; if TRUE, keratin entries will be removed. The default
is FALSE.}

\item{rm_outliers}{Logical; if TRUE, PSM outlier removals will be performed
for peptides with more than two identifying PSMs. Dixon's method will be
used when \eqn{2 < n \le 25} and Rosner's method will be used when \eqn{n >
25}. The default is FALSE.}

\item{annot_kinases}{Logical; if TRUE, proteins of human or mouse origins
will be annotated with their kinase attributes. The default is FALSE.}

\item{plot_rptr_int}{Logical; if TRUE, the distributions of reporter-ion
intensities will be plotted. The default is TRUE.}

\item{plot_log2FC_cv}{Logical; if TRUE, the distributions of the CV of peptide
\code{log2FC} will be plotted. The default is TRUE.}

\item{use_lowercase_aa}{Logical; if TRUE, modifications in amino acid
residues will be abbreviated with lower-case and/or \code{^_~}. See the
table below for details. The default is TRUE.}

\item{...}{\code{filter_}: Variable argument statements for the filtration of
data rows. Each statement contains to a list of logical expression(s). The
\code{lhs} needs to start with \code{filter_}. The logical condition(s) at
the \code{rhs} needs to be enclosed in \code{exprs} with round parenthesis.
For example, \code{pep_expect} is a column key present in \code{Mascot} PSM
exports and \code{filter_psms_at = exprs(pep_expect <= 0.1)} will remove PSM
entries with \code{pep_expect > 0.1}.}
}
\description{
TMT labeling efficiency
}
