% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{load_dbs}
\alias{load_dbs}
\title{Load Species-specific Databases}
\usage{
load_dbs(dat_dir, expt_smry = "expt_smry.csv")
}
\description{
A function loads a set of precompiled tables in protein accessions, GO data and KEGG pathways.
}
\examples{
load_dbs()

}
\seealso{
\code{\link{setup_expts}} for supported species.
}
