% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{load_dbs}
\alias{load_dbs}
\title{Loads species-specific Databases}
\usage{
load_dbs(gset_nms = NULL, species = NULL)
}
\arguments{
\item{gset_nms}{Character string or vector containing the shorthanded name(s),
full file path(s), or both, to gene sets for enrichment analysis. For
species among \code{"human", "mouse", "rat"}, the default of
\code{c("go_sets", "c2_msig", "kinsub")} will utilize terms from gene
ontology (\code{GO}), molecular signatures (\code{MSig}) and
kinase-substrate network (\code{PSP Kinase-Substrate}). Custom \code{GO},
\code{MSig} and other data bases at given species are also supported. See
also: \code{\link{prepGO}} for the preparation of custom \code{GO};
\code{\link{prepMSig}} for the preparation of custom \code{MSig}. For other
custom data bases, follow the same format of list as \code{GO} or
\code{MSig}.}

\item{species}{Character string; the name of a species.}
}
\description{
A function loads a set of precompiled gene sets of 
\href{http://current.geneontology.org/products/pages/downloads.html}{GO}
and
\href{http://software.broadinstitute.org/gsea/msigdb}{molecular signatures}.
}
\examples{
\donttest{load_dbs("go_sets", "human")}

}
\seealso{
\code{\link{load_expts}} for supported species.
}
