% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{load_mgfs}
\alias{load_mgfs}
\title{Helper in loading MGFs.}
\usage{
load_mgfs(
  mgf_path,
  min_mass = 500L,
  min_ms2mass = 110L,
  topn_ms2ions = 100L,
  ppm_ms1 = 20L,
  ppm_ms2 = 25L,
  index_ms2 = FALSE
)
}
\arguments{
\item{mgf_path}{The file path to a list of MGF files. There is no default and
the experimenters need to supply the files.}

\item{min_mass}{A minimum mass of precursors.}

\item{min_ms2mass}{A minimum m/z of MS2 ions.}

\item{index_ms2}{Logical; if TRUE, converts MS2 m/z values to indices.}
}
\description{
Helper in loading MGFs.
}
