% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular.R
\name{make_subcell_expr}
\alias{make_subcell_expr}
\title{Prepare the matrix of expression data.}
\usage{
make_subcell_expr(
  df = NULL,
  id = "gene",
  key_col = "log10Int",
  key_group = "group",
  levels_fraction = c("1K", "3K", "5K", "9K", "12K", "15K", "30K", "79K", "120K", "SN"),
  scale = TRUE,
  match_colnames_to_levels_fraction = TRUE
)
}
\arguments{
\item{df}{An input data frame.}

\item{id}{An identifier.}

\item{key_col}{The column key where values will be used for preparing
expression data.}

\item{key_group}{The column key where group information will be found for
preparing expression data.}

\item{levels_fraction}{The levels of subcellular fractions.}

\item{scale}{Logical; to scale values by rowSums or not.}

\item{match_colnames_to_levels_fraction}{Llogical; to match the column
names of the expression data with the keys defined in
\code{levels_fraction} or not.}
}
\description{
One subtype at a time.
}
