% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{makeLFQXYTS}
\alias{makeLFQXYTS}
\title{Prepare the XYTS matrices for LFQ}
\usage{
makeLFQXYTS(
  pep_seq_modzs = NULL,
  pep_exp_mzs = NULL,
  pep_exp_ints = NULL,
  pep_apex_rets = NULL,
  pep_apex_scans = NULL,
  pep_apex_fwhm = NULL,
  pep_apex_n = NULL,
  add_colnames = TRUE
)
}
\arguments{
\item{pep_seq_modzs}{List of peptide sequences with modifications and charge
states.}

\item{pep_exp_mzs}{List of m-over-z (X) values.}

\item{pep_exp_ints}{List of intensity (Y) values.}

\item{pep_apex_rets}{List of retention time (T) values.}

\item{pep_apex_scans}{List of apex scan numbers (S).}

\item{pep_apex_fwhm}{List of apex FWHM values.}

\item{pep_apex_n}{List of apex lengths (number of MS2 scans).}

\item{add_colnames}{Logical; add column names or not.}
}
\description{
Prepare the XYTS matrices for LFQ
}
