% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{map_essential}
\alias{map_essential}
\title{A helper function for mapping between gene sets and essential gene sets}
\usage{
map_essential(sig_sets)
}
\arguments{
\item{sig_sets}{A data frame containing the gene sets that are significant
under given criteria.}
}
\description{
A helper function for mapping between gene sets and essential gene sets
}
