% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{map_essential}
\alias{map_essential}
\title{A helper function for mapping btw gene sets and essential gene sets}
\usage{
map_essential(sig_sets, sig_greedy_terms)
}
\description{
A helper function for mapping btw gene sets and essential gene sets
}
