% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-go.R
\name{map_to_entrez}
\alias{map_to_entrez}
\title{Map uniprot or refseq to entrez (not currently used)}
\usage{
map_to_entrez(species, from)
}
\description{
Map uniprot or refseq to entrez (not currently used)
}
\examples{
\dontrun{
# available species: c("human", "mouse", "rat")
res <- map_to_entrez(species = "human", from = "egUNIPROT")
dir.create(file.path("~\\\\proteoQ\\\\dbs\\\\entrez\\\\to_unirpot"), recursive = TRUE, showWarnings = FALSE)
write.table(res, file.path("~\\\\proteoQ\\\\dbs\\\\entrez\\\\to_unirpot", paste0("uniprot_entrez_", species, ".txt")), sep = "\t", col.names = TRUE, row.names = FALSE)

res <- map_entrez(species, from = "egREFSEQ")
dir.create(file.path("~\\\\proteoQ\\\\dbs\\\\entrez\\\\to_refseq"), recursive = TRUE, showWarnings = FALSE)
write.table(res, file.path("~\\\\proteoQ\\\\dbs\\\\entrez\\\\to_refseq", paste0("refseq_entrez_", species, ".txt")), sep = "\t", col.names = TRUE, row.names = FALSE)
}
}
