% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_expt}
\alias{match_expt}
\title{Matches the file name containing the summary of TMT experiment}
\usage{
match_expt(fn_pars = "load_expts.txt")
}
\arguments{
\item{fn_pars.}{}
}
\description{
The default file name is \code{expt_smry.xlsx}. The \code{match_expt} matches
the file name provided by users.
}
