% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_frac}
\alias{match_frac}
\title{Matches the file name containing the information of analyte fractionation}
\usage{
match_frac(fn_pars = "load_expts.txt")
}
\arguments{
\item{fn_pars.}{}
}
\description{
The default file name is \code{frac_smry.xlsx}. The \code{match_frac} matches
the file name provided by users.
}
