% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{match_gspa_filename}
\alias{match_gspa_filename}
\title{Matches the name of GSPA result file}
\usage{
match_gspa_filename(anal_type = "GSPA", subdir = NULL)
}
\arguments{
\item{anal_type}{}
}
\description{
Matches the name of GSPA result file
}
