% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils_to_be_removed.R
\name{match_identifier}
\alias{match_identifier}
\title{Matches the current id to the id in normPep or normPrn}
\usage{
match_identifier(id = c("pep_seq", "pep_seq_mod", "prot_acc", "gene"))
}
\arguments{
\item{id.}{}
}
\description{
Matches the current id to the id in normPep or normPrn
}
