% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{match_logi_gv}
\alias{match_logi_gv}
\title{Match to a global logical variable}
\usage{
match_logi_gv(var)
}
\description{
Match to a global logical variable
}
\examples{
foo <- function(scale_log2r = FALSE) {
  match_logi_gv(scale_log2r)
}
foo()
}
