% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{match_scale_log2r}
\alias{match_scale_log2r}
\title{Match scale_log2r}
\usage{
match_scale_log2r(scale_log2r)
}
\description{
\code{match_scale_log2r} matches `scale_log2r` from the caller environment.
}
