% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{matchMS}
\alias{matchMS}
\title{Searches MS ions.}
\usage{
matchMS(
  out_path = "~/proteoQ/outs",
  mgf_path = file.path(out_path, "mgf"),
  fasta = c("~/proteoQ/dbs/fasta/uniprot/uniprot_hs_2020_05.fasta",
    "~/proteoQ/dbs/fasta/crap/crap.fasta"),
  acc_type = c("uniprot_acc", "other"),
  acc_pattern = NULL,
  fixedmods = c("TMT6plex (K)", "Carbamidomethyl (C)"),
  varmods = c("TMT6plex (N-term)", "Acetyl (Protein N-term)", "Oxidation (M)",
    "Deamidated (N)", "Gln->pyro-Glu (N-term = Q)"),
  include_insource_nl = FALSE,
  enzyme = c("trypsin"),
  maxn_fasta_seqs = 2e+05,
  maxn_vmods_setscombi = 64,
  maxn_vmods_per_pep = 5,
  maxn_sites_per_vmod = 3,
  maxn_vmods_sitescombi_per_pep = 32,
  min_len = 7,
  max_len = 100,
  max_miss = 2,
  type_ms2ions = "by",
  topn_ms2ions = 100,
  minn_ms2 = 6,
  ppm_ms1 = 20,
  ppm_ms2 = 25,
  ppm_reporters = 10,
  quant = c("none", "tmt6", "tmt10", "tmt11", "tmt16"),
  target_fdr = 0.01,
  fdr_type = c("psm", "peptide", "protein"),
  combine_tier_three = FALSE,
  digits = 5
)
}
\arguments{
\item{out_path}{The output path.}

\item{mgf_path}{The file path to a list of MGF files. There is no default and
the experimenters need to supply the files.}

\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. There is no default and the experimenters need to
supply the files.}

\item{acc_type}{Character string(s); the types of protein accessions in one
of c("uniprot_acc", "uniprot_id", "refseq_acc", "other"). For custom names,
the corresponding regular expressions need to be supplied via argument
\code{acc_pattern}.}

\item{acc_pattern}{Regular expression(s) describing the patterns to separate
the header lines of fasta entries. At the \code{NULL} default, the pattern
will be automated when \code{acc_type} are among c("uniprot_acc",
"uniprot_id", "refseq_acc", "other"). See also \link{load_fasta2} for
custom examples.}

\item{fixedmods}{A character vector of fixed modifications. See also
\link{parse_unimod} for grammars.}

\item{varmods}{A character vector of variable modifications.}

\item{include_insource_nl}{Logical Logical; if TRUE, includes MS1 precursor
masses with the losses of neutral species prior to MS2 fragmentation. The
default is FALSE. The setting at TRUE remains experimenting by allowing
additional masses in the universe of MS1 precursors. The offsets in NLs (by
adding back precursor masses) have not yet taken into account in MS2 ion
searches. A more systemically approach such as \code{open} MS1 masses might
be developed in the future.}

\item{enzyme}{A character string; the proteolytic specificity of the assumed
enzyme will be used to generate peptide sequences from proteins. The enzyme
is currently \code{trypsin}.}

\item{maxn_fasta_seqs}{Integer; the maximum number of protein sequences in
fasta files.}

\item{maxn_vmods_setscombi}{Integer; the maximum number of combinatorial variable
modifications and neutral losses.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{maxn_vmods_sitescombi_per_pep}{Integer; the maximum number of
combinatorial variable modifications per peptide sequence.}

\item{min_len}{Integer; the minimum length of peptides. Shorter peptides will
be excluded.}

\item{max_len}{Integer; the maximum length of peptides. Longer peptides will
be excluded.}

\item{max_miss}{The maximum number of mis-cleavages per peptide sequence.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{minn_ms2}{Integer; the minimum number of MS2 ions for consideration as
a hit.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{ppm_reporters}{The mass tolerance of MS2 reporter ions.}

\item{quant}{A quantitation method. The default is "none". Additional choices
include \code{tmt6} etc. For other multiplicities of \code{tmt}, use the
compatible higher plexes, for example, \code{tmt16} for \code{tmt12} etc.
and \code{tmt10} for \code{tmt8} etc.}

\item{target_fdr}{Numeric; the levels of false-discovery rate (FDR) at the
levels of PSMs or peptides. The same level applies further to protein FDR
and indicated under the column \code{prot_issig} in outputs. This results
in FDR controls of either PSM <-> protein or peptide <-> protein at a given
\code{target_fdr}.}

\item{fdr_type}{Character string; the type of FDR controlling. The value is
in one of c("psm", "peptide"). Note that protein FDR is in conjunction with
"psm" or "peptide". Separate protein FDR may be available in the future.}

\item{combine_tier_three}{Logical; if TRUE, combines tier-3 proteins in the
output of \code{psmQ.txt}. If FALSE, saves tier-3 proteins separately in
\code{psmT3.txt}.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Searches MS ions.
}
\seealso{
\link{load_fasta2} for setting the values of \code{acc_type} and
  \code{acc_pattern}. \link{parse_unimod} for the grammar of Unimod.
}
