% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{mcalc_monopep}
\alias{mcalc_monopep}
\title{Calculates multiply the molecular weight of a polypeptides ([MH]+).}
\usage{
mcalc_monopep(
  aa_seqs,
  aa_masses,
  include_insource_nl = FALSE,
  maxn_vmods_per_pep = 5,
  maxn_sites_per_vmod = 3,
  parallel = TRUE,
  digits = 5
)
}
\arguments{
\item{aa_seqs}{Character string; a vector of peptide sequences with
one-letter representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{include_insource_nl}{Logical Logical; if TRUE, includes MS1 precursor
masses with the losses of neutral species prior to MS2 fragmentation. The
default is FALSE. The setting at TRUE remains experimenting by allowing
additional masses in the universe of MS1 precursors. The offsets in NLs (by
adding back precursor masses) have not yet taken into account in MS2 ion
searches. A more systemically approach such as \code{open} MS1 masses might
be developed in the future.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{parallel}{Logical; if TRUE, performs parallel computation. The default
is TRUE.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
The calculations iterate through protein accessions and peptide sequences.
}
\examples{
\dontrun{
aa_masses_all <- calc_aamasses()

# Error if peptide sequences not 'compatible' with the given 'aa_masses'
x <- mcalc_monopep(list(protein_a = c("AAIMDWFDMGKEFSGNPNIK", "MAAIDWFDGKEFSGNPNIK"),
                        protein_b = c("AAIDWFDGKEMFSGNPNIK")),
                   aa_masses_all[[11]])
}
}
