% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{mcalc_monopep}
\alias{mcalc_monopep}
\title{Calculates multiply the molecular weight of a polypeptides ([MH]+).}
\usage{
mcalc_monopep(
  aa_seqs,
  aa_masses,
  maxn_vmods_per_pep = 5,
  maxn_sites_per_vmod = 3,
  parallel = TRUE,
  n_cores = parallel::detectCores(),
  cl = parallel::makeCluster(getOption("cl.cores", n_cores)),
  digits = 5
)
}
\arguments{
\item{aa_seqs}{Character string; a vector of peptide sequences with
one-letter representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{parallel}{Logical; if TRUE, performs parallel computation. The default
is TRUE.}

\item{n_cores}{Integer; the number of computer cores for uses.}

\item{cl}{The cluster nodes from \link[parallel]{makeCluster}.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
The calculations iterate through protein accessions and peptide sequences.
}
\examples{
\dontrun{
data(package = "proteoQ", aa_residues)
aa_masses <- aa_residues \%>\%
  dplyr::select(c("one_letter", "monoisotopic_da"))
aa_masses <- aa_masses$monoisotopic_da \%>\% `names<-`(aa_masses$one_letter)

x <- mcalc_monopep(list(protein_a = c("AAIDWFDGKEFSGNPIK", "MAAIDWFDGKEFSGNPIK"),
                        protein_b = c("AAIDWFDGKEFSGNPIK")),
                   aa_masses)

aa_masses_all <- calc_aamasses()

# Error if peptide sequences not 'compatible' with the given 'aa_masses'
x <- mcalc_monopep(list(protein_a = c("AAIMDWFDMGKEFSGNPNIK", "MAAIDWFDGKEFSGNPNIK"),
                        protein_b = c("AAIDWFDGKEMFSGNPNIK")),
                   aa_masses_all[[11]])
}
}
