% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{mcalc_monopep}
\alias{mcalc_monopep}
\title{Calculates multiply the molecular weight of a polypeptides ([MH]+).}
\usage{
mcalc_monopep(
  aa_seqs,
  aa_masses,
  include_insource_nl = FALSE,
  maxn_vmods_per_pep = 5,
  maxn_sites_per_vmod = 3,
  parallel = TRUE,
  n_cores = parallel::detectCores(),
  cl = parallel::makeCluster(getOption("cl.cores", n_cores)),
  digits = 5
)
}
\arguments{
\item{aa_seqs}{Character string; a vector of peptide sequences with
one-letter representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{include_insource_nl}{Logical Logical; if TRUE, includes MS1 precursor
  masses with the losses of neutral species prior to MS2 fragmentation. The
  default is FALSE.

  Note that there is no combination of neutral losses for fixed modifications
  at the precursor levels. Changes from fixed to variable modifications for
  complete combinations.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{parallel}{Logical; if TRUE, performs parallel computation. The default
is TRUE.}

\item{n_cores}{Integer; the number of computer cores for uses.}

\item{cl}{The cluster nodes from \link[parallel]{makeCluster}.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
The calculations iterate through protein accessions and peptide sequences.
}
\examples{
\dontrun{
aa_masses_all <- calc_aamasses()

# Error if peptide sequences not 'compatible' with the given 'aa_masses'
x <- mcalc_monopep(list(protein_a = c("AAIMDWFDMGKEFSGNPNIK", "MAAIDWFDGKEFSGNPNIK"),
                        protein_b = c("AAIDWFDGKEMFSGNPNIK")),
                   aa_masses_all[[11]])
}
}
