% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{mcPSM}
\alias{mcPSM}
\title{Median-centering normalization of PSM data}
\usage{
mcPSM(
  df = NULL,
  set_idx = 1L,
  injn_idx = 1L,
  mc_psm_by = "peptide",
  group_psm_by = "pep_seq",
  group_pep_by = "prot_acc",
  rm_allna = FALSE
)
}
\arguments{
\item{df}{A data frame containing the PSM table from database searches.}

\item{set_idx}{A set index.}

\item{injn_idx}{An injection index.}

\item{mc_psm_by}{A character string specifying the method in the median
centering of PSM \code{log2FC} across samples. At the \code{peptide}
default, the median description of PSMs (grouped by \code{pep_seq} or
\code{pep_seq_mod} according to \code{group_psm_by}) will be first
calculated and the offsets to zero (of logarithmic 2) will be used for the
centering of PSMs across samples. At \code{mc_psm_by = protein}, the median
description of PSMs (grouped by \code{prot_acc} or \code{gene} according to
\code{group_pep_by}) will be calculated and the corresponding offsets to
zero will be applied. At the \code{mc_psm_by = psm}, PSMs will be median
centered without grouping.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{rm_allna}{Logical; if TRUE, removes data rows that are exclusively NA
across ratio columns of \code{log2_R126} etc. The setting also applies to
\code{log2_R000} in LFQ.}
}
\description{
\code{mcPSM} adds fields of \code{log2_R, N_log2_R and N_I} to PSM tables.
}
