% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtests.R
\name{model_onechannel}
\alias{model_onechannel}
\title{Factorial model formula for interaction terms. All factors under one channel
in `label_scheme_sub`}
\usage{
model_onechannel(
  dfR = NULL,
  dfI = NULL,
  id,
  formula,
  label_scheme_sub,
  complete_cases = FALSE,
  impute_group_na = FALSE,
  method = "limma",
  padj_method = "BH",
  var_cutoff = 0.001,
  pval_cutoff = 1,
  logFC_cutoff = log2(1),
  ...
)
}
\arguments{
\item{dfI}{Intensity data frame.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{formula}{Language; the formula in linear modeling.}

\item{label_scheme_sub}{A data frame. Subset entries from \code{label_scheme}
for selected samples.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{impute_group_na}{Logical; if TRUE, impute (intensity) values that are
exclusively NA under a sample group. The sample grouping is defined under
\code{label_scheme}.}

\item{method}{Dummy argument to avoid incurring the corresponding argument in
\link[stats]{dist} by partial argument matches.}

\item{padj_method}{Character string; the method of multiple-test corrections
for uses with \link[stats]{p.adjust}. The default is "BH". See
?p.adjust.methods for additional choices.}

\item{var_cutoff}{Numeric; the cut-off in the variances of protein log2FC.
Entries with variances smaller than the threshold will be removed from GSVA.
The default is 0.5.}

\item{pval_cutoff}{Numeric value or vector; the cut-off in protein
significance \code{pVal}. Entries with \code{pVals} less significant than
the threshold will be excluded from enrichment analysis. The default is 0.05
for all formulas matched to or specified in argument \code{fml_nms}.
Formula-specific threshold is allowed by supplying a vector of cut-off
values.}

\item{logFC_cutoff}{Numeric value or vector; the cut-off in protein
\code{log2FC}. Entries with absolute \code{log2FC} smaller than the
threshold will be excluded from enrichment analysis. The default magnitude
is \code{log2(1.2)} for all formulas matched to or specified in argument
\code{fml_nms}. Formula-specific threshold is allowed by supplying a vector
of absolute values in \code{log2FC}.}

\item{...}{\code{filter_}: Variable argument statements for the row
filtration of data against the column keys in \code{Peptide.txt} for
peptides or \code{Protein.txt} for proteins. Each statement contains to a
list of logical expression(s). The \code{lhs} needs to start with
\code{filter_}. The logical condition(s) at the \code{rhs} needs to be
enclosed in \code{exprs} with round parenthesis. \cr \cr For example,
\code{pep_len} is a column key in \code{Peptide.txt}. The statement
\code{filter_peps_at = exprs(pep_len <= 50)} will remove peptide entries
with \code{pep_len > 50}. See also \code{\link{normPSM}}. \cr \cr
Additional parameters for plotting with \code{ggplot2}: \cr \code{xmin},
the minimum \eqn{x} at a log2 scale; the default is -2. \cr \code{xmax},
the maximum \eqn{x} at a log2 scale; the default is +2. \cr \code{xbreaks},
the breaks in \eqn{x}-axis at a log2 scale; the default is 1. \cr
\code{binwidth}, the binwidth of \code{log2FC}; the default is \eqn{(xmax -
xmin)/80}. \cr \code{ncol}, the number of columns; the default is 1. \cr
\code{width}, the width of plot; \cr \code{height}, the height of plot. \cr
\code{scales}, should the scales be fixed across panels; the default is
"fixed" and the alternative is "free".}
}
\description{
Factorial model formula for interaction terms. All factors under one channel
in `label_scheme_sub`
}
