% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-sigtests.R
\name{model_onechannel}
\alias{model_onechannel}
\title{Factorial model formula for interaction terms. All factors under one channel
in `label_scheme_sub`}
\usage{
model_onechannel(
  df,
  id,
  formula,
  label_scheme_sub,
  complete_cases,
  method,
  var_cutoff,
  pval_cutoff,
  logFC_cutoff,
  ...
)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{formula}{Language; the formula in linear modeling.}

\item{label_scheme_sub}{A data frame. Subset entries from \code{label_scheme}
for selected samples.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{method}{Character string; the method to assess the p-values of GSPA. The
default is \code{mean}. See also section \code{Details} for the
calculations.}

\item{var_cutoff}{Numeric; the cut-off in the variances of protein log2FC.
Entries with variances smaller than the threshold will be removed from GSVA.
The default is 0.5.}

\item{pval_cutoff}{Numeric value or vector; the cut-off in protein
significance \code{pVal}. Entries with \code{pVals} less significant than
the threshold will be excluded from enrichment analysis. The default is 0.05
for all formulas matched to or specified in argument \code{fml_nms}.
Formula-specific threshold is allowed by supplying a vector of cut-off
values.}

\item{logFC_cutoff}{Numeric value or vector; the cut-off in protein
\code{log2FC}. Entries with absolute \code{log2FC} smaller than the
threshold will be excluded from enrichment analysis. The default magnitude
is \code{log2(1.2)} for all formulas matched to or specified in argument
\code{fml_nms}. Formula-specific threshold is allowed by supplying a vector
of absolute values in \code{log2FC}.}

\item{...}{\code{filter_}: Variable argument statements for the row filtration
of data against the column keys in \code{Peptide.txt} for peptides or
\code{Protein.txt} for proteins. Each statement contains to a list of
logical expression(s). The \code{lhs} needs to start with \code{filter_}.
The logical condition(s) at the \code{rhs} needs to be enclosed in
\code{exprs} with round parenthesis. \cr \cr For example, \code{pep_len} is
a column key present in \code{Mascot} peptide tables of \code{Peptide.txt}.
The statement \code{filter_peps_at = exprs(pep_len <= 50)} will remove
peptide entries with \code{pep_len > 50}. See also \code{\link{normPSM}}.
\cr \cr Additional parameters for plotting with \code{ggplot2}: \cr
\code{xmin}, the minimum \eqn{x} at a log2 scale; the default is -2. \cr
\code{xmax}, the maximum \eqn{x} at a log2 scale; the default is +2. \cr
\code{xbreaks}, the breaks in \eqn{x}-axis at a log2 scale; the default is
1. \cr \code{binwidth}, the binwidth of \code{log2FC}; the default is
\eqn{(xmax - xmin)/80}. \cr \code{ncol}, the number of columns; the default
is 1. \cr \code{width}, the width of plot; \cr \code{height}, the height of
plot. \cr \code{scales}, should the scales be fixed across panels; the
default is "fixed" and the alternative is "free".}
}
\description{
Factorial model formula for interaction terms. All factors under one channel
in `label_scheme_sub`
}
