% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches2.R
\name{ms2match}
\alias{ms2match}
\title{Matches theoretical peptides (parallel by mgf chunks).}
\usage{
ms2match(
  mgf_path,
  aa_masses_all,
  out_path,
  mod_indexes,
  type_ms2ions,
  maxn_vmods_per_pep,
  maxn_sites_per_vmod,
  maxn_vmods_sitescombi_per_pep,
  minn_ms2,
  ppm_ms1,
  ppm_ms2,
  min_ms2mass,
  quant,
  ppm_reporters,
  fasta,
  acc_type,
  acc_pattern,
  topn_ms2ions,
  fixedmods,
  varmods,
  include_insource_nl,
  enzyme,
  maxn_fasta_seqs,
  maxn_vmods_setscombi,
  min_len,
  max_len,
  max_miss,
  digits
)
}
\arguments{
\item{mgf_path}{The file path to a list of MGF files. There is no default and
the experimenters need to supply the files.}

\item{aa_masses_all}{A list of amino acid lookups for all the combination of
fixed and variable modifications.}

\item{out_path}{An output path.}

\item{mod_indexes}{Integer; the indexes of fixed and/or variable
modifications.}

\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. There is no default and the experimenters need to
supply the files.}

\item{acc_pattern}{Regular expression(s) describing the patterns to separate
the header lines of fasta entries. At the \code{NULL} default, the pattern
will be automated when \code{acc_type} are among c("uniprot_acc",
"uniprot_id", "refseq_acc", "other"). See also \link{load_fasta2} for
custom examples.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{n_cores}{Integer; the number of computer cores.}
}
\description{
All files under `out_path` are removed if incur \code{calc_pepmasses} in the
upstream.
}
