% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm.R
\name{my_pheatmap}
\alias{my_pheatmap}
\title{A wrapper of pheatmap}
\usage{
my_pheatmap(
  mat,
  filename,
  annotation_col,
  annotation_row,
  color,
  annotation_colors,
  breaks,
  ...
)
}
\arguments{
\item{mat}{The same as in \link[pheatmap]{pheatmap}.}

\item{filename}{The output filename.}

\item{annotation_col}{The same as in \link[pheatmap]{pheatmap}.}

\item{annotation_row}{The same as in \link[pheatmap]{pheatmap}.}

\item{color}{The same as in \link[pheatmap]{pheatmap}.}

\item{annotation_colors}{The same as in \link[pheatmap]{pheatmap}.}

\item{breaks}{The same as in \link[pheatmap]{pheatmap}.}

\item{...}{Additional arguments for \link[pheatmap]{pheatmap}.}
}
\description{
A wrapper of pheatmap
}
