% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-hm.R
\name{my_pheatmap}
\alias{my_pheatmap}
\title{A wrapper around pheatmap}
\usage{
my_pheatmap(mat, filename, annotation_col, color, annotation_colors,
  breaks, ...)
}
\description{
A wrapper around pheatmap
}
