% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{my_union}
\alias{my_union}
\title{my union of named list 
names will be kept after the unification}
\usage{
my_union(x, y)
}
\description{
my union of named list 
names will be kept after the unification
}
