% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{n_TMT_sets}
\alias{n_TMT_sets}
\title{Finds the number of multiplex TMT experiments}
\usage{
n_TMT_sets(label_scheme_full)
}
\arguments{
\item{label_scheme_full}{The label_scheme with the probable inclusion of
different LCMS_inj under the same TMT_Set.}
}
\description{
Finds the number of multiplex TMT experiments
}
