% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{na_zeroIntensity}
\alias{na_zeroIntensity}
\title{Replace zero intensity with NA}
\usage{
na_zeroIntensity(df)
}
\arguments{
\item{df}{A list of file names.}
}
\description{
\code{na_zeroIntensity} replaces zero intensity with NA to avoid -Inf in
log10 transformation.
}
