% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{names_pos<-}
\alias{names_pos<-}
\title{Prefix form of colnames(x)[c(2, 5, ...)] for use in pipes}
\usage{
names_pos(x, pos) <- value
}
\arguments{
\item{x}{A data frame.}

\item{pos}{Numeric.  The index of coloumns for name change.}

\item{value}{Characters.  The new column names.}
}
\value{
The data frame with new names.
}
\description{
\code{names_pos<-} rename the columns at the indeces of \code{pos}.
}
