% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{newColnames}
\alias{newColnames}
\title{Make new column names}
\usage{
newColnames(
  TMT_Set = 1L,
  df,
  label_scheme,
  pattern = "[RI][0-9]{3}[NC]{0,1}",
  group_ids = NULL
)
}
\arguments{
\item{TMT_Set}{Integer; the index of TMT experiment.}

\item{df}{Data frame; log2FC data.}

\item{label_scheme}{Experiment summary.}

\item{pattern}{Regex pattern for capturing intensity and ratio columns.}

\item{group_ids}{A character vector; for example, c("heavy", "light") that
can be found from column \code{pep_group} in \code{df}; \code{group_ids} is
NULL if column \code{pep_group} is not in \code{df}.}
}
\description{
According to the Sample_ID(s) in label_scheme.
}
