% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-normalization.R
\name{normMulGau}
\alias{normMulGau}
\title{Data normalization}
\usage{
normMulGau(
  df,
  method_align,
  n_comp,
  seed = NULL,
  range_log2r,
  range_int,
  filepath,
  col_select = NULL,
  ...
)
}
\arguments{
\item{df}{An input data frame}
}
\value{
A data frame.
}
\description{
\code{normMulGau} normalizes \code{log2FC} under the assumption of multi
Gaussian kernels.
}
