% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-normalization.R
\name{normMulGau}
\alias{normMulGau}
\title{Data normalization}
\usage{
normMulGau(
  df,
  method_align,
  n_comp,
  seed = NULL,
  range_log2r,
  range_int,
  filepath,
  col_select = NULL,
  ...
)
}
\arguments{
\item{df}{An input data frame}

\item{method_align}{Character string indicating the method in aligning
\code{log2FC} across samples. \code{MC}: median-centering; \code{MGKernel}:
the kernel density defined by multiple Gaussian functions
(\code{\link[mixtools]{normalmixEM}}). At the \code{MC} default, the ratio
profiles of each sample will be aligned in that the medians of the
\code{log2FC} are zero. At \code{MGKernel}, the ratio profiles of each
sample will be aligned in that the \code{log2FC} at the maximums of kernel
density are zero.}

\item{n_comp}{Integer; the number of Gaussian components to be used with
\code{method_align = MGKernel}. A typical value is 2 or 3. The variable
\code{n_comp} overwrites the argument \code{k} in
\code{\link[mixtools]{normalmixEM}}.}

\item{seed}{Integer; a seed setting a starting point for reproducible
analyses.}

\item{range_log2r}{Numeric vector at length two. The argument specifies the
range of the \code{log2FC} for use in the scaling normalization of standard
deviation across samples. The default is between the 10th and the 90th
quantiles.}

\item{range_int}{Numeric vector at length two. The argument specifies the
range of the \code{intensity} of reporter ions for use in the scaling
normalization of standard deviation across samples. The default is between
the 5th and the 95th quantiles.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
At the \code{NULL} default, the column key of \code{Select} in
\code{expt_smry.xlsx} will be used. In the case of no samples being
specified under \code{Select}, the column key of \code{Sample_ID} will be
used. The non-empty entries under the ascribing column will be used in
indicated analysis.}

\item{...}{\code{filter_}: Variable argument statements for the row filtration
of data against the column keys in \code{Peptide.txt} for peptides or
\code{Protein.txt} for proteins. Each statement contains to a list of
logical expression(s). The \code{lhs} needs to start with \code{filter_}.
The logical condition(s) at the \code{rhs} needs to be enclosed in
\code{exprs} with round parenthesis. \cr \cr For example, \code{pep_len} is
a column key present in \code{Mascot} peptide tables of \code{Peptide.txt}.
The statement \code{filter_peps_at = exprs(pep_len <= 50)} will remove
peptide entries with \code{pep_len > 50}. See also \code{\link{normPSM}}.
\cr \cr Additional parameters for plotting with \code{ggplot2}: \cr
\code{xmin}, the minimum \eqn{x} at a log2 scale; the default is -2. \cr
\code{xmax}, the maximum \eqn{x} at a log2 scale; the default is +2. \cr
\code{xbreaks}, the breaks in \eqn{x}-axis at a log2 scale; the default is
1. \cr \code{binwidth}, the binwidth of \code{log2FC}; the default is
\eqn{(xmax - xmin)/80}. \cr \code{ncol}, the number of columns; the default
is 1. \cr \code{width}, the width of plot; \cr \code{height}, the height of
plot. \cr \code{scales}, should the scales be fixed across panels; the
default is "fixed" and the alternative is "free".}
}
\value{
A data frame.
}
\description{
\code{normMulGau} normalizes \code{log2FC} under the assumption of multi
Gaussian kernels.
}
