% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{normPep_Mplex}
\alias{normPep_Mplex}
\title{Combines peptide reports across multiple experiments.}
\usage{
normPep_Mplex(
  group_psm_by = "pep_seq_mod",
  group_pep_by = "prot_acc",
  engine = "mz",
  lfq_mbr = TRUE,
  use_duppeps = TRUE,
  duppeps_repair = "denovo",
  cut_points = Inf,
  omit_single_lfq = FALSE,
  use_mq_pep = FALSE,
  use_mf_pep = FALSE,
  rm_allna = FALSE,
  mbr_ret_tol = 25,
  max_mbr_fold = 4,
  ret_sd_tol = Inf,
  rm_ret_outliers = FALSE,
  ...
)
}
\arguments{
\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{engine}{The name of search engine.}

\item{lfq_mbr}{Logical; if TRUE, performs match-between-run (MBR) with Mzion
LFQ data. Also requires \code{ms1full_[rawfile].rds} at the same file-folder
level of \code{psmQ[...].txt}.}

\item{use_duppeps}{Logical; if TRUE, re-assigns double/multiple dipping
peptide sequences to the most likely proteins by majority votes.}

\item{duppeps_repair}{Not currently used (or only with \code{majority}).
 Character string; the method of reparing double-dipping peptide sequences
 upon data pooling.

 For instance, the same sequence of PEPTIDE may be assigned to protein
 accession PROT_ACC1 in data set 1 and PROT_ACC2 in data set 2. At the
 \code{denovo} default, the peptide to protein association will be
 re-established freshly. At the \code{majority} alternative, a majority rule
 will be applied for the re-assignments.}

\item{cut_points}{A named, numeric vector defines the cut points (knots) for
the median-centering of \code{log2FC} by sections. For example, at
\code{cut_points = c(mean_lint = seq(4, 7, .5))}, \code{log2FC} will be
binned according to the intervals of \eqn{-Inf, 4, 4.5, ..., 7, Inf} under
column \code{mean_lint} (mean log10 intensity) in the input data. The
default is \code{cut_points = Inf}, or equivalently \code{-Inf}, where the
\code{log2FC} under each sample will be median-centered as one piece. See
also \code{\link{prnHist}} for data binning in histogram visualization.}

\item{omit_single_lfq}{Logical; if TRUE, omits LFQ entries with single
measured values across all samples. The default is FALSE.}

\item{use_mq_pep}{Logical; if TRUE, uses the peptides.txt from MaxQuant.}

\item{use_mf_pep}{Logical; if TRUE, uses the peptides.txt from MSFragger.}

\item{rm_allna}{Logical; if TRUE, removes data rows that are exclusively NA
across ratio columns of \code{log2_R126} etc. The setting also applies to
\code{log2_R000} in LFQ.}

\item{mbr_ret_tol}{Retention time tolerance (in seconds) for LFQ-MBR.}

\item{max_mbr_fold}{The maximum absolute fold change in MBR.}

\item{ret_sd_tol}{Numeric; the tolerance in the variance of retention time
(w.r.t. measures in seconds). The thresholding applies to TMT data. The
default is \code{Inf}. Depends on the setting of LCMS gradients, a setting
of, e.g., 150 might be suitable.}

\item{rm_ret_outliers}{Logical; if TRUE, removes peptide entries with outlying
retention times across samples and/or LCMS series.}

\item{...}{\code{filter_}: Variable argument statements for the filtration of
data rows. Each statement contains to a list of logical expression(s). The
\code{lhs} needs to start with \code{filter_}. The logical condition(s) at
the \code{rhs} needs to be enclosed in \code{exprs} with round parenthesis.
For example, \code{pep_expect} is a column key present in \code{Mascot} PSM
exports and \code{filter_psms_at = exprs(pep_expect <= 0.1)} will remove
PSM entries with \code{pep_expect > 0.1}.}
}
\description{
Median summary of data from the same TMT or LFQ experiment at different LCMS
injections summed \code{pep_n_psm}, \code{prot_n_psm}, and \code{prot_n_pep}
after data merging no Z_log2_R yet available use \code{col_select =
expr(Sample_ID)} not \code{col_select} to get all Z_log2_R why: users may
specify \code{col_select} only partial to Sample_ID entries.
}
