% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-peptable.R
\name{normPep_Splex}
\alias{normPep_Splex}
\title{Peptide results for individual TMT experiments}
\usage{
normPep_Splex(id = "pep_seq_mod", method_psm_pep = "median")
}
\arguments{
\item{id}{The variable for summarising PSMs into peptides. PSMs with the same
value in \code{id} will be summarised into a single entry of peptide.}

\item{method_psm_pep}{The method for summarising \code{log2-ratios} from PSMs
to peptides. \code{log10-intensities} will be the weights if
\code{method_psm_pep = c("top.3", "weighted.mean")}.}
}
\value{
Results in \code{.txt} files for each of TMT experiments and LC/MS
  injections.
}
\description{
\code{normPep_Splex} prepares peptide data for each TMT experiment at
different LC/MS injections.
}
\examples{
	normPep_Splex(
		id = pep_seq_mod,
		method_psm_pep = "median"
	)

\dontrun{
}
}
