% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-peptable.R
\name{normPepSplex_mqpep}
\alias{normPepSplex_mqpep}
\title{Peptide reports for individual TMT experiments}
\usage{
normPepSplex_mqpep(id = "pep_seq_mod", fasta = fasta,
  pep_unique_by = "group", corrected_mq_int = TRUE,
  rm_mq_krts = TRUE, rm_mq_craps = TRUE, rm_mq_reverses = TRUE)
}
\arguments{
\item{id}{Character string; the variable for summarisation of \code{PSMs} into
peptides. The option \code{id = pep_seq} corresponds to the summarisation by
the primary sequences of peptides. The option \code{id = pep_seq_mod}
corresponds to the summarisation by both the primary sequences and variable
modifications of peptides. \code{PSMs} data with the same value in
\code{pep_seq} or \code{pep_seq_mod} will be summarised into a single entry
of peptide. Currently, the value of \code{id} will match automatically to
the value of \code{group_psm_by} in \code{normPSM}.}

\item{pep_unique_by}{A character string for use in the filtration of peptide
data from \code{MaxQuant}. The choice is in one of \code{c("group",
"protein", "none")} At the \code{group} default, only peptides unique by
protein groups will be kept. At \code{protein}, only peptides unique by
protein entries will be kept. At \code{none}, all peptides will be kept.}

\item{rm_mq_krts}{Logical; if TRUE, removes keratin entries from MaxQuant
peptide results.}

\item{rm_mq_craps}{Logical; if TRUE, removes \code{Potential contaminant}
entries from MaxQuant peptide results.}
}
\value{
Results in \code{.txt} files for each of TMT experiments and LC/MS
  injections.
}
\description{
\code{normPepSplex_mqpep} processes peptide data from MaxQuant peptide
table(s). It splits the original tables by TMT set numbers. Different LCMS
injections under the same TMT experiment will be handled as different
fractions. Therefore, the injection index is always "1".
}
