% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-prntable.R
\name{normPrn}
\alias{normPrn}
\title{Reports Protein Results}
\usage{
normPrn(id = c("prot_acc", "gene"), method_pep_prn = c("median",
  "mean", "weighted.mean", "top.3"), method_align = c("MC", "MGKernel"),
  range_log2r = c(20, 90), range_int = c(5, 95), n_comp = NULL,
  seed = NULL, fasta = NULL, col_refit = NULL, cache = TRUE, ...)
}
\arguments{
\item{id}{Character string; the variable to summarise peptides into proteins.
The option \code{prot_acc} corresponds to the summarisation by the accession
numbers or entry names of proteins. The option \code{gene} corresponds to
the summarisation by the gene names of proteins. At \code{id = gene}, data
under the same gene name but different acccesssion numbers or entry names
will be summarised into one entry.}

\item{method_pep_prn}{Character string; the method to summarise the
\code{log2FC} and the \code{intensity} of peptides by protein entries. The
descriptive statistics includes \code{c("mean", "median", "top.3",
"weighted.mean")}. The representative \code{log10-intensity} of reporter
ions at the peptide levels (from \code{\link{normPep}}) will be the weigth
when summarising \code{log2FC} with \code{top.3} or \code{weighted.mean}.}

\item{method_align}{Character string or a list of gene symbols; the method to
align the \code{log2FC} of peptide/protein entries across samples.
\code{MC}: median-centering; \code{MGKernel}: the kernal density defined by
multiple Gaussian functions (\code{\link[mixtools]{normalmixEM}}). At
\code{method_align = "MC"}, the ratio profiles of each sample will be
aligned in that the medians of the \code{log2FC} are zero. At
\code{method_align = "MGKernel"}, the \code{log2FC} will be aligned in that
the maximums of kernel density are zero. It is also possible to align the
\code{log2FC} to the median of a list of user-supplied genes:
\code{method_align = c("ACTB", "GAPDH", ...)}.}

\item{range_log2r}{Numeric vector at length two; the range of the
\code{log2FC} of peptide/protein entries for use in the scaling
normalization of standard deviation across samples. The default is between
the 20th and the 90th quantiles.}

\item{range_int}{Numeric vector at length two; the range of the
\code{intensity} of reporter ions for use in the scaling normalization of
standard deviation across samples. The default is between the 5th and the
95th quantiles.}

\item{n_comp}{Integer; the number of Gaussian components to be used with
\code{method_align = "MGKernel"}. A typical value is 2 or 3. The variable
\code{n_comp} overwrites the augument \code{k} in
\code{\link[mixtools]{normalmixEM}}.}

\item{seed}{Integer; a seed for reproducible fitting at \code{method_align =
MGKernel}.}

\item{fasta}{Character string(s); the file name(s) with prepended directory
path to the \code{fasta} database being used in MS/MS ion search.}

\item{col_refit}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_refit} will be
used in the refit of \code{log2FC} using multiple Gaussian kernels. The
density estimates from an earlier analyis will be kept for the remaining
samples. At the NULL default, the column key \code{Sample_ID} will be used,
which results in the refit of the \code{log2FC} for all samples.}

\item{cache}{Logical; if TRUE, use cache.}

\item{...}{Additional parameters for \code{\link[mixtools]{normalmixEM}}:\cr
\code{maxit}, the maximum number of iterations allowed; \cr \code{epsilon},
tolerance limit for declaring algorithm convergence.}
}
\value{
The primary output in "\code{~\\dat_dir\\Protein\\Protein.txt}".
}
\description{
\code{normPrn} summarises peptides into proteins and performs data
normalization.
}
\details{
The column keys are defined in
\code{\href{http://www.matrixscience.com/help/csv_headers.html}{Matrix
Science}} with the following additions:

\tabular{ll}{ \strong{Header}   \tab \strong{Descrption}\cr entrez \tab
\code{\href{https://www.ncbi.nlm.nih.gov/gene}{Protein Entrez ID}} \cr gene
\tab Protein gene name \cr kin_attr       \tab Optional; the attribute of
proteins being kinases \cr kin_class      \tab Optional; the classes of
kinases, e.g., TK, TKL... \cr kin_order      \tab Optional; the order of
"kin_class" from the \code{\href{http://kinase.com/human/kinome/}{kinase tree
diagram}} \cr n_pep \tab The number of significant peptide sequences matched
to a proposed protein \cr n_psm \tab The number of significant MS/MS queries
matched to a peptide sequence \cr pep_seq_mod    \tab One-letter
representation of peptide sequences: amino acid residue(s) with variable
modifications are in the lower cases; the acetylations of protein N-terminals
indicated by '_' and the flanking residues on the N- or C-terminal side of
peptides separated by '.', e.g. "-._mAsGVAVSDGVIK.V"\cr prot_cover     \tab
Protein sequence coverage calculated from all available raw files   \cr
raw_file \tab MS file name(s) where peptides or proteins were identified \cr
refseq_acc \tab \code{\href{https://www.ncbi.nlm.nih.gov/refseq/}{Protein
RefSeq accession number}} \cr uniprot_acc    \tab
\code{\href{https://www.uniprot.org/help/accession_numbers}{Protein UniProt
accession number}} \cr uniprot_id     \tab
\code{\href{https://www.uniprot.org/help/entry_name}{Protein UniProt entry
name}} \cr I... (...)     \tab Reporter-ion intensity calculated from the
descriptive statistics in \code{mothod_psm_pep} or \code{mothod_pep_prn} for
indicated samples \cr N_I... (...)   \tab Normalized \code{I... (...)}. The
calibration factors for the alignment of \code{log2FC} are used to scale the
reporter-ion intensity \cr log2_R (...)   \tab \code{log2FC} relative to
reference materials for indicated samples \cr N_log2_R (...) \tab Aligned
\code{log2_R (...)} according to \code{method_align} without scaling
normalization \cr Z_log2_R (...) \tab \code{N_log2_R (...)} with scaling
normalization \cr }
}
\examples{
\dontrun{
normPrn(
 id = gene,
 fasta = c("~\\\\proteoQ\\\\dbs\\\\refseq\\\\refseq_hs_2013_07.fasta", 
           "~\\\\proteoQ\\\\dbs\\\\refseq\\\\refseq_mm_2013_07.fasta"),
 method_pep_prn = median,
 method_align = MGKernel,
 range_log2r = c(5, 95),
 range_int = c(5, 95),
 n_comp = 2,
 seed = 749662,
 maxit = 200,
 epsilon = 1e-05
)
}
}
\seealso{
\code{\link{normPSM}} for PSMs and \code{\link{normPep}} for
 peptides.
}
\concept{aggregate functions}
