% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{normPSM}
\alias{normPSM}
\title{Reports PSM results}
\usage{
normPSM(dat_dir = NULL, expt_smry = "expt_smry.xlsx",
  frac_smry = "frac_smry.xlsx", rptr_intco = 1000, rm_craps = FALSE,
  rm_krts = FALSE, rm_outliers = FALSE, plot_violins = TRUE)
}
\arguments{
\item{rptr_intco}{Numeric; the threshold of reporter ion intensity.}

\item{rm_craps}{Logical; if TRUE, removes
\code{\href{https://www.thegpm.org/crap/}{cRAP}} proteins.}

\item{rm_krts}{Logical; if TRUE, removes keratin entries from the output.}

\item{rm_outliers}{Logical; if TRUE, performs outlier removals.}

\item{plot_violins}{Logical; if TRUE, prepares the violin plots of
reporter-ion intensities.}
}
\value{
Outputs under \code{C:\\my_direcotry\\PSM}. Primary results are in
 \code{TMTset1_LCMSinj1_PSM_N.txt, TMTset2_LCMSinj1_PSM_N.txt, ...} for
 respective TMT experiments and LC/MS injections.
}
\description{
\code{normPSM} summarises
\code{\href{https://www.ebi.ac.uk/pride/help/archive/search/tables}{PSM}}
results from \code{\href{https://en.wikipedia.org/wiki/Tandem_mass_tag}{TMT}}
experiments.
}
\details{
End users will export \code{PSM} data from
\code{\href{https://http://www.matrixscience.com/}{Mascot}} at a \code{.csv}
format. The header information should be included during the \code{csv}
export. A \code{.csv} filename should be defaulted by
\code{\href{https://http://www.matrixscience.com/}{Mascot}}: starting with the
letter \code{'F'}, followed by a six-digit number without space \code{(e.g.,
F004453.csv)}. The \code{PSM} files should be present under the folder,
\code{"~\\dat_dir"}, specified by the end users.

In each primary output file, "\code{...PSM_N.txt}", values under columns
\code{log2_R...} are relative to the \code{reference(s)} within each multiplex
TMT set or to the row means if no \code{reference(s)} are present. Values
under columns \code{N_log2_R...} are \code{log2_R...} with median-centering
alignment. Values under columns \code{N_I...} are normalized
\code{reporter-ion intensity}. Character strings under \code{pep_seq_mod}
denote peptide sequences with applicable variable modifications. See
\code{\link{normPrn}} for the description of column keys and normalization.
}
\examples{
# An examplary PSM file
system.file("extdata", "F012345.csv", package = "proteoQ")

dat_dir <- "C:\\\\my_directory"
normPSM(
  rptr_intco = 1000,
  rm_craps = FALSE,
  rm_krts = FALSE,
  rm_outliers = FALSE,
  plot_violins = TRUE
)

}
\seealso{
\code{\link{normPep}} for peptides and \code{\link{normPrn}} for
 proteins.
}
