% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{normPSM}
\alias{normPSM}
\title{Reports PSM results}
\usage{
normPSM(dat_dir = NULL, expt_smry = "expt_smry.xlsx",
  frac_smry = "frac_smry.xlsx", fasta = NULL,
  pep_unique_by = "group", corrected_int = FALSE, rm_reverses = TRUE,
  rptr_intco = 1000, rm_craps = FALSE, rm_krts = FALSE,
  rm_outliers = FALSE, plot_violins = TRUE, ...)
}
\arguments{
\item{dat_dir}{A character string to the working directory.}

\item{expt_smry}{A character string to the \code{.xlsx} file containing the
metadata of TMT experiments. The default is "expt_smry.xlsx".}

\item{frac_smry}{A character string to the \code{.xlsx} file containing
peptide fractionation summary. The default is "frac_smry.xlsx".}

\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search.}

\item{pep_unique_by}{A character string for use in the filtration of peptide
data from \code{MaxQuant}. The choice is in one of \code{c("group",
"protein", "none")} At the \code{group} default, only peptides unique by
protein groups will be kept. At \code{protein}, only peptides unique by
protein entries will be kept. At \code{none}, all peptides will be kept.}

\item{corrected_int}{Logical. If TRUE, values under columns "Reporter
intensity corrected" in MaxQuant PSM results (\code{msms.txt}) will be used.
Otherwise, "Reporter intensity" values without corrections will be used.}

\item{rm_reverses}{Logical; if TRUE, removes \code{Reverse} entries from
MaxQuant peptide results.}

\item{rptr_intco}{Numeric; the threshold of reporter ion intensity.}

\item{rm_craps}{Logical; if TRUE, removes
\code{\href{https://www.thegpm.org/crap/}{cRAP}} proteins.}

\item{rm_krts}{Logical; if TRUE, removes keratin entries from the output.}

\item{rm_outliers}{Logical; if TRUE, performs outlier removals.}

\item{plot_violins}{Logical; if TRUE, prepares the violin plots of
reporter-ion intensities.}

\item{...}{Additional parameters for row filtration of data.}
}
\value{
Outputs under \code{dat_dir\\PSM}. Primary results are in
 \code{TMTset1_LCMSinj1_PSM_N.txt, TMTset2_LCMSinj1_PSM_N.txt, ...} The
 indeces of TMT experiment and LC/MS injection are indicated in the file
 names.
}
\description{
\code{normPSM} reports
\code{\href{https://www.ebi.ac.uk/pride/help/archive/search/tables}{PSM}}
results from \code{\href{https://en.wikipedia.org/wiki/Tandem_mass_tag}{TMT}}
experiments.
}
\details{
The \code{PSM} files should be present under the working directory,
\code{"dat_dir"}, specified by end users.

In each primary output file, "\code{...PSM_N.txt}", values under columns
\code{log2_R...} are relative to the \code{reference(s)} within each multiplex
TMT set or to the row means if no \code{reference(s)} are present. Values
under columns \code{N_log2_R...} are \code{log2_R...} with median-centering
alignment. Values under columns \code{N_I...} are normalized
\code{reporter-ion intensity}. Character strings under \code{pep_seq_mod}
denote peptide sequences with applicable variable modifications.
}
\section{\code{Mascot}}{
 End users will export \code{PSM} data from
 \code{\href{https://http://www.matrixscience.com/}{Mascot}} at a \code{.csv}
 format. The header information should be included during the \code{.csv}
 export. The file name(s) should be defaulted by
 \code{\href{https://http://www.matrixscience.com/}{Mascot}}: starting with
 the letter \code{'F'}, followed by a six-digit number without space and
 ended with a \code{'.csv'} extension \code{(e.g., F004453.csv)}.

 See \code{\link{normPrn}} for the description of column keys in the output.
}

\section{\code{MaxQuant}}{
 End users will copy over \code{msms.txt}
 file(s) from \code{\href{https://www.maxquant.org/}{MaxQuant}} to the
 \code{dat_dir} directory. In the case of multiple \code{msms.txt} files for
 processing, the file names need to be compiled in that they all start with
 \code{'msms'} and end with a \code{'.txt'} extension.
}

\examples{
\dontrun{
# Mascot example
# set up a working directory
dat_dir <- c("C:\\\\The\\\\First\\\\Example")

# copy fasta
library(proteoQDA)
copy_refseq_hs("~\\\\proteoQ\\\\dbs\\\\refseq")
copy_refseq_mm("~\\\\proteoQ\\\\dbs\\\\refseq")

# copy Mascot PSM data
cptac_csv_1(dat_dir)

# copy "expt_smry.xlsx" and "frac_smry.xlsx"
cptac_expt_1(dat_dir)
cptac_frac_1(dat_dir)

# load experiments
library(proteoQ)
load_expts()

# process Mascot PSMs
normPSM(
  rptr_intco = 3000,
  rm_craps = FALSE,
  rm_krts = FALSE,
  rm_outliers = FALSE,
  plot_violins = TRUE, 
  
  # `dot-dot-dot` filtration of Mascot PSMs
  pep_rank = expr(pep_rank == 1), 
  pep_miss = expr(pep_miss <= 2), 
  pep_exp_z = expr(pep_exp_z == 2 | pep_exp_z == 3), 
)


# MaxQuant example
dat_dir <- c("C:\\\\The\\\\MQ\\\\PSM_Example")

# copy MaxQuant PSM data
library(proteoQDB)
cptac_mq_psm_1(dat_dir)

# copy "expt_smry.xlsx" and "frac_smry.xlsx"
cptac_expt_1(dat_dir)
cptac_frac_1(dat_dir)

# load experiments
library(proteoQ)
load_expts()

# process MaxQuant PSMs
normPSM(
  fasta = c("~\\\\proteoQ\\\\dbs\\\\refseq\\\\refseq_hs_2013_07.fasta",
            "~\\\\proteoQ\\\\dbs\\\\refseq\\\\refseq_mm_2013_07.fasta"),
  rptr_intco = 3000, 
  
  # `dot-dot-dot` filtration of MaxQuant PSMs
  Charge = expr(Charge == "2" | Charge == "3"), 
  `Oxidation (M)` = expr(`Oxidation (M)` == 1),
)
}

}
\seealso{
\code{\link{normPep}} for peptides and \code{\link{normPrn}} for
 proteins.
}
