% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_all_NA}
\alias{not_all_NA}
\title{Finds all-NA column(s)}
\usage{
not_all_NA(x)
}
\arguments{
\item{x}{A data frame of \code{log2FC} and \code{intensity}.}
}
\description{
\code{not_all_NA} identifies the column indexes with all NA values.
}
