% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_all_nan}
\alias{not_all_nan}
\title{Finds all-NaN column(s)}
\usage{
not_all_nan(x, ...)
}
\arguments{
\item{x}{A data frame of \code{log2FC} and \code{intensity}.}

\item{...}{The same in \code{sum}.}
}
\description{
Finds all-NaN column(s)
}
