% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_all_nan}
\alias{not_all_nan}
\title{Finds non-all-NaN column(s)}
\usage{
not_all_nan(x, ...)
}
\arguments{
\item{x}{A vector.}
}
\description{
Finds non-all-NaN column(s)
}
\examples{
\donttest{purrr::map_lgl(mtcars, not_all_nan, na.rm = TRUE)}
}
