% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_all_zero}
\alias{not_all_zero}
\title{Finds not-all-zero column(s)}
\usage{
not_all_zero(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A logical vector
}
\description{
\code{not_all_zero} identifies the column indexes with all NA values.
}
\examples{
\donttest{not_all_zero(mtcars)}
}
