% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{pad_mascot_channels}
\alias{pad_mascot_channels}
\title{Pad Mascot TMT channels to the highest plex}
\usage{
pad_mascot_channels(data, TMT_plex)
}
\arguments{
\item{data}{A PSM table from Mascot export}

\item{TMT_plex}{The multiplicity of TMT according to expt_smry.xlsx}
}
\description{
Pad Mascot TMT channels to the highest plex
}
