% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{pad_mf_channels}
\alias{pad_mf_channels}
\title{Pad MSFragger TMT channels to the highest plex.}
\usage{
pad_mf_channels(file, ...)
}
\arguments{
\item{file}{An intermediate PSM table.}

\item{...}{filter_dots.}
}
\description{
Pad MSFragger TMT channels to the highest plex.
}
