% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{pad_mq_channels}
\alias{pad_mq_channels}
\title{Pad MaxQuant TMT channels to the highest plex}
\usage{
pad_mq_channels(file)
}
\arguments{
\item{file}{A file name of PSM table from MaxQuant export}
}
\description{
Pad MaxQuant TMT channels to the highest plex
}
