% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{pad_sm_channels}
\alias{pad_sm_channels}
\title{Pads Spectrum Mill TMT channels to the highest plex}
\usage{
pad_sm_channels(file = NULL, ...)
}
\arguments{
\item{file}{An intermediate PSM table.}

\item{...}{filter_dots.}
}
\description{
Pads Spectrum Mill TMT channels to the highest plex
}
