% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_acc}
\alias{parse_acc}
\title{Determine the protein accession type from PSM tables}
\usage{
parse_acc(df)
}
\arguments{
\item{df}{An input data frame.}
}
\description{
Find the protein accession from a non-cRAP entry and parse it.
}
