% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{pep_mq_lfq}
\alias{pep_mq_lfq}
\title{Handling of MaxQuant peptide.txt}
\usage{
pep_mq_lfq(label_scheme, omit_single_lfq)
}
\description{
Fill back temporarily intensity values that are not filled in LFQ msms.txt.
}
