% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{pep_mq_lfq}
\alias{pep_mq_lfq}
\title{Handling of MaxQuant peptide.txt}
\usage{
pep_mq_lfq(label_scheme, omit_single_lfq)
}
\arguments{
\item{label_scheme}{Experiment summary}

\item{omit_single_lfq}{Logical; if TRUE, omits LFQ entries with single
measured values across all samples. The default is FALSE.}
}
\description{
Fill back intensity values that are not filled in LFQ msms.txt.
}
