% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{pep_to_prn}
\alias{pep_to_prn}
\title{Helper of Pep2Prn}
\usage{
pep_to_prn(
  id,
  method_pep_prn,
  use_unique_pep,
  gn_rollup,
  rm_outliers,
  rm_allna = FALSE,
  ...
)
}
\arguments{
\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{method_pep_prn}{Character string; the method to summarize the
\code{log2FC} and the \code{intensity} of peptides by protein entries. The
descriptive statistics includes \code{c("mean", "median", "weighted_mean",
"top_3_mean", "lfq_max", "lfq_top_2_sum", "lfq_top_3_sum", "lfq_all")} with
\code{median} being the default for TMT and \code{lfq_top_3_sum} for LFQ.
The representative \code{log10-intensity} of reporter (or LFQ) ions at the
peptide levels will be the weight when summarizing \code{log2FC} with
various \code{"top_n"} statistics or \code{"weighted_mean"}.}

\item{use_unique_pep}{Logical. If TRUE, only entries that are \code{TRUE} or
equal to \code{1} under the column \code{pep_isunique} in
\code{Peptide.txt} will be used, for summarizing the \code{log2FC} and the
\code{intensity} of peptides into protein values. The default is to use
unique peptides only. For \code{MaxQuant} data, the levels of uniqueness
are according to the \code{pep_unique_by} in \code{\link{normPSM}}. The
argument currently do nothing to \code{Spectrum Mill} data where both
unique and shared peptides will be kept.}

\item{gn_rollup}{Logical; if TRUE, rolls up protein accessions to gene names.}

\item{rm_outliers}{Logical; if TRUE, PSM outlier removals will be performed
for peptides with more than two identifying PSMs. Dixon's method will be
used when \eqn{2 < n \le 25} and Rosner's method will be used when \eqn{n >
25}. The default is FALSE.}

\item{rm_allna}{Logical; if TRUE, removes data rows that are exclusively NA
across ratio columns of \code{log2_R126} etc. The setting also applies to
\code{log2_R000} in LFQ.}

\item{...}{\code{filter_}: Variable argument statements for the filtration of
data rows. Each statement contains a list of logical expression(s). The
\code{lhs} needs to start with \code{filter_}. The logical condition(s) at
the \code{rhs} needs to be enclosed in \code{exprs} with round parenthesis.
For example, \code{pep_len} is a column key in \code{Peptide.txt}. The
statement of \code{filter_peps_at = exprs(pep_len <= 50)} will remove
peptide entries with \code{pep_len > 50}.}
}
\description{
Helper of Pep2Prn
}
