% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-histo.R
\name{pepHist}
\alias{pepHist}
\title{Visualizes the histograms of peptide data}
\usage{
pepHist(...)
}
\description{
Visualizes the histograms of peptide data
}
\examples{
# no scaling normalization
pepHist(
  scale_log2r = FALSE,
  show_curves = TRUE,
  show_vline = TRUE,
  xmin = -1,
  xmax = 1,
  ncol = 4
)

# scaling normalization
pepHist(
  scale_log2r = TRUE,
  show_curves = TRUE,
  show_vline = TRUE,
  xmin = -1,
  xmax = 1,
  ncol = 4
)

\dontrun{
pepHist(
  col_select = "a_non_existed_column_key"
)
}

}
\seealso{
\code{\link{proteoHist}} for parameters
}
