% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-nmf.R
\name{pepNMF}
\alias{pepNMF}
\title{NMF Analysis of Peptide \code{log2-ratios}}
\usage{
pepNMF(...)
}
\description{
NMF Analysis of Peptide \code{log2-ratios}
}
\seealso{
\code{\link{proteoNMF}} for parameters
}
