% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{peptideMBR}
\alias{peptideMBR}
\title{MBR of peptides}
\usage{
peptideMBR(
  ms1files,
  mbr_peps,
  mbr_rets,
  mbr_mzs,
  mbr_ys,
  base_name,
  dat_dir,
  mbr_ret_tol = 25,
  max_mbr_fold = 4,
  step = 1e-05
)
}
\arguments{
\item{ms1files}{The names of MS1 data files corresponding to a peptide table,
e.g., \code{TMTset1_LCMSinj1_Peptide_N.txt}.}

\item{mbr_peps}{Sequences of peptides for MBR.}

\item{mbr_rets}{Values of retention times for MBR.}

\item{mbr_mzs}{Values of m-over-zs for MBR.}

\item{mbr_ys}{Intensity values for MBR.}

\item{base_name}{The basename of a peptide table.}

\item{dat_dir}{The working data directory.}

\item{mbr_ret_tol}{The tolerance in MBR retention time in seconds.}

\item{max_mbr_fold}{The maximum absolute fold change in MBR.}

\item{step}{The mass error in \code{ppm / 1e6}.}
}
\description{
MBR of peptides
}
