% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanos.R
\name{plot_ibaq}
\alias{plot_ibaq}
\title{Rank plot of protein iBAQ.}
\usage{
plot_ibaq(
  outname = NULL,
  highlights = NULL,
  group_pep_by = "gene",
  col_group = NULL,
  subcellular = list(cytosol = c("Actb", "Acta1", "Actbl2", "Gapdh"), pm = c("Atp1a1"),
    er = c("Canx", "Calr", "Hspa5"), golgi = c("Golga2", "Golga1"), mito = c("Tomm20",
    "Vdac1"), lyso = c("Lamp1", "Lamp2", "Ctsd"), endosome = c("Eea1", "Rab5a"), nuclear
    = c("H4c1", "H2ax", "H1-2", "Npm1")),
  show_subcelluar = TRUE,
  ...
)
}
\arguments{
\item{outname}{The output file name.}

\item{highlights}{The protein names to be highlighted.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{col_group}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_group} will be
used for sample grouping in the indicated analysis. At the NULL default, the
column key \code{Group} will be used. No data annotation by groups will be
performed if the fields under the indicated group column is empty.}

\item{subcellular}{A list defining subcellular protein markers.}

\item{show_subcelluar}{Logical; show subcellular proteins or not.}

\item{...}{Additional arguments passed to \code{ggsave}.}
}
\description{
Rank plot of protein iBAQ.
}
