% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-trends.R
\name{plot_prnTrend}
\alias{plot_prnTrend}
\title{Trend visualization}
\usage{
plot_prnTrend(...)
}
\description{
Visualizes the trends of protein \code{log2-ratios}
}
\examples{
plot_prnTrend(
  scale_log2r = TRUE,
  col_order = Order,
  n_clust = 6
)

}
\seealso{
\code{\link{proteoTrend}} for parameters
}
