% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-histo.R
\name{plotHisto}
\alias{plotHisto}
\title{Plots histograms}
\usage{
plotHisto(df = NULL, id, label_scheme_sub, params, scale_log2r,
  complete_cases, show_curves, show_vline, scale_y, filepath = NULL,
  filename, theme, ...)
}
\description{
Plots histograms
}
