% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-histo.R
\name{plotHisto}
\alias{plotHisto}
\title{Plots histograms}
\usage{
plotHisto(df = NULL, id, label_scheme_sub, params, scale_log2r,
  pep_pattern, show_curves, show_vline, filepath = NULL, filename, ...)
}
\description{
Plots histograms
}
