% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm.R
\name{plotHM}
\alias{plotHM}
\title{Makes heat maps}
\usage{
plotHM(
  df,
  id,
  col_benchmark,
  label_scheme_sub,
  filepath,
  filename,
  scale_log2r,
  complete_cases,
  annot_cols = NULL,
  annot_colnames = NULL,
  annot_rows = annot_rows,
  xmin = -1,
  xmax = 1,
  xmargin = 0.1,
  p_dist_rows = 2,
  p_dist_cols = 2,
  hc_method_rows = "complete",
  hc_method_cols = "complete",
  x = NULL,
  p = NULL,
  method = NULL,
  diag = NULL,
  upper = NULL,
  annotation_col = NULL,
  annotation_row = NULL,
  clustering_method = NULL,
  rm_allna = TRUE,
  ...
)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{col_benchmark}{Not used.}

\item{label_scheme_sub}{A data frame. Subset entries from \code{label_scheme}
for selected samples.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{filename}{A representative file name to outputs. By default, the name(s)
will be determined automatically. For text files, a typical file extension
is \code{.txt}. For image files, they are typically saved via
\code{\link[ggplot2]{ggsave}} or \code{\link[pheatmap]{pheatmap}} where the
image type will be determined by the extension of the file name.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{annot_cols}{A character vector of column keys in \code{expt_smry.xlsx}.
The values under the selected keys will be used to color-code sample IDs on
the top of the indicated plot. The default is NULL without column
annotation.}

\item{annot_colnames}{A character vector of replacement name(s) to
\code{annot_cols}. The default is NULL without name replacement.}

\item{annot_rows}{A character vector of column keys that can be found from
input files of \code{Peptide.txt}, \code{Protein.txt} etc. The values
under the selected keys will be used to color-code peptides or proteins on
the side of the indicated plot. The default is NULL without row annotation.}

\item{xmin}{Numeric; the minimum x at a log2 scale. The default is -1.}

\item{xmax}{Numeric; the maximum  x at a log2 scale. The default is 1.}

\item{xmargin}{Numeric; the margin in heat scales. The default is 0.1.}

\item{p_dist_rows}{Numeric; the power of the Minkowski distance in the measures 
of row \code{\link[stats]{dist}} at \code{clustering_distance_rows = "minkowski"}. 
The default is 2.}

\item{p_dist_cols}{Numeric; similar to \code{p_dist_rows} but for column data.}

\item{hc_method_rows}{A character string; the same agglomeration method for 
\code{\link[stats]{hclust}} of data rows. The default is \code{complete}.}

\item{hc_method_cols}{A character string; similar to \code{hc_method_rows} 
but for column data.}

\item{x}{Dummy argument to avoid incurring the corresponding argument in
\link[stats]{dist} by partial argument matches.}

\item{p}{Dummy argument to avoid incurring the corresponding argument in
\link[stats]{dist} by partial argument matches.}

\item{method}{Dummy argument to avoid incurring the corresponding argument in
\link[stats]{dist} by partial argument matches.}

\item{diag}{Dummy argument to avoid incurring the corresponding argument in
\link[stats]{dist} by partial argument matches.}

\item{upper}{Dummy argument to avoid incurring the corresponding argument in
\link[stats]{dist} by partial argument matches.}

\item{annotation_col}{Dummy argument to avoid incurring the corresponding
argument in \link[pheatmap]{pheatmap}.}

\item{annotation_row}{Dummy argument to avoid incurring the corresponding
argument in \link[pheatmap]{pheatmap}.}

\item{clustering_method}{Dummy argument to avoid incurring the corresponding
argument in \link[pheatmap]{pheatmap}.}

\item{rm_allna}{Logical; if TRUE, removes data rows that are exclusively NA
across ratio columns of \code{log2_R126} etc. The setting also applies to
\code{log2_R000} in LFQ.}

\item{...}{\code{filter_}: Variable argument statements for the row filtration
against data in a primary file linked to \code{df}. Each statement contains
to a list of logical expression(s). The \code{lhs} needs to start with
\code{filter_}. The logical condition(s) at the \code{rhs} needs to be
enclosed in \code{exprs} with round parenthesis. For example, \code{pep_len}
is a column key in \code{Peptide.txt}. The statement \code{filter_peps_at =
exprs(pep_len <= 50)} will remove peptide entries with \code{pep_len > 50}.
See also \code{\link{pepHist}}, \code{\link{normPSM}}. \cr \cr
\code{arrange_}: Variable argument statements for the row ordering against
data in a primary file linked to \code{df}. The \code{lhs} needs to start
with \code{arrange_}. The expression(s) at the \code{rhs} needs to be
enclosed in \code{exprs} with round parenthesis. For example,
\code{arrange_peps_by = exprs(gene, prot_n_pep)} will arrange entries by
\code{gene}, then by \code{prot_n_pep}. \cr \cr Additional parameters for
plotting: \cr \code{width}, the width of plot \cr \code{height}, the height
of plot \cr \cr Additional arguments for \code{\link[pheatmap]{pheatmap}}:
\cr \code{cluster_rows, clustering_method, clustering_distance_rows}... \cr
\cr Notes about \code{pheatmap}: \cr \code{annotation_col} disabled; instead
use keys indicated in \code{annot_cols} \cr \code{annotation_row} disabled;
instead use keys indicated in \code{annot_rows} \cr \code{clustering_method}
breaks into \code{hc_method_rows} for row data and \code{hc_method_cols} for
column data \cr \code{clustering_distance_rows = "minkowski"} allowed
together with the powder of \code{p_dist_rows} and/or \code{p_dist_cols}}
}
\description{
Makes heat maps
}
