% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-informatics.R
\name{plotHM}
\alias{plotHM}
\title{Makes heat maps}
\usage{
plotHM(df, id, scale_log2r, col_benchmark, label_scheme_sub, filepath,
  filename, complete_cases, xmin = -1, xmax = 1, x_margin = 0.1,
  annot_cols = NULL, ...)
}
\description{
Makes heat maps
}
