% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-hm.R
\name{plotHM}
\alias{plotHM}
\title{Makes heat maps}
\usage{
plotHM(
  df,
  id,
  col_benchmark,
  label_scheme_sub,
  filepath,
  filename,
  scale_log2r,
  complete_cases,
  annot_cols = NULL,
  annot_colnames = NULL,
  annot_rows = annot_rows,
  xmin = -1,
  xmax = 1,
  xmargin = 0.1,
  p_dist_rows = 2,
  p_dist_cols = 2,
  hc_method_rows = "complete",
  hc_method_cols = "complete",
  ...
)
}
\description{
Makes heat maps
}
