% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-mds.R
\name{plotMDS}
\alias{plotMDS}
\title{Plots MDS}
\usage{
plotMDS(df, col_color = NULL, col_fill = NULL, col_shape = NULL,
  col_size = NULL, col_alpha = NULL, color_brewer = NULL,
  fill_brewer = NULL, size_manual = NULL, shape_manual = NULL,
  alpha_manual = NULL, label_scheme_sub = NULL, filepath = NULL,
  filename = NULL, show_ids = FALSE, ...)
}
\description{
Plots MDS
}
