% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-nmf.R
\name{plotNMFmeta}
\alias{plotNMFmeta}
\title{Plots coef results from NMF analysis}
\usage{
plotNMFmeta(df, id, r, label_scheme_sub, filepath, fn_prx, ...)
}
\description{
Plots coef results from NMF analysis
}
