% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-mds.R
\name{plotPCA}
\alias{plotPCA}
\title{Plots PCA}
\usage{
plotPCA(df, col_color = NULL, col_fill = NULL, col_shape = NULL,
  col_size = NULL, col_alpha = NULL, color_brewer = NULL,
  fill_brewer = NULL, size_manual = NULL, shape_manual = NULL,
  alpha_manual = NULL, label_scheme_sub = label_scheme_sub, prop_var,
  filepath, filename, show_ids, ...)
}
\description{
Plots PCA
}
