% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular.R
\name{plotSubcellular_sub}
\alias{plotSubcellular_sub}
\title{Plots trends at a given \code{n_clust}.}
\usage{
plotSubcellular_sub(
  df2 = NULL,
  custom_prefix = NULL,
  fn_prefix = NULL,
  fn_suffix = NULL,
  df = NULL,
  id = "gene",
  col_group = NULL,
  col_order = NULL,
  levels_subcellular = NULL,
  levels_subtype = NULL,
  complete_cases = FALSE,
  panel_ids = NULL,
  filepath = NULL,
  label_scheme_sub = NULL,
  theme = NULL,
  dots = NULL
)
}
\arguments{
\item{df2}{Character vector or string; the name(s) of secondary data file(s).
An informatic task, i.e. \code{anal_prnTrend(...)} against a primary
\code{df} generates secondary files such as
\code{Protein_Trend_Z_nclust6.txt} etc. See also \code{\link{prnHist}} for
the description of a primary \code{df}; \code{\link{normPSM}} for the lists
of \code{df} and \code{df2}.}

\item{custom_prefix}{A custom filename prefix.}

\item{fn_prefix}{A file name prefix.}

\item{fn_suffix}{A file name suffix.}

\item{df}{A data frame.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{col_group}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_group} will be
used for sample grouping in the indicated analysis. At the NULL default, the
column key \code{Group} will be used. No data annotation by groups will be
performed if the fields under the indicated group column is empty.}

\item{col_order}{Character string to a column key in \code{expt_smry.xlsx}.
Numeric values under which will be used for the left-to-right arrangement of
samples in graphic outputs or top-to-bottom arrangement in text outputs. At
the NULL default, the column key \code{Order} will be used. If values under
column \code{Order} are left blank, samples will be ordered by their names.}

\item{levels_subcellular}{The levels of subcellular fractions.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{panel_ids}{Panel IDs for plotting.}

\item{filepath}{An output file path.}

\item{label_scheme_sub}{A metadata subset.}

\item{theme}{A
\href{https://ggplot2.tidyverse.org/reference/ggtheme.html}{ggplot2} theme,
i.e., theme_bw(), or a custom theme. At the NULL default, a system theme
will be applied.}

\item{dots}{Variable arguments.}
}
\description{
Plots trends at a given \code{n_clust}.
}
