% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{plotTrend}
\alias{plotTrend}
\title{Plots trends}
\usage{
plotTrend(
  id = "gene",
  dat_dir = NULL,
  anal_type = "Trend",
  col_group = NULL,
  col_order = NULL,
  levels_group = NULL,
  label_scheme_sub = NULL,
  data_type = "ratio",
  pat = "Trend_[ONZ]_.*nclust\\\\d+\\\\.txt$",
  ext = "txt",
  n_clust = NULL,
  panel_ids = NULL,
  show_panel_ids = TRUE,
  group_data_by = "ratio",
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  df2 = NULL,
  filepath = NULL,
  filename = NULL,
  theme = NULL,
  ...
)
}
\arguments{
\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{anal_type}{The type of analysis. To differentiate from custom plots
such as results from svmProb.}

\item{col_group}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_group} will be
used for sample grouping in the indicated analysis. At the NULL default, the
column key \code{Group} will be used. No data annotation by groups will be
performed if the fields under the indicated group column is empty.}

\item{col_order}{Character string to a column key in \code{expt_smry.xlsx}.
Numeric values under which will be used for the left-to-right arrangement of
samples in graphic outputs or top-to-bottom arrangement in text outputs. At
the NULL default, the column key \code{Order} will be used. If values under
column \code{Order} are left blank, samples will be ordered by their names.}

\item{levels_group}{The factor levels of the values under the column 'group'
in trend outputs for ordering on the X-axis. If not NULL, the levels will
overwrite the order specified by the combination of \code{col_order} and
\code{group}.}

\item{label_scheme_sub}{A data frame. Subset entries from \code{label_scheme}
for selected samples.}

\item{data_type}{The type of data, either intensity or log2FC.}

\item{pat}{The pattern of secondary files (df2).}

\item{ext}{The filename extension.}

\item{n_clust}{Numeric vector; the cluster ID(s) corresponding to
\code{\link{anal_prnTrend}} for visualization. At the NULL default, all
available cluster IDs will be used.}

\item{panel_ids}{Panel IDs for plotting.}

\item{show_panel_ids}{Logical; show panel IDs or not.}

\item{scale_log2r}{Logical; at the TRUE default, input files with
\code{_Z[...].txt} in name will be used. Otherwise, files with
\code{_N[...].txt} in name will be taken. An error will be thrown if no
files are matched under given conditions.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{impute_na}{Logical; at the TRUE default, input files with
\code{_impNA[...].txt} in name will be loaded. Otherwise, files without
\code{_impNA} in name will be taken. An error will be thrown if no files are
matched under given conditions.}

\item{df2}{Character vector or string; the name(s) of secondary data file(s).
An informatic task, i.e. \code{anal_prnTrend(...)} against a primary
\code{df} generates secondary files such as
\code{Protein_Trend_Z_nclust6.txt} etc. See also \code{\link{prnHist}} for
the description of a primary \code{df}; \code{\link{normPSM}} for the lists
of \code{df} and \code{df2}.}

\item{filepath}{Use system default.}

\item{filename}{A representative file name to outputs. By default, it will be
determined automatically by the name of the current call.}

\item{theme}{A
\href{https://ggplot2.tidyverse.org/reference/ggtheme.html}{ggplot2} theme,
i.e., theme_bw(), or a custom theme. At the NULL default, a system theme
will be applied.}

\item{...}{\code{filter2_}: Variable argument statements for the row
filtration against data in secondary file(s) of
\code{[...]Protein_Trend_[...].txt}. See also \code{\link{prnGSPAHM}} for
the format of \code{filter2_} statements. \cr \cr \code{arrange2_}: Variable
argument statements for the row ordering against data in secondary file(s)
of \code{[...]Protein_Trend_[...].txt}. See also \code{\link{prnGSPAHM}} for
the format of \code{arrange2_} statements. \cr \cr Additional parameters for
use in \code{plot_} functions: \cr \code{ymin}, the minimum y at \code{log2}
scale; \cr \code{ymax}, the maximum y at \code{log2} scale; \cr
\code{ybreaks}, the breaks in y-axis at \code{log2} scale; \cr \code{nrow},
the number of rows; \cr \code{width}, the width of plot; \cr \code{height},
the height of plot; \cr \code{color}, the color of lines; \cr \code{alpha},
the transparency of lines.}
}
\description{
Plots trends
}
