% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-trends.R
\name{plotTrend}
\alias{plotTrend}
\title{Make Trend lines from fuzzy c-mean clustering}
\usage{
plotTrend(df, id, col_group, col_order, label_scheme_sub, n_clust,
  complete_cases, scale_log2r, filepath, filename, ...)
}
\description{
Make Trend lines from fuzzy c-mean clustering
}
