% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-trends.R
\name{plotTrend}
\alias{plotTrend}
\title{Plots trends}
\usage{
plotTrend(id, col_group, col_order, label_scheme_sub, n_clust, scale_log2r,
  impute_na, filepath, filename, ...)
}
\description{
Plots trends
}
