% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{plotTrend_sub}
\alias{plotTrend_sub}
\title{Plots trends at a given \code{n_clust}.}
\usage{
plotTrend_sub(
  df2,
  custom_prefix,
  fn_prefix,
  fn_suffix,
  df,
  id,
  complete_cases = FALSE,
  panel_ids = NULL,
  filepath = NULL,
  label_scheme_sub = NULL,
  theme = NULL,
  dots = NULL
)
}
\arguments{
\item{df2}{Character vector or string; the name(s) of secondary data file(s).
An informatic task, i.e. \code{anal_prnTrend(...)} against a primary
\code{df} generates secondary files such as
\code{Protein_Trend_Z_nclust6.txt} etc. See also \code{\link{prnHist}} for
the description of a primary \code{df}; \code{\link{normPSM}} for the lists
of \code{df} and \code{df2}.}

\item{custom_prefix}{A custom filename prefix.}

\item{fn_prefix}{A file name prefix.}

\item{fn_suffix}{A file name suffix.}

\item{df}{A data frame.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{panel_ids}{Panel IDs for plotting.}

\item{filepath}{An output file path.}

\item{label_scheme_sub}{A metadata subset.}

\item{theme}{A
\href{https://ggplot2.tidyverse.org/reference/ggtheme.html}{ggplot2} theme,
i.e., theme_bw(), or a custom theme. At the NULL default, a system theme
will be applied.}

\item{dots}{Variable arguments.}
}
\description{
Plots trends at a given \code{n_clust}.
}
