% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-volcanos.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Plot volcanos}
\usage{
plotVolcano(df = NULL, id = "gene", filepath = NULL,
  filename = NULL, adjP = FALSE, show_labels = TRUE,
  anal_type = "Volcano", pval_cutoff = 0.05,
  logFC_cutoff = log2(1.2), show_sig = "none",
  gset_nms = c("go_sets", "kegg_sets"), ...)
}
\description{
Plot volcanos
}
