% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{prep_fraction_scheme}
\alias{prep_fraction_scheme}
\title{Loads the information of analyte prefractionation}
\usage{
prep_fraction_scheme(dat_dir, filename)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{filename}{A representative file name to outputs. By default, the name(s)
will be determined automatically. For text files, a typical file extension
is \code{.txt}. For image files, they are typically saved via
\code{\link[ggplot2]{ggsave}} or \code{\link[pheatmap]{pheatmap}} where the
image type will be determined by the extension of the file name.}
}
\description{
Loads the information of analyte prefractionation
}
