% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{prepDM}
\alias{prepDM}
\title{Prepares data for analysis}
\usage{
prepDM(df, id, scale_log2r, sub_grp, type = "ratio", anal_type)
}
\arguments{
\item{sub_grp}{Numeric.  A list of sample IDs that will be used in subsequent
analysis.}

\item{anal_type}{Character string; the type of analysis that are preset for
method dispatch in function factories. The value will be determined
automatically. Examplary values include \code{anal_type = c("PCA",
"Corrplot", "EucDist", "GSPA", "Heatmap", "Histogram", "MDS", "Model",
"NMF", "Purge", "Trend", ...)}.}
}
\value{
A data frame tailored for subsequent analysis.
}
\description{
\code{prepDM} prepares a minimal adequate data frame for subsequent analysis.
}
\examples{
tempData <- prepDM(df, entrez, scale_log2r, sub_grp = label_scheme_sub$Sample_ID)

\dontrun{
}
}
