% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-informatics.R
\name{prepDM}
\alias{prepDM}
\title{Prpare data for analysis}
\usage{
prepDM(df, id, scale_log2r, sub_grp, type = "ratio", anal_type)
}
\arguments{
\item{df}{A data frame containing only numeric values.}

\item{id}{The name of unqiue identifiers.}

\item{scale_log2r}{Logical; if TRUE, rescales \code{log2-ratios} to the same scale.}

\item{sub_grp}{Numeric.  A list of sample IDs that will be used in subsequent analysis.}
}
\value{
A data frame tailored for subsequent analysis.
}
\description{
\code{prepDM} prepare a minimal data frame for subsequent analysis.
}
\details{
In the "label_scheme.csv" file,
}
\examples{
tempData <- prepDM(df, entrez, scale_log2r, sub_grp = label_scheme_sub$Sample_ID)

\dontrun{
}
}
