% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{prepMSig}
\alias{prepMSig}
\title{Preparation of Molecular Signatures}
\usage{
prepMSig(
  species = "human",
  msig_url = NULL,
  abbr_species = NULL,
  ortho_mart = switch(species, mouse = "mmusculus_gene_ensembl", rat =
    "rnorvegicus_gene_ensembl", human = "to_itself", "unknown"),
  db_path = "~/proteoQ/dbs/msig",
  filename = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{species}{Character string; the name of a species for the
\emph{conveninent} preparation of \code{MSig} data bases. The species
available for the convenience feature is in one of \code{c("human", "mouse",
"rat")} with \code{"human"} being the default. The argument is not required
for other species; instead, users will provide values under arguments
\code{ortho_mart} for the lookup of orthologs to human.}

\item{msig_url}{A URL to
\href{https://data.broadinstitute.org/gsea-msigdb/msigdb/release/7.0/}{MSig
}. At the \code{NULL} default, a \code{c2.all.v[...].entrez.gmt} data will
be used for all species. A valid web address is required for a custom data
base. For simplicity, only files with entrez IDs will be handled; 
files of \code{c2.all.v[...].symbols.gmt} will not be parsed.}

\item{abbr_species}{Two-letter character string; the abbreviated name of
species used with
\href{https://bioconductor.org/packages/release/data/annotation/html/org.Hs.eg.db.html}{org.Xx.eg.db}.
The value of \code{abbr_species} will be determined automatically if the
species is in one of \code{c("human", "mouse", "rat")}. Otherwise, for
example, users need to provide \code{abbr_species = Ce} for fetching the
\code{org.Ce.eg.db} package in the name space of proteoQ. 

For analysis against
\href{http://current.geneontology.org/products/pages/downloads.html}{gene
ontology} and \href{https://www.gsea-msigdb.org/gsea/index.jsp}{Molecular
Signatures}, the argument is further applied to differentiate the same
biological terms under different species; e.g., \code{GO~0072686 mitotic
spindle} becomes \code{hs_GO~0072686 mitotic spindle} for human and
\code{mm_GO~0072686 mitotic spindle} for mouse.}

\item{ortho_mart}{Character string; a dataset name from
\code{\link[biomaRt]{useMart}} and/or \code{\link[biomaRt]{listDatasets}}
for the lookup of orthologs to \code{human} genes. For species in
\code{c("human", "mouse", "rat")}, the value will be determined
automatically unless otherwise specified.}

\item{db_path}{Character string; the local path for database(s). The default
is \code{"~/proteoQ/dbs/msig"}.}

\item{filename}{Character string; An output file name. At the \code{NULL}
default, the name will be determined automatically at a given
\code{species}; i.e., \code{msig_hs.rds} for \code{human} data. The file is
saved as a \code{.rds} object for uses with \code{\link{prnGSPA}}.}

\item{overwrite}{Logical; if TRUE, overwrite the downloaded database(s). The
default is FALSE.}
}
\description{
\code{prepMSig} downloads and prepares data bases of
\href{https://www.gsea-msigdb.org/gsea/index.jsp}{Molecular Signatures}
(MSig) for enrichment analysis by gene sets.
}
\examples{
\donttest{
library(proteoQ)

## the default `MSig` is `c2.all`
# `human`; outputs under `db_path`
prepMSig()
head(readRDS(file.path("~/proteoQ/dbs/msig/msig_hs.rds")))

prnGSPA(
  gset_nms = file.path("~/proteoQ/dbs/msig/msig_hs.rds"), 
)

prepMSig(species = human, msig_url = "https://data.wikipathways.org/current/gmt/wikipathways-20250810-gmt-Homo_sapiens.gmt", filename = wk_hs.rds)
prepMSig(species = mouse, msig_url = "https://data.wikipathways.org/current/gmt/wikipathways-20250810-gmt-Mus_musculus.gmt", filename = wk_mm.rds)

# mouse
prepMSig(species = "mouse", msig_url = "https://data.broadinstitute.org/gsea-msigdb/msigdb/release/2025.1.Mm/msigdb.v2025.1.Mm.entrez.gmt", filename = msig_mm.rds)
head(readRDS(file.path("~/proteoQ/dbs/msig/msig_mm.rds")))

# `rat`
prepMSig(species = "rat", filename = msig_rn.rds)
head(readRDS(file.path("~/proteoQ/dbs/msig/msig_rn.rds")))

# `dog`; need `ortho_mart` for species other than `human`, `mouse` and `rat`
# (try `?biomaRt::useMart` for a list of marts)
prepMSig(
  # species = dog,
  abbr_species = Cf, 
  ortho_mart = cfamiliaris_gene_ensembl,
  filename = msig_cf.rds,
)

# also `dog`
prepMSig(
  species = my_dog,
  abbr_species = Cf, 
  ortho_mart = cfamiliaris_gene_ensembl,
  filename = msig_cf2.rds,
)

msig_cf <- readRDS(file.path("~/proteoQ/dbs/msig/msig_cf.rds"))
msig_cf2 <- readRDS(file.path("~/proteoQ/dbs/msig/msig_cf2.rds"))
identical(msig_cf, msig_cf2)

## use an `MSig`other than the default of `c2.all`
prepMSig(
  msig_url = "https://data.broadinstitute.org/gsea-msigdb/msigdb/release/7.0/c2.cgp.v7.0.entrez.gmt",
  species = human,
  filename = c2_cgp_hs.rds,
)

prepMSig(
  msig_url = "https://data.broadinstitute.org/gsea-msigdb/msigdb/release/7.0/c2.cgp.v7.0.entrez.gmt",
  species = dog,
  ortho_mart = cfamiliaris_gene_ensembl,
  filename = c2_cgp_cf.rds,
)
}

\dontrun{
# enrichment analysis with custom `MSig`
prnGSPA(
  gset_nms = c("~/proteoQ/dbs/msig/msig_hs.rds",
               "~/proteoQ/dbs/msig/msig_mm.rds"),
)
}

}
