% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{prepTrend}
\alias{prepTrend}
\title{Prepare data for trend analysis.}
\usage{
prepTrend(
  dat_dir = NULL,
  df = NULL,
  id = NULL,
  col_group = NULL,
  col_order = NULL,
  col_fraction = NULL,
  col_subtype = NULL,
  label_scheme_sub = NULL,
  impute_group_na = TRUE,
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  anal_type = "Trend",
  p_outlier = 0.05,
  group_renorm_by = NULL,
  int_to_long = TRUE,
  ...
)
}
\arguments{
\item{dat_dir}{The current working directory.}

\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{col_group}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_group} will be
used for sample grouping in the indicated analysis. At the NULL default, the
column key \code{Group} will be used. No data annotation by groups will be
performed if the fields under the indicated group column is empty.}

\item{col_order}{Character string to a column key in \code{expt_smry.xlsx}.
Numeric values under which will be used for the left-to-right arrangement of
samples in graphic outputs or top-to-bottom arrangement in text outputs. At
the NULL default, the column key \code{Order} will be used. If values under
column \code{Order} are left blank, samples will be ordered by their names.}

\item{col_fraction}{Only applicable to subcellular data analysis for
  assessing fraction purity etc. Character string to a column key in
  \code{expt_smry.xlsx} indicating subcellular fractions. No calculation of
  subcellular location score at the \code{NULL} default.

  Suggest, e.g., the corresponding column key \code{Fraction} in metadata.}

\item{col_subtype}{Only applicable to subcellular data analysis for assessing
sub-group fraction purity etc. Character string to a column key in
\code{expt_smry.xlsx} indicating sample subtypes linked to subcellular
fractions. This is needed, e.g., when each subcellular fraction contain
multiple subtypes of control, treated etc. and purity will be asssessed
independently.}

\item{label_scheme_sub}{A data frame. Subset entries from \code{label_scheme}
for selected samples.}

\item{impute_group_na}{Logical; if TRUE, impute (intensity) values that are
exclusively NA under a sample group. The sample grouping is defined under
\code{label_scheme}.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE. At
\code{scale_log2r = NA}, the raw \code{log2FC} without normalization will
be used.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used. The default is FALSE.}

\item{anal_type}{Character string; the type of analysis that are preset for
method dispatch in function factories. The value will be determined
automatically. Exemplary values include \code{anal_type = c("PCA",
"Corrplot", "EucDist", "GSPA", "Heatmap", "Histogram", "MDS", "Model",
"NMF", "Purge", "Trend", "LDA", ...)}.}

\item{p_outlier}{Type-I error in outlier assessment. Set p = 1.0 to skip
outlier removals.}

\item{int_to_long}{Logical; if TRUE, pivot intensities to a long form.}

\item{...}{\code{filter_}: Variable argument statements for the row
filtration against data in a primary file linked to \code{df}. See also
\code{\link{normPSM}} for the format of \code{filter_} statements. \cr \cr
\code{arrange_}: Variable argument statements for the row ordering against
data in a primary file linked to \code{df}. See also \code{\link{prnHM}}
for the format of \code{arrange_} statements. \cr \cr Additional arguments
for \link[e1071]{cmeans}, \link[stats]{kmeans}, \link[cluster]{clara},
\link[cluster]{pam}. Note that \code{centers} in \link[e1071]{cmeans} or
\link[stats]{kmeans} is replaced with \code{n_clust}. The same applies to
\code{k} in \link[cluster]{clara} or \link[cluster]{pam}. \cr With
\link[e1071]{cmeans}, \code{m} is according to Schwaemmle and Jensen if not
provided; \cr \code{x} is disabled with input data being determined
automatically.}
}
\description{
Prepare data for trend analysis.
}
