% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplots.R
\name{pepCorr_logFC}
\alias{pepCorr_logFC}
\alias{pepCorr_logInt}
\alias{prnCorr_logFC}
\alias{prnCorr_logInt}
\title{Correlation plots}
\usage{
pepCorr_logFC(
  col_select = NULL,
  col_order = NULL,
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  df = NULL,
  filepath = NULL,
  filename = NULL,
  cor_method = "pearson",
  digits = 2L,
  ...
)

pepCorr_logInt(
  col_select = NULL,
  col_order = NULL,
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  df = NULL,
  filepath = NULL,
  filename = NULL,
  cor_method = "pearson",
  digits = 2L,
  ...
)

prnCorr_logFC(
  col_select = NULL,
  col_order = NULL,
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  df = NULL,
  filepath = NULL,
  filename = NULL,
  cor_method = "pearson",
  digits = 2L,
  ...
)

prnCorr_logInt(
  col_select = NULL,
  col_order = NULL,
  scale_log2r = TRUE,
  complete_cases = FALSE,
  impute_na = FALSE,
  df = NULL,
  filepath = NULL,
  filename = NULL,
  cor_method = "pearson",
  digits = 2L,
  ...
)
}
\arguments{
\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
At the \code{NULL} default, the column key of \code{Select} in
\code{expt_smry.xlsx} will be used. In the case of no samples being
specified under \code{Select}, the column key of \code{Sample_ID} will be
used. The non-empty entries under the ascribing column will be used in
indicated analysis.}

\item{col_order}{Character string to a column key in \code{expt_smry.xlsx}.
Numeric values under which will be used for the left-to-right arrangement of
samples in graphic outputs or top-to-bottom arrangement in text outputs. At
the NULL default, the column key \code{Order} will be used. If values under
column \code{Order} are left blank, samples will be ordered by their names.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE. At
\code{scale_log2r = NA}, the raw \code{log2FC} without normalization will
be used.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used. The default is FALSE.}

\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{filename}{A representative file name to outputs. By default, the
name(s) will be determined automatically. For text files, a typical file
extension is \code{.txt}. For image files, they are typically saved via
\code{\link[ggplot2]{ggsave}} or \code{\link[pheatmap]{pheatmap}} where the
image type will be determined by the extension of the file name.}

\item{cor_method}{A character string indicating which correlation coefficient
is to be computed. One of \code{"pearson"} (default), \code{"kendall"}, or
\code{"spearman"}.}

\item{digits}{The number of decimal places in correlation values to be
displayed.}

\item{...}{\code{filter_}: Variable argument statements for the row filtration
against data in a primary file linked to \code{df}. See also
\code{\link{normPSM}} for the format of \code{filter_} statements. \cr \cr
Additional parameters for plotting: \cr \code{width}, the width of plot \cr
\code{height}, the height of plot \cr \code{xmin}, the minimum \eqn{x} of
logFC or intensity \cr \code{xmax}, the maximum \eqn{x} of logFC data or
intensity data \cr \code{xbreaks}, the breaks on \eqn{x} axis; the same
breaks will be applied to \eqn{y} axis.}
}
\value{
Correlation plots.
}
\description{
\code{pepCorr_logFC} plots correlation for peptide \code{logFC}. data.

\code{pepCorr_logInt} plots correlation of the \code{log10} intensity of ions
for peptide data.

\code{prnCorr_logFC} plots correlation for protein \code{logFC}.

\code{prnCorr_logInt} plots correlation of the \code{log10} intensity of ions
for protein data.
}
\details{
With TMT experiments, the same polypeptide may be triggered for MS2 any where
between the baseline and the apex levels during a peak elution. The direct
comparison of reporter-ion intensities between plex-es might have little
meaning.
}
\examples{
\donttest{
# ===================================
# Correlation
# ===================================

## !!!require the brief working example in `?load_expts`

## global option
scale_log2r <- TRUE

# peptide log2FC with sample ID ordering
#(no more than 40 samples for visualization)
pepCorr_logFC(
  col_select = BI,
  col_order = Order, 
  width = 25,
  height = 25,
  filter_peps_by = exprs(pep_n_psm >= 3),
  filename = bi_npsm3.png,
)

# protein log2FC
prnCorr_logFC(
  col_select = W2,
  col_order = Order, 
  width = 40,
  height = 40,
  filter_prots = exprs(prot_n_pep >= 2),
  filename = w2_npep2.png,
)


## Not run: 
# at most 40 samples
pepCorr_logFC(
  col_order = Order, 
  width = 40,
  height = 40,
  filter_peps_by = exprs(pep_n_psm >= 3),
  filename = too_many_cols.png,
)

# interplex comparison of peptide intensity
# (modest correlation in interplex reporter-ion intensity at data-dependant acquistion)
pepCorr_logInt(
  width = 10,
  height = 10,
  filter_peps_by = exprs(pep_n_psm >= 3),
  filename = pepcorr_int_npsm3.png,
)

# protein intensity
prnCorr_logInt(
  width = 10,
  height = 10,
  filter_prots_by = exprs(prot_n_pep >= 5),
  filename = prncorr_int_npep5.png,
)
## End(Not run)
}

}
\seealso{
\emph{Metadata} \cr \code{\link{load_expts}} for metadata preparation
 and a reduced working example in data normalization \cr

 \emph{Data normalization} \cr \code{\link{normPSM}} for extended examples in
 PSM data normalization \cr \code{\link{PSM2Pep}} for extended examples in
 PSM to peptide summarization \cr \code{\link{mergePep}} for extended
 examples in peptide data merging \cr \code{\link{standPep}} for extended
 examples in peptide data normalization \cr \code{\link{Pep2Prn}} for
 extended examples in peptide to protein summarization \cr
 \code{\link{standPrn}} for extended examples in protein data normalization.
 \cr \code{\link{purgePSM}} and \code{\link{purgePep}} for extended examples
 in data purging \cr \code{\link{pepHist}} and \code{\link{prnHist}} for
 extended examples in histogram visualization. \cr \code{\link{extract_raws}}
 and \code{\link{extract_psm_raws}} for extracting MS file names \cr

 \emph{Variable arguments of `filter_...`} \cr \code{\link{contain_str}},
 \code{\link{contain_chars_in}}, \code{\link{not_contain_str}},
 \code{\link{not_contain_chars_in}}, \code{\link{start_with_str}},
 \code{\link{end_with_str}}, \code{\link{start_with_chars_in}} and
 \code{\link{ends_with_chars_in}} for data subsetting by character strings
 \cr

 \emph{Missing values} \cr \code{\link{pepImp}} and \code{\link{prnImp}} for
 missing value imputation \cr

 \emph{Informatics} \cr \code{\link{pepSig}} and \code{\link{prnSig}} for
 significance tests \cr \code{\link{pepVol}} and \code{\link{prnVol}} for
 volcano plot visualization \cr \code{\link{prnGSPA}} for gene set enrichment
 analysis by protein significance pVals \cr \code{\link{gspaMap}} for mapping
 GSPA to volcano plot visualization \cr \code{\link{prnGSPAHM}} for heat map
 and network visualization of GSPA results \cr \code{\link{prnGSVA}} for gene
 set variance analysis \cr \code{\link{prnGSEA}} for data preparation for
 online GSEA. \cr \code{\link{pepMDS}} and \code{\link{prnMDS}} for MDS
 visualization \cr \code{\link{pepPCA}} and \code{\link{prnPCA}} for PCA
 visualization \cr \code{\link{pepLDA}} and \code{\link{prnLDA}} for LDA
 visualization \cr \code{\link{pepHM}} and \code{\link{prnHM}} for heat map
 visualization \cr \code{\link{pepCorr_logFC}}, \code{\link{prnCorr_logFC}},
 \code{\link{pepCorr_logInt}} and \code{\link{prnCorr_logInt}}  for
 correlation plots \cr \code{\link{anal_prnTrend}} and
 \code{\link{plot_prnTrend}} for trend analysis and visualization \cr
 \code{\link{anal_pepNMF}}, \code{\link{anal_prnNMF}},
 \code{\link{plot_pepNMFCon}}, \code{\link{plot_prnNMFCon}},
 \code{\link{plot_pepNMFCoef}}, \code{\link{plot_prnNMFCoef}} and
 \code{\link{plot_metaNMF}} for NMF analysis and visualization \cr

 \emph{Custom databases} \cr \code{\link{Uni2Entrez}} for lookups between
 UniProt accessions and Entrez IDs \cr \code{\link{Ref2Entrez}} for lookups
 among RefSeq accessions, gene names and Entrez IDs \cr \code{\link{prepGO}}
 for
 \code{\href{http://current.geneontology.org/products/pages/downloads.html}{gene
  ontology}} \cr \code{\link{prepMSig}} for
 \href{https://data.broadinstitute.org/gsea-msigdb/msigdb/release/7.0/}{molecular
  signatures} \cr \code{\link{prepString}} and \code{\link{anal_prnString}}
 for STRING-DB \cr

 \emph{Column keys in PSM, peptide and protein outputs} \cr
 system.file("extdata", "psm_keys.txt", package = "proteoQ") \cr
 system.file("extdata", "peptide_keys.txt", package = "proteoQ") \cr
 system.file("extdata", "protein_keys.txt", package = "proteoQ") \cr
}
