% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gage.R
\name{prnGAGE}
\alias{prnGAGE}
\title{GAGE analysis}
\usage{
prnGAGE(id = gene, scale_log2r = FALSE, df = NULL, filepath = NULL,
  filename = NULL, impute_na = TRUE, complete_cases = FALSE,
  method = "limma", gset_nm = "go_sets", var_cutoff = 0.001,
  pval_cutoff = 0.05, logFC_cutoff = log2(1.1), ...)
}
\description{
GAGE analysis
}
