% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{prnGSPA}
\alias{prnGSPA}
\title{GSPA of protein data}
\usage{
prnGSPA(id = gene, scale_log2r = TRUE, df = NULL, filepath = NULL,
  filename = NULL, impute_na = TRUE, complete_cases = FALSE,
  gset_nm = "go_sets", var_cutoff = 0.5, pval_cutoff = 0.01,
  logFC_cutoff = log2(1.2), min_size = 10, ...)
}
\arguments{
\item{id}{Character string to indicate the type of data. Peptide data will be
used at \code{id = pep_seq} or \code{pep_seq_mod}, and protein data at
\code{id = prot_acc} or \code{gene}.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation for all samples.}

\item{df}{The name of input data file. By default, it will be determined by
the value of \code{id}.}

\item{filepath}{Use system default.}

\item{filename}{Use system default.}

\item{impute_na}{Logical; if TRUE, imputes missing values.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used for visualization.}

\item{gset_nm}{Character vector of the name(s) of gene sets. The currently
available data sets include \code{c("go_sets", "kegg_sets", "c2_msig")}.}

\item{var_cutoff}{Not currently used.}

\item{pval_cutoff}{The cut-off in significance \code{pVal}. Protein entries
with \code{pVals} smaller than the threshold will be removed from GSPA.}

\item{logFC_cutoff}{The cut-off in \code{log2FC}. Protein entries with
\code{log2FC} smaller than the threshold will be removed from GSPA.}

\item{min_size}{Minimum number of protein entries for a gene set test.}

\item{...}{Contrasts for linear modeling. The syntax starts with a tilde,
followed by the name of an available column key in \code{expt_smry.xlsx} and
square brackets. The contrast groups are then quoted with multiple contrast
groups separated by commas.}
}
\description{
\code{prnGSPA} performs the Gene Set Probability Analysis (GSPA) aganist
protein \code{log2FC}
}
\details{
The formula(s) of contrast(s) used in \code{\link{prnSig}} will be taken by
default.
}
\examples{
prnGSPA(
  scale_log2r = TRUE, 
  impute_na = FALSE, 
  pval_cutoff = 5E-2, 
  gset_nm = c("go_sets", "kegg_sets"), 
)

\dontrun{
}

}
