% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{prnGSPA}
\alias{prnGSPA}
\title{GSPA of protein data}
\usage{
prnGSPA(id = gene, scale_log2r = TRUE, df = NULL, filepath = NULL,
  filename = NULL, impute_na = TRUE, complete_cases = FALSE,
  gset_nm = "go_sets", var_cutoff = 0.5, pval_cutoff = 0.01,
  logFC_cutoff = log2(1.2), min_size = 10, ...)
}
\arguments{
\item{id}{Character string to indicate the type of data. Peptide data will be
used at \code{id = pep_seq} or \code{pep_seq_mod}, and protein data at
\code{id = prot_acc} or \code{gene}.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation for all samples.}

\item{df}{The name of input data file. By default, it will be determined by
the value of \code{id}.}

\item{filepath}{Use system default.}

\item{filename}{Use system default.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used.}

\item{gset_nm}{Character vector containing the name(s) of gene sets for
enrichment analysis. The currently supported data sets are
\code{c("go_sets", "kegg_sets", "c2_msig")}.}

\item{var_cutoff}{Not currently used.}

\item{pval_cutoff}{The cut-off in significance \code{pVal}. Protein entries
with \code{pVals} smaller than the threshold will be removed from GSPA.}

\item{logFC_cutoff}{The cut-off in \code{log2FC}. Protein entries with
\code{log2FC} smaller than the threshold will be removed from GSPA.}

\item{min_size}{Minimum number of protein entries for consideration of a gene
set test.}

\item{...}{To be defined.}
}
\description{
\code{prnGSPA} performs the analysis of Gene Set Probability Asymmetricity
(GSPA) aganist protein \code{log2FC} data
}
\details{
The significance pVals of proteins are first obtained from
\code{\link{prnSig}}, which involves moderated t-test using
\code{\link[limma]{eBayes}}. The geometric mean of pVals are then each
calculated for the groups of up or down regulated proteins. The quotient of
the two pVals is used to prepsent the significance of gene set enrichment. The
arguments \code{pval_cutoff} and \code{logFC_cutoff} are used to filter out
low influence genes.

The formula(s) of contrast(s) used in \code{\link{prnSig}} will be taken by
default.
}
\examples{
prnGSPA(
  scale_log2r = TRUE,
  impute_na = FALSE,
  pval_cutoff = 5E-2,
  gset_nm = c("go_sets", "kegg_sets"),
)

\dontrun{
}

}
