% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{prnGSPAHM}
\alias{prnGSPAHM}
\title{Heat map visualization of GSPA results}
\usage{
prnGSPAHM(annot_cols = NULL, annot_colnames = NULL,
  annot_rows = NULL, ...)
}
\arguments{
\item{annot_cols}{A character vector of column keys that can be found in
\code{essmap_.*.txt}. The values under the selected keys will be used to
color-code enrichment terms on the top of heat maps. The default is NULL
without column annotation.}

\item{annot_colnames}{A character vector of replacement name(s) to
\code{annot_cols}. The default is NULL without name replacement.}

\item{annot_rows}{A character vector of column keys that can be found from
\code{essmeta_.*.txt} . The values under the selected keys will be used to
color-code essential terms on the side of heat maps. The default is NULL
without row annotation.}

\item{...}{\code{filter_}: Logical expression(s) for the row filtration of
data; also see \code{\link{normPSM}}. \cr \cr Additional arguments for
\code{\link[pheatmap]{pheatmap}}, i.e., \code{fontsize }... \cr \cr Note
arguments disabled from \code{pheatmap}: \cr \code{annotation_col}; instead
use keys indicated in \code{annot_cols} \cr \code{annotation_row}; instead
use keys indicated in \code{annot_rows}}
}
\description{
\code{prnGSPAHM} visualizes distance heat maps between essential and all gene
sets.
}
\details{
The list of gene sets and the associative quality metrics of \code{size} and
\code{ess_size} are assessed after data filtration with the criteria specified
by arguments \code{pval_cutoff} and \code{logFC_cutoff}, as well as optional
varargs of \code{filter_}.
}
\section{\code{Protein_GSPA_...txt}}{


 \tabular{ll}{ \strong{Key}   \tab \strong{Descrption}\cr term \tab a gene
 set term \cr is_essential \tab a logicial indicator of gene set essentiality
 \cr size \tab the number of IDs under a \code{term} \cr ess_size \tab the
 number of IDs that can be found under a corresponding essential set \cr
 contrast \tab a contrast of sample groups \cr p_val \tab significance p
 values \cr q_val \tab \code{p_val} with \code{BH} adjustment of multiple
 tests \cr log2fc \tab the fold change of a gene set at logarithmic base of 2
 \cr }
}

\section{\code{essmap_Protein_GSPA_...txt}}{


 \tabular{ll}{ \strong{Key}   \tab \strong{Descrption}\cr term \tab a gene
 set term \cr ess_term \tab an essential gene set term \cr size \tab the
 number of IDs under a \code{term} with matches to an \code{ess_term} \cr
 ess_size \tab the number of essential IDs under a \code{term} with matches
 to an \code{ess_term} \cr fraction \tab a fraction of matches in IDs between
 a \code{term} and a \code{ess_term} \cr distance \tab 1 - \code{fraction}
 \cr idx \tab a numeric index of \code{term} \cr ess_idx \tab a numeric index
 of \code{ess_term} \cr }
}

\examples{
# ===================================
# Heat maps of GSPA
# ===================================

## !!!require the brief working example in `?load_expts`

## global option
scale_log2r <- TRUE

## prerequisites in significance and enrichment tests
# (see also ?prnSig, ?prnGSPA)
pepSig(
  impute_na = FALSE, 
  W2_bat = ~ Term["(W2.BI.TMT2-W2.BI.TMT1)", 
                  "(W2.JHU.TMT2-W2.JHU.TMT1)", 
                  "(W2.PNNL.TMT2-W2.PNNL.TMT1)"], # batch effects
  W2_loc = ~ Term_2["W2.BI-W2.JHU", 
                    "W2.BI-W2.PNNL", 
                    "W2.JHU-W2.PNNL"], # location effects
  W16_vs_W2 = ~ Term_3["W16-W2"], 
)

prnSig(impute_na = FALSE)

prnGSPA(
  pval_cutoff = 5E-2,
  logFC_cutoff = log2(1.2),
  gspval_cutoff = 5E-2,
  gset_nms = c("go_sets", "kegg_sets"),
  impute_na = FALSE,
)

# ===================================
# Distance heat maps of gene sets
# (also interactive networks)
# ===================================
# a `term` is a subset of an `ess_term` if the distance is zero
prnGSPAHM(
  filter_by = exprs(distance <= .6),
  annot_cols = "ess_idx",
  annot_colnames = "Eset index",
  annot_rows = "ess_size",
  filename = show_some_redundancy.png,
)

# human terms only
prnGSPAHM(
  filter_num = exprs(distance <= .95),
  filter_sp = exprs(start_with_str("hs", term)),
  annot_cols = "ess_idx",
  annot_colnames = "Eset index",
  filename = show_more_connectivity.png,
)

# custom color palette
prnGSPAHM(
  annot_cols = c("ess_idx", "ess_size"),
  annot_colnames = c("Eset index", "Size"),
  filter_by = exprs(distance <= .95),
  color = colorRampPalette(c("blue", "white", "red"))(100),
  filename = "custom_colors.png",
)


}
\seealso{
\code{\link{load_expts}} for a reduced working example in data normalization \cr

 \code{\link{normPSM}} for extended examples in PSM data normalization \cr
 \code{\link{PSM2Pep}} for extended examples in PSM to peptide summarization \cr 
 \code{\link{mergePep}} for extended examples in peptide data merging \cr 
 \code{\link{standPep}} for extended examples in peptide data normalization \cr
 \code{\link{Pep2Prn}} for extended examples in peptide to protein summarization \cr
 \code{\link{standPrn}} for extended examples in protein data normalization. \cr 
 \code{\link{purgePSM}} and \code{\link{purgePep}} for extended examples in data purging \cr
 \code{\link{pepHist}} and \code{\link{prnHist}} for extended examples in histogram visualization. \cr 
 \code{\link{extract_raws}} and \code{\link{extract_psm_raws}} for extracting MS file names \cr 
 
 \code{\link{contain_str}}, \code{\link{contain_chars_in}}, \code{\link{not_contain_str}}, 
 \code{\link{not_contain_chars_in}}, \code{\link{start_with_str}}, 
 \code{\link{end_with_str}}, \code{\link{start_with_chars_in}} and 
 \code{\link{ends_with_chars_in}} for data subsetting by character strings \cr 
 
 \code{\link{pepImp}} and \code{\link{prnImp}} for missing value imputation \cr 
 \code{\link{pepSig}} and \code{\link{prnSig}} for significance tests \cr 
 \code{\link{pepVol}} and \code{\link{prnVol}} for volcano plot visualization \cr 
 
 \code{\link{prnGSPA}} for gene set enrichment analysis by protein significance pVals \cr 
 \code{\link{gspaMap}} for mapping GSPA to volcano plot visualization \cr 
 \code{\link{prnGSPAHM}} for heat map and network visualization of GSPA results \cr 
 \code{\link{prnGSVA}} for gene set variance analysis \cr 
 \code{\link{prnGSEA}} for data preparation for online GSEA. \cr 
 
 \code{\link{pepMDS}} and \code{\link{prnMDS}} for MDS visualization \cr 
 \code{\link{pepPCA}} and \code{\link{prnPcA}} for PCA visualization \cr 
 \code{\link{pepHM}} and \code{\link{prnHM}} for heat map visualization \cr 
 \code{\link{pepCorr_logFC}}, \code{\link{prnCorr_logFC}}, \code{\link{pepCorr_logInt}} and 
 \code{\link{prnCorr_logInt}}  for correlation plots \cr 
 
 \code{\link{anal_prnTrend}} and \code{\link{plot_prnTrend}} for protein trend analysis and visualization \cr 
 \code{\link{anal_pepNMF}}, \code{\link{anal_prnNMF}}, \code{\link{plot_pepNMFCon}}, 
 \code{\link{plot_prnNMFCon}}, \code{\link{plot_pepNMFCoef}}, \code{\link{plot_prnNMFCoef}} and 
 \code{\link{plot_metaNMF}} for protein NMF analysis and visualization \cr 
 
 \code{\link{dl_stringdbs}} and \code{\link{anal_prnString}} for STRING-DB
}
