% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{prnGSPAHM}
\alias{prnGSPAHM}
\title{Heat map visualization of GSPA results}
\usage{
prnGSPAHM(annot_cols = NULL, annot_colnames = NULL,
  annot_rows = NULL, ...)
}
\arguments{
\item{annot_cols}{A character vector of column keys that can be found in
\code{essmap_.*.csv}. The values under the selected keys will be used to
color-code enrichment terms on the top of heat maps.}

\item{annot_colnames}{A character vector of replacement name(s) to
\code{annot_cols}. The default is NULL with no replacement.}

\item{annot_rows}{A character vector of column keys that can be found from
\code{essmeta_.*.csv} . The values under the selected keys will be used to
color-code essential terms on the side of heat maps.}

\item{...}{\code{filter_}: Logical expression(s) for the row filtration of
data; also see \code{\link{normPSM}}. \cr \cr Additional arguments for
\code{\link[pheatmap]{pheatmap}}, i.e., \code{fontsize }... \cr Note
arguments disabled from \code{pheatmap}: \cr \code{annotation_col}; instead
use keys indicated in \code{annot_cols} \cr \code{annotation_row}; instead
use keys indicated in \code{annot_rows}}
}
\description{
\code{prnGSPAHM} visualizes distance heat maps between essential and all gene
sets.
}
\details{
The list of gene sets and the associative quality metrics of \code{size} and
\code{ess_size} are assessed after data filtration with the criteria specified
by arguments \code{pval_cutoff} and \code{logFC_cutoff}, as well as optional
varargs of \code{filter_}.
}
\section{\code{Protein_GSPA_...csv}}{


 \tabular{ll}{ \strong{Key}   \tab \strong{Descrption}\cr term \tab a gene
 set term \cr is_essential \tab a logicial indicator of gene set essentiality
 \cr size \tab the number of IDs under a \code{term} \cr ess_size \tab the
 number of IDs that can be found under a corresponding essential set \cr
 contrast \tab a contrast of sample groups \cr p_val \tab significance p
 values \cr q_val \tab \code{p_val} with \code{BH} adjustment of multiple
 tests \cr log2fc \tab the fold change of a gene set at logarithmic base of 2
 \cr }
}

\section{\code{essmap_Protein_GSPA_...csv}}{


 \tabular{ll}{ \strong{Key}   \tab \strong{Descrption}\cr term \tab a gene
 set term \cr ess_term \tab an essential gene set term \cr size \tab the
 number of IDs under a \code{term} with matches to an \code{ess_term} \cr
 ess_size \tab the number of essential IDs under a \code{term} with matches
 to an \code{ess_term} \cr fraction \tab a fraction of matches in IDs between
 a \code{term} and a \code{ess_term} \cr distance \tab 1 - \code{fraction}
 \cr idx \tab a numeric index of \code{term} \cr ess_idx \tab a numeric index
 of \code{ess_term} \cr }
}

\seealso{
\code{\link{prnGSPA}} for gene set enrichment analysis.
}
