% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gsva.R
\name{prnGSVA}
\alias{prnGSVA}
\title{GSVA of protein data}
\usage{
prnGSVA(scale_log2r = TRUE, df = NULL, filepath = NULL,
  filename = NULL, impute_na = TRUE, complete_cases = FALSE,
  lm_method = "limma", gset_nms = "go_sets", var_cutoff = 0.5,
  pval_cutoff = 1e-04, logFC_cutoff = log2(1.1), ...)
}
\arguments{
\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE.}

\item{df}{The name of input data file. By default, it will be determined
automatically by the value of \code{id}.}

\item{filepath}{Use system default.}

\item{filename}{Use system default.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used. The default is FALSE.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{lm_method}{Character string indicating the linear modeling method for
significance assessment of GSVA enrichment scores}

\item{var_cutoff}{Numeric; the cut-off in the variances of \code{log2FC}.
Entries with variances smaller than the threshold will be removed from
linear modeling. The default is 1E-3.}

\item{pval_cutoff}{Numeric; the cut-off in significance \code{pVal}. Entries
with \code{pVals} smaller than the threshold will be removed from multiple
test corrections. The default is \code{1} without pVal cut-offs.}

\item{logFC_cutoff}{Numeric; the cut-off in \code{log2FC}. Entries with
\code{log2FC} smaller than the threshold will be removed from multiple test
corrections. The default is at \code{log2(1)}  without fold change cut-offs.}

\item{...}{Parameters for \code{\link[pheatmap]{pheatmap}}}

\item{gset_nm}{Character vector of the name(s) of gene sets. The currently
available data sets include c("go_sets", "kegg_sets", "c2_msig").}

\item{Parameters}{for \code{\link[GSVA]{gsva}}}
}
\description{
\code{prnGSVA} performs the GSVA aganist protein \code{log2FC}
}
\details{
The formula(s) of contrast(s) used in \code{\link{prnSig}} will be taken by
default.
}
\examples{
prnGSVA(
 scale_log2r = TRUE, 
 impute_na = FALSE, 
 min.sz = 10, 
 verbose = FALSE, 
 parallel.sz = 0, 
 mx.diff = TRUE, 
 gset_nms = c("go_sets", "kegg_sets"), 
)

}
