% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-histo.R
\name{prnHist}
\alias{prnHist}
\title{Histograms of Protein Data}
\usage{
prnHist(...)
}
\description{
Histograms of Protein Data
}
\examples{
# no scaling normalization
prnHist(
  scale_log2r = FALSE, 
  show_curves = TRUE,
  show_vline = TRUE,
  ncol = 4
)

# scaling normalization
prnHist(
  scale_log2r = TRUE, 
  show_curves = TRUE,
  show_vline = TRUE,
  ncol = 4
)

# direct call to proteoHist()
proteoHist(
  id = prot_acc, 
  scale_log2r = TRUE, 
  show_curves = TRUE,
  show_vline = TRUE,
  ncol = 4
)

\dontrun{
prnHist(
  col_select = "a_non_existed_column_key"
)
}

}
\seealso{
\code{\link{proteoHist}} for parameters
}
