% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-histo.R
\name{prnHist}
\alias{prnHist}
\title{Visualizes the histograms of protein data}
\usage{
prnHist(...)
}
\description{
Visualizes the histograms of protein data
}
\examples{
# no scaling normalization
prnHist(
  scale_log2r = FALSE,
  show_curves = TRUE,
  show_vline = TRUE,
  xmin = -1,
  xmax = 1,
  ncol = 4
)

# scaling normalization
prnHist(
  scale_log2r = TRUE,
  show_curves = TRUE,
  show_vline = TRUE,
  xmin = -1,
  xmax = 1,
  ncol = 4
)

\dontrun{
prnHist(
  col_select = "a_non_existed_column_key"
)
}

}
\seealso{
\code{\link{proteoHist}} for parameters
}
