% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-mds.R
\name{prnPCA}
\alias{prnPCA}
\title{PCA plots}
\usage{
prnPCA(...)
}
\description{
PCA plots of protein data
}
\examples{
prnPCA(
  scale_log2r = TRUE,
  col_color = Color,
  col_shape = Shape,
  show_ids = TRUE,
)

\dontrun{
prnPCA(
  col_color = "column_key_not_existed",
  col_shape = "also_a_missing_column_key"
)
}

}
\seealso{
\code{\link{proteoPCA}} for parameters
}
