% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-sigtests.R
\name{prnSig}
\alias{prnSig}
\title{Significance Tests of Protein \code{log2-ratios}}
\usage{
prnSig(...)
}
\description{
Significance Tests of Protein \code{log2-ratios}
}
\seealso{
\code{\link{proteoSigtest}} for parameters
}
