% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-volcanos.R
\name{prnVol}
\alias{prnVol}
\title{Volcano Plots of Protein \code{log2-ratios}}
\usage{
prnVol(...)
}
\description{
Volcano Plots of Protein \code{log2-ratios}
}
\seealso{
\code{\link{proteoVolcano}} for parameters
}
