% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{probco_bypeplen}
\alias{probco_bypeplen}
\title{Helper of \link{calc_pepfdr}.}
\usage{
probco_bypeplen(len, td, fdr_type, target_fdr)
}
\arguments{
\item{len}{Numeric; the length of peptides.}

\item{td}{A target-decoy pair.}

\item{fdr_type}{Character string; the type of FDR controlling. The value is
in one of c("psm", "peptide"). Note that protein FDR is in conjunction with
"psm" or "peptide". Separate protein FDR may be available in the future.}

\item{target_fdr}{Numeric; the levels of false-discovery rate (FDR) at the
levels of PSMs or peptides. The same level applies further to protein FDR
and indicated under the column \code{prot_issig} in outputs. This results
in FDR controls of either PSM <-> protein or peptide <-> protein at a given
\code{target_fdr}.}
}
\description{
Calculates the probability cut-off for target-decoy pairs at a given peptide
length.
}
