% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{probco_bypeplen}
\alias{probco_bypeplen}
\title{Helper of \link{calc_pepfdr}.}
\usage{
probco_bypeplen(len, td, fdr_type, target_fdr, out_path)
}
\arguments{
\item{len}{Numeric; the length of peptides.}

\item{td}{A target-decoy pair.}

\item{fdr_type}{Character string; the type of FDR controlling. The value is
in one of c("psm", "peptide", "protein"). Note that \code{fdr_type =
protein} is equivalent to \code{fdr_type = psm} or \code{peptide} with the
additional filtration of data at \code{prot_tier == 1}.}

\item{target_fdr}{Numeric; the levels of false-discovery rate (FDR) at the
levels of PSMs, peptides or proteins. See also argument \code{fdr_type}.}

\item{out_path}{The output path.}
}
\description{
Calculates the probability cut-off for target-decoy pairs at a given peptide
length.
}
