% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{proc_mgf_chunks_i}
\alias{proc_mgf_chunks_i}
\title{Helper of \link{proc_mgf_chunks}.}
\usage{
proc_mgf_chunks_i(file, topn_ms2ions = 100, ret_range = c(0, Inf))
}
\arguments{
\item{file}{A mgf chunk (chunk_1.mgf etc.) with prepending file path.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches.}

\item{ret_range}{The range of retention time in seconds.}
}
\description{
Helper of \link{proc_mgf_chunks}.
}
