% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{proc_mgfs}
\alias{proc_mgfs}
\title{Helper in processing MGF entries in chunks.}
\usage{
proc_mgfs(
  lines,
  topn_ms2ions = 100L,
  ret_range = c(0L, Inf),
  ppm_ms2 = 25L,
  min_ms2mass = 110L,
  index_ms2 = FALSE
)
}
\arguments{
\item{lines}{MGF lines.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{ret_range}{The range of retention time in seconds.}
}
\description{
Helper in processing MGF entries in chunks.
}
