% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-go.R
\name{proc_obo}
\alias{proc_obo}
\title{Helper to save `obo` without header}
\usage{
proc_obo(
  db_path,
  fn_obo,
  type = c("biological_process", "cellular_component", "molecular_function")
)
}
\description{
Helper to save `obo` without header
}
