% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-mds.R
\name{proteoEucDist}
\alias{proteoEucDist}
\title{Visualizes Euclidean distances}
\usage{
proteoEucDist(id = gene, col_select = NULL, col_group = NULL,
  col_color = NULL, col_fill = NULL, col_shape = NULL,
  col_size = NULL, col_alpha = NULL, scale_log2r = FALSE,
  adjEucDist = FALSE, annot_cols = NULL, df = NULL,
  filepath = NULL, filename = NULL, ...)
}
\arguments{
\item{id}{Character string to indicate the type of data. Peptide data will be
used at column \code{id = "pep_seq"} or \code{"pep_seq_mod"}, and protein
data at \code{id = "prot_acc"} or \code{"gene"}.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2-ratios} to the same
scale of standard deviation for all samples.}

\item{adjEucDist}{Logical; if TRUE, adjusts the inter-plex Euclidean distance
by \eqn{1/sqrt(2)}. The option \code{adjEucDist = TRUE} may be suitable when
\code{reference samples} in each TMT plex undergo approximately the same
sample handling process as the rest of the samples. For instance,
\code{reference samples} were split at the levels of protein lysates.
Typically, \code{adjEucDist = FALSE} if \code{reference samples} were split
near the end of a sample handling process, for instance, at the stages
immediately before or after TMT labeling.}

\item{annot_cols}{Column names available in \code{expt_smry.xlsx}. Values under
the selected columns will be used to color-code sample IDs on the top of
\code{EucDist} plots.}

\item{df}{The filename of input data. By default, it will be determined by the
value of \code{id}.}

\item{filepath}{The filepath to output results. By default, it will be
determined by the name of the current function \code{call} and the value of
\code{id}.}

\item{filename}{A representative filename to output images. By default, it
will be determined by the names of the current \code{call}.}

\item{...}{Parameters inherited from \code{\link[pheatmap]{pheatmap}}}
}
\value{
Plots of distance matrices.
}
\description{
\code{proteoEucDist} visualizes the Euclidean distances of protein or peptide
data.
}
\details{
An Euclidean distance matrix of \code{log2-ratios} is returned by
\code{\link[base]{dist}}.
}
