% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gsea.R
\name{proteoGSEA}
\alias{proteoGSEA}
\alias{prnGSEA}
\title{GSEA of protein data}
\usage{
proteoGSEA(id = gene, df = NULL, filepath = NULL, filename = NULL,
  gset_nms = "go_sets", scale_log2r = TRUE, impute_na = NULL,
  complete_cases = FALSE, var_cutoff = 0.5, pval_cutoff = 0.05,
  logFC_cutoff = log2(1.2), gspval_cutoff = 0.05, min_size = 10,
  max_size = Inf, min_greedy_size = 1, fml_nms = NULL,
  task = "anal", ...)

prnGSEA(...)
}
\arguments{
\item{id}{Character string indicating the type of data. The value will be
determined automatically.}

\item{df}{The name of input data file. By default, it will be determined
automatically by the value of \code{id}.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{filename}{A file name to output results. By default, it will be
determined automatically by the name of the calling function and the value
of id in the call.}

\item{gset_nms}{Not currently used.}

\item{scale_log2r}{Not used for full volcano plots by \code{pepVol} or 
\code{prnVol}; matched to the value in the latest call to \code{\link{prnGSPA}} 
for \code{gspaMap}.}

\item{impute_na}{Logical. At the NULL default, the TRUE or FALSE will match
the choice in \code{\link{pepSig}} for peptide and \code{\link{prnSig}} for
protein data.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{var_cutoff}{Numeric value or vector; the cut-off in the variance of
protein \code{log2FC} across samples. Entries with \code{variances} less
than the threshold will be removed for enrichment analysis. The default is
0.5.}

\item{pval_cutoff}{Numeric value or vector; the cut-off in protein
significance \code{pVal}. Entries with \code{pVals} less significant than
the threshold will be ignored during enrichment analysis. The default is
0.05 for all formulae matched to or specified in argument \code{fml_nms}.
Formula-specific threshold is allowed by supplying a vector of cut-off
values.}

\item{logFC_cutoff}{Numeric value or vector; the cut-off in protein
\code{log2FC}. The default magnitude is \code{log2(1.2)} for all formulae
matched to or specified in argument \code{fml_nms}. Entries with absolute
\code{log2FC} smaller than the threshold will be ignored during enrichment
analysis. Formula-specific threshold is allowed by supplying a vector of
absolute values in \code{log2FC}.}

\item{gspval_cutoff}{Not currently used.}

\item{min_size}{Not currently used.}

\item{max_size}{Not currently used.}

\item{min_greedy_size}{Not currently used.}

\item{fml_nms}{Character string or vector; the formula name(s). By default,
the formula(e) will match those used in \code{\link{pepSig}} or
\code{\link{prnSig}}.}

\item{...}{\code{filter_}: Logical expression(s) for the row filtration of
data; also see \code{\link{normPSM}}.}
}
\description{
\code{proteoGSEA} prepares data for the analysis of
\code{\href{http://software.broadinstitute.org/gsea/index.jsp}{GSEA}} aganist
protein \code{log2FC} data. Users should avoid calling the method directly,
but instead use the following wrappers.

\code{prnGSEA} is a wrapper of \code{\link{proteoGSEA}} for gene set
enrichment analysis.
}
\details{
The arguments \code{var_cutoff}, \code{pval_cutoff} and \code{logFC_cutoff}
are used to filter out low influence genes. Additional subsetting of data via
the \code{vararg} approach of \code{filter_} is feasible.

The outputs include \code{Protein_GSEA.gct} and \code{protein_GSEA.cls} for
samples indicated in file \code{Protein_pVals.txt} or
\code{Protein_impNA_pVals.txt}. These outputs can be used with online
\code{\href{http://software.broadinstitute.org/gsea/index.jsp}{GSEA}}.

The current GSEA may not support the comparisons between two grouped
conditions, i.e.,  (grpA + grpB) versus (grpC + grpD). The \code{prnGSEA}
utility further breaks the input data into pairs of groups according to the
formulae and contrasts defined in \code{pepSig} or \code{prnSig}. The
phenotype labels are then reformed in reflection of the original group names,
weights and directions, i.e., \code{0.5xgrpA&0.5xgrpB	-0.5xgrpC&-0.5xgrpD}.
The corresponding \code{.gct} and \code{.cls} files can be used with the
online or the github version of R-GSEA.
}
\examples{
# ===================================
# GSEA
# ===================================

## !!!require the brief working example in `?load_expts`

# global option
scale_log2r <- TRUE

## prerequisites in significance tests
# (see also ?prnSig)
pepSig(
  impute_na = FALSE, 
  W2_bat = ~ Term["(W2.BI.TMT2-W2.BI.TMT1)", 
                  "(W2.JHU.TMT2-W2.JHU.TMT1)", 
                  "(W2.PNNL.TMT2-W2.PNNL.TMT1)"], # batch effects
  W2_loc = ~ Term_2["W2.BI-W2.JHU", 
                    "W2.BI-W2.PNNL", 
                    "W2.JHU-W2.PNNL"], # location effects
  W16_vs_W2 = ~ Term_3["W16-W2"], 
)

prnSig(impute_na = FALSE)

# all human proteins
prnGSEA(
  var_cutoff = 0, 
  pval_cutoff = 1, 
  logFC_cutoff = log2(1), 
  filter_by_sp = exprs(species == "human"), 
)

# prefiltration by variances, pVals and logFCs
prnGSEA(
  var_cutoff = 0.5,     
  pval_cutoff = 5E-2,
  logFC_cutoff = log2(1.2),
  filter_by_sp = exprs(species == "human"), 
)

# cases that are complete with no missing values
prnGSEA(
  var_cutoff = 0.5,     
  pval_cutoff = 5E-2,
  logFC_cutoff = log2(1.2),
  complete_cases = TRUE, 
  filter_by_sp = exprs(species == "human"), 
)

# customized thresholds for the corresponding formulae in `pepSig` or `prnSig()`
prnGSEA(
  fml_nms = c("W2_bat", "W2_loc", "W16_vs_W2"),
  var_cutoff = c(0, 0.2, 0.5), 
  pval_cutoff = c(5E-2, 5E-2, 1E-5),
  logFC_cutoff = c(log2(1.1), log2(1.1), log2(1.2)),
  filter_by_sp = exprs(species == "human"), 
)

}
\seealso{
\code{\link{load_expts}} for a reduced working example in data normalization \cr

 \code{\link{normPSM}} for extended examples in PSM data normalization \cr
 \code{\link{PSM2Pep}} for extended examples in PSM to peptide summarization \cr 
 \code{\link{mergePep}} for extended examples in peptide data merging \cr 
 \code{\link{standPep}} for extended examples in peptide data normalization \cr
 \code{\link{Pep2Prn}} for extended examples in peptide to protein summarization \cr
 \code{\link{standPrn}} for extended examples in protein data normalization. \cr 
 \code{\link{purgePSM}} and \code{\link{purgePep}} for extended examples in data purging \cr
 \code{\link{pepHist}} and \code{\link{prnHist}} for extended examples in histogram visualization. \cr 
 \code{\link{extract_raws}} and \code{\link{extract_psm_raws}} for extracting MS file names \cr 
 
 \code{\link{contain_str}}, \code{\link{contain_chars_in}}, \code{\link{not_contain_str}}, 
 \code{\link{not_contain_chars_in}}, \code{\link{start_with_str}}, 
 \code{\link{end_with_str}}, \code{\link{start_with_chars_in}} and 
 \code{\link{ends_with_chars_in}} for data subsetting by character strings \cr 
 
 \code{\link{pepImp}} and \code{\link{prnImp}} for missing value imputation \cr 
 \code{\link{pepSig}} and \code{\link{prnSig}} for significance tests \cr 
 \code{\link{pepVol}} and \code{\link{prnVol}} for volcano plot visualization \cr 
 
 \code{\link{prnGSPA}} for gene set enrichment analysis by protein significance pVals \cr 
 \code{\link{gspaMap}} for mapping GSPA to volcano plot visualization \cr 
 \code{\link{prnGSPAHM}} for heat map and network visualization of GSPA results \cr 
 \code{\link{prnGSVA}} for gene set variance analysis \cr 
 \code{\link{prnGSEA}} for data preparation for online GSEA. \cr 
 
 \code{\link{pepMDS}} and \code{\link{prnMDS}} for MDS visualization \cr 
 \code{\link{pepPCA}} and \code{\link{prnPcA}} for PCA visualization \cr 
 \code{\link{pepHM}} and \code{\link{prnHM}} for heat map visualization \cr 
 \code{\link{pepCorr_logFC}}, \code{\link{prnCorr_logFC}}, \code{\link{pepCorr_logInt}} and 
 \code{\link{prnCorr_logInt}}  for correlation plots \cr 
 
 \code{\link{anal_prnTrend}} and \code{\link{plot_prnTrend}} for protein trend analysis and visualization \cr 
 \code{\link{anal_pepNMF}}, \code{\link{anal_prnNMF}}, \code{\link{plot_pepNMFCon}}, 
 \code{\link{plot_prnNMFCon}}, \code{\link{plot_pepNMFCoef}}, \code{\link{plot_prnNMFCoef}} and 
 \code{\link{plot_metaNMF}} for protein NMF analysis and visualization \cr 
 
 \code{\link{dl_stringdbs}} and \code{\link{getStringDB}} for STRING-DB
}
