% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{proteoGSPA}
\alias{proteoGSPA}
\alias{prnGSPA}
\title{GSPA of protein data}
\usage{
proteoGSPA(id = gene, scale_log2r = TRUE, df = NULL,
  filepath = NULL, filename = NULL, impute_na = NULL,
  complete_cases = FALSE, gset_nms = "go_sets", var_cutoff = 0.5,
  pval_cutoff = 0.05, logFC_cutoff = log2(1.2), gspval_cutoff = 0.05,
  min_size = 10, max_size = Inf, min_greedy_size = 1,
  fml_nms = NULL, task = "anal", ...)

prnGSPA(...)
}
\arguments{
\item{id}{Character string indicating the type of data. The value will be
determined automatically.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE.}

\item{df}{The name of input data file. By default, it will be determined
automatically by the value of \code{id}.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{filename}{A file name to output results. By default, it will be
determined automatically by the name of the calling function and the value
of id in the call.}

\item{impute_na}{Logical. At the NULL default, the TRUE or FALSE will match
the choice in \code{\link{pepSig}} for peptide and \code{\link{prnSig}} for
protein data.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{gset_nms}{Character string or vector containing the name(s) of gene sets
for enrichment analysis. The possible values are in \code{c("go_sets",
"kegg_sets")}. Note that the currently supported species are human, mouse
and rat.}

\item{var_cutoff}{Not currently used.}

\item{pval_cutoff}{Numeric value or vector; the cut-off in protein
significance \code{pVal}. Entries with \code{pVals} less significant than
the threshold will be ignored during enrichment analysis.}

\item{logFC_cutoff}{Numeric value or vector; the cut-off in protein
\code{log2FC}. Entries with absolute \code{log2FC} smaller than the
threshold will be ignored during enrichment analysis.}

\item{gspval_cutoff}{Numeric value or vector; the cut-off in gene-set
significance \code{pVal}. Only enrichment terms with \code{pVals} more
significant than the threshold will be reported.}

\item{min_size}{Numeric value or vector; minimum number of protein entries for
consideration in gene set tests. The number is the sum of up or
down-expressed proteins after data filtration by \code{pval_cutoff},
\code{logFC_cutoff} or varargs expressions under \code{filter_}.}

\item{max_size}{Numeric value or vector; maximum number of protein entries for
consideration in gene set tests. The number is the sum of up or
down-expressed proteins after data filtration by \code{pval_cutoff},
\code{logFC_cutoff} or varargs expressions under \code{filter_}. The default
in infinite. For samples that are vastly different between contrasts,
essential gene sets may be blackholed with trivial terms such as cell parts,
molecular functions et al. In cases like this, a smaller \code{max_size} may
be considered.}

\item{min_greedy_size}{Numeric value or vector; minimum number of unique
protein entries for a gene set to be considered essential.}

\item{fml_nms}{Character string or vector; the formula name(s). By default,
the names will match those used in \code{\link{pepSig}} or
\code{\link{prnSig}}.}

\item{...}{\code{filter_}: Logical expression(s) for the row filtration of
data; also see \code{\link{normPSM}}.}
}
\description{
\code{prnGSPA} performs the analysis of Gene Set Probability Asymmetricity
(GSPA) aganist protein \code{log2FC} data. Users should avoid call the method
directly, but instead use the following wrappers.

\code{prnGSPA} is a wrapper of \code{\link{proteoGSPA}} for enrichment analysis
}
\details{
The significance \code{pVals} of proteins are first obtained from
\code{\link{prnSig}}, which involves linear modelings using
\code{\link[limma]{eBayes}} or \code{\link[lmerTest]{lmer}}. The geometric
mean of \code{pVals} are then each calculated for the groups of up or down
regulated proteins. The quotient of the two \code{pVals} is used to prepsent
the significance of gene set enrichment. The arguments \code{pval_cutoff} and
\code{logFC_cutoff} are used to filter out low influence genes. Additional
subsetting of data via the \code{vararg} approach of \code{filter_} is
feasible.

The formula(e) of contrast(s) used in \code{\link{prnSig}} will be taken by
default.
}
