% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-gspa.R
\name{proteoGSPA}
\alias{proteoGSPA}
\alias{prnGSPA}
\alias{prnGSPAHM}
\title{GSPA of protein data}
\usage{
proteoGSPA(id = gene, scale_log2r = TRUE, df = NULL,
  filepath = NULL, filename = NULL, impute_na = TRUE,
  complete_cases = FALSE, gset_nms = "go_sets", var_cutoff = 0.5,
  pval_cutoff = 0.05, logFC_cutoff = log2(1.2), gspval_cutoff = 0.01,
  min_size = 10, task = "anal", ...)

prnGSPA(...)

prnGSPAHM(annot_cols = NULL, annot_colnames = NULL, ...)
}
\arguments{
\item{id}{Character string to indicate the type of data. The value will be
determined automatically. Peptide data will be used at \code{id = pep_seq}
or \code{pep_seq_mod}, and protein data will be used at \code{id = prot_acc}
or \code{gene}.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples.}

\item{df}{The name of input data file. By default, it will be determined
automatically by the value of \code{id}.}

\item{filepath}{Use system default.}

\item{filename}{Use system default.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used.}

\item{gset_nms}{Character vector containing the name(s) of gene sets for
enrichment analysis. The possible values are in \code{c("go_sets",
"kegg_sets", "c2_msig")}.}

\item{var_cutoff}{Not currently used.}

\item{pval_cutoff}{The cut-off in protein significance \code{pVal}. Entries
with \code{pVals} smaller than the threshold will be ignored.}

\item{logFC_cutoff}{The cut-off in protein \code{log2FC}. Entries with
\code{log2FC} smaller than the threshold will be ignored.}

\item{gspval_cutoff}{The cut-off in significance \code{pVal} of gene sets.
Gene sets with \code{pVals} smaller than the threshold will be ignored.}

\item{min_size}{Minimum number of protein entries for consideration of a gene
set test.}

\item{...}{\code{filter_}: Logical expression(s) for the row filtration of
data; also see \code{\link{normPSM}}.}

\item{annot_cols}{For use in heat map visualization of gene sets. A character
vector of column keys in \code{essmap_Protein_GSPA_...csv}. The values under
the selected keys will be used to color-code sample IDs on the top of the
indicated plot.}

\item{annot_colnames}{For use in heat map visualization of gene sets. A
character vector of replacement name(s) to annot_cols.}
}
\description{
\code{prnGSPA} performs the analysis of Gene Set Probability Asymmetricity
(GSPA) aganist protein \code{log2FC} data

\code{prnGSPA} is a wrapper of \code{\link{proteoGSPA}}

\code{prnGSPAHM} is a wrapper of \code{\link{proteoGSPA}}
}
\details{
The significance \code{pVals} of proteins are first obtained from
\code{\link{prnSig}}, which involves linear modelings using
\code{\link[limma]{eBayes}} or \code{\link[lmerTest]{lmer}}. The geometric
mean of \code{pVals} are then each calculated for the groups of up or down
regulated proteins. The quotient of the two \code{pVals} is used to prepsent
the significance of gene set enrichment. The arguments \code{pval_cutoff} and
\code{logFC_cutoff} are used to filter out low influence genes.

The formula(s) of contrast(s) used in \code{\link{prnSig}} will be taken by
default.

\code{prnGSPAHM} visualizes the distance heat map between essential and all
gene sets. The distance is \eqn{1 - x} where \eqn{x} is the percent overlap
for ids under a gene set.
}
\examples{
prnGSPA(
  scale_log2r = TRUE,
  impute_na = FALSE,
  pval_cutoff = 5E-2,
  gspval_cutoff = 5E-3,
  gset_nms = c("go_sets", "kegg_sets"),

  filter_by_npep = exprs(prot_n_pep >= 2),
  # `filename(s)` will be automated,
)


# distances of GSPA terms
prnGSPAHM(
  filter_by = exprs(count >= 8, percent >= .20),
  annot_cols = "source",
  annot_colnames = "Essential set index",
  
  fontsize = 16,
  fontsize_row = 20,
  fontsize_col = 20,
  fontsize_number = 8,
  border_color = "grey60",
  cellwidth = 24,
  cellheight = 24,
)

# custom colors
# `source` and `count` are column keys in `essmap_Protein_GSPA...csv`
prnGSPAHM(
  annot_cols = c("source", "count"),
  annot_colnames = c("Essential set index", "Count"),	
  filter_by = exprs(count >= 8, percent >= .20),
  
  fontsize = 16,
  fontsize_row = 20,
  fontsize_col = 20,
  fontsize_number = 8,
  border_color = "grey60",
  cellwidth = 24,
  cellheight = 24,
  color = colorRampPalette(c("blue", "white", "red"))(100),
  filename = "custom_colors.png"
)
\dontrun{
}

}
