% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-histo.R
\name{proteoHist}
\alias{proteoHist}
\alias{pepHist}
\alias{prnHist}
\title{Visualizes histograms}
\usage{
proteoHist(id = c("pep_seq", "pep_seq_mod", "prot_acc", "gene"),
  col_select = NULL, scale_log2r = FALSE, pep_pattern = "zzz",
  show_curves = TRUE, show_vline = TRUE, new_fit = FALSE,
  df = NULL, filepath = NULL, filename = NULL, ...)

pepHist(...)

prnHist(...)
}
\arguments{
\item{id}{Character string to indicate the type of data. Peptide data will be
used at \code{id = pep_seq} or \code{pep_seq_mod}, and protein data at
\code{id = prot_acc} or \code{gene}.}

\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_select} will be
included in the indicated analysis. At the NULL default, the column key
\code{Select} will be used.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation for all samples.}

\item{pep_pattern}{Character string containing one-letter representation of
amino acids. At the "zzz" default, all peptides will be used. Letters in the
character string are case sensitive. For example, \code{pep_pattern = "m"}
will extract peptides with oxidized methiones and \code{pep_pattern = "_"}
with N-terminal acetylation.}

\item{show_curves}{Logical; if TRUE, shows the fitted curves.}

\item{show_vline}{Logical; if TRUE, shows the vertical lines at \code{x = 0}.}

\item{new_fit}{Not currently used.}

\item{df}{The name of input data file. By default, it will be determined by
the value of \code{id}.}

\item{filepath}{A file path to output results. By default, it will be
determined by the name of the calling \code{function} and the value of
\code{id} in the \code{call}.}

\item{filename}{A representative file name to output image(s). By default, it
will be determined by the name of the current \code{call}. The image(s) are
saved via \code{\link[ggplot2]{ggsave}} and the image type will be
automatically determined by the extension of the file name. A \code{.png}
format will be used at default or an unrecognized file extension.}

\item{...}{Additional parameters for plotting: \cr \code{xmin}, the minimum
\eqn{x} at a log2 scale; the default is -2. \cr \code{xmax}, the maximum
\eqn{x} at a log2 scale; the default is +2. \cr \code{x_breaks}, the breaks
in \eqn{x}-axis at a log2 scale; the default is 1. \cr \code{binwidth}, the
binwidth of \code{log2FC}; the default is \eqn{(xmax - xmin)/80}. \cr
\code{ncol}, the number of columns; the default is 1. \cr \code{width}, the
width of plot; \cr \code{height}, the height of plot.}
}
\value{
The histograms of \code{log2FC} under
 \code{~\\dat_dir\\Peptide\\Histogram} or
 \code{~\\dat_dir\\Protein\\Histogram}.
}
\description{
\code{proteoHist} prepares the histogram visualization of \code{log2FC} for
proteins or peptides data.

\code{pepHist} is a wrapper function of \code{\link{proteoHist}} for peptide
data

\code{prnHist} is a wrapper function of \code{\link{proteoHist}} for protein
data
}
\details{
In the histograms, the \code{log2FC} under each TMT channel are color-coded by
their contributing reporter-ion intensity.

The function matches the current \code{id} to those in the latest \code{call}
to \code{\link{normPep}} or \code{\link{normPrn}}.  For example, if
\code{pep_seq} was used in \code{\link{normPep()}}, the current \code{id =
pep_seq_mod} will be matched to \code{id = pep_seq}.
}
\examples{
# without scaling normalization
pepHist(
  scale_log2r = FALSE,
  xmin = -1,
  xmax = 1,
  ncol = 5
)

# with scaling normalization
pepHist(
  scale_log2r = TRUE,
  xmin = -1,
  xmax = 1,
  ncol = 5
)

\dontrun{
pepHist(
  col_select = "a_non_existed_column_key"
)
}

# no scaling normalization
prnHist(
  ncol = 10
)

# scaling normalization
prnHist(
  scale_log2r = TRUE,
  ncol = 10
)


}
