% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-histo.R
\name{proteoHist}
\alias{proteoHist}
\title{Visualizes histograms}
\usage{
proteoHist(id = c("pep_seq", "pep_seq_mod", "prot_acc", "gene"),
  col_select = NULL, scale_log2r = TRUE, show_curves = TRUE,
  show_vline = TRUE, new_fit = FALSE, df = NULL, filepath = NULL,
  filename = NULL, ...)
}
\arguments{
\item{id}{Character string to indicate the type of data. Peptide data will be
used at \code{id = pep_seq} or \code{pep_seq_mod}, and protein data at
\code{id = prot_acc} or \code{gene}.}

\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
The default key is \code{Select}. Samples corresponding to non-empty entries
under \code{col_select} will be included in the indicated analysis.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2-ratios} to the same
scale of standard deviation for all samples.}

\item{show_curves}{Logical; if TRUE, shows the fitted curves.}

\item{show_vline}{Logical; if TRUE, shows the vertical lines at \code{x = 0}.}

\item{new_fit}{Not currently used.}

\item{df}{The filename of input data. By default, it will be determined by the
value of \code{id}.}

\item{filepath}{The filepath to output results. By default, it will be
determined by the name of the current function \code{call} and the value of
\code{id}.}

\item{filename}{A representative filename to output images. By default, it
will be determined by the names of the current \code{call}. The images are
saved via \code{\link[ggplot2]{ggsave}} and the type of images will be
determined by the extension of filenames with the default being \code{.png}.}

\item{...}{Additional parameters for plotting: \cr \code{xmin}, the minimum x;
\cr \code{xmax}, the maximum x; \cr \code{x_breaks}, the breaks in x-axis;
\cr \code{binwidth}, the binwidth of \code{log2-ratios}; \cr \code{ncol},
the number of columns; \cr \code{width}, the width of plot; \cr
\code{height}, the height of plot.}
}
\value{
The histograms of \code{log2-ratios} under
 "\code{C:\\my_directory\\Histogram}".
}
\description{
\code{proteoHist} prepares the histogram visualization of \code{log2-ratios}
for proteins or peptides data.
}
\details{
In the histograms, the \code{log2-ratios} under each TMT channel are
color-coded by their contributing reporter-ion intensity.

The function matches the current \code{id} to those in the latest \code{calls}
to \code{\link{normPep}} or \code{\link{normPrn}}.  For example, if
\code{pep_seq} was used in \code{\link{normPep()}}, the current \code{id =
pep_seq_mod} will be matched to \code{id = pep_seq}.
}
\examples{
# no scaling normalization
proteoHist(
  id = pep_seq_mod,
  scale_log2r = FALSE,
  show_curves = TRUE,
  show_vline = TRUE,
  xmin = -1,
  xmax = 1,
  ncol = 4
)

# scaling normalization
proteoHist(
  id = pep_seq_mod,
  col_select = Select,
  scale_log2r = TRUE,
  show_curves = TRUE,
  show_vline = TRUE,
  ncol = 4
)

\dontrun{
proteoHist(
  id = pep_seq_mod,
  col_select = "a_non_existed_column_key"
)
}
}
