% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-hm.R
\name{proteoHM}
\alias{proteoHM}
\alias{pepHM}
\alias{prnHM}
\title{Visualization of heat maps}
\usage{
proteoHM(id = gene, col_select = NULL, col_benchmark = NULL,
  scale_log2r = TRUE, impute_na = FALSE, complete_cases = FALSE,
  df = NULL, filepath = NULL, filename = NULL, xmin = -1,
  xmax = 1, x_margin = 0.1, annot_cols = NULL,
  annot_colnames = NULL, ...)

pepHM(...)

prnHM(...)
}
\arguments{
\item{id}{Character string to indicate the type of data. Peptide data will be
used at \code{id = pep_seq} or \code{pep_seq_mod}, and protein data at
\code{id = prot_acc} or \code{gene}.}

\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under the column key will be
included in the indicated analysis. At the NULL default, the column key will
be \code{Select}.}

\item{col_benchmark}{Not used.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation for all samples.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used.}

\item{df}{The name of input data file. By default, it will be determined by
the value of \code{id}.}

\item{filepath}{A file path to output results. By default, it will be
determined by the name of the calling \code{function} and the value of
\code{id} in the \code{call}.}

\item{filename}{A representative file name to output image(s). By default, it
will be determined by the name of the current \code{call}. The image(s) are
saved via \code{\link[ggplot2]{ggsave}} where the image type will be
determined by the extension of the file name. A \code{.png} format will be
used at default or an unrecognized file extension.}

\item{annot_cols}{A character vector of column names in \code{expt_smry.xlsx}.
Values under the selected columns will be used to color-code sample IDs on
the top of the indicated plot.}

\item{annot_colnames}{A character vector of replacement name(s) to
\code{annot_cols}.}

\item{...}{More parameters for plotting: \cr \code{xmin}, the minimum \eqn{x}
at a log2 scale; the default is -1 \cr \code{xmax}, the maximum \eqn{x} at a
log2 scale; the default is +1 \cr \code{x_margin}, the margin in heat
scales; the default is 0.1 \cr \code{width}, the width of plot \cr
\code{height}, the height of plot \cr additional arguments for
\code{\link[pheatmap]{pheatmap}}.}
}
\value{
Heat maps.
}
\description{
\code{proteoHM} visualizes the heat maps of protein or peptide \code{log2FC}.

\code{pepHM} is a wrapper function of \code{\link{proteoHM}} for peptide data

\code{prnHM} is a wrapper function of \code{\link{proteoHM}} for protein data
}
\details{
Data columns with complete missing values will be removed prior to
hierarchical column clustering. Data rows without non-missing pairs will
result in NA distances in inter-row dissimilarities
(\code{\link[stats]{dist}}). At \code{complet_cases = TRUE}, the data subset
that are complete without missing values will be used. At \code{impute_na =
TRUE}, all data rows will be used with NA imputation (\code{\link{imputeNA}}).
At the default of \code{complet_cases = FALSE} and \code{impute_na = FALSE},
NA distances will be arbitrarily replaced with the mean value in the
row-disance matrix for hierarchical row clustering.

To avoid memory failure, row aggregation using the \code{kmeans_k} option
(\code{\link[pheatmap]{pheatmap}}) may be considered for large data sets.

The function matches the current \code{id} to those in the latest \code{calls}
to \code{\link{normPep}} or \code{\link{normPrn}}.  For example, if
\code{pep_seq} was used in \code{\link{normPep}}, the current \code{id =
pep_seq_mod} will be matched to \code{id = pep_seq}.
}
\examples{
prnHM(
 scale_log2r = TRUE,
 xmin = -1,
 xmax = 1,
 x_margin = 0.1,
 annot_cols = c("Peptide_Yield", "TMT_Set", "Group"),
 annot_colnames = c("Peptide_Yield", "Batch", "Phenotype"),
 cluster_rows = TRUE,
 cutree_rows = 6,
 show_rownames = FALSE,
 show_colnames = TRUE,
 fontsize_row = 3,
 cellwidth = 14,
 width = 10,
 height = 12
)

prnHM(
 scale_log2r = TRUE,
 annot_cols = c("Group"),
 cluster_rows = TRUE,
 clustering_distance_rows  = "maximum",
 cutree_rows = 6,
 show_rownames = FALSE,
 show_colnames = TRUE,
 fontsize_row = 3,
 cellwidth = 14,
 width = 10,
 height = 12
)

prnHM(
 scale_log2r = TRUE,
 annot_cols = c("Group"),
 cluster_rows = FALSE, # no row clustering
 clustering_distance_rows  = "maximum",
 cutree_rows = 6, # will be overruled at 'cluster_rows = FALSE'
 show_rownames = FALSE,
 show_colnames = TRUE,
 fontsize_row = 3,
 cellwidth = 14,
 width = 10,
 height = 12
)

\dontrun{
}
}
