% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-mds.R
\name{proteoMDS}
\alias{proteoMDS}
\title{Visualizes MDS plots}
\usage{
proteoMDS(id = gene, col_select = NULL, col_group = NULL,
  col_color = NULL, col_fill = NULL, col_shape = NULL,
  col_size = NULL, col_alpha = NULL, scale_log2r = TRUE,
  adjEucDist = FALSE, classical = TRUE, show_ids = TRUE,
  annot_cols = NULL, df = NULL, filepath = NULL, filename = NULL,
  ...)
}
\arguments{
\item{id}{Character string to indicate the type of data. Peptide data will be
used at column \code{id = "pep_seq"} or \code{"pep_seq_mod"}, and protein
data at \code{id = "prot_acc"} or \code{"gene"}.}

\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
The default key is \code{Select}. Samples corresponding to non-empty entries
under \code{col_select} will be included in the indicated analysis.}

\item{col_group}{Not currently used.}

\item{col_color}{Character string to a column key in \code{expt_smry.xlsx}.
The default key is \code{Color}. Values under which will be used for the
\code{color} aesthetics in \code{ggplot2}.}

\item{col_fill}{Character string to a column key in \code{expt_smry.xlsx}.
The default key is \code{Fill}. Values under which will be used for the
\code{fill} aesthetics in \code{ggplot2}.}

\item{col_shape}{Character string to a column key in \code{expt_smry.xlsx}.
The default key is \code{Shape}. Values under which will be used for the
\code{shape} aesthetics in \code{ggplot2}.}

\item{col_size}{Character string to a column key in \code{expt_smry.xlsx}.
The default key is \code{Size}. Values under which will be used for the
\code{size} aesthetics in \code{ggplot2}.}

\item{col_alpha}{Character string to a column key in \code{expt_smry.xlsx}.
The default key is \code{Alpha}. Values under which will be used for the
\code{alpha} aesthetics in \code{ggplot2}.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2-ratios} to the same
scale of standard deviation for all samples.}

\item{adjEucDist}{Logical; if TRUE, adjusts the inter-plex Euclidean distance
by \eqn{1/sqrt(2)}. The option \code{adjEucDist = TRUE} may be suitable when
\code{reference samples} in each TMT plex undergo approximately the same
sample handling process as the rest of the samples. For instance,
\code{reference samples} were split at the levels of protein lysates.
Typically, \code{adjEucDist = FALSE} if \code{reference samples} were split
near the end of a sample handling process, for instance, at the stages
immediately before or after TMT labeling.}

\item{classical}{Logical; performs metric MDS at TRUE and non-metric MDS at
FALSE.}

\item{show_ids}{Logical; if TRUE, shows the sample IDs in \code{MDS/PCA}
plots.}

\item{annot_cols}{Not currently used.}

\item{df}{The filename of input data. By default, it will be determined by the
value of \code{id}.}

\item{filepath}{The filepath to output results. By default, it will be
determined by the name of the current function \code{call} and the value of
\code{id}.}

\item{filename}{A representative filename to output images. By default, it
will be determined by the names of the current \code{call}.}

\item{...}{Parameters inherited from \code{ggsave}: \cr \code{width}, the
width of plot; \cr \code{height}, the height of plot \cr \code{...}}
}
\value{
MDS plots.
}
\description{
\code{proteoMDS} visualizes the results from multidimensional scaling (MDS).
}
\details{
An Euclidean distance matrix of \code{log2-ratios} is returned by
\code{\link[base]{dist}}, followed by a metric (\code{\link[stats]{cmdscale}})
or non-metric (\code{\link[MASS]{isoMDS}}) MDS. The default is metric MDS with
the input dissimilarities being euclidean distances.
}
