% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-mds.R
\name{proteoMDS}
\alias{proteoMDS}
\alias{pepMDS}
\alias{prnMDS}
\title{Visualization of MDS plots}
\usage{
proteoMDS(id = gene, col_select = NULL, col_group = NULL,
  col_color = NULL, col_fill = NULL, col_shape = NULL,
  col_size = NULL, col_alpha = NULL, scale_log2r = TRUE,
  adjEucDist = FALSE, classical = TRUE, show_ids = TRUE,
  annot_cols = NULL, df = NULL, filepath = NULL, filename = NULL,
  ...)

pepMDS(...)

prnMDS(...)
}
\arguments{
\item{id}{Character string to indicate the type of data. Peptide data will be
used at \code{id = pep_seq} or \code{pep_seq_mod}, and protein data at
\code{id = prot_acc} or \code{gene}.}

\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under the column key will be
included in the indicated analysis. At the NULL default, the column key will
be \code{Select}.}

\item{col_group}{Not used.}

\item{col_color}{Character string to a column key in \code{expt_smry.xlsx}.
Values under which will be used for the \code{color} aesthetics in plots. At
the NULL default, the column key \code{Color} will be used.}

\item{col_fill}{Character string to a column key in \code{expt_smry.xlsx}.
Values under which will be used for the \code{fill} aesthetics in plots. At
the NULL default, the column key \code{Fill} will be used.}

\item{col_shape}{Character string to a column key in \code{expt_smry.xlsx}.
Values under which will be used for the \code{shape} aesthetics in plots. At
the NULL default, the column key \code{Shape} will be used.}

\item{col_size}{Character string to a column key in \code{expt_smry.xlsx}.
Values under which will be used for the \code{size} aesthetics in plots. At
the NULL default, the column key \code{Size} will be used.}

\item{col_alpha}{Character string to a column key in \code{expt_smry.xlsx}.
Values under which will be used for the \code{alpha} (transparency)
aesthetics in plots. At the NULL default, the column key \code{Alpha} will
be used.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation for all samples.}

\item{adjEucDist}{Logical; if TRUE, adjusts the inter-plex Euclidean distance
by \eqn{1/sqrt(2)}. The option \code{adjEucDist = TRUE} may be suitable when
\code{reference samples} from each TMT plex undergo approximately the same
sample handling process as the samples of interest. For instance,
\code{reference samples} were split at the levels of protein lysates.
Typically, \code{adjEucDist = FALSE} if \code{reference samples} were split
near the end of a sample handling process, for instance, at the stages
immediately before or after TMT labeling.}

\item{classical}{Logical; performs metric MDS at TRUE and non-metric MDS at
FALSE.}

\item{show_ids}{Logical; if TRUE, shows the sample IDs in \code{MDS/PCA}
plots.}

\item{annot_cols}{Not used.}

\item{df}{The name of input data file. By default, it will be determined by
the value of \code{id}.}

\item{filepath}{A file path to output results. By default, it will be
determined by the name of the calling \code{function} and the value of
\code{id} in the \code{call}.}

\item{filename}{A representative file name to output image(s). By default, it
will be determined by the name of the current \code{call}. The image(s) are
saved via \code{\link[ggplot2]{ggsave}} where the image type will be
determined by the extension of the file name. A \code{.png} format will be
used at default or an unrecognized file extension.}

\item{...}{Parameters for \code{ggsave}: \cr \code{width}, the width of plot;
\cr \code{height}, the height of plot \cr \code{...}}
}
\value{
MDS plots.
}
\description{
\code{proteoMDS} visualizes the results from multidimensional scaling (MDS).

\code{pepMDS} is a wrapper of \code{\link{proteoMDS}} for peptide data

\code{prnMDS} is a wrapper of \code{\link{proteoMDS}} for protein data
}
\details{
An Euclidean distance matrix of \code{log2FC} is returned by
\code{\link[base]{dist}}, followed by a metric (\code{\link[stats]{cmdscale}})
or non-metric (\code{\link[MASS]{isoMDS}}) MDS. The default is metric MDS with
the input dissimilarities being euclidean distances.

The function matches the current \code{id} to those in the latest \code{calls}
to \code{\link{normPep}} or \code{\link{normPrn}}.  For example, if
\code{pep_seq} was used in \code{\link{normPep()}}, the current \code{id =
pep_seq_mod} will be matched to \code{id = pep_seq}.
}
\examples{
pepMDS(
  scale_log2r = FALSE,
  col_select = Select
)

prnMDS(
  scale_log2r = TRUE,
  col_color = Color,
  col_shape = Shape,
  show_ids = TRUE,
)

\dontrun{
prnMDS(
  col_color = "column_key_not_existed",
  col_shape = "another_missing_column_key"
)
}

}
