% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-mds.R
\name{proteoMDS}
\alias{proteoMDS}
\title{MDS Plots}
\usage{
proteoMDS(id = gene, col_select = NULL, col_group = NULL,
  col_color = NULL, col_fill = NULL, col_shape = NULL,
  col_size = NULL, col_alpha = NULL, scale_log2r = FALSE,
  adjEucDist = FALSE, classical = TRUE, show_ids = TRUE,
  annot_cols = NULL, df = NULL, filepath = NULL, filename = NULL,
  ...)
}
\arguments{
\item{id}{Character string to indicate the type of data (a column ID). Peptide
data will be used at column \code{id = "pep_seq"} or \code{"pep_seq_mod"},
and protein data at \code{id = "prot_acc"} or \code{"gene"}.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2-ratios} to the same
scale of standard deviation for all samples.}

\item{adjEucDist}{Logical; if TRUE, adjusts the inter-plex Euclidean distance
by \eqn{1/sqrt(2)}. The option \code{adjEucDist = TRUE} may be suitable when
\code{reference samples} in each TMT plex undergo approximately the same
sample handling process as the rest of the samples. For instance,
\code{reference samples} were split at the levels of protein lysates.
Typically, \code{adjEucDist = FALSE} if \code{reference samples} were split
near the end of a sample handling process, for instance, at the stages
immediately before or after TMT labeling.}

\item{classical}{Logical; performs metric MDS at TRUE and non-metric MDS at
FALSE.}

\item{show_ids}{Logical; if TRUE, shows the sample IDs in MDS and PCA plots.}

\item{annot_cols}{Column names available in \code{expt_smry.csv}. Values
under the selected columns will be used to color-code sample IDs on the top
of \code{EucDist} plots.}

\item{df}{The filename of input data. By default, it will be determined by the
value of \code{id}.}

\item{filepath}{The filepath to output results. By default, it will be
determined by the name of the current function \code{call} and the value of
\code{id}.}

\item{filename}{A representative filename to output images. By default, it
will be determined by the names of the current \code{call}.}

\item{...}{More parameters for plotting:

 \code{mds_color}, character string to a column ID in \code{Label
 scheme.csv}; values under which will be used to color-code data points in
 \code{MDS/PCA} plots.

 \code{mds_shape}, character string to a column ID in \code{Label
 scheme.csv}; values under which will be used to shape-code data points in
 \code{MDS/PCA} plots.

 \code{mds_alpha}, character string to a column ID in \code{Label
 scheme.csv}; values under which will be used to control the transparency of
 data points in \code{MDS/PCA} plots.}
}
\value{
The plots of MDS, PCA and Euclidean distance.
}
\description{
\code{proteoMDS} visualizes the results from multidimensional scaling (MDS)
and principal component analysis (PCA).
}
\details{
An Euclidean distance matrix of \code{log2-ratios} is returned by
\code{\link[base]{dist}}, followed by a metric (\code{\link[stats]{cmdscale}})
or non-metric (\code{\link[MASS]{isoMDS}}) MDS. The default is metric MDS with
the input dissimilarities being euclidean distances.

\code{log2-ratios} are used in PCA (\code{\link[stats]{prcomp}}).
}
\examples{
head(label_scheme)

proteoMDS(
  id = pep_seq_mod,
  scale_log2r = TRUE,
  col_color = Color,
  col_shape = Shape,
  show_ids = TRUE,
)

proteoMDS(
  id = pep_seq_mod,
  scale_log2r = TRUE,

  # parameters for MDS
  adjEucDist = FALSE,

  annot_cols = c("Peptide_Yield", "TMT_Set", "MDS_Group"),

  display_numbers = TRUE,
  number_color = "grey30",
  number_format = "\%.2f",

  clustering_distance_rows = "euclidean",
  clustering_distance_cols = "euclidean",

  fontsize = 16,
  fontsize_row = 20,
  fontsize_col = 20,
  fontsize_number = 16,

  cluster_rows = TRUE,
  show_rownames = TRUE,
  show_colnames = TRUE,
  border_color = "grey60",
  cellwidth = 50,
  cellheight = 50
)

\dontrun{
proteoMDS(
  id = pep_seq_mod,
  col_color = "column_key_not_in_label_scheme", 
  col_shape = "also_a_missing_column_key", 
  annot_cols = c("bad_column_key", "yet_another_bad_column_key")
)
}
}
