% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-nmf.R
\name{proteoNMF}
\alias{proteoNMF}
\alias{anal_pepNMF}
\alias{anal_prnNMF}
\alias{plot_prnNMFCon}
\alias{plot_prnNMFCoef}
\alias{plot_metaNMF}
\title{NMF analysis}
\usage{
proteoNMF(id = c("pep_seq", "pep_seq_mod", "prot_acc", "gene"),
  col_select = NULL, col_group = NULL, scale_log2r = TRUE,
  impute_na = TRUE, complete_cases = FALSE, df = NULL,
  filepath = NULL, filename = NULL, task = "anal", r = NULL,
  nrun = 200, ...)

anal_pepNMF(...)

anal_prnNMF(...)

plot_prnNMFCon(annot_cols = NULL, annot_colnames = NULL, ...)

plot_prnNMFCoef(annot_cols = NULL, annot_colnames = NULL, ...)

plot_metaNMF(annot_cols = NULL, annot_colnames = NULL, ...)
}
\arguments{
\item{id}{Character string to indicate the type of data. Peptide data will be
used at \code{id = pep_seq} or \code{pep_seq_mod}, and protein data at
\code{id = prot_acc} or \code{gene}.}

\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_select} will be
included in the indicated analysis. At the NULL default, the column key
\code{Select} will be used.}

\item{col_group}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_group} will be
used for sample grouping in the indicated analysis. At the NULL default, the
column key \code{Group} will be used.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation for all samples.}

\item{impute_na}{Logical; if TRUE, data will the imputation of missing values
will be used.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used.}

\item{df}{The name of input data file. By default, it will be determined by
the value of \code{id}.}

\item{filepath}{Use system default.}

\item{filename}{Use system default.}

\item{task}{Character string; a switch for different tasks in a functional
factory.}

\item{r}{Numeric vector; the factorization rank(s) in
(\code{\link[NMF]{nmf}}).}

\item{nrun}{Numeric; the number of runs to perform (\code{\link[NMF]{nmf}}).
The default is c(4:8).}

\item{...}{In \code{anal_} functions: additional arguments inherited from
\code{\link[NMF]{nmf}}; in \code{plot_} functions: additional arguments
inherited from \code{proteoEucDist}.}

\item{annot_cols}{A character vector of column names in \code{expt_smry.xlsx}.
Values under the selected columns will be used to color-code sample IDs on the
top of a \code{EucDist} plot.}

\item{annot_colnames}{A character vector of replacement name(s) to
\code{annot_cols}.}
}
\value{
NMF classification and visualization of \code{log2FC}.
}
\description{
Analyzes and visualizes the NMF clustering of peptide or protein \code{log2FC}

\code{anal_pepNMF} is a wrapper function of \code{\link{proteoNMF}} for the
NMF analysis of peptide data

\code{anal_prnNMF} is a wrapper function of \code{\link{proteoNMF}} for the
NMF analysis of protein data

\code{plot_prnNMFCon} is a wrapper function of \code{\link{proteoNMF}} for the
visualization of the consensus heat map of protein data

\code{plot_prnNMFCoef} is a wrapper function of \code{\link{proteoNMF}} for the
visualization of the coefficient heat map of protein data

\code{plot_metaNMF} is a wrapper function of \code{\link{proteoNMF}} for the
visualization of the metagene heat maps of protein data
}
\details{
The option of \code{complete_cases} will be forced to \code{TRUE} at
\code{impute_na = FALSE}
}
\examples{
library(NMF)

# peptide data NMF at single r(ank)
anal_pepNMF(
  scale_log2r = TRUE,
  col_group = Group, # optional a priori knowledge of sample groups
  r = 6,
  nrun = 200
)

# protein data NMF at multiple ranks
anal_prnNMF(
  impute_na = FALSE, 
  scale_log2r = TRUE, 
  col_group = Group,
  r = c(5:8), 
  nrun = 200
)


# protein consensus heat maps at specific ranks
plot_prnNMFCon(
  r = c(3, 5), 
  annot_cols = c("Color", "Alpha", "Shape"), 
  annot_colnames = c("Lab", "Batch", "WHIM"), 
  width = 10, 
  height = 10
)

# protein consensus heat maps at all ranks
plot_prnNMFCon(
  impute_na = FALSE, 
  annot_cols = c("Color", "Alpha", "Shape"), 
  annot_colnames = c("Lab", "Batch", "WHIM"), 
  width = 10, 
  height = 10
)


# protein coefficient heat maps at all ranks
plot_prnNMFCoef(
  annot_cols = c("Color", "Alpha", "Shape"), 
  annot_colnames = c("Lab", "Batch", "WHIM"), 
  width = 10, 
  height = 10
)

# metagenes heat maps at all ranks; additional arguments are from `pheatmap`
plot_metaNMF(
  annot_cols = c("Color", "Alpha", "Shape"),
  annot_colnames = c("Lab", "Batch", "WHIM"),

  fontsize = 8,
  fontsize_col = 5
)

}
