% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-nmf.R
\name{proteoNMF}
\alias{proteoNMF}
\title{NMF analysis}
\usage{
proteoNMF(id = c("pep_seq", "pep_seq_mod", "prot_acc", "gene"),
  col_select = NULL, scale_log2r = TRUE, impute_na = TRUE,
  complete_cases = FALSE, df = NULL, filepath = NULL,
  filename = NULL, anal_type = "NMF", task = "anal", r = 4,
  nrun = 200, ...)
}
\arguments{
\item{id}{Character string to indicate the type of data. Peptide data will be
used at \code{id = pep_seq} or \code{pep_seq_mod}, and protein data at
\code{id = prot_acc} or \code{gene}.}

\item{col_select}{Character string to a column key in \code{expt_smry.xlsx}.
The default key is \code{Select}. Samples corresponding to non-empty entries
under \code{col_select} will be included in the indicated analysis.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2-ratios} to the same
scale of standard deviation for all samples.}

\item{impute_na}{Logical; if TRUE, imputes missing values.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used for visualization.}

\item{df}{The filename of input data. By default, it will be determined by the
value of \code{id}.}

\item{filepath}{The filepath to output results. By default, it will be
determined by the names of the current functional \code{call}.}

\item{filename}{A representative filename to output images. By default, it
will be determined by the names of the current \code{call}.}

\item{r}{Numeric; the factorization rank (\code{\link[NMF]{nmf}}).}

\item{nrun}{Numeric; the number of runs to perform (\code{\link[NMF]{nmf}}).}

\item{...}{Additional parameters for plotting: \cr \code{xmin}, the minimum x; \cr
\code{xmax}, the maximum x; \cr \code{x_margin}, the margin in heat
scales;\cr \code{annot_cols}, the column keys in \code{expt_smry.xlsx} for
use in the color coding of samples;\cr \code{width_}, the width of plot; \cr
\code{height_}, the height of plot; \cr additional arguments inherited from
\code{\link[pheatmap]{pheatmap}}.}
}
\value{
NMF plots.
}
\description{
Analyzes and visualizes the NMF clustering of peptide or protein
\code{log2-ratios}
}
\details{
The option of \code{complete_cases} will be forced to \code{TRUE} at
\code{impute_na = FALSE}
}
