% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ.R
\docType{package}
\name{proteoQ}
\alias{proteoQ-package}
\alias{proteoQ}
\title{proteoQ: A package for processing mass spectrometry data.}
\description{
The proteoQ provides four categories of functions in data preprocessing, (2)
annotation, (3) quality assurance assessments and (4) informatic analysis.
The companion package, mzion, performs database searches.
}
\section{Utilities in data preprocessing}{
 normPSM -> (purgePSM) -> PSM2Pep ->
  mergePep -> standPep -> (purgePep) -> standPrn
}

\section{Utilities in custom database preparation}{
 prepGO, prepMSig,
  prepString, Uni2Entrez, Ref2Entrez
}

\section{Utilities in data QA and QC}{
 pepHist, prnHist, pepCorr_logFC,
  pepCorr_logInt, prnCorr_logFC, prnCorr_logInt,
}

\section{Utilities in informatic analysis}{
 pepMDS, prnMDS, pepPCA, prnPCA,
  pepLDA, prnLDA, pepEucDist, prnEucDist, pepHM, prnHM, anal_pepNMF,
  anal_prnNMF, plot_pepNMFCon, plot_prnNMFCon, plot_pepNMFCoef,
  plot_prnNMFCoef, plot_metaNMF, anal_prnTrend, plot_prnTrend, pepSig,
  prnSig, pepVol, prnVol, prnGSPA, prnGSPAHM, gspaMap, anal_prnString,
  pepImp, prnImp, prnGSVA, prnGSEA, cluego
}

\author{
\strong{Maintainer}: Qiang Zhang \email{qzhang503@gmail.com} (\href{https://orcid.org/0000-0001-7046-4378}{ORCID})

}
