% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ.R
\docType{package}
\name{proteoQ}
\alias{proteoQ}
\alias{proteoQ-package}
\title{proteoQ: A package for processing mass spectrometry data using tandam mass tags 
(\url{https://en.wikipedia.org/wiki/Tandem_mass_tag}).}
\description{
The proteoQ package provides three categories of functions in data preprocessing, 
  quality assurance assessments and informatic analysis.
}
\section{proteoQ functions in data preprocessing}{
 normPSM, normPep, normPrn
}

\section{proteoQ functions in data QA}{
 pepHist, prnHist, pepCorr, prnCorr,
  pepMA, prnMA
}

\section{proteoQ functions in informatic analysis}{
 pepMDS, prnMDS, pepPCA,
  prnPCA, pepEucDist, prnEucDist, pepHM, prnHM, pepNMF, prnNMF, prnTrend,
  pepSig, prnSig, prnGAGE, prnGSVA, pepVol, prnVol, gsvaMap, gageMap
}

