% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-sigtests.R
\name{proteoSigtest}
\alias{proteoSigtest}
\alias{pepSig}
\alias{prnSig}
\title{Significance tests}
\usage{
proteoSigtest(df = NULL, id = gene, scale_log2r = TRUE,
  filepath = NULL, filename = NULL, impute_na = TRUE,
  complete_cases = FALSE, method = "limma", var_cutoff = 0.001,
  pval_cutoff = 1, logFC_cutoff = log2(1), ...)

pepSig(...)

prnSig(...)
}
\arguments{
\item{df}{The name of input data file. By default, it will be determined
automatically by the value of \code{id}.}

\item{id}{Character string to indicate the type of data. The value will be
determined automatically. Peptide data will be used at \code{id = pep_seq}
or \code{pep_seq_mod}, and protein data will be used at \code{id = prot_acc}
or \code{gene}.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{filename}{A file name to output results. The default is
\code{Peptide_pVals.txt} for peptides and \code{Protein_pVals} for proteins.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used. The default is FALSE.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{method}{Character string; the method of linear modeling. The default is
\code{limma}. At \code{method = lm}, the \code{lm()} in base R will be used
for models without random effects and the \code{\link[lmerTest]{lmer}} will
be used for models with random effects.}

\item{var_cutoff}{Numeric; the cut-off in the variances of \code{log2FC}.
Entries with variances smaller than the threshold will be removed from
linear modeling. The default is 1E-3.}

\item{pval_cutoff}{Numeric; the cut-off in significance \code{pVal}. Entries
with \code{pVals} smaller than the threshold will be removed from multiple
test corrections. The default is \code{1} without pVal cut-offs.}

\item{logFC_cutoff}{Numeric; the cut-off in \code{log2FC}. Entries with
\code{log2FC} smaller than the threshold will be removed from multiple test
corrections. The default is at \code{log2(1)}  without fold change cut-offs.}

\item{...}{User-defined formulae for linear modeling. The syntax starts with a
tilde, followed by the name of an available column key in
\code{expt_smry.xlsx} and square brackets. The contrast groups are then
quoted with multiple contrast groups separated by commas. Additive random
effects are indicated by \code{+ (1|col_key_1) + (1|col_key_2)}... \cr \cr
\code{filter_}: Logical expression(s) for the row filtration of data; also
see \code{\link{normPSM}}.}
}
\value{
The primary output is
 \code{~\\dat_dir\\Peptide\\Model\\Peptide_pVals.txt} for peptide data or
 \code{~\\dat_dir\\Protein\\Model\\Protein_pVals.txt} for protein data. \cr
 \cr At \code{impute_na = TRUE}, the corresponding outputs are
 \code{Peptide_impNA_pvals.txt} or \code{Protein_impNA_pvals.txt}.
}
\description{
\code{proteoSigtest} performs significance tests aganist peptide or protein
\code{log2FC}. Users should avoid call the method directly, but instead use
the following wrappers.

\code{pepSig} is a wrapper of \code{\link{proteoSigtest}} for the
significance tests of peptide data

\code{prnSig} is a wrapper of \code{\link{proteoSigtest}} for the significance
tests of protein data
}
\details{
In general, special characters of \code{+} or \code{-} should be avoided from
contrast terms. Occasionally, such as in biological studies, it may be
convenient to use \code{A+B} to denote a condition of combined treatment of
\code{A} and \code{B} . In the case, one can put the term(s) containing
\code{+} or \code{-} into a pair of pointy brackets. The syntax in the
following hypothetical example will compare the effects of \code{A}, \code{B},
\code{A+B} and the average of \code{A} and \code{B} to control \code{C}:

\code{prnSig(fml = ~ Term["A - C", "B - C", "<A + B> - C", "(A + B)/2 - C"])}

Note that \code{<A + B>} stands for one sample and \code{(A + B)} has two
samples in it.
}
\examples{
# ===================================
# Significance tests
# ===================================

## !!!require the brief working example in `?load_expts`

# global option
scale_log2r <- TRUE

# peptide significance tests
# (`Term`, `Term_2`, `Term_3` are column keys in `expt_smry.xlsx`)
pepSig(
  impute_na = FALSE, 
  W2_bat = ~ Term["(W2.BI.TMT2-W2.BI.TMT1)", 
                  "(W2.JHU.TMT2-W2.JHU.TMT1)", 
                  "(W2.PNNL.TMT2-W2.PNNL.TMT1)"], # batch effects
  W2_loc = ~ Term_2["W2.BI-W2.JHU", 
                    "W2.BI-W2.PNNL", 
                    "W2.JHU-W2.PNNL"], # location effects
  W16_vs_W2 = ~ Term_3["W16-W2"], 
)

# protein significance tests
# (formulae matched to `pepSig` by default)
prnSig(impute_na = FALSE)

# averaged batch effect
prnSig(
  impute_na = FALSE,
  W2_loc_2 = ~ Term["(W2.BI.TMT2+W2.BI.TMT1)/2 - (W2.JHU.TMT2+W2.JHU.TMT1)/2"], 
)


# impute NA
# (suggested for models with random effects)
pepImp(m = 2, maxit = 2)
prnImp(m = 5, maxit = 5)

# single random effect
prnSig(
  impute_na = TRUE,
  W2_vs_W16_fix = ~ Term_3["W16-W2"], # fixed
  W2_vs_W16_mix = ~ Term_3["W16-W2"] + (1|TMT_Set), # one fixed and one random
)

# one to multiple random effect
prnSig(
  impute_na = TRUE,
  method = lm,
  W2_vs_W16_fix = ~ Term_3["W16-W2"], # one fixed
  W2_vs_W16_mix = ~ Term_3["W16-W2"] + (1|TMT_Set), # one fixed and one random
  W2_vs_W16_mix_2 = ~ Term_3["W16-W2"] + (1|TMT_Set) + (1|Color), # one fixed and two randoms
)

}
\seealso{
\code{\link{load_expts}} for a reduced working example in data
 normalization \cr \code{\link{normPSM}} for extended examples in PSM data
 normalization \cr \code{\link{PSM2Pep}} for extended examples in PSM to
 peptide summarization \cr \code{\link{mergePep}} for extended examples in
 peptide data merging \cr \code{\link{standPep}} for extended examples in
 peptide data normalization \cr \code{\link{Pep2Prn}} for extended examples
 in peptide to protein summarization \cr \code{\link{standPrn}} for extended
 examples in protein data normalization. \cr \code{\link{pepHist}} and
 \code{\link{prnHist}} for extended examples in histogram visualization. \cr
 \code{\link{purgePSM}} and \code{\link{purgePep}} for extended examples in
 data purging \cr \code{\link{contain_str}}, \code{\link{contain_chars_in}},
 \code{\link{not_contain_str}}, \code{\link{not_contain_chars_in}},
 \code{\link{start_with_str}}, \code{\link{end_with_str}},
 \code{\link{start_with_chars_in}} and \code{\link{ends_with_chars_in}} for
 data subsetting by character strings \cr \code{\link{pepImp}} and
 \code{\link{prnImp}} for missing value imputation \cr \code{\link{pepSig}}
 and \code{\link{prnSig}} for significance tests \cr \code{\link{pepHM}} and
 \code{\link{prnHM}} for heat map visualization \cr \code{\link{pepMDS}} and
 \code{\link{prnMDS}} for MDS visualization \cr \code{\link{pepPCA}} and
 \code{\link{prnPcA}} for PCA visualization \cr
}
