% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-volcanos.R
\name{proteoVolcano}
\alias{proteoVolcano}
\title{Significance tests}
\usage{
proteoVolcano(id = "gene", anal_type = "Volcano", df = NULL,
  scale_log2r = TRUE, filepath = NULL, filename = NULL,
  impute_na = TRUE, adjP = FALSE, show_labels = TRUE,
  use_gagep = TRUE, pval_cutoff = 0.05, show_sig = NULL, ...)
}
\arguments{
\item{id}{The name of a unique id (see \code{\link[proteoQ]{MDS}}).}

\item{scale_log2r}{Logical; if TRUE, rescales \code{log2-ratios} to the same
scale of standard deviation for all samples.}
}
\value{
Images stored under the file folders that are associated to
  \code{id}, \code{anal_type}.
}
\description{
\code{proteoSigtest} produces heat maps.  It is a wrapper round the call
\code{info_anal(..., anal_type = "Heatmap")}.
}
\details{
reads the data from either "\code{~\\Direcotry\\Peptide\\Peptide All.txt}" at
\code{id = pep_seq_mod},
}
