% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-volcanos.R
\name{proteoVolcano}
\alias{proteoVolcano}
\alias{pepVol}
\alias{prnVol}
\alias{gspaMap}
\title{Volcano plot visualization}
\usage{
proteoVolcano(id = "gene", anal_type = "Volcano", df = NULL,
  scale_log2r = TRUE, filepath = NULL, filename = NULL,
  impute_na = TRUE, adjP = FALSE, show_labels = TRUE,
  pval_cutoff = 0.05, logFC_cutoff = log2(1.2), show_sig = "none",
  gset_nm = "go_sets", ...)

pepVol(...)

prnVol(...)

gspaMap(...)
}
\arguments{
\item{id}{Character string to indicate the type of data. Peptide data will be
used at \code{id = pep_seq} or \code{pep_seq_mod}, and protein data at
\code{id = prot_acc} or \code{gene}.}

\item{anal_type}{The type of analysis for method dispatch in functional
factories; Values include \code{anal_type = c("MDS", "PCA", "EucDist",
"MA", "Heatmap", "Histogram", "Corrplot", "Trend", "NMF", "Model", ...)}.
Most likely users will ignore the argument as its value is pre-set in
wrapper functions.}

\item{df}{The name of input data file. By default, it will be determined by
the value of \code{id}.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation for all samples.}

\item{filepath}{A file path to output results. By default, it will be
determined by the name of the calling \code{function} and the value of
\code{id} in the \code{call}.}

\item{filename}{A representative file name to output image(s). By default, it
will be determined by the name of the current \code{call}. The image(s) are
saved via \code{\link[ggplot2]{ggsave}} and the image type will be
automatically determined by the extension of the file name. A \code{.png}
format will be used at default or an unrecognized file extension.}

\item{impute_na}{Logical; if TRUE, data will the imputation of missing values
will be used.}

\item{adjP}{Logical; if TRUE, use Benjamini-Hochberg pVals for protein/peptide
entries in volcano plot(s).}

\item{show_labels}{Logical; if TRUE, shows the labels of top twenty entries.}

\item{pval_cutoff}{Numeric. \code{Gene sets} with enrichment \code{pVals} more
significant than the threshold will be used for volcano plot visualization.
The argument is not used in \code{prnVol} and \code{pepVol}.}

\item{logFC_cutoff}{Numeric. \code{Gene sets} with enrichment \code{log2FC}
greater than the threshold will be used for volcano plot visualization. The
cut-off is in a logarithmic base of 2, not in a linear scale. The argument
is not used in \code{prnVol} and \code{pepVol}.}

\item{show_sig}{Character string indicating the type of significance values
to be shown on volcano plots. The default is \code{"none"}. Additional
choices are from \code{c("pVal", "qVal")} where \code{pVal} or \code{qVal}
will be shown, respectively, in the facet grid of plots.}

\item{gset_nm}{Character vector containing the name(s) of gene sets for
enrichment analysis. The currently supported data sets are
\code{c("go_sets", "kegg_sets", "c2_msig")}.}

\item{...}{Additional parameters for plotting: \cr \code{xco}, the cut-off
lines of fold changes at position \code{x}; the default is at \eqn{-1.2} and
\eqn{+1.2}. \cr \code{yco}, the cut-off line of \code{pVal} at position
\code{y}; the default is \eqn{0.05}. \cr \code{width}, the width of plot;
\cr \code{height}, the height of plot.}
}
\description{
\code{proteoVolcano} provides the volcano plot visualization of
peptide/protein data or protein sub groups in the context of gene sets.

\code{pepVol} is a wrapper function of \code{\link{proteoVolcano}} for peptide data

\code{prnVol} is a wrapper function of \code{\link{proteoVolcano}} for protein data

\code{gspaMap} is a wrapper function of \code{\link{proteoVolcano}} for
mapping of protein data by gene sets.
}
\examples{
# protein volcano plot(s)
prnVol(
  scale_log2r = TRUE,
  show_labels = TRUE,
  xco = 1.2,
  yco = 0.01,
)

# protein volcano plot subsets by gene sets
gspaMap(
  scale_log2r = TRUE,
  show_labels = TRUE,
  pval_cutoff = 5E-3,
  logFC_cutoff = log2(1.2),
  # gset_nm = c("go_sets", "kegg_sets"),
  show_sig = p,
  yco = 0.01,
)

\dontrun{
}

}
