% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-volcanos.R
\name{proteoVolcano}
\alias{proteoVolcano}
\alias{pepVol}
\alias{prnVol}
\alias{gspaMap}
\title{Volcano plot visualization}
\usage{
proteoVolcano(id = "gene", anal_type = "Volcano", df = NULL,
  scale_log2r = TRUE, filepath = NULL, filename = NULL,
  fml_nms = NULL, impute_na = NULL, adjP = FALSE,
  show_labels = TRUE, pval_cutoff = 0.05, logFC_cutoff = log2(1.2),
  show_sig = "none", gset_nms = c("go_sets", "kegg_sets"), ...)

pepVol(...)

prnVol(...)

gspaMap(...)
}
\arguments{
\item{id}{Character string to indicate the type of data. The value will be
determined automatically. Peptide data will be used at \code{id = pep_seq}
or \code{pep_seq_mod}, and protein data will be used at \code{id = prot_acc}
or \code{gene}.}

\item{anal_type}{Character string; the type of analysis that are preset for
method dispatch in function factories. The value will be determined
automatically. Examplary values include \code{anal_type = c("PCA",
"Corrplot", "EucDist", "GSPA", "Heatmap", "Histogram", "MDS", "Model",
"NMF", "Purge", "Trend", ...)}.}

\item{df}{The name of input data file. By default, it will be determined
automatically after matching the data type with an \code{id} among
\code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}.}

\item{scale_log2r}{Not used for full volcano plots by \code{pepVol} or 
\code{prnVol}; matched to the value in the latest call to \code{\link{prnGSPA}} 
for \code{gspaMap}.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{filename}{A file name to output results. By default, it will be
determined automatically by the name of the calling function and the value
of id in the call.}

\item{fml_nms}{Character string or vector; the formula name(s). By default,
the formula(e) will match those used in \code{\link{pepSig}} or
\code{\link{prnSig}}.}

\item{impute_na}{Logical. At the NULL default, the TRUE or FALSE will match
the choice in \code{\link{pepSig}} for peptide and \code{\link{prnSig}} for
protein data.}

\item{adjP}{Logical; if TRUE, use Benjamini-Hochberg pVals in volcano plots.
The default is FALSE.}

\item{show_labels}{Logical; if TRUE, shows the labels of top twenty entries.
The default is TRUE.}

\item{pval_cutoff}{Numeric value or vector for uses with
\code{\link{gspaMap}}. \code{Gene sets} with enrichment \code{pVals} less
significant than the threshold will be excluded from volcano plot
visualization. The default signficance is 0.05 for all formulae matched to
or specified in argument \code{fml_nms}. Formula-specific threshold is
allowed by supplying a vector of cut-off values.}

\item{logFC_cutoff}{Numeric value or vector for uses with
\code{\link{gspaMap}}. \code{Gene sets} with absolute enrichment
\code{log2FC} less than the threshold will be excluded from volcano plot
visualization. The default magnitude is \code{log2(1.2) } for all formulae
matched to or specified in argument \code{fml_nms}. Formula-specific
threshold is allowed by supplying a vector of absolute values in
\code{log2FC}.}

\item{show_sig}{Character string indicating the type of significance values to
be shown with \code{\link{gspaMap}}. The default is \code{"none"}.
Additional choices are from \code{c("pVal", "qVal")} where \code{pVal} or
\code{qVal} will be shown, respectively, in the facet grid of the plots.}

\item{gset_nms}{Character vector containing the name(s) of gene sets for uses
with \code{\link{gspaMap}}. By default, the names will match those used in
\code{\link{prnGSPA}}.}

\item{...}{\code{filter_}: Logical expression(s) for the row filtration of
data; also see \code{\link{normPSM}}. \cr \cr Additional parameters for
plotting: \cr \code{xco}, the cut-off lines of fold changes at position
\code{x}; the default is at \eqn{-1.2} and \eqn{+1.2}. \cr \code{yco}, the
cut-off line of \code{pVal} at position \code{y}; the default is \eqn{0.05}.
\cr \code{width}, the width of plot; \cr \code{height}, the height of plot.
\cr \code{nrow}, the number of rows in a plot.}
}
\description{
\code{proteoVolcano} visualizes the volcano plots of peptide or protein data,
or protein subgroups under the same gene sets. Users should avoid call the
method directly, but instead use the following wrappers.

\code{pepVol} is a wrapper of \code{\link{proteoVolcano}} for peptide data

\code{prnVol} is a wrapper of \code{\link{proteoVolcano}} for protein data

\code{gspaMap} is a wrapper of \code{\link{proteoVolcano}} for
mapping of protein data by gene sets.
}
\examples{
# ===================================
# Volcano plots
# ===================================

## !!!require the brief working example in `?load_expts`

## global option
scale_log2r <- TRUE

## for all peptides or proteins
# all peptides
pepVol()

# all proteins
prnVol(
  show_labels = TRUE,
  xco = 1.2,
  yco = 0.01,
)

# kinases and prot_n_pep >= 2
prnVol(
  show_labels = TRUE,
  xco = 1.2,
  yco = 0.01,
  filter_prots_by_kin = exprs(kin_attr, prot_n_pep >= 2),
  filename = "kin_npep2.png"
)


## protein subgroups by gene sets
# prerequisite analysis of GSPA
prnGSPA(
  impute_na = FALSE,
  pval_cutoff = 5E-2,
  logFC_cutoff = log2(1.2),
  gspval_cutoff = 5E-2,
  gset_nms = c("go_sets", "kegg_sets"),
)

# filtered by proteins with two or more identifying peptides for visualization
gspaMap(
  pval_cutoff = 5E-3,
  logFC_cutoff = log2(1.2),
  gset_nms = c("go_sets"),
  show_sig = pVal,
  show_labels = TRUE,
  yco = 0.01,
  filter_prots_by_npep = exprs(prot_n_pep >= 2),
  # `filename`(s) will be automated, i.e., by gene-set names
)

# customized thresholds for the corresponding formulae in `pepSig` or `prnSig()`
# (may be suitable with the examplary differences in `W16_vs_W2` being much greater than `W2_bat`...)
gspaMap(
  fml_nms = c("W2_bat", "W2_loc", "W16_vs_W2"),
  pval_cutoff = c(5E-2, 5E-2, 1E-10),
  logFC_cutoff = log2(1.2),
  
  show_sig = pVal,
  show_labels = TRUE,
  yco = 0.05,
  filter_prots = exprs(prot_n_pep >= 2),
)

}
\seealso{
\code{\link{load_expts}} for a reduced working example in data normalization \cr

 \code{\link{normPSM}} for extended examples in PSM data normalization \cr
 \code{\link{PSM2Pep}} for extended examples in PSM to peptide summarization \cr 
 \code{\link{mergePep}} for extended examples in peptide data merging \cr 
 \code{\link{standPep}} for extended examples in peptide data normalization \cr
 \code{\link{Pep2Prn}} for extended examples in peptide to protein summarization \cr
 \code{\link{standPrn}} for extended examples in protein data normalization. \cr 
 \code{\link{purgePSM}} and \code{\link{purgePep}} for extended examples in data purging \cr
 \code{\link{pepHist}} and \code{\link{prnHist}} for extended examples in histogram visualization. \cr 
 \code{\link{extract_raws}} and \code{\link{extract_psm_raws}} for extracting MS file names \cr 
 
 \code{\link{contain_str}}, \code{\link{contain_chars_in}}, \code{\link{not_contain_str}}, 
 \code{\link{not_contain_chars_in}}, \code{\link{start_with_str}}, 
 \code{\link{end_with_str}}, \code{\link{start_with_chars_in}} and 
 \code{\link{ends_with_chars_in}} for data subsetting by character strings \cr 
 
 \code{\link{pepImp}} and \code{\link{prnImp}} for missing value imputation \cr 
 \code{\link{pepSig}} and \code{\link{prnSig}} for significance tests \cr 
 \code{\link{pepVol}} and \code{\link{prnVol}} for volcano plot visualization \cr 
 
 \code{\link{prnGSPA}} for gene set enrichment analysis by protein significance pVals \cr 
 \code{\link{gspaMap}} for mapping GSPA to volcano plot visualization \cr 
 \code{\link{prnGSPAHM}} for heat map and network visualization of GSPA results \cr 
 \code{\link{prnGSVA}} for gene set variance analysis \cr 
 \code{\link{prnGSEA}} for data preparation for online GSEA. \cr 
 
 \code{\link{pepMDS}} and \code{\link{prnMDS}} for MDS visualization \cr 
 \code{\link{pepPCA}} and \code{\link{prnPcA}} for PCA visualization \cr 
 \code{\link{pepHM}} and \code{\link{prnHM}} for heat map visualization \cr 
 \code{\link{pepCorr_logFC}}, \code{\link{prnCorr_logFC}}, \code{\link{pepCorr_logInt}} and 
 \code{\link{prnCorr_logInt}}  for correlation plots \cr 
 
 \code{\link{anal_prnTrend}} and \code{\link{plot_prnTrend}} for trend analysis and visualization \cr 
 \code{\link{anal_pepNMF}}, \code{\link{anal_prnNMF}}, \code{\link{plot_pepNMFCon}}, 
 \code{\link{plot_prnNMFCon}}, \code{\link{plot_pepNMFCoef}}, \code{\link{plot_prnNMFCoef}} and 
 \code{\link{plot_metaNMF}} for NMF analysis and visualization \cr 
 
 \code{\link{dl_stringdbs}} and \code{\link{anal_prnString}} for STRING-DB
}
