% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{PSM2Pep}
\alias{PSM2Pep}
\title{Interim peptide tables}
\usage{
PSM2Pep(method_psm_pep = c("median", "mean", "weighted.mean", "top.3"),
  ...)
}
\arguments{
\item{method_psm_pep}{Character string; the method to summarise the
\code{log2FC} and the \code{intensity} of \code{PSMs} by peptide entries.
The descriptive statistics includes \code{c("mean", "median", "top.3",
"weighted.mean")} with \code{median} being the default. The
\code{log10-intensity} of reporter ions at the \code{PSMs} levels will be
the weight when summarising \code{log2FC} with \code{"top.3"} or
\code{"weighted.mean"}.}

\item{...}{Not currently used.}
}
\value{
Tables under \code{PSM} folder for each TMT experiment and LC/MS
 series: \code{TMTset1_LCMSinj1_PSM_N.txt}, \code{TMTset1_LCMSinj2_PSM_N.txt}...
}
\description{
\code{PSM2Pep} summarises
\code{\href{https://www.ebi.ac.uk/pride/help/archive/search/tables}{PSMs}} to
peptides by individual TMT experiments and LC/MS series.
}
\details{
In general, fields other than \code{log2FC} and \code{intensity} are
summarized with median statistics. One exception is with \code{pep_expect} in
Mascot or \code{PEP} in MaxQuant where geometric mean is applied.
}
\examples{
# ===================================
# PSMs to peptides
# ===================================

## !!!require the brief working example in `?load_expts`

# default median statistics
PSM2Pep()

# mean statistics
PSM2Pep(method_psm_pep = mean)

}
\seealso{
\code{\link{load_expts}} for a reduced working example in data normalization \cr

 \code{\link{normPSM}} for extended examples in PSM data normalization \cr
 \code{\link{PSM2Pep}} for extended examples in PSM to peptide summarization \cr 
 \code{\link{mergePep}} for extended examples in peptide data merging \cr 
 \code{\link{standPep}} for extended examples in peptide data normalization \cr
 \code{\link{Pep2Prn}} for extended examples in peptide to protein summarization \cr
 \code{\link{standPrn}} for extended examples in protein data normalization. \cr 
 \code{\link{purgePSM}} and \code{\link{purgePep}} for extended examples in data purging \cr
 \code{\link{pepHist}} and \code{\link{prnHist}} for extended examples in histogram visualization. \cr 
 \code{\link{extract_raws}} and \code{\link{extract_psm_raws}} for extracting MS file names \cr 
 
 \code{\link{contain_str}}, \code{\link{contain_chars_in}}, \code{\link{not_contain_str}}, 
 \code{\link{not_contain_chars_in}}, \code{\link{start_with_str}}, 
 \code{\link{end_with_str}}, \code{\link{start_with_chars_in}} and 
 \code{\link{ends_with_chars_in}} for data subsetting by character strings \cr 
 
 \code{\link{pepImp}} and \code{\link{prnImp}} for missing value imputation \cr 
 \code{\link{pepSig}} and \code{\link{prnSig}} for significance tests \cr 
 \code{\link{pepVol}} and \code{\link{prnVol}} for volcano plot visualization \cr
}
