% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{PSM2Pep}
\alias{PSM2Pep}
\title{Interim peptide tables}
\usage{
PSM2Pep(
  method_psm_pep = c("median", "mean", "weighted_mean", "top_3_mean", "lfq_max",
    "lfq_top_2_sum", "lfq_top_3_sum", "lfq_all"),
  parallel = TRUE,
  ...
)
}
\arguments{
\item{method_psm_pep}{Character string; the method to summarize the
\code{log2FC} and the \code{intensity} of \code{PSMs} by peptide entries.
The descriptive statistics includes \code{c("mean", "median",
"weighted_mean", "top_3_mean", "lfq_max", "lfq_top_2_sum", "lfq_top_3_sum",
"lfq_all")} with \code{median} being the default for TMT and
\code{lfq_top_2_sum} for LFQ. The \code{log10-intensity} of reporter (or
LFQ) ions at the \code{PSMs} levels will be the weight when summarizing
\code{log2FC} with various \code{"top_n"} statistics or
\code{"weighted_mean"}.}

\item{parallel}{Logical; if TRUE, performs parallel computation. The default
is TRUE.}

\item{...}{\code{filter_}: Variable argument statements for the filtration of
data rows. See also \code{\link{normPSM}}.}
}
\value{
Tables under \code{Peptide} folder for each TMT experiment and LC/MS
 series: \code{TMTset1_LCMSinj1_Peptide_N.txt} etc.
}
\description{
\code{PSM2Pep} summarizes
\href{https://www.ebi.ac.uk/pride/help/archive/search/tables}{PSMs} to
peptides by individual TMT (or LFQ) experiments and LC/MS series.
}
\details{
In general, fields other than \code{log2FC} and \code{intensity} are
summarized with median statistics. One exception is with \code{pep_expect} in
Mascot or \code{PEP} in MaxQuant where geometric mean is applied.
}
\examples{
\donttest{
# ===================================
# PSMs to peptides
# ===================================

## !!!require the brief working example in `?load_expts`

# default median statistics
PSM2Pep()

# mean statistics
PSM2Pep(method_psm_pep = mean)  

## Not run: 
# cut-offs in precursor intensity
# (error if `pep_tot_int` are all NAs)
PSM2Pep(filter_ms1int = rlang::exprs(pep_tot_int >= 1E4)) 
## End(Not run)
}
}
\seealso{
\emph{Metadata} \cr 
 \code{\link{load_expts}} for metadata preparation and a reduced working example in data normalization \cr

 \emph{Data normalization} \cr 
 \code{\link{normPSM}} for extended examples in PSM data normalization \cr
 \code{\link{PSM2Pep}} for extended examples in PSM to peptide summarization \cr 
 \code{\link{mergePep}} for extended examples in peptide data merging \cr 
 \code{\link{standPep}} for extended examples in peptide data normalization \cr
 \code{\link{Pep2Prn}} for extended examples in peptide to protein summarization \cr
 \code{\link{standPrn}} for extended examples in protein data normalization. \cr 
 \code{\link{purgePSM}} and \code{\link{purgePep}} for extended examples in data purging \cr
 \code{\link{pepHist}} and \code{\link{prnHist}} for extended examples in histogram visualization. \cr 
 \code{\link{extract_raws}} and \code{\link{extract_psm_raws}} for extracting MS file names \cr 
 
 \emph{Variable arguments of `filter_...`} \cr 
 \code{\link{contain_str}}, \code{\link{contain_chars_in}}, \code{\link{not_contain_str}}, 
 \code{\link{not_contain_chars_in}}, \code{\link{start_with_str}}, 
 \code{\link{end_with_str}}, \code{\link{start_with_chars_in}} and 
 \code{\link{ends_with_chars_in}} for data subsetting by character strings \cr 
 
 \emph{Missing values} \cr 
 \code{\link{pepImp}} and \code{\link{prnImp}} for missing value imputation \cr 
 
 \emph{Informatics} \cr 
 \code{\link{pepSig}} and \code{\link{prnSig}} for significance tests \cr 
 \code{\link{pepVol}} and \code{\link{prnVol}} for volcano plot visualization \cr 
 \code{\link{prnGSPA}} for gene set enrichment analysis by protein significance pVals \cr 
 \code{\link{gspaMap}} for mapping GSPA to volcano plot visualization \cr 
 \code{\link{prnGSPAHM}} for heat map and network visualization of GSPA results \cr 
 \code{\link{prnGSVA}} for gene set variance analysis \cr 
 \code{\link{prnGSEA}} for data preparation for online GSEA. \cr 
 \code{\link{pepMDS}} and \code{\link{prnMDS}} for MDS visualization \cr 
 \code{\link{pepPCA}} and \code{\link{prnPCA}} for PCA visualization \cr 
 \code{\link{pepLDA}} and \code{\link{prnLDA}} for LDA visualization \cr 
 \code{\link{pepHM}} and \code{\link{prnHM}} for heat map visualization \cr 
 \code{\link{pepCorr_logFC}}, \code{\link{prnCorr_logFC}}, \code{\link{pepCorr_logInt}} and 
 \code{\link{prnCorr_logInt}}  for correlation plots \cr 
 \code{\link{anal_prnTrend}} and \code{\link{plot_prnTrend}} for trend analysis and visualization \cr 
 \code{\link{anal_pepNMF}}, \code{\link{anal_prnNMF}}, \code{\link{plot_pepNMFCon}}, 
 \code{\link{plot_prnNMFCon}}, \code{\link{plot_pepNMFCoef}}, \code{\link{plot_prnNMFCoef}} and 
 \code{\link{plot_metaNMF}} for NMF analysis and visualization \cr

\emph{Custom databases} \cr 
 \code{\link{Uni2Entrez}} for lookups between UniProt accessions and Entrez IDs \cr 
 \code{\link{Ref2Entrez}} for lookups among RefSeq accessions, gene names and Entrez IDs \cr 
 \code{\link{prepGO}} for \code{\href{http://current.geneontology.org/products/pages/downloads.html}{gene 
 ontology}} \cr 
 \code{\link{prepMSig}} for \href{https://data.broadinstitute.org/gsea-msigdb/msigdb/release/7.0/}{molecular 
 signatures} \cr 
 \code{\link{prepString}} and \code{\link{anal_prnString}} for STRING-DB \cr
 
 \emph{Column keys in PSM, peptide and protein outputs} \cr 
 system.file("extdata", "psm_keys.txt", package = "proteoQ") \cr
 system.file("extdata", "peptide_keys.txt", package = "proteoQ") \cr
 system.file("extdata", "protein_keys.txt", package = "proteoQ") \cr
}
\concept{custom database preparation}
