% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{psm_msplit}
\alias{psm_msplit}
\title{Parallel splits of PSM tables}
\usage{
psm_msplit(
  df = NULL,
  nm = NULL,
  fn_lookup = NULL,
  dat_dir = NULL,
  plot_rptr_int = TRUE,
  TMT_plex = 10L
)
}
\arguments{
\item{df}{A list of data frames.}

\item{nm}{Names of the \code{df}'s.}

\item{fn_lookup}{A lookup of file names.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{plot_rptr_int}{Logical; if TRUE, the distributions of reporter-ion
intensities will be plotted. The default is TRUE. The argument is also
applicable to the precursor intensity with MaxQuant LFQ.}

\item{TMT_plex}{Numeric; the multiplexity of TMT, i.e., 10, 11 etc.}
}
\description{
Parallel splits of PSM tables
}
