% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{psm_to_pep}
\alias{psm_to_pep}
\title{Helper of PSM2Pep}
\usage{
psm_to_pep(
  file,
  dat_dir,
  label_scheme_full,
  group_psm_by,
  group_pep_by,
  method_psm_pep
)
}
\arguments{
\item{file}{The name of a PSM file.}

\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{method_psm_pep}{Character string; the method to summarize the
\code{log2FC} and the \code{intensity} of \code{PSMs} by peptide entries.
The descriptive statistics includes \code{c("mean", "median",
"weighted_mean", "top_3_mean", "lfq_max", "lfq_top_2_sum", "lfq_top_3_sum",
"lfq_all")} with \code{median} being the default for TMT and
\code{lfq_top_2_sum} for LFQ. The \code{log10-intensity} of reporter (or
LFQ) ions at the \code{PSMs} levels will be the weight when summarizing
\code{log2FC} with various \code{"top_n"} statistics or
\code{"weighted_mean"}.}
}
\description{
Helper of PSM2Pep
}
