% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{psmC2Q}
\alias{psmC2Q}
\title{From \code{psmQ} to \code{psmQ}.}
\usage{
psmC2Q(out, combine_tier_three = FALSE, fdr_type = "psm", out_path)
}
\arguments{
\item{out}{A result from \code{psmQ} with additional filtration.}

\item{combine_tier_three}{Logical; if TRUE, combines all protein results to
the output of \code{psmQ.txt}. Outputs under the option TRUE are often
comparable to Mascot outputs with FDR controls at the levels of PSMs or
peptides.}

\item{fdr_type}{Character string; the type of FDR controlling. The value is
in one of c("psm", "peptide", "protein"). Note that \code{fdr_type =
protein} is equivalent to \code{fdr_type = psm} or \code{peptide} with the
additional filtration of data at \code{prot_tier == 1}.}

\item{out_path}{The output path.}
}
\description{
From \code{psmQ} to \code{psmQ}.
}
