% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{psmC2Q}
\alias{psmC2Q}
\title{From \code{psmQ} to \code{psmQ}.}
\usage{
psmC2Q(out, combine_tier_three = FALSE, out_path)
}
\arguments{
\item{out}{A result from \code{psmQ} with additional filtrations.}

\item{combine_tier_three}{Logical; if TRUE, combines tier-3 proteins in the
output of \code{psmQ.txt}. If FALSE, saves tier-3 proteins separately in
\code{psmT3.txt}.}

\item{out_path}{The output path.}
}
\description{
From \code{psmQ} to \code{psmQ}.
}
