% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-purge.R
\name{purge_by_cv}
\alias{purge_by_cv}
\title{Filter data groups by a CV cut-off}
\usage{
purge_by_cv(df, id, max_cv, keep_ohw = TRUE)
}
\arguments{
\item{df}{The name of input data file. By default, it will be determined
automatically by the value of \code{id}.}

\item{id}{Character string to indicate the type of data. The value will be
determined automatically. Peptide data will be used at \code{id = pep_seq}
or \code{pep_seq_mod}, and protein data will be used at \code{id = prot_acc}
or \code{gene}.}

\item{max_cv}{Numeric; the cut-off in maximum CV. Values above the threshold
will be replaced with NA.}
}
\description{
\code{purge_by_cv} replaces the data entries at \code{group CV > max_cv} to
NA.
}
