% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\name{purge_by_cv}
\alias{purge_by_cv}
\title{Filter data groups by a CV cut-off}
\usage{
purge_by_cv(df, id, max_cv, keep_ohw = TRUE, min_n = 1, rm_allna = FALSE)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{max_cv}{Numeric; the cut-off in maximum CV. Values above the threshold
will be replaced with NA. The default is NULL with no data trimming by max
CV.}

\item{keep_ohw}{Logical; if TRUE, keep one-hit-wonders with unknown CV. The
default is TRUE. The argument is softly depreciated; instead use
\code{min_n}. Equivalency: \code{keep_ohw = TRUE <=> min_n == 1; keep_ohw =
FALSE <=> min_n == 2}.}

\item{min_n}{Positive integer. When calling from \code{purgePSM}, peptide
entries in PSM tables with the number of identifying PSMs smaller than
\code{min_n} will be replaced with NA. When calling from \code{purgePep},
protein entries in peptide tables with the number of identifying peptides
smaller than \code{min_n} will be replaced with NA.}

\item{rm_allna}{Logical; if TRUE, removes data rows that are exclusively NA
across ratio columns of \code{log2_R126} etc. The setting also applies to
\code{log2_R000} in LFQ.}
}
\description{
\code{purge_by_cv} replaces the data entries at \code{group CV > max_cv} to
NA.
}
