% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-purge.R
\name{purge_by_n}
\alias{purge_by_n}
\title{Filter data groups by a minimal number of observations (n_obs)}
\usage{
purge_by_n(df, id, min_n)
}
\arguments{
\item{df}{The name of input data file. By default, it will be determined
automatically after matching the data type with an \code{id} among 
\code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}.}

\item{id}{Character string to indicate the type of data. The value will be
determined automatically. Peptide data will be used at \code{id = pep_seq}
or \code{pep_seq_mod}, and protein data will be used at \code{id = prot_acc}
or \code{gene}.}

\item{min_n}{Positive integer. When calling from \code{purgePSM}, peptide
entries in PSM tables with the number of identifying PSMs smaller than
\code{min_n} will be replaced with NA. When calling from \code{purgePep},
protein entries in peptide tables with the number of identifying peptides
smaller than \code{min_n} will be replaced with NA.}
}
\description{
\code{purge_by_n} replaces the data entries at \code{group n_obs < min_n} to
NA.
}
