% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-purge.R
\name{purge_by_qt}
\alias{purge_by_qt}
\title{Filter data groups by quantiles of CV}
\usage{
purge_by_qt(df, id, pt_cv = NULL, keep_ohw = TRUE)
}
\arguments{
\item{df}{The name of input data file. By default, it will be determined
automatically by the value of \code{id}.}

\item{id}{Character string to indicate the type of data. The value will be
determined automatically. Peptide data will be used at \code{id = pep_seq}
or \code{pep_seq_mod}, and protein data will be used at \code{id = prot_acc}
or \code{gene}.}

\item{pt_cv}{Numeric between 0 and 1; the percentile of CV. Values above the
percentile threshold will be replaced with NA.}
}
\description{
\code{purge_by_qt} replaces the data entries at \code{CV > quantile} with NA.
}
