% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\name{purge_by_qt}
\alias{purge_by_qt}
\title{Filters data groups by quantiles of CV}
\usage{
purge_by_qt(df, id, pt_cv = NULL, min_n = 1L, rm_allna = FALSE)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{pt_cv}{Numeric between 0 and 1; the percentile of CV. Values above the
percentile threshold will be replaced with NA. The default is NULL with no
data trimming by CV percentile. The precedence in data purging is
\code{pt_cv} \eqn{\ge} \code{max_cv} \eqn{\ge} \code{min_n}.}
}
\description{
\code{purge_by_qt} replaces the data entries at \code{CV > quantile} with NA.
}
