% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-purge.R
\name{purgePep}
\alias{purgePep}
\title{Purge peptide data}
\usage{
purgePep(dat_dir = NULL, max_cv = NULL, min_n = 1, ...)
}
\arguments{
\item{dat_dir}{A character string to the working directory.}

\item{max_cv}{Numeric; the cut-off in CV. The CV are from the ascribing PSMs
for the filtration of peptide entries or ascribing peptides for the
filtration of protein entries.}

\item{min_n}{Positive integer. When calling from \code{purgePSM}, peptide
entries in PSM tables with the number of identifying PSMs smaller than
\code{min_n} will be replaced with NA. When calling from \code{purgePep},
protein entries in peptide tables with the number of identifying peptides
smaller than \code{min_n} will be replaced with NA.}

\item{...}{Additional parameters for plotting: \cr \code{ymax}, the maximum
\eqn{y} at a log2 scale; the default is +0.6. \cr \code{y_breaks}, the
breaks in \eqn{y}-axis at a log2 scale; the default is 0.2. \cr
\code{width}, the width of plot. \cr \code{height}, the height of plot.}
}
\description{
\code{purgePep} removes \code{protein} entries from peptide table(s) by
selection criteria.
}
\details{
The CV of proteins are calculated from contributing peptides at the basis of
per TMT experiment per LCMS run.
}
