% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-purge.R
\name{purgePSM}
\alias{purgePSM}
\title{Purge PSM data}
\usage{
purgePSM(dat_dir = NULL, max_cv = NULL, min_n = 1, ...)
}
\arguments{
\item{dat_dir}{A character string to the working directory.}

\item{max_cv}{Numeric; the cut-off in CV. The CV are from the ascribing PSMs
for the filtration of peptides or ascribing peptides for the filtration of
proteins.}

\item{min_n}{Positive integer. When calling from \code{purgePSM}, peptide
entries in PSM tables with the number of identifying PSMs smaller than
\code{min_n} will be replaced with NA. When calling from \code{purgePep},
protein entries in peptide tables with the number of identifying peptides
smaller than \code{min_n} will be replaced with NA.}

\item{...}{Yet to be defined.}
}
\description{
\code{purgePSM} removes \code{peptide} entries from PSM tables by selection
criteria.
}
\details{
The CV of peptides are calculated from contributing PSMs. The calculation is
per TMT experiment and per LCMS run.
}
