% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-purge.R
\name{purgePSM}
\alias{purgePSM}
\title{Purge PSM data}
\usage{
purgePSM(dat_dir = NULL, pt_cv = NULL, max_cv = NULL,
  adjSD = FALSE, keep_ohw = TRUE, ...)
}
\arguments{
\item{dat_dir}{A character string to the working directory.}

\item{pt_cv}{Numeric between 0 and 1; the percentile of CV. Values above the
percentile threshold will be replaced with NA.}

\item{max_cv}{Numeric; the cut-off in maximum CV. Values above the threshold
will be replaced with NA.}

\item{adjSD}{Logical; if TRUE, adjust the standard deviation in relative to
the width of ratio profiles.}

\item{keep_ohw}{Logical; if TRUE, keep one-hit-wonders with unknown CV. The
default is TRUE.}

\item{...}{Additional parameters for plotting: \cr \code{ymax}, the maximum
\eqn{y} at a log2 scale; the default is +0.6. \cr \code{ybreaks}, the breaks
in \eqn{y}-axis at a log2 scale; the default is 0.2. Note: \code{y_breaks}
depreciated. \cr \code{width}, the width of plot. \cr \code{height}, the
height of plot. \cr \code{flip_coord}, logical; if TRUE, flip \code{x} and
\code{y} axis.}
}
\description{
\code{purgePSM} removes \code{peptide} entries from PSM tables by selection
criteria.
}
\details{
The CV of peptides are calculated from contributing PSMs at the basis of per
TMT experiment per series of LC/MS. Note that greater CV may be encountered
for samples that are more different to reference material(s).
}
