% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{quick_leftjoin}
\alias{quick_leftjoin}
\title{Left-joining of two data frames.}
\usage{
quick_leftjoin(x, y, by = NULL)
}
\arguments{
\item{x}{The left data frame.}

\item{y}{The right data frame.}

\item{by}{The key.}
}
\description{
Left-joining of two data frames.
}
\examples{
\donttest{
df1 <- data.frame(A = c("a", "b", "c"), B = c(1, 1, 1))
df2 <- data.frame(A = c("a", "c", "d"), C = c(2, 2, "3"))

x1 <- quick_leftjoin(df1, df2, by = "A")
x1 <- x1[, c("A", "B", "C")]
rownames(x1) <- seq_len(nrow(x1))

x2 <- dplyr::left_join(df1, df2, by = "A")

stopifnot(identical(x1, x2))
}
}
