% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{readMGF}
\alias{readMGF}
\title{Reads MGF files in chunks.}
\usage{
readMGF(
  filepath = "~/proteoQ/mgfs",
  min_mass = 500L,
  min_ms2mass = 110L,
  topn_ms2ions = 100L,
  ret_range = c(0, Inf),
  ppm_ms1 = 20L,
  ppm_ms2 = 25L,
  index_ms2 = FALSE,
  out_path = file.path(filepath, "mgf_queries.rds")
)
}
\arguments{
\item{filepath}{The file path to a list of mgf files.}

\item{min_mass}{Numeric; the minimum mass of MS1 species. The value needs to
match the one in  \link{binTheoPeps}.}

\item{min_ms2mass}{A minimum m/z of MS2 ions.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{ret_range}{The range of retention time in seconds.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{index_ms2}{Logical; if TRUE, converts MS2 m/z values to indices.}

\item{out_path}{An output path.}
}
\description{
Reads MGF files in chunks.
}
\examples{
\donttest{
mgf_queries <- readMGF()
}
}
