% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reorder_files}
\alias{reorder_files}
\title{Re-order file names}
\usage{
reorder_files(filelist)
}
\arguments{
\item{filelist}{A list of file names.}
}
\description{
\code{reorder_files} re-orders file names by TMT set numbers then by LCMS
injection numbers.
}
