% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{reorderCols}
\alias{reorderCols}
\title{Re-order columns in a data frame}
\usage{
reorderCols(df, endColIndex, col_to_rn)
}
\arguments{
\item{df}{A data frame.}

\item{endColIndex}{the indeces of columns to be moved to the end of
\code{df}.}

\item{col_to_rn}{the column identifier where the values under that column
will be used as row names.}
}
\description{
\code{reorderCols} re-orders columns in a data frame.
}
