% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_trivial_with_na}
\alias{replace_trivial_with_na}
\title{Replace a trivial vector with NA.}
\usage{
replace_trivial_with_na(x)
}
\arguments{
\item{x}{A numeric vecotr}
}
\description{
Replace a trivial vector with NA.
}
