% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{rm_RToutliers}
\alias{rm_RToutliers}
\title{Remove retention outliers}
\usage{
rm_RToutliers(
  pep_seq_modzs,
  pep_exp_mzs,
  pep_exp_ints,
  pep_apex_rets,
  pep_apex_scans,
  qt = 0.95
)
}
\arguments{
\item{pep_seq_modzs}{List of peptide sequences with modifications and charge
states.}

\item{pep_exp_mzs}{List of moverz values.}

\item{pep_exp_ints}{List of intensity values.}

\item{pep_apex_rets}{List of retention time values.}

\item{pep_apex_scans}{List of apex scan numbers.}

\item{qt}{The quantile for considering retention-time outliers.}
}
\description{
Remove retention outliers
}
