% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{rm_RToutliers}
\alias{rm_RToutliers}
\title{Cross-file removals of retention outliers}
\usage{
rm_RToutliers(
  pep_seq_modzs = NULL,
  pep_exp_mzs = NULL,
  pep_exp_ints = NULL,
  pep_apex_rets = NULL,
  pep_apex_scans = NULL,
  pep_apex_fwhm = NULL,
  pep_apex_n = NULL,
  mx = NULL,
  my = NULL,
  mt = NULL,
  ms = NULL,
  mf = NULL,
  mn = NULL,
  qt_rt = 0.95,
  ok_unv = FALSE,
  ignore_n2_outliers = FALSE,
  rt_tol = 0,
  by_fwhm = FALSE,
  fwhm_co = 0.5,
  qt_fwhm = 0.02,
  method = 0L
)
}
\arguments{
\item{pep_seq_modzs}{List of peptide sequences with modifications and charge
states.}

\item{pep_exp_mzs}{List of m-over-z values. Each list element corresponds to
entries in a \code{df}.}

\item{pep_exp_ints}{List of intensity values.}

\item{pep_apex_rets}{List of retention time values.}

\item{pep_apex_scans}{List of apex scan numbers.}

\item{pep_apex_fwhm}{List of apex FWHM values.}

\item{pep_apex_n}{List of apex lengths (number of MS2 scans).}

\item{mx}{The matrix of X values.}

\item{my}{The matrix of Y values.}

\item{mt}{The matrix of T values.}

\item{ms}{The matrix of Scan values.}

\item{mf}{The matrix of FWHM values.}

\item{mn}{The matrix of N (numbers of scans) values.}

\item{qt_rt}{The quantile for considering retention-time outliers.}

\item{ok_unv}{Logical; is the universe ready or need to be compiled.}

\item{ignore_n2_outliers}{Logical; ignore discrepant retention times at a
two-sample scenario, for example, to retain a high-quality subset for RT
calibrations.}

\item{rt_tol}{Not yet used. The tolerance in retention times between two
samples.}

\item{by_fwhm}{Additional outlier removals by FWHM values.}

\item{fwhm_co}{The cut-off in FWHM values.}

\item{qt_fwhm}{The quantile for considering FWHM outliers.}

\item{method}{A method for outlier removals.}
}
\description{
Nullify X, Y, T(ime) and S(can) values if found as RT outliers (> 95%
quantile of SD).
}
