% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{rmPSMHeaders}
\alias{rmPSMHeaders}
\title{Removes PSM headers}
\usage{
rmPSMHeaders(parallel = TRUE)
}
\arguments{
\item{parallel}{Logical; if TRUE, performs parallel computation. The default
is TRUE.}
}
\value{
Intermediate PSM table(s).
}
\description{
\code{rmPSMHeaders} removes the header of PSM from
\href{https://http://www.matrixscience.com/}{Mascot} outputs. It also
removes the spacer columns in the fields of ratio and intensity values.
}
