% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{Rosner_outliers}
\alias{Rosner_outliers}
\title{Outlier removals with Rosner's method}
\usage{
Rosner_outliers(x)
}
\arguments{
\item{x}{A matrix or data.frame.}
}
\description{
Outlier removals with Rosner's method
}
