% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rptr_violin}
\alias{rptr_violin}
\title{Violin plots of reporter-ion intensity per TMT_Set and LCMS_injection}
\usage{
rptr_violin(df, filepath, width, height)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{width}{The width of a plot.}

\item{height}{The height of a plot.}
}
\description{
Violin plots of reporter-ion intensity per TMT_Set and LCMS_injection
}
