% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{scoreLDA}
\alias{scoreLDA}
\title{Scores LDA}
\usage{
scoreLDA(
  df,
  id,
  label_scheme_sub,
  anal_type,
  scale_log2r,
  center_features,
  scale_features,
  type,
  col_group,
  folds,
  out_file,
  ...
)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{label_scheme_sub}{A data frame. Subset entries from \code{label_scheme}
for selected samples.}

\item{anal_type}{Character string; the type of analysis that are preset for
method dispatch in function factories. The value will be determined
automatically. Exemplary values include \code{anal_type = c("PCA",
"Corrplot", "EucDist", "GSPA", "Heatmap", "Histogram", "MDS", "Model",
"NMF", "Purge", "Trend", "LDA", ...)}.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE.}

\item{center_features}{Logical; if TRUE, adjusts log2FC to center zero by
features (proteins or peptides). The default is TRUE. Note the difference to
data alignment with \code{method_align} in \code{\link{standPrn}}
or \code{\link{standPep}} where log2FC are aligned by observations
(samples).}

\item{scale_features}{Logical; if TRUE, adjusts log2FC to the same scale of
variance by features (protein or peptide entries). The default is TRUE. Note
the difference to data scaling with \code{scale_log2r} where log2FC are
scaled by observations (samples).}

\item{type}{Character string indicating the type of PCA by either
\emph{observations} or \emph{features}. At the \code{type = obs} default,
observations (samples) are in rows and features (peptides or proteins) in
columns for \code{\link[stats]{prcomp}}. The principal components are then
plotted by observations. Alternatively at \code{type = feats}, features
(peptides or proteins) are in rows and observations (samples) are in
columns. The principal components are then plotted by features.}

\item{col_group}{Character string to a column key in \code{expt_smry.xlsx}.
Samples corresponding to non-empty entries under \code{col_group} will be
used for sample grouping in the indicated analysis. At the NULL default, the
column key \code{Group} will be used. No data annotation by groups will be
performed if the fields under the indicated group column is empty.}

\item{folds}{Not currently used. Integer; the degree of folding data into
subsets. The default is one without data folding.}

\item{out_file}{A file path object to an output file.}

\item{...}{\code{filter_}: Variable argument statements for the row filtration
against data in a primary file linked to \code{df}. See also
\code{\link{normPSM}} for the format of \code{filter_} statements. \cr \cr
Arguments passed to \code{\link[stats]{prcomp}}: \code{rank.}, \code{tol}
etc. At \code{type = obs}, argument \code{scale} becomes
\code{scale_features} and \code{center} matches \code{center_features}. At
\code{type = feats}, the setting of \code{scale_log2r} will be applied for
data scaling and data centering be automated by
\code{\link{standPep}} or \code{\link{standPrn}}. \cr \cr
Additional arguments for \code{ggsave}: \cr \code{width}, the width of plot;
\cr \code{height}, the height of plot \cr \code{...}}
}
\description{
Scores LDA
}
