% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-purge.R
\name{sd_lgl_cleanup}
\alias{sd_lgl_cleanup}
\title{Update data groups of `log2_R` etc. by logical matrix `lgl_log2_sd`}
\usage{
sd_lgl_cleanup(df)
}
\description{
Update data groups of `log2_R` etc. by logical matrix `lgl_log2_sd`
}
