% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{sd_violin}
\alias{sd_violin}
\title{Violin plots of CV per TMT_Set and LCMS_injection}
\usage{
sd_violin(df = NULL, id = NULL, filepath = NULL, width = NULL,
  height = NULL, type = "log2_R", adjSD = FALSE, is_psm = FALSE,
  col_select = NULL, col_order = NULL, theme = NULL, ...)
}
\description{
Violin plots of CV per TMT_Set and LCMS_injection
}
