% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-go.R
\name{set_db_outname}
\alias{set_db_outname}
\title{Helper to set the output file name for a data base}
\usage{
set_db_outname(filename = NULL, species = "human", siganature)
}
\description{
Helper to set the output file name for a data base
}
