% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setHMColor}
\alias{setHMColor}
\title{Customizes the colors in column annotation}
\usage{
setHMColor(annotation_col)
}
\arguments{
\item{annotation_col}{The same as in \link[pheatmap]{pheatmap}.}
}
\description{
Customizes the colors in column annotation
}
