% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-sigtests.R
\name{sigTest}
\alias{sigTest}
\title{Perform significance tests}
\usage{
sigTest(df, id, label_scheme_sub, scale_log2r, complete_cases, impute_na,
  filepath, filename, method, var_cutoff, pval_cutoff, logFC_cutoff,
  data_type, anal_type, ...)
}
\description{
Perform significance tests
}
