% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtests.R
\name{sigTest}
\alias{sigTest}
\title{Perform significance tests}
\usage{
sigTest(
  df,
  id,
  label_scheme_sub,
  scale_log2r,
  complete_cases,
  impute_na,
  filepath,
  filename,
  method,
  padj_method,
  var_cutoff,
  pval_cutoff,
  logFC_cutoff,
  data_type,
  anal_type,
  ...
)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{label_scheme_sub}{A data frame. Subset entries from \code{label_scheme}
for selected samples.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{impute_na}{Logical; if TRUE, data with the imputation of missing values
will be used. The default is FALSE.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{filename}{A representative file name to outputs. By default, the name(s)
will be determined automatically. For text files, a typical file extension
is \code{.txt}. For image files, they are typically saved via
\code{\link[ggplot2]{ggsave}} or \code{\link[pheatmap]{pheatmap}} where the
image type will be determined by the extension of the file name.}

\item{method}{Dummy argument to avoid incurring the corresponding argument in
\link[stats]{dist} by partial argument matches.}

\item{padj_method}{Character string; the method of multiple-test corrections
for uses with \link[stats]{p.adjust}. The default is "BH". See
?p.adjust.methods for additional choices.}

\item{var_cutoff}{Numeric; the cut-off in the variances of protein log2FC.
Entries with variances smaller than the threshold will be removed from GSVA.
The default is 0.5.}

\item{pval_cutoff}{Numeric value or vector; the cut-off in protein
significance \code{pVal}. Entries with \code{pVals} less significant than
the threshold will be excluded from enrichment analysis. The default is 0.05
for all formulas matched to or specified in argument \code{fml_nms}.
Formula-specific threshold is allowed by supplying a vector of cut-off
values.}

\item{logFC_cutoff}{Numeric value or vector; the cut-off in protein
\code{log2FC}. Entries with absolute \code{log2FC} smaller than the
threshold will be excluded from enrichment analysis. The default magnitude
is \code{log2(1.2)} for all formulas matched to or specified in argument
\code{fml_nms}. Formula-specific threshold is allowed by supplying a vector
of absolute values in \code{log2FC}.}

\item{data_type}{The type of data being either \code{Peptide} or \code{Protein}.}

\item{anal_type}{Character string; the type of analysis that are preset for
method dispatch in function factories. The value will be determined
automatically. Exemplary values include \code{anal_type = c("PCA",
"Corrplot", "EucDist", "GSPA", "Heatmap", "Histogram", "MDS", "Model",
"NMF", "Purge", "Trend", "LDA", ...)}.}

\item{...}{\code{filter_}: Variable argument statements for the row filtration
of data against the column keys in \code{Peptide.txt} for peptides or
\code{Protein.txt} for proteins. Each statement contains to a list of
logical expression(s). The \code{lhs} needs to start with \code{filter_}.
The logical condition(s) at the \code{rhs} needs to be enclosed in
\code{exprs} with round parenthesis. \cr \cr For example, \code{pep_len} is
a column key in \code{Peptide.txt}. The statement \code{filter_peps_at =
exprs(pep_len <= 50)} will remove peptide entries with \code{pep_len > 50}.
See also \code{\link{normPSM}}. \cr \cr Additional parameters for plotting
with \code{ggplot2}: \cr \code{xmin}, the minimum \eqn{x} at a log2 scale;
the default is -2. \cr \code{xmax}, the maximum \eqn{x} at a log2 scale; the
default is +2. \cr \code{xbreaks}, the breaks in \eqn{x}-axis at a log2
scale; the default is 1. \cr \code{binwidth}, the binwidth of \code{log2FC};
the default is \eqn{(xmax - xmin)/80}. \cr \code{ncol}, the number of
columns; the default is 1. \cr \code{width}, the width of plot; \cr
\code{height}, the height of plot. \cr \code{scales}, should the scales be
fixed across panels; the default is "fixed" and the alternative is "free".}
}
\description{
Perform significance tests
}
