% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{simple_label_scheme}
\alias{simple_label_scheme}
\title{Simplifies label schemes from \code{label_scheme_full}.}
\usage{
simple_label_scheme(dat_dir, label_scheme_full)
}
\description{
Removes duplicated sample entries under different LC/MS injections.
}
