% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{single_mq_prntable}
\alias{single_mq_prntable}
\title{Check the single presence of MaxQuant proteinGroups[...].txt.}
\usage{
single_mq_prntable(dat_dir)
}
\arguments{
\item{dat_dir}{The working directory.}
}
\description{
Check the single presence of MaxQuant proteinGroups[...].txt.
}
