% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{splitPSM}
\alias{splitPSM}
\title{Splits PSM tables}
\usage{
splitPSM(group_psm_by = "pep_seq", group_pep_by = "prot_acc",
  fasta = NULL, rm_craps = FALSE, rm_krts = FALSE,
  rptr_intco = 1000, annot_kinases = FALSE, plot_rptr_int = TRUE,
  ...)
}
\arguments{
\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search.}

\item{rm_craps}{Logical; if TRUE, removes
\code{\href{https://www.thegpm.org/crap/}{cRAP}} proteins.}

\item{rm_krts}{Logical; if TRUE, removes keratin entries.}

\item{rptr_intco}{Numeric; the threshold of reporter ion intensity.}

\item{annot_kinases}{Logical; if TRUE, annotates kinase attributes of
proteins.}

\item{plot_rptr_int}{Logical; if TRUE, prepares the violin plots of
reporter-ion intensities.}
}
\description{
\code{splitPSM} splits the PSM outputs after \code{rmPSMHeaders()}. It
separates PSM data by TMT experiment and LC/MS injection.
}
\details{
Arguments \code{group_psm_by} and \code{group_pep_by} are used to update
\code{prot_matches_sig} and \code{prot_sequences_sig} after data merge.
}
