% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{splitPSM}
\alias{splitPSM}
\title{Splits PSM tables}
\usage{
splitPSM(
  group_psm_by = "pep_seq",
  group_pep_by = "prot_acc",
  fasta = NULL,
  entrez = NULL,
  rm_craps = FALSE,
  rm_krts = FALSE,
  rptr_intco = 1000,
  annot_kinases = FALSE,
  plot_rptr_int = TRUE,
  use_lowercase_aa = TRUE,
  ...
)
}
\arguments{
\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search. There is no default and users need to provide the
correct file path(s) and name(s).}

\item{entrez}{Character string(s) to the name(s) of entrez file(s) with
prepended directory path. At the \code{NULL} default, a convenience lookup
is available for species among \code{c("human", "mouse", "rat")}. For other
species, users need to provide the file path(s) and name(s) for the lookup
table(s). See also \code{\link{Uni2Entrez}} and \code{\link{Ref2Entrez}}
for preparing custom entrez files.}

\item{rm_craps}{Logical; if TRUE,
\href{https://www.thegpm.org/crap/}{cRAP} proteins will be removed.
The default is FALSE.}

\item{rm_krts}{Logical; if TRUE, keratin entries will be removed. The default
is FALSE.}

\item{rptr_intco}{Numeric; the threshold of reporter ion intensity. The
default is 1,000.}

\item{annot_kinases}{Logical; if TRUE, proteins of human or mouse origins
will be annotated with their kinase attributes. The default is FALSE.}

\item{plot_rptr_int}{Logical; if TRUE, the distributions of reporter-ion
intensities will be plotted. The default is TRUE.}

\item{use_lowercase_aa}{Logical; if TRUE, modifications in amino acid
residues will be abbreviated with lower-case and/or \code{^_~}. See the
table below for details. The default is TRUE.}

\item{...}{Not currently used.}
}
\description{
\code{splitPSM} splits the PSM outputs after \code{rmPSMHeaders()}. It
separates PSM data by TMT experiment and LC/MS injection.
}
\details{
Arguments \code{group_psm_by} and \code{group_pep_by} are used to update
\code{prot_matches_sig} and \code{prot_sequences_sig} after data merge.
}
