% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{splitPSM}
\alias{splitPSM}
\title{Splits PSM tables}
\usage{
splitPSM(rptr_intco = 1000, rm_craps = FALSE, crap = "cRAP",
  plot_violins = TRUE)
}
\arguments{
\item{rptr_intco}{Numeric; the threshold of reporter ion intensity.}

\item{rm_craps}{Logical; if TRUE, removes
\code{\href{https://www.thegpm.org/crap/}{cRAP}} proteins.}

\item{crap}{Character string; the name of cRAP database, which can be found
from the header of PSM outputs: "Database,1::cRAP 2::SwissProt".}

\item{plot_violins}{Logical; if TRUE, prepares the violin plots of
reporter-ion intensities.}
}
\description{
\code{splitPSM} splits the PSM outputs after \code{rmPSMHeaders()}. It
separates PSM data by TMT experiment and LC/MS injection.
}
\examples{
splitPSM(rptr_intco = 1000, rm_craps = TRUE, plot_violins = TRUE)

}
