% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmtable.R
\name{splitPSM_mf}
\alias{splitPSM_mf}
\title{Splits MSFragger PSM tables}
\usage{
splitPSM_mf(
  group_psm_by = "pep_seq",
  group_pep_by = "prot_acc",
  fasta = NULL,
  entrez = NULL,
  pep_unique_by = "group",
  scale_rptr_int = FALSE,
  rm_craps = FALSE,
  rm_krts = FALSE,
  rm_allna = FALSE,
  purge_phosphodata = TRUE,
  annot_kinases = FALSE,
  plot_rptr_int = TRUE,
  rptr_intco = 0,
  rptr_intrange = c(0, 100),
  use_lowercase_aa = TRUE,
  parallel = TRUE,
  ...
)
}
\arguments{
\item{group_psm_by}{A character string specifying the method in PSM grouping.
At the \code{pep_seq} default, descriptive statistics will be calculated
based on the same \code{pep_seq} groups. At the \code{pep_seq_mod}
alternative, peptides with different variable modifications will be treated
as different species and descriptive statistics will be calculated based on
the same \code{pep_seq_mod} groups.}

\item{group_pep_by}{A character string specifying the method in peptide
grouping. At the \code{prot_acc} default, descriptive statistics will be
calculated based on the same \code{prot_acc} groups. At the \code{gene}
alternative, proteins with the same gene name but different accession
numbers will be treated as one group.}

\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search. There is no default and users need to provide the
correct file path(s) and name(s).}

\item{entrez}{Character string(s) to the name(s) of entrez file(s) with
prepended directory path. At the \code{NULL} default, a convenience lookup
is available for species among \code{c("human", "mouse", "rat")}. For other
species, users need to provide the file path(s) and name(s) for the lookup
table(s). See also \code{\link{Uni2Entrez}} and \code{\link{Ref2Entrez}} for
preparing custom entrez files.}

\item{pep_unique_by}{A character string for annotating the uniqueness of
peptides. At the \code{group} default, the uniqueness of peptides is by
groups with the collapses of same-set or sub-set proteins. At a more
stringent criterion of \code{protein}, the uniqueness of peptides is by
protein entries without grouping. On the other extreme of choice
\code{none}, all peptides are treated as unique. A new column of
\code{pep_isunique} with corresponding logical TRUE or FALSE will be added
to the PSM reports. Note that the choice of \code{none} is only for
convenience, as the same may be achieved by setting \code{use_unique_pep =
FALSE} in \link{Pep2Prn}.}

\item{scale_rptr_int}{Logical; if TRUE, scales (up) MS2 reporter-ion
intensities by MS1 precursor intensity: \eqn{I_{MS1}*(I_{x}/\sum I_{MS2})}.
\eqn{I_{MS1}}, MS1 precursor intensity; \eqn{I_{MS2}}, MS2 reporter-ion
intensity; \eqn{I_{x}}, MS2 reporter-ion intensity under TMT channel
\eqn{x}. Note that the scaling will not affect \code{log2FC}.}

\item{rm_craps}{Logical; if TRUE, \href{https://www.thegpm.org/crap/}{cRAP}
proteins will be removed. The default is FALSE.}

\item{rm_krts}{Logical; if TRUE, keratin entries will be removed. The default
is FALSE.}

\item{rm_allna}{Logical; if TRUE, removes data rows that are exclusively NA
across ratio columns of \code{log2_R126} etc. The setting also applies to
\code{log2_R000} in LFQ.}

\item{purge_phosphodata}{Logical; if TRUE and phosphorylation present as
variable modification(s), entries without phosphorylation will be removed.
The default is TRUE.}

\item{annot_kinases}{Logical; if TRUE, proteins of human or mouse origins will
be annotated with their kinase attributes. The default is FALSE.}

\item{plot_rptr_int}{Logical; if TRUE, the distributions of reporter-ion
intensities will be plotted. The default is TRUE. The argument is also
applicable to the precursor intensity with MaxQuant LFQ.}

\item{rptr_intco}{Numeric; the threshold of reporter-ion intensity (TMT:
\code{I126} etc.; LFQ: \code{I000}) being considered non-trivial. The
default is 0 without cut-offs. The data nullification will not be applied
synchronously to the precursor intensity (\code{pep_tot_int}) under the same
PSM query. To guard against odds such as higher MS2 reporter-ion intensities
than their contributing MS1 precursor intensity, employs for example
\code{filter_... = rlang::exprs(pep_tot_int >= my_ms1_cutoff)} during
\link{PSM2Pep}. The rule of thumb is that \code{pep_tot_int} is a single
column; thus the corresponding data filtration against it may be readily
achieved without introducing new arguments. By contrast, \code{rptr_intco}
applies to a set of columns, \code{I126} etc.; it might be slightly more
involved/laborious when applying suitable statements of \code{filter_}
varargs.}

\item{rptr_intrange}{Numeric vector at length two. The argument specifies the
range of reporter-ion intensities (TMT: \code{I126} etc.; LFQ: \code{I000})
being considered non-trivial. The default is between 0 and 100 percentile
without cut-offs. While argument \code{rptr_intco} employs a universal
cut-off across samples by absolute values, \code{range_int} provides an
alternative means of sample-specific thresholding of intensities by
percentiles. The data nullification will not be applied synchronously to the
precursor intensity under the same PSM query.}

\item{use_lowercase_aa}{Logical; if TRUE, modifications in amino acid residues
will be abbreviated with lower-case and/or \code{^_~}. See the table below
for details. The default is TRUE.}

\item{...}{Not currently used.}
}
\description{
\code{splitPSM_mf} splits the PSM outputs by TMT experiment and LC/MS
injection.
}
