% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{splitPSM_mq}
\alias{splitPSM_mq}
\title{Splits PSM tables}
\usage{
splitPSM_mq(group_psm_by = "pep_seq", group_pep_by = "prot_acc",
  fasta = NULL, pep_unique_by = "group", corrected_int = FALSE,
  rptr_intco = 1000, rm_craps = FALSE, rm_reverses = TRUE,
  annot_kinases = FALSE, plot_rptr_int = TRUE, ...)
}
\arguments{
\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search.}

\item{pep_unique_by}{A character string for the characterization of peptide
uniqueness from \code{MaxQuant} PSMs. The choice is one of \code{c("group",
"protein", "none")}. At the \code{group} default, the uniqueness of peptides
is by protein groups. At a more strigent criterion of \code{protein}, the
uniqueness of peptides is by protein entries. A new column of
\code{pep_isunique} will be added accordingly to the PSM reports. At a
trivia criterion of \code{none}, all values under \code{pep_unique_by} will
be set to \code{1}.}

\item{corrected_int}{Logical. If TRUE, values under columns "Reporter
intensity corrected" in MaxQuant PSM results (\code{msms.txt}) will be used.
Otherwise, "Reporter intensity" values without corrections will be used.}

\item{rptr_intco}{Numeric; the threshold of reporter ion intensity.}

\item{rm_craps}{Logical; if TRUE, removes
\code{\href{https://www.thegpm.org/crap/}{cRAP}} proteins.}

\item{rm_reverses}{Logical; if TRUE, removes \code{Reverse} entries from
MaxQuant peptide results.}

\item{annot_kinases}{Logical; if TRUE, annotates kinase attributes of
proteins.}

\item{plot_rptr_int}{Logical; if TRUE, prepares the violin plots of
reporter-ion intensities.}
}
\description{
\code{splitPSM_mq} splits the PSM outputs after \code{rmPSMHeaders()}. It
separates PSM data by TMT experiment and LC/MS injection.
}
\details{
Different to \code{splitPSM} used in \code{Mascot} processes,
\code{pep_seq_mod}, \code{prot_n_psm}, \code{prot_n_pep} and \code{pep_n_psm}
are calculated later in \code{annotPSM_mq}. This is suitable mainly because
there is no columns like \code{prot_matches_sig} and \code{prot_sequences_sig}
need to be updated after data merging in \code{splitPSM_mq}.
}
