% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{splitPSM_mq}
\alias{splitPSM_mq}
\title{Splits PSM tables}
\usage{
splitPSM_mq(fasta = NULL, pep_unique_by = "group",
  corrected_int = FALSE, rptr_intco = 1000, rm_craps = FALSE,
  rm_reverses = TRUE, annot_kinases = FALSE, plot_rptr_int = TRUE,
  ...)
}
\arguments{
\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search.}

\item{pep_unique_by}{A character string for use in the filtration of peptide
data from \code{MaxQuant}. The choice is in one of \code{c("group",
"protein", "none")} At the \code{group} default, only peptides unique by
protein groups will be kept. At \code{protein}, only peptides unique by
protein entries will be kept. At \code{none}, all peptides will be kept.}

\item{corrected_int}{Logical. If TRUE, values under columns "Reporter
intensity corrected" in MaxQuant PSM results (\code{msms.txt}) will be used.
Otherwise, "Reporter intensity" values without corrections will be used.}

\item{rptr_intco}{Numeric; the threshold of reporter ion intensity.}

\item{rm_craps}{Logical; if TRUE, removes
\code{\href{https://www.thegpm.org/crap/}{cRAP}} proteins.}

\item{rm_reverses}{Logical; if TRUE, removes \code{Reverse} entries from
MaxQuant peptide results.}

\item{annot_kinases}{Logical; if TRUE, annotates kinase attributes of
proteins.}

\item{plot_rptr_int}{Logical; if TRUE, prepares the violin plots of
reporter-ion intensities.}
}
\description{
\code{splitPSM} splits the PSM outputs after \code{rmPSMHeaders()}. It
separates PSM data by TMT experiment and LC/MS injection.
}
\examples{
splitPSM(rptr_intco = 1000, rm_craps = TRUE, plot_rptr_int = TRUE)

}
