% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{splitPSM_sm}
\alias{splitPSM_sm}
\title{Splits PSM tables}
\usage{
splitPSM_sm(fasta = NULL, rptr_intco = 1000, rm_craps = FALSE,
  plot_rptr_int = TRUE, ...)
}
\arguments{
\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search.}

\item{rptr_intco}{Numeric; the threshold of reporter ion intensity.}

\item{rm_craps}{Logical; if TRUE, removes
\code{\href{https://www.thegpm.org/crap/}{cRAP}} proteins.}

\item{plot_rptr_int}{Logical; if TRUE, prepares the violin plots of
reporter-ion intensities.}
}
\description{
\code{splitPSM} splits the PSM outputs after \code{rmPSMHeaders()}. It
separates PSM data by TMT experiment and LC/MS injection.
}
\examples{
splitPSM_sm(rptr_intco = 1000, rm_craps = TRUE, plot_rptr_int = TRUE)

}
