% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-psmtable.R
\name{splitPSM_sm}
\alias{splitPSM_sm}
\title{Splits PSM tables from \code{Spectrum Mill}}
\usage{
splitPSM_sm(
  group_psm_by = "pep_seq",
  group_pep_by = "prot_acc",
  fasta = NULL,
  rm_craps = FALSE,
  rm_krts = FALSE,
  rptr_intco = 1000,
  annot_kinases = FALSE,
  plot_rptr_int = TRUE,
  ...
)
}
\arguments{
\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The \code{fasta} database(s) need to match those
used in MS/MS ion search. There is no default and users need to provide the
correct file path(s) and name(s).}

\item{rm_craps}{Logical; if TRUE,
\code{\href{https://www.thegpm.org/crap/}{cRAP}} proteins will be removed.
The default is FALSE.}

\item{rm_krts}{Logical; if TRUE, keratin entries will be removed. The default
is FALSE.}

\item{rptr_intco}{Numeric; the threshold of reporter ion intensity. The
default is 1,000.}

\item{annot_kinases}{Logical; if TRUE, proteins of human or mouse origins
will be annotated with their kinase attributes. The default is FALSE.}

\item{plot_rptr_int}{Logical; if TRUE, the distributions of reporter-ion
intensities will be plotted. The default is TRUE.}
}
\description{
\code{splitPSM_sm} splits the PSM outputs from \code{Spectrum Mill}. It
separates PSM data by TMT experiment and LC/MS injection.
}
