% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptable.R
\name{spreadPepNums}
\alias{spreadPepNums}
\title{Spreads peptide numbers.}
\usage{
spreadPepNums(
  df,
  dat_dir = NULL,
  basenames,
  tmt_plex = 0L,
  ok_lfq = FALSE,
  group_psm_by = "pep_seq_mod",
  group_pep_by = "gene",
  engine = "mz"
)
}
\arguments{
\item{df}{A data frame of peptide table, ordered by ascending \code{TMT_Set}
and \code{LCMS_Injection}.}

\item{dat_dir}{The working directory.}

\item{basenames}{Names of peptide table files.}

\item{tmt_plex}{The multiplicity of TMT; zero for LFQ.}

\item{ok_lfq}{Logical Capable of LFQ or not.}

\item{group_psm_by}{Group PSMs by.}

\item{group_pep_by}{Groups peptides by.}

\item{engine}{The search engine.}
}
\description{
Spreads fields of numeric values: sd_log2_R, log2_R, log2_R, I, N_I by TMT
sets.
}
\details{
Also works for LFQ as each sample corresponds to a TMT set.

For single SILAC sample, the values of log2Ratios spreads into
\emph{MULTIPLE} columns of heavy, light etc. Despite, log2Ratios remains NA,
just like regular single-sample LFQ. The log2Ratios will be later calculated
with \link{calcLFQPepNums} that are based on intensity values.
}
