% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{stringTest}
\alias{stringTest}
\title{String analysis}
\usage{
stringTest(
  df = NULL,
  id = gene,
  label_scheme_sub = NULL,
  db_nms = NULL,
  score_cutoff = 0.7,
  scale_log2r = TRUE,
  complete_cases = FALSE,
  filepath = NULL,
  filename = NULL,
  ...
)
}
\arguments{
\item{df}{The name of a primary data file. By default, it will be determined
automatically after matching the types of data and analysis with an
\code{id} among \code{c("pep_seq", "pep_seq_mod", "prot_acc", "gene")}. A
primary file contains normalized peptide or protein data and is among
\code{c("Peptide.txt", "Peptide_pVal.txt", "Peptide_impNA_pVal.txt",
"Protein.txt", "Protein_pVal.txt", "protein_impNA_pVal.txt")}. For analyses
require the fields of significance p-values, the \code{df} will be one of
\code{c("Peptide_pVal.txt", "Peptide_impNA_pVal.txt", "Protein_pVal.txt",
"protein_impNA_pVal.txt")}.}

\item{id}{Character string; one of \code{pep_seq}, \code{pep_seq_mod},
\code{prot_acc} and \code{gene}.}

\item{label_scheme_sub}{A data frame. Subset entries from \code{label_scheme}
for selected samples.}

\item{db_nms}{Character string(s) to the name(s) of STRING database(s) with
prepended directory path. The \code{STRING} database(s) need to match those
generated from \code{\link{prepString}}. There is no default and users need
to provide the correct file path(s) and name(s).}

\item{score_cutoff}{Numeric; the threshold in the \code{combined_score} of
protein-protein interaction. The default is 0.7.}

\item{scale_log2r}{Logical; if TRUE, adjusts \code{log2FC} to the same scale
of standard deviation across all samples. The default is TRUE.}

\item{complete_cases}{Logical; if TRUE, only cases that are complete with no
missing values will be used. The default is FALSE.}

\item{filepath}{A file path to output results. By default, it will be
determined automatically by the name of the calling function and the value
of \code{id} in the \code{call}.}

\item{filename}{A representative file name to outputs. By default, the name(s)
will be determined automatically. For text files, a typical file extension
is \code{.txt}. For image files, they are typically saved via
\code{\link[ggplot2]{ggsave}} or \code{\link[pheatmap]{pheatmap}} where the
image type will be determined by the extension of the file name.}

\item{...}{\code{filter_}: Variable argument statements for the row filtration
of data against the column keys in \code{Peptide.txt} for peptides or
\code{Protein.txt} for proteins. Each statement contains to a list of
logical expression(s). The \code{lhs} needs to start with \code{filter_}.
The logical condition(s) at the \code{rhs} needs to be enclosed in
\code{exprs} with round parenthesis. \cr \cr For example, \code{pep_len} is
a column key in \code{Peptide.txt}. The statement \code{filter_peps_at =
exprs(pep_len <= 50)} will remove peptide entries with \code{pep_len > 50}.
See also \code{\link{normPSM}}. \cr \cr Additional parameters for plotting
with \code{ggplot2}: \cr \code{xmin}, the minimum \eqn{x} at a log2 scale;
the default is -2. \cr \code{xmax}, the maximum \eqn{x} at a log2 scale; the
default is +2. \cr \code{xbreaks}, the breaks in \eqn{x}-axis at a log2
scale; the default is 1. \cr \code{binwidth}, the binwidth of \code{log2FC};
the default is \eqn{(xmax - xmin)/80}. \cr \code{ncol}, the number of
columns; the default is 1. \cr \code{width}, the width of plot; \cr
\code{height}, the height of plot. \cr \code{scales}, should the scales be
fixed across panels; the default is "fixed" and the alternative is "free".}
}
\description{
The input \code{df} contains pVal fields.
}
\details{
The argument \code{scale_log2r} is not used in that both `_N` and `_Z`
columns from primary \code{df} will be kept. The argument \code{species} is
used for the generation of separate outputs by \code{species}.
}
