% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-string.R
\name{stringTest}
\alias{stringTest}
\title{String analysis}
\usage{
stringTest(df, id, col_group, col_order, label_scheme_sub,
  db_path = "~\\\\proteoQ\\\\dbs\\\\string", score_cutoff = 0.7,
  complete_cases, scale_log2r, filepath, filename, ...)
}
\description{
The input contains pVal fields
}
