% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanos.R
\name{summ_venn}
\alias{summ_venn}
\title{Venn counts.}
\usage{
summ_venn(df, id, contrast_groups, filepath, direction, fml_nm, max_size = 20L)
}
\arguments{
\item{df}{A data frame.}

\item{id}{ID.}

\item{contrast_groups}{Contrast groups}

\item{filepath}{A file path.}

\item{direction}{A direction of up or down.}

\item{fml_nm}{Formula name.}

\item{max_size}{The maximum number of contrast groups.}
}
\description{
Venn counts.
}
