% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular.R
\name{threshold_subcell_by_int}
\alias{threshold_subcell_by_int}
\title{Threshold subcellular locations by intensity.}
\usage{
threshold_subcell_by_int(df, col_subcellular = NULL, fct = 1)
}
\arguments{
\item{df}{A data frame.}

\item{col_subcellular}{The column key indicating subcellular locations.}

\item{fct}{A multiplification factor in intensity thresholding.}
}
\description{
Threshold subcellular locations by intensity.
}
