% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-params.R
\name{TMT_plex}
\alias{TMT_plex}
\title{Finds the multiplexity of TMT labels}
\usage{
TMT_plex(label_scheme_full)
}
\arguments{
\item{label_scheme_full}{The data frame returned by \code{\link{load_expts}},
including multiple LCMS series.}
}
\description{
\code{TMT_plex} returns the multiplexity of TMT labels.
}
