% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-utils.R
\name{TMT_top_n}
\alias{TMT_top_n}
\title{Calculate top.3}
\usage{
TMT_top_n(x, id, ...)
}
\arguments{
\item{x}{A data frame of \code{log2-ratios} and \code{intensity}.}

\item{id}{The variable to summarise \code{log2-ratios}.}
}
\description{
\code{TMT_wt_mean} calculates the weighted mean of \code{log2-ratios} and
\code{intensity}.
}
\examples{
df_num <- TMT_top_n(df, prot_acc, na.rm = TRUE)

}
