% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoQ-datasets.R
\docType{data}
\name{uniprot_entrez_hs}
\alias{uniprot_entrez_hs}
\title{Lookups between UniProt accessions and Entrez IDs for human proteins.}
\format{A data frame with 33194 rows and 2 variables:
\describe{
  \item{entrez}{Entrez ID}
  \item{uniprot_acc}{UniProt accession number}
}}
\usage{
uniprot_entrez_hs
}
\description{
A dataset containing human \code{entrez} IDs and \code{UniProt} accessions.
}
\keyword{datasets}
