% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{uniprot_entrez_rn}
\alias{uniprot_entrez_rn}
\title{Lookups between UniProt accessions and Entrez IDs for rat proteins.}
\format{
A data frame with 23097 rows and 4 variables:
\describe{
  \item{uniprot_acc}{UniProt accession number}
  \item{gene}{Gene name}
  \item{entrez}{Entrez ID}
  \item{species}{Species name}
}
}
\usage{
uniprot_entrez_rn
}
\description{
A dataset containing rat \code{entrez} IDs and \code{UniProt} accessions.
}
\keyword{datasets}
