% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{write_fasta}
\alias{write_fasta}
\title{Writes fasta}
\usage{
write_fasta(fasta_db, file)
}
\arguments{
\item{fasta_db}{A list of protein entries from \code{\link{read_fasta}}.}

\item{file}{A character string to the name of a protein fasta file.}
}
\description{
Writes a fasta file.
}
\examples{
\donttest{
fasta_db <- read_fasta(file = "~/proteoQ/dbs/fasta/uniprot/uniprot_hs_2020_05.fasta")
write_fasta(fasta_db, "~/proteoQ/examples/my.fasta")
}

}
